### 2025-09-19 cb392d4
Update affiliations for authors in paper.md
Not sure exactly what the issue was, but hoping fixing the ordering will help?
