### 2025-09-19 df06cc5
Update DOIs in paper.bib file
