+++ paper.bib
@@ -70,7 +70,7 @@
   volume={16},
   pages={1498},
   year={2025},
-  doi={10.1038/s41467-025-26498-z}
+  doi={10.1038/s41467-024-55655-3}
 }
 
 @article{vriza2023selfdriving,
@@ -80,7 +80,8 @@
   volume={35},
   number={8},
   pages={3046--3056},
-  year={2023}
+  year={2023},
+  doi={10.1021/acs.chemmater.2c03593}
 }
 
 @article{vescovi2023modular,
