+++ example_lab/example_modules/platereader.py
@@ -68,22 +68,22 @@ class PlateReaderNode(RestNode):
     @action
     def read_plate(
         self,
-    ) -> ActionResult:
+    ) -> int:
         """Run a command on the plate reader."""
         time.sleep(5)
-        return ActionSucceeded(data={"example_data": {"example": "data"}})
+        return 5
 
     @action
     def create_plate_file(
         self,
-    ) -> ActionResult:
+    ) -> Path:
         """Run a command on the plate reader."""
 
         with (Path.home() / "test.txt").open("w") as f:
             self.logger.log_info(f.write("test"))
-        path = str(Path.home() / "test.txt")
+        path = Path.home() / "test.txt"
 
-        return ActionSucceeded(files={"example_file": path})
+        return path
 
     def get_location(self) -> AdminCommandResponse:
         """Get location for the plate reader"""
