### 2025-09-22 ddcfc5c
working on return annotation
