+++ example_lab/example_modules/platereader.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import Any, Optional
 
 from madsci.client.event_client import EventClient
-from madsci.common.types.action_types import ActionResult, ActionSucceeded
+from madsci.common.types.action_types import ActionFiles, ActionResult, ActionSucceeded
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.node_types import RestNodeConfig
 from madsci.node_module.helpers import action
@@ -41,7 +41,9 @@ class PlateReaderInterface:
         )
         time.sleep(2)  # Simulate command execution
 
-
+class PlateFiles(ActionFiles):
+        file_path_1: Path
+        file_path_2: Path
 class PlateReaderNode(RestNode):
     """A fake plate reader node module for testing."""
 
@@ -84,6 +86,22 @@ class PlateReaderNode(RestNode):
         path = Path.home() / "test.txt"
 
         return path
+    
+    @action
+    def create_plate_files(
+        self,
+    ) -> PlateFiles:
+        """Run a command on the plate reader."""
+
+        with (Path.home() / "test.txt").open("w") as f:
+            self.logger.log_info(f.write("test"))
+        path1 = Path.home() / "test.txt"
+        with (Path.home() / "test2.txt").open("w") as f:
+            self.logger.log_info(f.write("test2"))
+        path2 = Path.home() / "test2.txt"
+    
+    
+        return PlateFiles(file_path_1=path1, file_path_2=path2)
 
     def get_location(self) -> AdminCommandResponse:
         """Get location for the plate reader"""
