+++ src/madsci_common/madsci/common/types/action_types.py
@@ -1,15 +1,28 @@
 """Types for MADSci Actions."""
-
+from __future__ import annotations
 from datetime import datetime
 from enum import Enum
-from typing import Any, Literal, Optional, Union
+from pathlib import Path
+from typing import Any, Literal, Optional, TypeAlias, Union, List, Dict
+from typing_extensions import Self
+
 
 from madsci.common.types.base_types import Error, MadsciBaseModel, PathLike
 from madsci.common.types.datapoint_types import DataPoint
 from madsci.common.utils import localnow, new_ulid_str
 from pydantic import Field
 from pydantic.functional_validators import field_validator, model_validator
+from typing_extensions import TypeAliasType
 
+from pydantic import TypeAdapter
+from ulid import ULID
+
+Json = TypeAliasType(
+    'Json',
+    'Union[dict[str, Json], list[Json], str, int, float, bool, None]',  
+)
+
+ta = TypeAdapter(Json)
 
 class ActionStatus(str, Enum):
     """Status for a step of a workflow"""
@@ -81,104 +94,144 @@ class ActionRequest(MadsciBaseModel):
 
     def succeeded(
         self,
-        data: dict[str, Any] = {},
-        files: dict[str, PathLike] = {},
+        json_data: Optional[Json | ActionJSON] = None,
+        files: Optional[PathLike | ActionFiles] = None,
         errors: Union[Error, list[Error], str] = [],
     ) -> "ActionSucceeded":
         """Create an ActionSucceeded response"""
         return ActionSucceeded(
             action_id=self.action_id,
             errors=errors,
-            data=data,
+            json_data=json_data,
             files=files,
         )
 
     def running(
         self,
-        data: dict[str, Any] = {},
-        files: dict[str, PathLike] = {},
+        json_data: Optional[Json | ActionJSON] = None,
+        files: Optional[PathLike | ActionFiles] = None,
         errors: Union[Error, list[Error], str] = [],
     ) -> "ActionRunning":
         """Create an ActionRunning response"""
         return ActionRunning(
             action_id=self.action_id,
             errors=errors,
-            data=data,
+            json_data=json_data,
             files=files,
         )
 
     def not_ready(
         self,
         errors: Union[Error, list[Error], str] = [],
-        data: dict[str, Any] = {},
-        files: dict[str, PathLike] = {},
+        json_data: Optional[Json | ActionJSON] = None,
+        files: Optional[PathLike | ActionFiles] = None
     ) -> "ActionNotReady":
         """Create an ActionNotReady response"""
         return ActionNotReady(
             action_id=self.action_id,
             errors=errors,
-            data=data,
+            json_data=json_data,
             files=files,
         )
 
     def cancelled(
         self,
         errors: Union[Error, list[Error], str] = [],
-        data: dict[str, Any] = {},
-        files: dict[str, PathLike] = {},
+        json_data: Optional[Json | ActionJSON] = None,
+        files: Optional[PathLike | ActionFiles] = None
     ) -> "ActionCancelled":
         """Create an ActionCancelled response"""
         return ActionCancelled(
             action_id=self.action_id,
             errors=errors,
-            data=data,
+            json_daa=json_data,
             files=files,
         )
 
     def paused(
         self,
         errors: Union[Error, list[Error], str] = [],
-        data: dict[str, Any] = {},
-        files: dict[str, PathLike] = {},
+        json_data: Optional[Json | ActionJSON] = None,
+        files: Optional[PathLike | ActionFiles] = None
     ) -> "ActionResult":
         """Create an ActionResult response"""
         return ActionResult(
             action_id=self.action_id,
             status=ActionStatus.PAUSED,
             errors=errors,
-            data=data,
+            data=json_data,
[…]
