### 2025-09-22 4813059
Fix affiliation formatting for Ian T. Foster
