### 2025-09-22 c167955
Breakout transfer planning into separate system component
