+++ src/madsci_location_manager/tests/test_location_server.py
@@ -533,7 +533,7 @@ def test_transfer_graph_construction(transfer_setup):
     manager = transfer_setup["manager"]
 
     # Build the transfer graph
-    graph = manager._build_transfer_graph()
+    graph = manager.transfer_planner._transfer_graph
 
     # Expected edges based on shared representations:
     # pickup <-> processing (robotarm_1)
@@ -615,21 +615,21 @@ def test_can_transfer_between_locations(transfer_setup):
     )
 
     # Test compatible transfers
-    assert manager._can_transfer_between_locations(
+    assert manager.transfer_planner._can_transfer_between_locations(
         pickup_location, processing_location, robot_template
     )
-    assert manager._can_transfer_between_locations(
+    assert manager.transfer_planner._can_transfer_between_locations(
         pickup_location, storage_location, conveyor_template
     )
 
     # Test incompatible transfers
-    assert not manager._can_transfer_between_locations(
+    assert not manager.transfer_planner._can_transfer_between_locations(
         processing_location, storage_location, robot_template
     )
-    assert not manager._can_transfer_between_locations(
+    assert not manager.transfer_planner._can_transfer_between_locations(
         pickup_location, isolated_location, robot_template
     )
-    assert not manager._can_transfer_between_locations(
+    assert not manager.transfer_planner._can_transfer_between_locations(
         pickup_location, processing_location, nonexistent_template
     )
 
@@ -639,7 +639,7 @@ def test_shortest_transfer_path_direct(transfer_setup):
     manager = transfer_setup["manager"]
 
     # Test direct path between connected locations
-    path = manager._find_shortest_transfer_path(
+    path = manager.transfer_planner.find_shortest_transfer_path(
         transfer_setup["locations"]["pickup"], transfer_setup["locations"]["processing"]
     )
 
@@ -655,7 +655,7 @@ def test_shortest_transfer_path_no_connection(transfer_setup):
     manager = transfer_setup["manager"]
 
     # Test path to isolated location (should return None)
-    path = manager._find_shortest_transfer_path(
+    path = manager.transfer_planner.find_shortest_transfer_path(
         transfer_setup["locations"]["pickup"], transfer_setup["locations"]["isolated"]
     )
 
@@ -667,7 +667,7 @@ def test_shortest_transfer_path_same_location(transfer_setup):
     manager = transfer_setup["manager"]
 
     # Test same location (should return empty path)
-    path = manager._find_shortest_transfer_path(
+    path = manager.transfer_planner.find_shortest_transfer_path(
         transfer_setup["locations"]["pickup"], transfer_setup["locations"]["pickup"]
     )
 
@@ -680,7 +680,7 @@ def test_shortest_transfer_path_multi_hop(transfer_setup):
 
     # Add another location that creates a longer path
     # This would test processing -> pickup -> storage route
-    path = manager._find_shortest_transfer_path(
+    path = manager.transfer_planner.find_shortest_transfer_path(
         transfer_setup["locations"]["processing"],
         transfer_setup["locations"]["storage"],
     )
@@ -700,27 +700,31 @@ def test_multi_leg_transfer_workflow_step_count(transfer_setup):
     manager = transfer_setup["manager"]
 
     # Test direct transfer (1 hop)
-    direct_path = manager._find_shortest_transfer_path(
+    direct_path = manager.transfer_planner.find_shortest_transfer_path(
         transfer_setup["locations"]["pickup"], transfer_setup["locations"]["processing"]
     )
     assert direct_path is not None
     assert len(direct_path) == 1
 
[…]
