+++ src/madsci_location_manager/madsci/location_manager/transfer_planner.py
@@ -1,0 +1,302 @@
+"""Transfer planning functionality for the Location Manager."""
+
+import heapq
+from typing import Optional
+
+from madsci.common.types.location_types import (
+    Location,
+    LocationManagerDefinition,
+    TransferGraphEdge,
+    TransferStepTemplate,
+)
+from madsci.common.types.step_types import StepDefinition
+from madsci.common.types.workflow_types import (
+    WorkflowDefinition,
+    WorkflowMetadata,
+    WorkflowParameters,
+)
+from madsci.location_manager.location_state_handler import LocationStateHandler
+
+
+class TransferPlanner:
+    """Handles transfer planning and graph operations for the Location Manager."""
+
+    def __init__(
+        self,
+        state_handler: LocationStateHandler,
+        definition: LocationManagerDefinition,
+    ) -> None:
+        """Initialize the TransferPlanner.
+
+        Args:
+            state_handler: LocationStateHandler instance for accessing location data
+            definition: LocationManagerDefinition containing transfer capabilities
+        """
+        self.state_handler = state_handler
+        self.definition = definition
+        self._transfer_graph = self._build_transfer_graph()
+
+    def _build_transfer_graph(self) -> dict[tuple[str, str], TransferGraphEdge]:
+        """
+        Build transfer graph based on location representations and transfer templates.
+
+        Returns:
+            Dict mapping (source_id, dest_id) tuples to TransferGraphEdge objects
+        """
+        transfer_graph = {}
+
+        if not self.definition.transfer_capabilities:
+            return transfer_graph
+
+        locations = self.state_handler.get_locations()
+
+        for template in self.definition.transfer_capabilities.transfer_templates:
+            # Find all location pairs that can use this transfer template
+            for source_location in locations:
+                for dest_location in locations:
+                    if source_location.location_id == dest_location.location_id:
+                        continue  # Skip self-transfers
+
+                    if self._can_transfer_between_locations(
+                        source_location, dest_location, template
+                    ):
+                        edge = TransferGraphEdge(
+                            source_location_id=source_location.location_id,
+                            destination_location_id=dest_location.location_id,
+                            transfer_template=template,
+                            cost=template.cost_weight or 1.0,
+                        )
+                        transfer_graph[
+                            (source_location.location_id, dest_location.location_id)
+                        ] = edge
+
+        return transfer_graph
+
+    def _can_transfer_between_locations(
+        self, source: Location, dest: Location, template: TransferStepTemplate
+    ) -> bool:
+        """
+        Check if transfer is possible between two locations using a template.
+
+        Based on simple representation key matching: if both locations have
+        representations for the template's node_name, transfer is possible.
+        """
+        if not source.representations or not dest.representations:
+            return False
+
+        return (
+            template.node_name in source.representations
+            and template.node_name in dest.representations
+        )
+
+    def find_shortest_transfer_path(
+        self, source_id: str, dest_id: str
+    ) -> Optional[list[TransferGraphEdge]]:
+        """
+        Find shortest path using Dijkstra's algorithm with edge weights.
+
+        Args:
+            source_id: Source location ID
+            dest_id: Destination location ID
+
+        Returns:
+            List of edges representing the transfer path, or None if no path exists
+        """
+        if source_id == dest_id:
+            return []  # No transfer needed
+
+        # Dijkstra's algorithm
+        distances = {source_id: 0}
+        previous = {}
+        unvisited = [(0, source_id)]
+        visited = set()
+
+        while unvisited:
+            current_distance, current_location = heapq.heappop(unvisited)
+
+            if current_location in visited:
+                continue
+
+            visited.add(current_location)
+
+            if current_location == dest_id:
+                # Reconstruct path
+                path = []
+                current = dest_id
[…]

[…]
