### 2025-09-22 1cf3eb5
Fix index for University of Chicago entry
