### 2025-09-22 8130e23
Add transfer_resource back in
