+++ manager.yaml
@@ -1,0 +1,6 @@
+name: Default LocationManager
+description: null
+manager_id: 01K5S7H10JQNWTRNV3SVN9TXT0
+manager_type: location_manager
+locations: []
+transfer_capabilities: null

[…]
