### 2025-09-22 e670cd2
Try indices first in affiliation
