### 2025-09-22 6608b54
Validate transfer_resource on example lab
