### 2025-09-22 80c03f6
Validate transfer_resource on example lab
