### 2025-09-22 1ce20f2
Remove outdated lookup/reference
