### 2025-09-22 33ec6ce
Cleanup vestigial workcell location references
