+++ src/madsci_workcell_manager/tests/test_workcell_actions.py
@@ -6,9 +6,12 @@ import pytest
 from madsci.common.exceptions import LocationNotFoundError
 from madsci.common.types.action_types import ActionStatus
 from madsci.common.types.context_types import MadsciContext
+from madsci.common.types.resource_types.server_types import ResourceHierarchy
 from madsci.workcell_manager.workcell_actions import (
+    _find_resource_and_transfer,
     _resolve_location_identifier,
     transfer,
+    transfer_resource,
     wait,
 )
 
@@ -363,3 +366,284 @@ class TestResolveLocationIdentifier:
 
         # The function catches exceptions internally and raises "not found" instead of connection error
         assert "not found by ID or name" in str(exc_info.value)
+
+
+class TestTransferResourceAction:
+    """Test cases for the transfer_resource action."""
+
+    @patch("madsci.workcell_manager.workcell_actions.get_current_madsci_context")
+    def test_transfer_resource_missing_resource_server_url(self, mock_context):
+        """Test that transfer_resource fails when resource server URL is not configured."""
+        mock_context.return_value = MadsciContext(
+            location_server_url="http://localhost:8006/", resource_server_url=None
+        )
+
+        result = transfer_resource("resource_123", "dest_loc")
+
+        assert result.status == ActionStatus.FAILED
+        assert "Resource server URL not configured" in result.errors[0].message
+
+    @patch("madsci.workcell_manager.workcell_actions.get_current_madsci_context")
+    def test_transfer_resource_missing_location_server_url(self, mock_context):
+        """Test that transfer_resource fails when location server URL is not configured."""
+        mock_context.return_value = MadsciContext(
+            resource_server_url="http://localhost:8003/", location_server_url=None
+        )
+
+        result = transfer_resource("resource_123", "dest_loc")
+
+        assert result.status == ActionStatus.FAILED
+        assert "Location server URL not configured" in result.errors[0].message
+
+    @patch("madsci.workcell_manager.workcell_actions.get_current_madsci_context")
+    @patch("madsci.workcell_manager.workcell_actions._find_resource_and_transfer")
+    def test_transfer_resource_invalid_resource(
+        self, mock_find_and_transfer, mock_context
+    ):
+        """Test that transfer_resource fails when resource doesn't exist."""
+        mock_context.return_value = MadsciContext(
+            resource_server_url="http://localhost:8003/",
+            location_server_url="http://localhost:8006/",
+        )
+
+        # Mock helper function to return error message
+        mock_find_and_transfer.return_value = (
+            "Failed to verify resource 'invalid_resource': Resource not found"
+        )
+
+        result = transfer_resource("invalid_resource", "dest_loc")
+
+        assert result.status == ActionStatus.FAILED
+        assert "Failed to verify resource" in result.errors[0].message
+
+    @patch("madsci.workcell_manager.workcell_actions.get_current_madsci_context")
+    @patch("madsci.workcell_manager.workcell_actions._find_resource_and_transfer")
+    def test_transfer_resource_not_found_at_location(
+        self, mock_find_and_transfer, mock_context
+    ):
+        """Test that transfer_resource fails when resource is not found at any location."""
+        mock_context.return_value = MadsciContext(
+            resource_server_url="http://localhost:8003/",
+            location_server_url="http://localhost:8006/",
+        )
+
+        # Mock helper function to return error message
+        mock_find_and_transfer.return_value = (
+            "Resource 'resource_123' not found at any location"
+        )
+
+        result = transfer_resource("resource_123", "dest_loc")
+
+        assert result.status == ActionStatus.FAILED
+        assert (
+            "Resource 'resource_123' not found at any location"
[…]
