+++ src/madsci_workcell_manager/tests/test_workcell_engine.py
@@ -13,6 +13,7 @@ from madsci.common.types.action_types import (
     ActionStatus,
     ActionSucceeded,
 )
+from madsci.common.types.context_types import MadsciContext
 from madsci.common.types.datapoint_types import (
     FileDataPoint,
     ObjectStorageDataPoint,
@@ -79,7 +80,24 @@ def state_handler(redis_server: Redis) -> WorkcellStateHandler:
 @pytest.fixture
 def engine(state_handler: WorkcellStateHandler) -> Engine:
     """Fixture for creating an Engine instance."""
-    with warnings.catch_warnings():
+    # Create a mock context with all required URLs for LocationClient
+    mock_context = MadsciContext(
+        lab_server_url="http://localhost:8000/",
+        event_server_url="http://localhost:8001/",
+        experiment_server_url="http://localhost:8002/",
+        data_server_url="http://localhost:8004/",
+        resource_server_url="http://localhost:8003/",
+        workcell_server_url="http://localhost:8005/",
+        location_server_url="http://localhost:8006/",
+    )
+
+    with (
+        warnings.catch_warnings(),
+        patch(
+            "madsci.client.location_client.get_current_madsci_context",
+            return_value=mock_context,
+        ),
+    ):
         warnings.simplefilter("ignore", UserWarning)
         return Engine(state_handler=state_handler, data_client=DataClient())
 
