### 2025-09-22 d6836cb
Update location manager docs
