+++ src/madsci_location_manager/madsci/location_manager/__init__.py
@@ -1,3 +1,1 @@
 """LocationManager - Dedicated service for location management in MADSci."""
-
-__version__ = "0.1.0"
