+++ src/madsci_location_manager/README.md
@@ -1,27 +1,43 @@
 # MADSci Location Manager
 
-The Location Manager is a dedicated microservice for managing laboratory locations in the MADSci ecosystem. It provides centralized location management functionality that was previously embedded in the WorkcellManager.
+The Location Manager is a dedicated microservice for managing laboratory locations in the MADSci ecosystem. It provides centralized location management functionality including location CRUD operations, resource attachments, node-specific representations, and transfer planning capabilities.
 
 ## Features
 
 - **Location CRUD Operations**: Create, read, update, and delete locations
-- **Resource Attachment**: Attach resources to specific locations
-- **Representations**: Manage node-specific representations for locations
+- **Resource Attachment**: Attach resources to specific locations with automatic resource creation from templates
+- **Node-Specific Representations**: Manage node-specific representations for locations to enable flexible integration
+- **Transfer Planning**: Plan multi-step transfers between locations using transfer templates and graph algorithms
+- **Resource Hierarchy Queries**: Query resource hierarchies for resources attached to locations
 - **Redis State Management**: Persistent state storage using Redis
 - **RESTful API**: Clean REST endpoints for all location operations
 
 ## API Endpoints
 
+### Location Management
 - `GET /locations` - List all locations
 - `POST /location` - Create a new location
-- `GET /location/{location_id}` - Get a specific location
+- `GET /location` - Get a location by query parameters (location_id or name)
+- `GET /location/{location_id}` - Get a specific location by ID
 - `DELETE /location/{location_id}` - Delete a location
-- `POST /location/{location_id}/set_representation/{node_name}` - Set representations for a node
+- `POST /location/{location_id}/set_representation/{node_name}` - Set representations for a location and node
 - `POST /location/{location_id}/attach_resource` - Attach a resource to a location
+
+### Transfer Planning
+- `POST /transfer/plan` - Plan a transfer workflow from source to destination location
+- `GET /transfer/graph` - Get the current transfer graph as adjacency list
+
+### Resource Queries
+- `GET /location/{location_id}/resources` - Get resource hierarchy for resources at a location
+
+### System Endpoints
+- `GET /health` - Health check endpoint
+- `GET /definition` - Get Location Manager definition and configuration
+- `GET /` - Root endpoint returning manager definition
 
 ## Configuration
 
-The Location Manager uses environment variables for configuration:
+The Location Manager uses environment variables with the `LOCATION_` prefix:
 
 - `LOCATION_MANAGER_ID` - Unique identifier for this manager instance
 - `LOCATION_SERVER_HOST` - Server host (default: localhost)
@@ -35,9 +51,11 @@ The Location Manager uses environment variables for configuration:
 ### Starting the Server
 
 ```python
-from madsci.location_manager.location_server import main
+from madsci.location_manager.location_server import LocationManager
 
-main()
+# Create and run the manager
+manager = LocationManager()
+manager.run_server()
 ```
 
 ### Using the Client
@@ -45,23 +63,78 @@ main()
 ```python
 from madsci.client.location_client import LocationClient
 
+# Initialize client
 client = LocationClient("http://localhost:8006")
+
+# Basic location operations
 locations = client.get_locations()
+location = client.get_location("location_id")
+location_by_name = client.get_location_by_name("location_name")
+
+# Resource operations
+client.attach_resource("location_id", "resource_id")
+resources = client.get_location_resources("location_id")
+
+# Transfer planning
+transfer_graph = client.get_transfer_graph()
+workflow = client.plan_transfer("source_id", "dest_id")
+
+# Node representations
+client.set_representations("location_id", "node_name", {"key": "value"})
 ```
+
+## Key Components
+
+### LocationManager
[…]
