### 2025-09-22 3d109d4
Update LocationManager README
