+++ src/madsci_location_manager/README.md
@@ -127,14 +127,3 @@ The Location Manager integrates with:
 - **Workcell Manager**: For workflow location validation and transfer execution
 - **Event Manager**: For logging and ownership context
 - **UI Dashboard**: For location management interface and transfer visualization
-
-## Development
-
-This package follows MADSci's established patterns:
-
-- Inherits from `AbstractManagerBase` for consistent manager behavior
-- Uses Redis for state management via `LocationStateHandler`
-- Follows the standard manager port allocation (8006)
-- Implements standard health and definition endpoints
-- Uses ULID for all ID generation
-- Supports ownership context throughout all operations
