### 2025-09-22 1caa00a
Add 'allow_transfers' field to locations
