+++ src/madsci_location_manager/tests/test_location_server.py
@@ -1258,3 +1258,245 @@ def test_get_location_by_id_query_parameter(redis_server: Redis):
     error_data = response.json()
     assert "not found" in error_data["detail"].lower()
     assert "nonexistent_id" in error_data["detail"]
+
+
+# Non-transfer location tests
+def test_non_transfer_location_creation(redis_server: Redis):
+    """Test creating locations with allow_transfers=False."""
+    non_transfer_location = LocationDefinition(
+        location_name="non_transfer_station",
+        location_id=new_ulid_str(),
+        description="A station that doesn't allow transfers",
+        allow_transfers=False,
+    )
+
+    transfer_location = LocationDefinition(
+        location_name="transfer_station",
+        location_id=new_ulid_str(),
+        description="A station that allows transfers",
+        allow_transfers=True,
+    )
+
+    definition = LocationManagerDefinition(
+        name="Test Manager with Non-transfer Locations",
+        manager_id=new_ulid_str(),
+        locations=[non_transfer_location, transfer_location],
+    )
+
+    settings = LocationManagerSettings(
+        manager_id="test_non_transfer_manager",
+        redis_host=redis_server.connection_pool.connection_kwargs["host"],
+        redis_port=redis_server.connection_pool.connection_kwargs["port"],
+    )
+
+    manager = LocationManager(settings=settings, definition=definition)
+    manager.state_handler._redis_connection = redis_server
+    client = TestClient(manager.create_server())
+
+    # Check that locations are created with correct allow_transfers values
+    response = client.get(f"/location?location_id={non_transfer_location.location_id}")
+    assert response.status_code == 200
+    location_data = response.json()
+    assert location_data["allow_transfers"] is False
+
+    response = client.get(f"/location?location_id={transfer_location.location_id}")
+    assert response.status_code == 200
+    location_data = response.json()
+    assert location_data["allow_transfers"] is True
+
+
+def test_non_transfer_location_excluded_from_graph(redis_server: Redis):
+    """Test that non-transfer locations are excluded from transfer graph construction."""
+    # Create transfer templates
+    robot_template = TransferStepTemplate(
+        node_name="robot_arm",
+        action="transfer",
+        source_argument_name="source_location",
+        target_argument_name="target_location",
+        cost_weight=1.0,
+    )
+
+    transfer_capabilities = LocationTransferCapabilities(
+        transfer_templates=[robot_template]
+    )
+
+    # Create locations - some allow transfers, others don't
+    transfer_loc1 = LocationDefinition(
+        location_name="transfer_station_1",
+        location_id=new_ulid_str(),
+        representations={"robot_arm": {"position": [1, 0, 0]}},
+        allow_transfers=True,
+    )
+
+    transfer_loc2 = LocationDefinition(
+        location_name="transfer_station_2",
+        location_id=new_ulid_str(),
+        representations={"robot_arm": {"position": [2, 0, 0]}},
+        allow_transfers=True,
+    )
+
+    non_transfer_loc = LocationDefinition(
+        location_name="non_transfer_station",
+        location_id=new_ulid_str(),
+        representations={"robot_arm": {"position": [0, 1, 0]}},
+        allow_transfers=False,
+    )
+
+    definition = LocationManagerDefinition(
+        name="Test Manager with Non-transfer Graph",
+        manager_id=new_ulid_str(),
+        locations=[transfer_loc1, transfer_loc2, non_transfer_loc],
+        transfer_capabilities=transfer_capabilities,
+    )
+
+    settings = LocationManagerSettings(
+        manager_id="test_non_transfer_graph_manager",
+        redis_host=redis_server.connection_pool.connection_kwargs["host"],
+        redis_port=redis_server.connection_pool.connection_kwargs["port"],
+    )
[…]
