+++ src/madsci_location_manager/tests/test_location_server.py
@@ -12,6 +12,7 @@ from madsci.common.types.location_types import (
     LocationManagerSettings,
     LocationTransferCapabilities,
     TransferStepTemplate,
+    TransferTemplateOverrides,
 )
 from madsci.common.types.resource_types.server_types import ResourceHierarchy
 from madsci.common.types.workflow_types import WorkflowDefinition, WorkflowParameters
@@ -1500,3 +1501,515 @@ def test_location_allow_transfers_default():
     )
 
     assert location_false.allow_transfers is False
+
+
+# Override Transfer Template Tests
+
+
+@pytest.fixture
+def override_transfer_setup(redis_server: Redis):
+    """Create a location manager with override transfer templates for testing."""
+    # Create standard transfer templates
+    default_robot_template = TransferStepTemplate(
+        node_name="robot_default",
+        action="transfer",
+        source_argument_name="source_location",
+        target_argument_name="target_location",
+        cost_weight=5.0,
+    )
+
+    default_conveyor_template = TransferStepTemplate(
+        node_name="conveyor_default",
+        action="move",
+        source_argument_name="from_location",
+        target_argument_name="to_location",
+        cost_weight=3.0,
+    )
+
+    # Create override templates
+    special_robot_template = TransferStepTemplate(
+        node_name="robot_special",
+        action="special_transfer",
+        source_argument_name="source_location",
+        target_argument_name="target_location",
+        cost_weight=1.0,
+    )
+
+    fast_conveyor_template = TransferStepTemplate(
+        node_name="conveyor_fast",
+        action="fast_move",
+        source_argument_name="from_location",
+        target_argument_name="to_location",
+        cost_weight=0.5,
+    )
+
+    # Create override configuration
+    overrides = TransferTemplateOverrides(
+        source_overrides={
+            "station_a": [special_robot_template],
+        },
+        destination_overrides={
+            "station_c": [fast_conveyor_template],
+        },
+        pair_overrides={
+            "station_a": {
+                "station_b": [special_robot_template, fast_conveyor_template],
+            },
+        },
+    )
+
+    transfer_capabilities = LocationTransferCapabilities(
+        transfer_templates=[default_robot_template, default_conveyor_template],
+        override_transfer_templates=overrides,
+    )
+
+    # Create locations with representations that support all templates
+    station_a = LocationDefinition(
+        location_name="station_a",
+        location_id=new_ulid_str(),
+        description="Station A with all device access",
+        representations={
+            "robot_default": {"position": [1, 0, 0]},
+            "robot_special": {"position": [1, 0, 0]},
+            "conveyor_default": {"belt_position": 0},
+            "conveyor_fast": {"belt_position": 0},
+        },
+    )
+
+    station_b = LocationDefinition(
+        location_name="station_b",
+        location_id=new_ulid_str(),
+        description="Station B with all device access",
+        representations={
+            "robot_default": {"position": [2, 0, 0]},
+            "robot_special": {"position": [2, 0, 0]},
+            "conveyor_default": {"belt_position": 1},
+            "conveyor_fast": {"belt_position": 1},
+        },
+    )
+
+    station_c = LocationDefinition(
+        location_name="station_c",
+        location_id=new_ulid_str(),
+        description="Station C with all device access",
+        representations={
+            "robot_default": {"position": [3, 0, 0]},
+            "robot_special": {"position": [3, 0, 0]},
+            "conveyor_default": {"belt_position": 2},
+            "conveyor_fast": {"belt_position": 2},
+        },
+    )
+
+    station_d = LocationDefinition(
+        location_name="station_d",
+        location_id=new_ulid_str(),
[…]

[…]
