### 2025-09-23 19b7aeb
adding example modules
