+++ src/madsci_node_module/madsci/node_module/helpers.py
@@ -10,6 +10,7 @@ from zipfile import ZipFile
 import logging
 
 
+from madsci.common.types.datapoint_types import FileDataPoint, ObjectStorageDataPoint, ValueDataPoint
 import regex
 from madsci.common.types.action_types import (
     ActionDatapoints,
@@ -141,14 +142,7 @@ def get_named_input(main_string: str, plural: str) -> list[str]:
                     raise ValueError(string1.capitalize() + string2)
     return result_list
 
-
-def parse_results(func: Callable) -> list[ActionResultDefinition]:
-    """get the resulting data from an Action"""
-    returned = inspect.signature(func).return_annotation
-    print(returned)
-    
-    if returned is inspect.Signature.empty:
-        return []
+def parse_result(returned: Any):
     if returned is Path:
         return [FileActionResultDefinition(result_label="file")]
     if issubclass(returned, ActionFiles):
@@ -168,16 +162,35 @@ def parse_results(func: Callable) -> list[ActionResultDefinition]:
         return [JSONActionResultDefinition(result_label=key) for key in returned.__annotations__]
     if issubclass(returned, ActionDatapoints):
         for key, value in returned.__annotations__.items():
-            if value not in [str]:
+            if value not in [FileDataPoint, ValueDataPoint, ObjectStorageDataPoint]:
                 raise ValueError(
-                    f"All fields in an ActionDatapoints subclass must be of type str, int, float, bool, dict, or list but field {key} is of type {value}",
+                    f"All fields in an ActionDatapoints subclass must be datapoints but field {key} is of type {value}",
                 )
         return [DatapointActionResultDefinition(result_label=key) for key in returned.__annotations__]
     if returned not in [str, int, float, bool, dict, list]:
         raise ValueError(
             f"Action return type must be a subclass of ActionFiles, ActionJSON, ActionDatapoints, Path, str, int, float, bool, dict, or list but got {returned}",
         )
-    return [JSONActionResultDefinition(result_label="data", description=returned.__name__)]
+    return [JSONActionResultDefinition(result_label="data", data_type=returned.__name__)]
+
+
+def parse_results(func: Callable) -> list[ActionResultDefinition]:
+    """get the resulting data from an Action"""
+    returned = inspect.signature(func).return_annotation
+    
+    if returned is inspect.Signature.empty or returned is None:
+        return []
+    print(f"Parsing return type: {returned}")
+    print(returned is tuple)
+    if getattr(returned, "__origin__", None) is tuple:
+        result_definitions = []
+        for result in returned.__args__:
+            result_definitions.extend(parse_result(result))
+    else:
+        result_definitions = parse_result(returned)
+    return result_definitions
+            
+    
 
 def action_response_to_headers(action_response: ActionResult) -> dict[str, str]:
     """Converts the response to a dictionary of headers"""
