### 2025-09-23 3e70aa1
Account for current resource during transfer planning
