### 2025-09-23 5e3bfd1
Add some basic CRUD endpoints; cleanup dashboard
