### 2025-09-23 299f84b
Resource Hierarchies properly recursive
