### 2025-09-23 ecbe05a
Ensure GET_LOCATION admin action still works
