+++ example_lab/example_modules/node_definitions/advanced_example_node.node.info.yaml
@@ -1,8 +1,8 @@
-node_name: liquidhandler_1
-node_id: 01JYFEHVSV20D60Z88RVERJ75N
-node_description: A (fake) liquidhandler for testing purposes.
+node_name: advanced_example_node.node
+node_id: 01K5W0F6WKCRCMYVWD2XQBNA6V
+node_description: null
 node_type: device
-module_name: liquidhandler
+module_name: advanced_example_node
 module_version: 0.0.1
 capabilities:
   get_info: true
@@ -24,42 +24,49 @@ capabilities:
   - unlock
 node_url: null
 actions:
-  run_command:
-    name: run_command
-    description: Run a command on the liquid handler. Shows returning both JSON and
-      file data.
-    args:
-      command:
-        name: command
-        description: ''
-        argument_type: str
-        required: true
-        default: null
-    locations: {}
-    files: {}
-    results:
-      command:
-        result_label: command
-        description: null
-        result_type: json
-      log_file:
-        result_label: log_file
-        description: null
-        result_type: file
-    blocking: false
-    asynchronous: true
-  run_protocol:
-    name: run_protocol
+  return_none:
+    name: return_none
     description: Run a protocol on the liquid handler
     args: {}
     locations: {}
-    files:
-      protocol:
-        name: protocol
-        description: ''
-        argument_type: file
-        required: true
-        default: null
+    files: {}
+    results: {}
+    blocking: false
+    asynchronous: true
+  return_json:
+    name: return_json
+    description: Return a JSON object
+    args: {}
+    locations: {}
+    files: {}
+    results:
+      data:
+        result_label: data
+        description: null
+        result_type: json
+        data_type: int
+    blocking: false
+    asynchronous: true
+  return_dict:
+    name: return_dict
+    description: Return a JSON object
+    args: {}
+    locations: {}
+    files: {}
+    results:
+      data:
+        result_label: data
+        description: null
+        result_type: json
+        data_type: dict
+    blocking: false
+    asynchronous: true
+  return_file:
+    name: return_file
+    description: Return a file
+    args: {}
+    locations: {}
+    files: {}
     results:
       file:
         result_label: file
@@ -67,13 +74,94 @@ actions:
         result_type: file
     blocking: false
     asynchronous: true
+  return_file_and_json:
+    name: return_file_and_json
+    description: Return a file
+    args: {}
+    locations: {}
+    files: {}
+    results:
+      data:
+        result_label: data
+        description: null
+        result_type: json
+        data_type: int
+      file:
+        result_label: file
+        description: null
+        result_type: file
+    blocking: false
+    asynchronous: true
+  return_labeled_json_values:
+    name: return_labeled_json_values
+    description: Return labeled JSON values
+    args: {}
+    locations: {}
+    files: {}
+    results:
+      example_value_1:
+        result_label: example_value_1
+        description: null
+        result_type: json
+        data_type: int
+      example_value_2:
+        result_label: example_value_2
+        description: null
+        result_type: json
+        data_type: int
+    blocking: false
+    asynchronous: true
+  return_labeled_file_values:
+    name: return_labeled_file_values
+    description: Return labeled file values
+    args: {}
+    locations: {}
+    files: {}
+    results:
+      log_file_1:
+        result_label: log_file_1
+        description: null
+        result_type: file
+      log_file_2:
+        result_label: log_file_2
+        description: null
+        result_type: file
+    blocking: false
+    asynchronous: true
+  return_labeled_file_and_json_values:
+    name: return_labeled_file_and_json_values
+    description: Return labeled file and json values
+    args: {}
+    locations: {}
+    files: {}
+    results:
+      example_value_1:
[…]

[…]
