+++ src/madsci_location_manager/tests/test_location_server.py
@@ -672,7 +672,7 @@ def test_shortest_transfer_path_direct(transfer_setup):
     assert path is not None
     assert len(path) == 1
     assert path[0].source_location_id == transfer_setup["locations"]["pickup"]
-    assert path[0].destination_location_id == transfer_setup["locations"]["processing"]
+    assert path[0].target_location_id == transfer_setup["locations"]["processing"]
     assert path[0].transfer_template.node_name == "robotarm_1"
 
 
@@ -716,9 +716,9 @@ def test_shortest_transfer_path_multi_hop(transfer_setup):
     assert path is not None
     assert len(path) == 2
     assert path[0].source_location_id == transfer_setup["locations"]["processing"]
-    assert path[0].destination_location_id == transfer_setup["locations"]["pickup"]
+    assert path[0].target_location_id == transfer_setup["locations"]["pickup"]
     assert path[1].source_location_id == transfer_setup["locations"]["pickup"]
-    assert path[1].destination_location_id == transfer_setup["locations"]["storage"]
+    assert path[1].target_location_id == transfer_setup["locations"]["storage"]
 
 
 def test_multi_leg_transfer_workflow_step_count(transfer_setup):
@@ -773,11 +773,11 @@ def test_multi_leg_transfer_workflow_node_ordering(transfer_setup):
 
     # Verify path structure matches expected route
     assert path[0].source_location_id == transfer_setup["locations"]["processing"]
-    assert path[0].destination_location_id == transfer_setup["locations"]["pickup"]
+    assert path[0].target_location_id == transfer_setup["locations"]["pickup"]
     assert path[0].transfer_template.node_name == "robotarm_1"
 
     assert path[1].source_location_id == transfer_setup["locations"]["pickup"]
-    assert path[1].destination_location_id == transfer_setup["locations"]["storage"]
+    assert path[1].target_location_id == transfer_setup["locations"]["storage"]
     assert path[1].transfer_template.node_name == "conveyor"
 
     # Generate workflow and verify step ordering
@@ -915,7 +915,7 @@ def test_plan_transfer_endpoint_direct(transfer_setup):
         "/transfer/plan",
         params={
             "source_location_id": transfer_setup["locations"]["pickup"],
-            "destination_location_id": transfer_setup["locations"]["processing"],
+            "target_location_id": transfer_setup["locations"]["processing"],
         },
     )
 
@@ -939,7 +939,7 @@ def test_plan_transfer_endpoint_multi_hop(transfer_setup):
         "/transfer/plan",
         params={
             "source_location_id": transfer_setup["locations"]["processing"],
-            "destination_location_id": transfer_setup["locations"]["storage"],
+            "target_location_id": transfer_setup["locations"]["storage"],
         },
     )
 
@@ -980,7 +980,7 @@ def test_plan_transfer_endpoint_no_path(transfer_setup):
         "/transfer/plan",
         params={
             "source_location_id": transfer_setup["locations"]["pickup"],
-            "destination_location_id": transfer_setup["locations"]["isolated"],
+            "target_location_id": transfer_setup["locations"]["isolated"],
         },
     )
 
@@ -998,7 +998,7 @@ def test_plan_transfer_endpoint_invalid_location(transfer_setup):
         "/transfer/plan",
         params={
             "source_location_id": "invalid_location_id",
-            "destination_location_id": transfer_setup["locations"]["processing"],
+            "target_location_id": transfer_setup["locations"]["processing"],
         },
     )
 
@@ -1014,14 +1014,14 @@ def test_plan_transfer_endpoint_invalid_location(transfer_setup):
         "/transfer/plan",
         params={
             "source_location_id": transfer_setup["locations"]["pickup"],
[…]
