### 2025-09-24 3c20cb3
Cancelled status now displays
