### 2025-09-24 951b493
Minor updates on utilization methods | Removed timout on the API call
