+++ src/madsci_client/madsci/client/event_client.py
@@ -115,7 +115,7 @@ class EventClient:
         if self.event_server:
             response = requests.get(
                 str(self.event_server) + "events",
-                timeout=10,
+                timeout=120,
                 params={"number": number, "level": level},
             )
             if not response.ok:
@@ -276,10 +276,9 @@ class EventClient:
                     params["save_to_file"] = "true"
                     params["output_path"] = output_path
 
-            response = requests.get(
+            response = requests.get(  # noqa
                 str(self.event_server) + "utilization/periods",
                 params=params,
-                timeout=30,
             )
             if not response.ok:
                 self.logger.error(
@@ -341,10 +340,9 @@ class EventClient:
                     params["save_to_file"] = "true"
                     params["output_path"] = output_path
 
-            response = requests.get(
+            response = requests.get(  # noqa
                 str(self.event_server) + "utilization/sessions",
                 params=params,
-                timeout=60,
             )
 
             if not response.ok:
@@ -401,10 +399,9 @@ class EventClient:
                     params["save_to_file"] = "true"
                     params["output_path"] = output_path
 
-            response = requests.get(
+            response = requests.get(  # noqa
                 str(self.event_server) + "utilization/users",
                 params=params,
-                timeout=60,
             )
 
             if not response.ok:
