+++ README.md
@@ -4,43 +4,44 @@
 
 | Name                                                                                          |    Stmts |     Miss |   Cover |   Missing |
 |---------------------------------------------------------------------------------------------- | -------: | -------: | ------: | --------: |
-| src/madsci\_client/madsci/client/\_\_init\_\_.py                                              |       14 |        0 |    100% |           |
+| src/madsci\_client/madsci/client/\_\_init\_\_.py                                              |       15 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/data\_client.py                                              |      126 |       31 |     75% |71, 87-95, 105-106, 115-125, 174-182, 189, 232-233, 252-253, 313, 330, 346-353 |
 | src/madsci\_client/madsci/client/event\_client.py                                             |      266 |       30 |     89% |64, 88-89, 270-273, 281-284, 289, 347, 395-398, 407-410, 415, 440, 445-446, 455-456, 469, 496, 500-504 |
 | src/madsci\_client/madsci/client/experiment\_client.py                                        |       57 |        0 |    100% |           |
+| src/madsci\_client/madsci/client/location\_client.py                                          |      144 |       66 |     54% |57, 74, 91-93, 103, 113, 131-142, 160-171, 223-235, 255-266, 294-306, 364-376, 424-435, 463-482 |
 | src/madsci\_client/madsci/client/node/\_\_init\_\_.py                                         |        4 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/node/abstract\_node\_client.py                               |       29 |        4 |     86% |26, 29, 95-96 |
 | src/madsci\_client/madsci/client/node/rest\_node\_client.py                                   |      115 |        2 |     98% |   33, 131 |
-| src/madsci\_client/madsci/client/resource\_client.py                                          |      590 |      183 |     69% |99, 124, 146, 173-176, 184, 226-229, 246, 277, 308-312, 325-340, 366, 398-401, 450-457, 463, 478, 489-490, 526, 552, 597-598, 633-635, 675-676, 713-714, 746, 778-779, 811-812, 844-845, 877-878, 907-908, 936-941, 969-971, 1020-1030, 1056-1059, 1107-1117, 1134, 1137-1148, 1176-1183, 1203-1206, 1244-1268, 1283-1289, 1336-1364, 1406-1444, 1473-1490, 1574-1578, 1592, 1615-1616, 1624-1628, 1644-1645, 1656 |
-| src/madsci\_client/madsci/client/workcell\_client.py                                          |      287 |       48 |     83% |123, 152, 180, 195, 244-248, 254, 276, 280, 304-309, 417-438, 465-493, 532, 540, 548, 553-554, 658, 687 |
+| src/madsci\_client/madsci/client/resource\_client.py                                          |      613 |      186 |     70% |100, 125, 147, 174-177, 185, 227-230, 247, 309-313, 326-341, 367, 399-402, 451-458, 464, 477, 488-489, 525, 551, 596-597, 632-634, 674-675, 712-713, 745, 777-778, 810-811, 843-844, 876-877, 906-907, 935-940, 968-970, 1019-1029, 1055-1058, 1106-1116, 1133, 1136-1147, 1175-1182, 1202-1205, 1248-1267, 1282-1288, 1335-1361, 1403-1441, 1470-1487, 1571-1575, 1589, 1612-1613, 1621-1625, 1641-1642, 1653, 1687, 1700, 1713-1723 |
[…]
