+++ src/madsci_workcell_manager/madsci/workcell_manager/workcell_server.py
@@ -368,7 +368,7 @@ class WorkcellManager(
             except Exception as e:
                 traceback.print_exc()
                 raise HTTPException(status_code=422, detail=str(e)) from e
-            return state_handler.save_workflow_definition(
+            return self.state_handler.save_workflow_definition(
                 workflow_definition=wf_def,
             )
         except HTTPException as e:

[…]
