+++ src/madsci_common/madsci/common/types/action_types.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 from datetime import datetime
 from enum import Enum
 from pathlib import Path
-from typing import Any, Literal, Union
+from typing import Any, Literal, Optional, Union
 
-from madsci.common.types.base_types import Error, MadsciBaseModel, PathLike
+from madsci.common.types.base_types import Error, MadsciBaseModel
 from madsci.common.types.datapoint_types import DataPoint
 from madsci.common.utils import localnow, new_ulid_str
 from pydantic import Field, TypeAdapter
@@ -65,7 +65,7 @@ class ActionRequest(MadsciBaseModel):
         default_factory=dict,
     )
     """Arguments for the action"""
-    files: dict[str, PathLike] = Field(
+    files: dict[str, Path] = Field(
         title="Action Files",
         description="Files sent along with the action.",
         default_factory=dict,
@@ -74,9 +74,9 @@ class ActionRequest(MadsciBaseModel):
 
     def failed(
         self,
-        errors: Error | list[Error] | str = [],
+        errors: Union[Error, list[Error], str] = [],
         data: dict[str, Any] = {},
-        files: dict[str, PathLike] = {},
+        files: dict[str, Path] = {},
     ) -> ActionFailed:
         """Create an ActionFailed response"""
         # * Convert errors to a list of errors if they are a single error or a string
@@ -93,9 +93,9 @@ class ActionRequest(MadsciBaseModel):
 
     def succeeded(
         self,
-        json_data: Json | ActionJSON | None = None,
-        files: PathLike | ActionFiles | None = None,
-        errors: Error | list[Error] | str = [],
+        json_data: Optional[Union[Json, ActionJSON]] = None,
+        files: Optional[Union[Path, ActionFiles]] = None,
+        errors: Union[Error, list[Error], str] = [],
     ) -> ActionSucceeded:
         """Create an ActionSucceeded response"""
         return ActionSucceeded(
@@ -107,9 +107,9 @@ class ActionRequest(MadsciBaseModel):
 
     def running(
         self,
-        json_data: Json | ActionJSON | None = None,
-        files: PathLike | ActionFiles | None = None,
-        errors: Error | list[Error] | str = [],
+        json_data: Optional[Union[Json, ActionJSON]] = None,
+        files: Optional[Union[Path, ActionFiles]] = None,
+        errors: Union[Error, list[Error], str] = [],
     ) -> ActionRunning:
         """Create an ActionRunning response"""
         return ActionRunning(
@@ -121,9 +121,9 @@ class ActionRequest(MadsciBaseModel):
 
     def not_ready(
         self,
-        errors: Error | list[Error] | str = [],
-        json_data: Json | ActionJSON | None = None,
-        files: PathLike | ActionFiles | None = None,
+        errors: Union[Error, list[Error], str] = [],
+        json_data: Optional[Union[Json, ActionJSON]] = None,
+        files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionNotReady:
         """Create an ActionNotReady response"""
         return ActionNotReady(
@@ -135,9 +135,9 @@ class ActionRequest(MadsciBaseModel):
 
     def cancelled(
         self,
-        errors: Error | list[Error] | str = [],
-        json_data: Json | ActionJSON | None = None,
-        files: PathLike | ActionFiles | None = None,
+        errors: Union[Error, list[Error], str] = [],
+        json_data: Optional[Union[Json, ActionJSON]] = None,
+        files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionCancelled:
         """Create an ActionCancelled response"""
         return ActionCancelled(
[…]
