+++ src/madsci_common/madsci/common/types/action_types.py
@@ -59,7 +59,7 @@ class ActionRequest(MadsciBaseModel):
         description="The name of the action to perform.",
     )
     """Name of the action to perform"""
-    args: dict[str, Any] | None = Field(
+    args: Optional[dict[str, Any]] = Field(
         title="Action Arguments",
         description="Arguments for the action.",
         default_factory=dict,
@@ -263,17 +263,17 @@ class ActionResult(MadsciBaseModel):
         description="The data generated by the step.",
         default=None,
     )
-    files: Path | ActionFiles | None = Field(
+    files: Optional[Union[Path, ActionFiles]] = Field(
         title="Step Files",
         description="A dictionary of files produced by the step.",
         default=None,
     )
-    datapoints: ActionDatapoints | None = Field(
+    datapoints: Optional[ActionDatapoints] = Field(
         title="Data Points",
         description="A dictionary of datapoints sent to the data manager by the step.",
         default=None,
     )
-    history_created_at: datetime | None = Field(
+    history_created_at: Optional[datetime] = Field(
         title="History Created At",
         description="The time the history was updated.",
         default_factory=localnow,
@@ -367,24 +367,28 @@ class ActionDefinition(MadsciBaseModel):
             return ""
         return v
 
-    args: dict[str, ArgumentDefinition] | list[ArgumentDefinition] = Field(
+    args: Union[dict[str, ArgumentDefinition], list[ArgumentDefinition]] = Field(
         title="Action Arguments",
         description="The arguments of the action.",
         default_factory=dict,
     )
-    locations: (
-        dict[str, LocationArgumentDefinition] | list[LocationArgumentDefinition]
-    ) = Field(
+    locations: Union[
+        dict[str, LocationArgumentDefinition], list[LocationArgumentDefinition]
+    ] = Field(
         title="Action Location Arguments",
         description="The location arguments of the action.",
         default_factory=dict,
     )
-    files: dict[str, FileArgumentDefinition] | list[FileArgumentDefinition] = Field(
-        title="Action File Arguments",
-        description="The file arguments of the action.",
-        default_factory=dict,
+    files: Union[dict[str, FileArgumentDefinition], list[FileArgumentDefinition]] = (
+        Field(
+            title="Action File Arguments",
+            description="The file arguments of the action.",
+            default_factory=dict,
+        )
     )
-    results: dict[str, ActionResultDefinitions] | list[ActionResultDefinitions] = Field(
+    results: Union[
+        dict[str, ActionResultDefinitions], list[ActionResultDefinitions]
+    ] = Field(
         title="Action Results",
         description="The results of the action.",
         default_factory=dict,
@@ -470,7 +474,7 @@ class ArgumentDefinition(MadsciBaseModel):
         title="Argument Required",
         description="Whether the argument is required.",
     )
-    default: Any | None = Field(
+    default: Optional[Any] = Field(
         title="Argument Default",
         description="The default value of the argument.",
         default=None,
@@ -504,7 +508,7 @@ class ActionResultDefinition(MadsciBaseModel):
         title="Result Label",
         description="The label of the result.",
     )
-    description: str | None = Field(
+    description: Optional[str] = Field(
         title="Result Description",
         description="A description of the result.",
         default=None,
@@ -543,7 +547,7 @@ class JSONActionResultDefinition(ActionResultDefinition):
         description="The type of the result.",
         default="json",
     )
-    data_type: str | None = Field(
+    data_type: Optional[str] = Field(
         title="Data Type",
         description="The type of the data.",
         default=None,
