### 2025-09-25 0f6e397
Enable redis persistence, remove vestigial workcell_directory method
