### 2025-09-25 3ddd4aa
Minor bugs in location types, workcell manager startup command
