### 2025-09-26 a248823
fixed 2 tests
