### 2025-09-26 eb46604
fixed 2 tests
