+++ src/madsci_workcell_manager/madsci/workcell_manager/workcell_actions.py
@@ -1,6 +1,6 @@
 """Built-in actions for the Workcell Manager, which don't require a node to be specified."""
 
-import time
+import asyncio
 from typing import Any, Optional, Union
 
 from madsci.client.location_client import LocationClient
@@ -20,9 +20,10 @@ from madsci.common.types.action_types import (
 )
 
 
-def wait(seconds: Union[int, float]) -> None:
+async def wait(seconds: Union[int, float]) -> ActionResult:
     """Waits for a specified number of seconds"""
-    time.sleep(seconds)
+    await asyncio.sleep(seconds)
+    return ActionSucceeded()
 
 
 class WorkcellTransferJSON(ActionJSON):
@@ -42,7 +43,7 @@ class WorkcellTransferJSON(ActionJSON):
 
 def transfer(  # noqa: C901
     source: str, target: str, await_completion: bool = True
-) -> ActionJSON:
+) -> ActionResult:
     """
     Transfer a single discrete object between locations.
 
@@ -106,7 +107,7 @@ def transfer(  # noqa: C901
             if not await_completion:
                 # Return immediately after successful enqueueing
                 result = ActionSucceeded(
-                    json_data=WorkcellTransferJSON.model_validate(
+                    data=WorkcellTransferJSON.model_validate(
                         {
                             "message": f"Transfer workflow enqueued from {source} to {target}",
                             "workflow_id": workflow.workflow_id,
@@ -119,7 +120,7 @@ def transfer(  # noqa: C901
             # Check final workflow status after completion
             elif workflow.status.completed:
                 result = ActionSucceeded(
-                    data=WorkcellTransferJSON.model_validate(
+                    json_data=WorkcellTransferJSON.model_validate(
                         {
                             "message": f"Transfer completed from {source} to {target}",
                             "workflow_id": workflow.workflow_id,
@@ -136,15 +137,19 @@ def transfer(  # noqa: C901
                     else ""
                 )
                 description = workflow.status.description or "Unknown error"
-                raise Exception(f"Transfer workflow failed{step_info}: {description}")
+                result = ActionFailed(
+                    errors=[f"Transfer workflow failed{step_info}: {description}"]
+                )
             elif workflow.status.cancelled:
-                raise Exception("Transfer workflow was cancelled")
+                result = ActionFailed(errors=["Transfer workflow was cancelled"])
             else:
-                raise Exception(
-                    f"Transfer workflow ended in unexpected state: {workflow.status.model_dump()}"
+                result = ActionFailed(
+                    errors=[
+                        f"Transfer workflow ended in unexpected state: {workflow.status.model_dump()}"
+                    ]
                 )
     except Exception as e:
-        raise Exception(f"Unexpected error in transfer: {e}") from None
+        result = ActionFailed(errors=[f"Unexpected error in transfer: {e}"])
     return result
 
 
@@ -187,7 +192,7 @@ def _resolve_location_identifier(
 
 def transfer_resource(
     resource_id: str, target: str, await_completion: bool = True
-) -> WorkcellTransferJSON:
+) -> ActionResult:
     """
     Transfer a specific resource from its current location to a target.
 
@@ -207,13 +212,17 @@ def transfer_resource(
 
         # Validate context configuration
         if not context.location_server_url:
-            raise Exception(
-                "Location server URL not configured. Set location_server_url in context."
+            return ActionFailed(
+                errors=[
+                    "Location server URL not configured. Set location_server_url in context."
+                ]
             )
 
         if not context.resource_server_url:
-            raise Exception(
-                "Resource server URL not configured. Set resource_server_url in context."
+            return ActionFailed(
+                errors=[
[…]
