### 2025-09-26 2779c93
gradually fixing pytests
