### 2025-09-26 2646c07
gradually fixing pytests
