### 2025-09-26 3fc3cf1
gradually fixing pytests
