### 2025-09-26 837e7fc
gradually fixing pytests
