### 2025-09-26 2502121
gradually fixing pytests
