### 2025-09-26 935bc83
gradually fixing pytests
