+++ src/madsci_workcell_manager/tests/test_workcell_engine.py
@@ -198,19 +198,11 @@ def test_run_single_step_with_update_parameters(
     engine: Engine, state_handler: WorkcellStateHandler
 ) -> None:
     """Test running a step in a workflow."""
-    step = Step(
-        name="Test Step 1",
-        action="test_action",
-        node="node1",
-        args={},
-        data_labels={"test": "test_label"},
-    )
+    step = Step(name="Test Step 1", action="test_action", node="node1", args={})
     workflow = Workflow(
         name="Test Workflow",
         parameters=WorkflowParameters(
-            feed_forward=[
-                ParameterFeedForwardJson(key="test_param", step=0, label="test_label")
-            ]
+            feed_forward=[ParameterFeedForwardJson(key="test_param", step=0)]
         ),
         steps=[step],
         status=WorkflowStatus(running=True),
@@ -224,8 +216,7 @@ def test_run_single_step_with_update_parameters(
         "madsci.workcell_manager.workcell_engine.find_node_client"
     ) as mock_client:
         mock_client.return_value.send_action.return_value = ActionResult(
-            status=ActionStatus.SUCCEEDED,
-            json_data=UpdateParamJSON.model_validate({"test": "test_value"}),
+            status=ActionStatus.SUCCEEDED, json_data="test_value"
         )
         thread = engine.run_step(workflow.workflow_id)
         thread.join()
@@ -342,7 +333,7 @@ def test_handle_data_and_files_with_data(
 
     with patch.object(engine.data_client, "submit_datapoint") as mock_submit:
         updated_result = engine.handle_data_and_files(step, workflow, action_result)
-        assert "data" in updated_result.datapoints
+        assert "json_data" in updated_result.datapoints.model_dump()
         mock_submit.assert_called_once()
         submitted_datapoint = mock_submit.call_args[0][0]
         assert isinstance(submitted_datapoint, ValueDataPoint)
@@ -377,11 +368,11 @@ def test_handle_data_and_files_with_files(
         patch("pathlib.Path.exists", return_value=True),
     ):
         updated_result = engine.handle_data_and_files(step, workflow, action_result)
-        assert "files" in updated_result.datapoints
+        assert "file" in updated_result.datapoints.model_dump()
         mock_submit.assert_called_once()
         submitted_datapoint = mock_submit.call_args[0][0]
         assert isinstance(submitted_datapoint, FileDataPoint)
-        assert submitted_datapoint.label == "files"
+        assert submitted_datapoint.label == "file"
         assert submitted_datapoint.path == "/path/to/file"
 
 
