### 2025-09-26 90895d2
Update test_workcell_engine.py
