### 2025-09-26 ab777e5
Update test_workcell_engine.py
