### 2025-09-27 a0ac16d
Fix names/docs in advanced_example_node
