### 2025-09-27 063ce9f
Cleanup old node definition/info artifacts
