+++ example_lab/example_modules/node_definitions/advanced_example_node.node.info.yaml
@@ -1,230 +1,0 @@
-node_name: advanced_example_node.node
-node_id: 01K5W0F6WKCRCMYVWD2XQBNA6V
-node_description: null
-node_type: device
-module_name: advanced_example_node
-module_version: 0.0.1
-capabilities:
-  get_info: true
-  get_state: true
-  get_status: true
-  send_action: true
-  get_action_result: true
-  get_action_history: true
-  action_files: true
-  send_admin_commands: true
-  set_config: true
-  get_resources: false
-  get_log: true
-  admin_commands:
-  - get_location
-  - lock
-  - reset
-  - shutdown
-  - unlock
-node_url: null
-actions:
-  return_none:
-    name: return_none
-    description: Run a protocol on the liquid handler
-    args: {}
-    locations: {}
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-  return_json:
-    name: return_json
-    description: Return a JSON object
-    args: {}
-    locations: {}
-    files: {}
-    results:
-      data:
-        result_label: data
-        description: null
-        result_type: json
-        data_type: int
-    blocking: false
-    asynchronous: true
-  return_dict:
-    name: return_dict
-    description: Return a JSON object
-    args: {}
-    locations: {}
-    files: {}
-    results:
-      data:
-        result_label: data
-        description: null
-        result_type: json
-        data_type: dict
-    blocking: false
-    asynchronous: true
-  return_file:
-    name: return_file
-    description: Return a file
-    args: {}
-    locations: {}
-    files: {}
-    results:
-      file:
-        result_label: file
-        description: null
-        result_type: file
-    blocking: false
-    asynchronous: true
-  return_file_and_json:
-    name: return_file_and_json
-    description: Return a file
-    args: {}
-    locations: {}
-    files: {}
-    results:
-      data:
-        result_label: data
-        description: null
-        result_type: json
-        data_type: int
-      file:
-        result_label: file
-        description: null
-        result_type: file
-    blocking: false
-    asynchronous: true
-  return_labeled_json_values:
-    name: return_labeled_json_values
-    description: Return labeled JSON values
-    args: {}
-    locations: {}
-    files: {}
-    results:
-      example_value_1:
-        result_label: example_value_1
-        description: null
-        result_type: json
-        data_type: int
-      example_value_2:
-        result_label: example_value_2
-        description: null
-        result_type: json
-        data_type: int
-    blocking: false
-    asynchronous: true
-  return_labeled_file_values:
-    name: return_labeled_file_values
-    description: Return labeled file values
-    args: {}
-    locations: {}
-    files: {}
-    results:
-      log_file_1:
-        result_label: log_file_1
-        description: null
-        result_type: file
-      log_file_2:
-        result_label: log_file_2
-        description: null
-        result_type: file
-    blocking: false
-    asynchronous: true
-  return_labeled_file_and_json_values:
-    name: return_labeled_file_and_json_values
-    description: Return labeled file and json values
-    args: {}
-    locations: {}
-    files: {}
-    results:
-      example_value_1:
-        result_label: example_value_1
-        description: null
-        result_type: json
-        data_type: int
-      example_value_2:
-        result_label: example_value_2
-        description: null
-        result_type: json
-        data_type: int
-      log_file_1:
-        result_label: log_file_1
-        description: null
-        result_type: file
-      log_file_2:
-        result_label: log_file_2
-        description: null
-        result_type: file
-    blocking: false
-    asynchronous: true
-config:
-  node_definition: node_definitions/advanced_example_node.node.yaml
-  node_info_path: null
-  update_node_files: true
-  status_update_interval: 2.0
-  state_update_interval: 2.0
-  node_url: http://localhost:2003/
-  uvicorn_kwargs: {}
-  device_number: 0
-config_schema:
[…]
