+++ example_lab/example_modules/default.node.info.yaml
@@ -1,151 +1,0 @@
-node_name: default.node
-node_id: 01K52C4N6420DR92V1JDHD26AM
-node_description: null
-node_type: device
-module_name: liquid_handler_node
-module_version: 0.0.1
-capabilities:
-  get_info: true
-  get_state: true
-  get_status: true
-  send_action: true
-  get_action_result: true
-  get_action_history: true
-  action_files: true
-  send_admin_commands: true
-  set_config: true
-  get_resources: false
-  get_log: true
-  admin_commands:
-  - get_location
-  - lock
-  - reset
-  - shutdown
-  - unlock
-node_url: null
-actions:
-  run_command:
-    name: run_command
-    description: Run a command on the liquid handler. Shows returning both JSON and
-      file data.
-    args:
-      command:
-        name: command
-        description: ''
-        argument_type: str
-        required: true
-        default: null
-    locations: {}
-    files: {}
-    results:
-      command:
-        result_label: command
-        description: null
-        result_type: json
-      log_file:
-        result_label: log_file
-        description: null
-        result_type: file
-    blocking: false
-    asynchronous: true
-  run_protocol:
-    name: run_protocol
-    description: Run a protocol on the liquid handler
-    args: {}
-    locations: {}
-    files:
-      protocol:
-        name: protocol
-        description: ''
-        argument_type: file
-        required: true
-        default: null
-    results:
-      file:
-        result_label: file
-        description: null
-        result_type: file
-    blocking: false
-    asynchronous: true
-  quick_action:
-    name: quick_action
-    description: A quick action that returns nothing
-    args: {}
-    locations: {}
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-config:
-  node_definition: default.node.yaml
-  node_info_path: null
-  update_node_files: true
-  status_update_interval: 2.0
-  state_update_interval: 2.0
-  node_url: http://localhost:2000/
-  uvicorn_kwargs: {}
-  device_number: 0
-config_schema:
-  description: Configuration for the liquid handler node module.
-  properties:
-    node_definition:
-      anyOf:
-      - type: string
-      - format: path
-        type: string
-      - type: 'null'
-      default: default.node.yaml
-      description: Path to the node definition file to use. If set, the node will
-        load the definition from this file on startup. Otherwise, a default configuration
-        will be created.
-      title: Node Definition File
-    node_info_path:
-      anyOf:
-      - type: string
-      - format: path
-        type: string
-      - type: 'null'
-      default: null
-      description: Path to export the generated node info file. If not set, will use
-        the node name and the node_definition's path.
-      title: Node Info Path
-    update_node_files:
-      default: true
-      description: Whether to update the node definition and info files on startup.
-        If set to False, the node will not update the files even if they are out of
-        date.
-      title: Update Node Files
-      type: boolean
-    status_update_interval:
-      anyOf:
-      - type: number
-      - type: 'null'
-      default: 2.0
-      description: The interval in seconds at which the node should update its status.
-      title: Status Update Interval
-    state_update_interval:
-      anyOf:
-      - type: number
-      - type: 'null'
-      default: 2.0
-      description: The interval in seconds at which the node should update its state.
-      title: State Update Interval
-    node_url:
-      default: http://127.0.0.1:2000/
-      description: The URL used to communicate with the node. This is the base URL
-        for the REST API.
-      format: uri
-      minLength: 1
-      title: Node URL
-      type: string
-    uvicorn_kwargs:
-      additionalProperties: true
-      description: Configuration for the Uvicorn server that runs the REST API.
-      title: Uvicorn Configuration
-      type: object
-    device_number:
-      default: 0
-      title: Device Number
-      type: integer
-  title: LiquidHandlerConfig
-  type: object
