+++ example_lab/example_modules/node_definitions/advanced_example_node.node.yaml
@@ -1,7 +1,0 @@
-node_name: advanced_example_node.node
-node_id: 01K5W0F6WKCRCMYVWD2XQBNA6V
-node_description: null
-node_type: device
-module_name: advanced_example_node
-module_version: 0.0.1
-capabilities: null
