### 2025-09-29 4d48bcb
Minor cleanup
