+++ example_lab/managers/example_workcell.yaml
@@ -1,53 +1,0 @@
-workcell_name: Example Workcell
-manager_type: workcell_manager
-workcell_id: 01JK7069DXAKAMWE0PDX6EY1PC
-description: An example workcell, consisiting of a (fake) liquid handler, a (fake)
-  plate reader, and a (fake) robot arm.
-nodes:
-  liquidhandler_1: http://localhost:2000/
-  robotarm_1: http://localhost:2001/
-  platereader_1: http://localhost:2002/
-  advanced_example_node: http://localhost:2003/
-locations:
-- location_name: location_1
-  location_id: 01JQCRA5HNG94DMQ9APVV5R1B1
-  description: null
-  lookup:
-    robotarm_1:
-    - 1
-    - 2
-    - 3
-    - 4
-  resource_definition:
-    resource_name: slot_1
-    resource_name_prefix: null
-    resource_class: ''
-    base_type: slot
-    resource_description: null
-    owner: {}
-    custom_attributes: null
-    capacity: 1
-    default_children: null
-    default_child_template: null
-    default_child_quantity: null
-- location_name: location_2
-  location_id: 01JR157336G2GXT79RQWZ1PE9F
-  description: null
-  lookup:
-    robotarm_1:
-    - 1
-    - 2
-    - 3
-    - 4
-  resource_definition:
-    resource_name: slot_2
-    resource_name_prefix: null
-    resource_class: ''
-    base_type: slot
-    resource_description: null
-    owner: {}
-    custom_attributes: null
-    capacity: 1
-    default_children: null
-    default_child_template: null
-    default_child_quantity: null
