### 2025-09-29 cbc53bd
Fix flaky data server query test
