### 2025-09-30 2b9d414
Switching from middleware to routes-based approach
