+++ src/madsci_node_module/madsci/node_module/rest_node_module.py
@@ -1,15 +1,13 @@
 """REST-based Node Module helper classes."""
 
-import asyncio
-import inspect
-import json
 import os
 import signal
 import tempfile
 import time
 from contextlib import asynccontextmanager
 from pathlib import Path
-from typing import Any, Callable, Optional, Union
+from typing import Any, Callable, Optional
+from zipfile import ZipFile
 
 from fastapi import Request, Response
 from fastapi.applications import FastAPI
@@ -19,10 +17,14 @@ from fastapi.routing import APIRouter
 from madsci.client.node.rest_node_client import RestNodeClient
 from madsci.common.ownership import global_ownership_info
 from madsci.common.types.action_types import (
-    ActionDatapoints,
-    ActionJSON,
+    ActionFiles,
     ActionRequest,
     ActionResult,
+    ActionStatus,
+    create_action_request_model,
+    create_action_result_model,
+    extract_file_parameters,
+    extract_file_result_definitions,
 )
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.base_types import Error
@@ -39,9 +41,8 @@ from madsci.common.utils import new_ulid_str
 from madsci.node_module.abstract_node_module import (
     AbstractNode,
 )
-from madsci.node_module.helpers import ActionResultWithFiles
 from pydantic import AnyUrl
-from starlette.middleware.base import BaseHTTPMiddleware
+from starlette.responses import FileResponse
 
 
 class RestNode(AbstractNode):
@@ -85,83 +86,6 @@ class RestNode(AbstractNode):
 
         return upload_files
 
-    def _handle_pure_return(self, response: ActionResult) -> Response:
-        """Handle pure return values from actions."""
-        if isinstance(response.json_data, ActionJSON):
-            response.json_data = response.json_data.model_dump(mode="json")
-        if isinstance(response.datapoints, ActionDatapoints):
-            response.datapoints = response.datapoints.model_dump(mode="json")
-        if response.json_data is not None and response.datapoints is not None:
-            response = (response.json_data, response.datapoints)
-        elif response.json_data is not None:
-            response = response.json_data
-        elif response.datapoints is not None:
-            response = response.datapoints
-        else:
-            response = {}
-        return Response(content=json.dumps(response), media_type="application/json")
-
-    def check_action_mode(
-        self, request: Request, json_data: dict[str, Any]
-    ) -> tuple[bool, ActionRequest]:
-        """Check if action is run from swaggerdocs or the client"""
-        if "action_name" in json_data:
-            if "args" in json_data:
-                json_data["args"] = json.loads(json_data["args"])
-            action_request = ActionRequest(**json_data)
-            pure_return = False
-        else:
-            action_name = Path(request.url.components.path).name
-            pure_return = True
-            action_request = ActionRequest(action_name=action_name, args=json_data)
-        return pure_return, action_request
-
-    class ActionMiddleware(BaseHTTPMiddleware):
-        """Middleware to handle action requests. allows use both through client and Swaggerdocs"""
-
-        rest_node_module: Optional["RestNode"] = None
-
-        async def dispatch(self, request: Request, call_next: Callable) -> Any:
-            """Middleware function to handle action requests."""
-            if "action" in str(request.url) and request.method == "POST":
-                files = []
-                json_data = request.query_params.__dict__["_dict"]
-                pure_return, action_request = self.rest_node_module.check_action_mode(
-                    request, json_data
-                )
-                files = await self.rest_node_module._get_request_files(request)
-                response = self.rest_node_module.run_action(
-                    action_request.action_name,
-                    json.dumps(action_request.args),
-                    files,
-                    action_request.action_id,
-                )
-                if pure_return:
-                    while (
[…]
