### 2025-09-30 4c731f2
Add create_dynamic_model helper
