### 2025-09-30 f1f7386
Simplifying json and datapoint result handling
