+++ example_lab/example_modules/advanced_example_module.py
@@ -4,19 +4,12 @@ from pathlib import Path
 
 from madsci.common.types.action_types import (
     ActionFiles,
-    ActionJSON,
 )
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.node_types import RestNodeConfig
 from madsci.node_module.helpers import action
 from madsci.node_module.rest_node_module import RestNode
-
-
-class ExampleJSONData(ActionJSON):
-    """Example of returned JSON data with labeled values"""
-
-    example_value_1: str
-    example_value_2: int
+from pydantic import BaseModel, Field
 
 
 class ExampleFileData(ActionFiles):
@@ -24,6 +17,28 @@ class ExampleFileData(ActionFiles):
 
     log_file_1: Path
     log_file_2: Path
+
+
+class AnalysisResult(BaseModel):
+    """Example custom pydantic model for analysis results"""
+
+    sample_id: str = Field(description="Unique identifier for the sample")
+    concentration: float = Field(description="Measured concentration in mg/mL")
+    ph_level: float = Field(description="pH level of the sample")
+    temperature: float = Field(description="Temperature in Celsius during measurement")
+    quality_score: int = Field(description="Quality score from 0-100", ge=0, le=100)
+    notes: str = Field(default="", description="Additional notes about the analysis")
+
+
+class ExperimentMetadata(BaseModel):
+    """Example complex pydantic model with nested data"""
+
+    experiment_id: str = Field(description="Unique experiment identifier")
+    start_time: str = Field(description="ISO timestamp when experiment started")
+    conditions: dict[str, float] = Field(description="Experimental conditions")
+    parameters: list[str] = Field(description="List of measured parameters")
+    status: str = Field(description="Current experiment status")
+    results_summary: dict[str, int] = Field(description="Summary of results count")
 
 
 class AdvancedExampleConfig(RestNodeConfig):
@@ -81,10 +96,10 @@ class AdvancedExampleNode(RestNode):
         return 5, path
 
     @action
-    def return_labeled_json_values(self) -> ExampleJSONData:
+    def return_labeled_json_values(self) -> dict:
         """Return labeled JSON values"""
         self.logger.log("returning labeled json")
-        return ExampleJSONData(example_value_1="test", example_value_2=5)
+        return {"example_value_1": "test", "example_value_2": 5}
 
     @action
     def return_labeled_file_values(self) -> ExampleFileData:
@@ -102,7 +117,7 @@ class AdvancedExampleNode(RestNode):
     @action
     def return_labeled_file_and_json_values(
         self,
-    ) -> tuple[ExampleJSONData, ExampleFileData]:
+    ) -> tuple[dict, ExampleFileData]:
         """Return labeled file and json values"""
 
         with (Path.home() / "test1.txt").open("w") as f:
@@ -112,9 +127,61 @@ class AdvancedExampleNode(RestNode):
             self.logger.log_info(f.write("test2"))
         path2 = Path.home() / "test2.txt"
 
-        return ExampleJSONData(
-            example_value_1="test", example_value_2=5
-        ), ExampleFileData(log_file_1=path1, log_file_2=path2)
+        return {"example_value_1": "test", "example_value_2": 5}, ExampleFileData(
+            log_file_1=path1, log_file_2=path2
+        )
+
+    @action
+    def return_analysis_result(self, sample_id: str = "SAMPLE_001") -> AnalysisResult:
+        """Return a custom pydantic model with analysis results"""
+        self.logger.log(f"returning analysis result for {sample_id}")
+        return AnalysisResult(
+            sample_id=sample_id,
+            concentration=15.75,
+            ph_level=7.2,
+            temperature=22.5,
+            quality_score=92,
[…]
