### 2025-10-01 c5633cf
Simplifying openapi schema generation
