+++ src/madsci_node_module/madsci/node_module/rest_node_module.py
@@ -6,7 +6,7 @@ import tempfile
 import time
 from contextlib import asynccontextmanager
 from pathlib import Path
-from typing import Any, Callable, Optional
+from typing import Any, Callable, Optional, Type
 from zipfile import ZipFile
 
 from fastapi import Request, Response
@@ -21,8 +21,10 @@ from madsci.common.types.action_types import (
     ActionRequest,
     ActionResult,
     ActionStatus,
+    DatapointActionResultDefinition,
+    FileActionResultDefinition,
+    JSONActionResultDefinition,
     create_action_request_model,
-    create_action_result_model,
     extract_file_parameters,
     extract_file_result_definitions,
 )
@@ -41,12 +43,12 @@ from madsci.common.utils import new_ulid_str
 from madsci.node_module.abstract_node_module import (
     AbstractNode,
 )
-from pydantic import AnyUrl
+from pydantic import AnyUrl, BaseModel, Field, create_model
 from starlette.responses import FileResponse
 
 
 class RestNode(AbstractNode):
-    """REST-based node implementation and helper class. Inherit from this class to create a new REST-based node class."""
+    """REST-based node implementation with better OpenAPI documentation and result handling."""
 
     rest_api = None
     """The REST API server for the node."""
@@ -133,8 +135,74 @@ class RestNode(AbstractNode):
         action_id: str,
     ) -> ActionResult:
         """Get the result of an action on the node."""
+        return super().get_action_result(action_id)
+
+    def get_action_result_dict(
+        self,
+        action_id: str,
+    ) -> dict[str, Any]:
+        """Get the result of an action on the node as a dictionary for API responses."""
         action_response = super().get_action_result(action_id)
-        return ActionResult.model_validate(action_response)
+
+        if isinstance(action_response, dict):
+            return self._process_dict_response(action_response)
+        return self._process_action_result_response(action_response)
+
+    def _process_dict_response(self, response: dict) -> dict[str, Any]:
+        """Process a response that's already a dictionary."""
+        result_dict = response.copy()
+
+        # Ensure status is a string
+        if hasattr(result_dict.get("status"), "value"):
+            result_dict["status"] = result_dict["status"].value
+
+        # Handle files serialization if present
+        files_field = result_dict.get("files")
+        if files_field is not None:
+            result_dict["files"] = self._serialize_files_field(files_field)
+
+        return result_dict
+
+    def _process_action_result_response(
+        self, action_response: ActionResult
+    ) -> dict[str, Any]:
+        """Process an ActionResult object response."""
+        result_dict = {
+            "action_id": action_response.action_id,
+            "status": action_response.status.value
+            if hasattr(action_response.status, "value")
+            else action_response.status,
+            "errors": [
+                error.model_dump() if hasattr(error, "model_dump") else error
+                for error in action_response.errors
+            ],
+            "json_result": action_response.json_result,
+            "datapoints": action_response.datapoints.model_dump()
+            if action_response.datapoints
+            else None,
+        }
+
+        # Handle files with proper serialization
+        if action_response.files:
+            result_dict["files"] = self._serialize_files_field(action_response.files)
+        else:
+            result_dict["files"] = None
+
+        return result_dict
+
+    def _serialize_files_field(self, files_field: Any) -> Any:
+        """Serialize files field to JSON-compatible format."""
+        if isinstance(files_field, Path):
+            return str(files_field)
+        if hasattr(files_field, "model_dump"):
+            files_dict = files_field.model_dump()
+            for key, value in files_dict.items():
+                if isinstance(value, Path):
+                    files_dict[key] = str(value)
+            return files_dict
+        if isinstance(files_field, str):
[…]
