+++ src/madsci_node_module/tests/test_rest_node_module.py
@@ -1,12 +1,14 @@
 """Automated pytest unit tests for the RestNode class."""
 
 import io
+import tempfile
 import time
 from pathlib import Path
 
 import pytest
 from fastapi.testclient import TestClient
 from madsci.common.types.action_types import (
+    ActionFiles,
     ActionRequest,
     ActionResult,
     ActionStatus,
@@ -15,6 +17,8 @@ from madsci.common.types.action_types import (
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.event_types import Event
 from madsci.common.types.node_types import NodeDefinition, NodeInfo, NodeStatus
+from madsci.node_module.helpers import action
+from pydantic import BaseModel, Field
 from ulid import ULID
 
 from madsci_node_module.tests.test_node import TestNode, TestNodeConfig
@@ -765,3 +769,457 @@ def test_mixed_pydantic_and_file_result_processing(test_node: TestNode) -> None:
         content = result.files.read_text()
         assert "mixed_test_456" in content
         assert "raw_data" in content
+
+
+# Additional tests for enhanced OpenAPI documentation and result handling features
+
+
+class TestActionFiles(ActionFiles):
+    """Test ActionFiles subclass for testing labeled file returns."""
+
+    log_file: Path
+    data_file: Path
+
+
+class TestResultModel(BaseModel):
+    """Test Pydantic model for testing custom return types."""
+
+    sample_id: str = Field(description="Sample identifier")
+    value: float = Field(description="Measured value")
+    status: str = Field(description="Processing status")
+
+
+class EnhancedTestNode(TestNode):
+    """Extended test node for validating enhanced REST endpoint generation."""
+
+    @action
+    def return_int(self) -> int:
+        """Returns a simple integer."""
+        return 42
+
+    @action
+    def return_dict(self) -> dict:
+        """Returns a dictionary."""
+        return {"key": "value", "number": 123}
+
+    @action
+    def return_file(self) -> Path:
+        """Returns a single file."""
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".txt", delete=False) as f:
+            f.write("test content")
+            return Path(f.name)
+
+    @action
+    def return_labeled_files(self) -> TestActionFiles:
+        """Returns multiple labeled files."""
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".txt", delete=False) as f:
+            f.write("log content")
+            log_file = Path(f.name)
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".csv", delete=False) as f:
+            f.write("data content")
+            data_file = Path(f.name)
+        return TestActionFiles(log_file=log_file, data_file=data_file)
+
+    @action
+    def return_custom_model(self) -> TestResultModel:
+        """Returns a custom Pydantic model."""
+        return TestResultModel(sample_id="SAMPLE_001", value=15.75, status="completed")
+
+    @action
+    def return_mixed(self) -> tuple[dict, Path]:
+        """Returns both JSON data and a file."""
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".txt", delete=False) as f:
+            f.write("mixed content")
+            return {"result": "success"}, Path(f.name)
+
+    @action
+    def take_file_input(self, input_file: Path, param: str = "default") -> dict:
+        """Action that takes a file input."""
+        return {"processed": param, "file_size": input_file.stat().st_size}
+
+
+@pytest.fixture
+def enhanced_test_node():
+    """Create an enhanced test node instance."""
+    node_definition = NodeDefinition(
+        node_name="Enhanced Test Node",
+        module_name="enhanced_test_node",
[…]
