+++ src/madsci_node_module/tests/test_rest_node_module.py
@@ -3,6 +3,7 @@
 import io
 import tempfile
 import time
+from datetime import datetime
 from pathlib import Path
 
 import pytest
@@ -12,6 +13,8 @@ from madsci.common.types.action_types import (
     ActionRequest,
     ActionResult,
     ActionStatus,
+    RestActionResult,
+    create_dynamic_model,
     extract_file_parameters,
 )
 from madsci.common.types.admin_command_types import AdminCommandResponse
@@ -1167,20 +1170,21 @@ class TestEnhancedActionResultTypeMapping:
 
     def test_labeled_files_return_structure(self, enhanced_client):
         """Test that ActionFiles subclass creates proper files structure."""
-        response = enhanced_client.post("/action/return_labeled_files", json={})
-        action_id = response.json()["action_id"]
+        with enhanced_client as client:
+            time.sleep(0.5)  # Wait for node to be ready
 
-        response = enhanced_client.post(
-            f"/action/return_labeled_files/{action_id}/start"
-        )
-        result = response.json()
+            response = client.post("/action/return_labeled_files", json={})
+            action_id = response.json()["action_id"]
 
-        assert "files" in result
-        files = result["files"]
-        assert "log_file" in files
-        assert "data_file" in files
-        assert files["log_file"].endswith(".txt")
-        assert files["data_file"].endswith(".csv")
+            response = client.post(f"/action/return_labeled_files/{action_id}/start")
+            result = response.json()
+
+            assert "files" in result
+            files = result["files"]
+            # Files are now returned as file keys (list) rather than dict
+            assert isinstance(files, list)
+            assert "log_file" in files
+            assert "data_file" in files
 
     def test_mixed_return_maps_to_both_fields(self, enhanced_client):
         """Test that tuple return populates both json_result and files."""
@@ -1223,3 +1227,269 @@ class TestEnhancedBackwardCompatibility:
         assert response.status_code == 200
         status = response.json()
         assert status in ["not_started", "running", "succeeded", "failed", "unknown"]
+
+
+# Test models to match the proposal example
+class ReadSensorValue(BaseModel):
+    """Result of reading a sensor, returned by the python action function"""
+
+    value: int
+    """Value from the sensor"""
+    timestamp: datetime
+    """Timestamp for when the sensor read occurred"""
+
+
+class TestNewActionResults(BaseModel):
+    """Test custom pydantic model for new action results"""
+
+    sensor_id: str = Field(description="Sensor identifier")
+    temperature: float = Field(description="Temperature measurement")
+    humidity: float = Field(description="Humidity measurement")
+    status: str = Field(description="Sensor status")
+
+
+class TestProposalExampleActionResultHandling:
+    """Test the action result handling features as described in the proposal."""
+
+    @pytest.fixture
+    def proposal_test_client(self) -> TestClient:
+        """Create a test client with actions that match the proposal examples."""
+
+        class ProposalTestNode(TestNode):
+            """Test node with actions matching the proposal."""
+
+            @action
+            def read_sensor(self) -> ReadSensorValue:
+                """Implementation matching the proposal example - returns simple ReadSensorValue."""
+                return ReadSensorValue(value=10, timestamp=datetime.now())
+
+            @action
+            def get_temperature_reading(self) -> TestNewActionResults:
+                """Returns a complex pydantic model."""
+                return TestNewActionResults(
+                    sensor_id="TEMP_001",
+                    temperature=23.5,
+                    humidity=45.2,
+                    status="active",
+                )
+
+            @action
+            def return_simple_int(self) -> int:
+                """Returns a simple integer."""
+                return 42
+
+            @action
+            def return_simple_string(self) -> str:
+                """Returns a simple string."""
+                return "hello world"
+
+            @action
+            def return_dict_data(self) -> dict:
[…]
