+++ src/madsci_common/madsci/common/types/action_types.py
@@ -6,21 +6,13 @@ import inspect
 from datetime import datetime
 from enum import Enum
 from pathlib import Path
-from typing import Any, Literal, Optional, Union
+from typing import Annotated, Any, Literal, Optional, Union
 
 from madsci.common.types.base_types import Error, MadsciBaseModel
 from madsci.common.utils import localnow, new_ulid_str
-from pydantic import Field, TypeAdapter, create_model
+from pydantic import Field, create_model
 from pydantic.functional_validators import field_validator, model_validator
 from pydantic.types import Discriminator, Tag
-from typing_extensions import Annotated, TypeAliasType
-
-Json = TypeAliasType(
-    "Json",
-    "Union[dict[str, Json], list[Json], str, int, float, bool, None]",
-)
-
-ta = TypeAdapter(Json)
 
 
 class ActionStatus(str, Enum):
@@ -89,7 +81,7 @@ class ActionRequest(MadsciBaseModel):
     def succeeded(
         self,
         errors: Union[Error, list[Error], str] = [],
-        json_result: Optional[Union[Json, ActionJSON]] = None,
+        json_result: Any = None,
         files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionSucceeded:
         """Create an ActionSucceeded response"""
@@ -103,7 +95,7 @@ class ActionRequest(MadsciBaseModel):
     def running(
         self,
         errors: Union[Error, list[Error], str] = [],
-        json_result: Optional[Union[Json, ActionJSON]] = None,
+        json_result: Any = None,
         files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionRunning:
         """Create an ActionRunning response"""
@@ -117,7 +109,7 @@ class ActionRequest(MadsciBaseModel):
     def not_ready(
         self,
         errors: Union[Error, list[Error], str] = [],
-        json_result: Optional[Union[Json, ActionJSON]] = None,
+        json_result: Any = None,
         files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionNotReady:
         """Create an ActionNotReady response"""
@@ -131,7 +123,7 @@ class ActionRequest(MadsciBaseModel):
     def cancelled(
         self,
         errors: Union[Error, list[Error], str] = [],
-        json_result: Optional[Union[Json, ActionJSON]] = None,
+        json_result: Any = None,
         files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionCancelled:
         """Create an ActionCancelled response"""
@@ -145,7 +137,7 @@ class ActionRequest(MadsciBaseModel):
     def paused(
         self,
         errors: Union[Error, list[Error], str] = [],
-        json_result: Optional[Union[Json, ActionJSON]] = None,
+        json_result: Any = None,
         files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionResult:
         """Create an ActionResult response"""
@@ -160,7 +152,7 @@ class ActionRequest(MadsciBaseModel):
     def not_started(
         self,
         errors: Union[Error, list[Error], str] = [],
-        json_result: Optional[Union[Json, ActionJSON]] = None,
+        json_result: Any = None,
         files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionResult:
         """Create an ActionResult response"""
@@ -175,7 +167,7 @@ class ActionRequest(MadsciBaseModel):
     def unknown(
         self,
         errors: Union[Error, list[Error], str] = [],
-        json_result: Optional[Union[Json, ActionJSON]] = None,
+        json_result: Any = None,
         files: Optional[Union[Path, ActionFiles]] = None,
     ) -> ActionResult:
[…]

[…]
