### 2025-10-02 43e7b72
Simplify file return
