+++ src/madsci_node_module/madsci/node_module/rest_node_module.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from typing import Any, Callable, Optional, Type, get_origin, get_type_hints
 from zipfile import ZipFile
 
-from fastapi import Request, Response
+from fastapi import HTTPException, Request, Response
 from fastapi.applications import FastAPI
 from fastapi.background import BackgroundTasks
 from fastapi.datastructures import UploadFile
@@ -399,60 +399,40 @@ class RestNode(AbstractNode):
         action_response = super().get_action_result(action_id)
 
         if not action_response.files:
-            raise ValueError(f"Action {action_id} has no files")
-
-        if isinstance(action_response.files, Path):
-            # Single file - return as is (no ZIP needed)
-            return FileResponse(
-                path=action_response.files,
-                headers={"content-type": "application/octet-stream"},
+            if action_response.status != ActionStatus.SUCCEEDED:
+                raise HTTPException(
+                    status_code=400,
+                    detail=f"Action {action_id} did not succeed (status: {action_response.status}). Cannot download files from failed action.",
+                )
+            raise HTTPException(
+                status_code=404,
+                detail=f"Action {action_id} completed successfully but produced no file results",
             )
-        if isinstance(action_response.files, ActionFiles):
-            # Multiple files - create ZIP
-            files_dict = action_response.files.model_dump()
 
-            with tempfile.NamedTemporaryFile(
-                suffix=".zip", delete=False
-            ) as temp_zip_file:
-                with ZipFile(temp_zip_file.name, "w") as zip_file:
-                    for _, file_path in files_dict.items():
+        # Always create a ZIP, even for single files for consistency
+        with tempfile.NamedTemporaryFile(suffix=".zip", delete=False) as temp_zip_file:
+            with ZipFile(temp_zip_file.name, "w") as zip_file:
+                if isinstance(action_response.files, Path):
+                    # Single file - add to ZIP with appropriate name
+                    if action_response.files.exists():
+                        zip_file.write(action_response.files, "file")
+                elif isinstance(action_response.files, ActionFiles):
+                    # Multiple files - add all to ZIP with their labels as names
+                    files_dict = action_response.files.model_dump()
+                    for file_label, file_path in files_dict.items():
                         if Path(file_path).exists():
-                            zip_file.write(file_path, Path(file_path).name)
+                            # Use the file label as the filename in the ZIP
+                            file_extension = Path(file_path).suffix
+                            zip_filename = f"{file_label}{file_extension}"
+                            zip_file.write(file_path, zip_filename)
+                else:
+                    raise ValueError("Invalid file response")
 
-                return FileResponse(
-                    path=temp_zip_file.name,
-                    filename=f"{action_id}_files.zip",
-                    headers={"content-type": "application/zip"},
-                )
-        else:
-            raise ValueError("Invalid file response")
-
-    def get_action_file(
-        self, _action_name: str, action_id: str, file_key: str
-    ) -> FileResponse:
-        """Get a specific file from an action."""
-        action_response = super().get_action_result(action_id)
-
-        if not action_response.files:
-            raise ValueError(f"Action {action_id} has no files")
-
-        if isinstance(action_response.files, Path):
-            # Single file
-            if file_key == "file":
-                return FileResponse(
-                    path=action_response.files,
-                    headers={"content-type": "application/octet-stream"},
-                )
-        elif isinstance(action_response.files, ActionFiles):
-            # Multiple files
-            files_dict = action_response.files.model_dump()
-            if file_key in files_dict:
-                file_path = Path(files_dict[file_key])
-                return FileResponse(
-                    path=file_path, headers={"content-type": "application/octet-stream"}
-                )
-
[…]

[…]
