### 2025-10-02 6895ca3
Rigorously testing OpenAPI schema generation for results
