### 2025-10-02 f8ce3cd
Adding a bunch of testing for action argument schema
