+++ example_lab/example_modules/advanced_example_module.py
@@ -1,11 +1,13 @@
 """A module for testing advanced features of MADSci."""
 
 from pathlib import Path
+from typing import Annotated, Optional, Union
 
 from madsci.common.types.action_types import (
     ActionFiles,
 )
 from madsci.common.types.admin_command_types import AdminCommandResponse
+from madsci.common.types.location_types import LocationArgument
 from madsci.common.types.node_types import RestNodeConfig
 from madsci.node_module.helpers import action
 from madsci.node_module.rest_node_module import RestNode
@@ -39,6 +41,20 @@ class ExperimentMetadata(BaseModel):
     parameters: list[str] = Field(description="List of measured parameters")
     status: str = Field(description="Current experiment status")
     results_summary: dict[str, int] = Field(description="Summary of results count")
+
+
+class SampleProcessingRequest(BaseModel):
+    """Example complex input model for testing pydantic argument handling"""
+
+    sample_ids: list[str] = Field(description="List of sample identifiers to process")
+    processing_type: str = Field(description="Type of processing to perform")
+    parameters: dict[str, Union[str, int, float]] = Field(
+        description="Processing parameters"
+    )
+    priority: int = Field(description="Processing priority (1-10)", ge=1, le=10)
+    notify_on_completion: bool = Field(
+        default=False, description="Whether to send notifications"
+    )
 
 
 class AdvancedExampleConfig(RestNodeConfig):
@@ -182,6 +198,319 @@ class AdvancedExampleNode(RestNode):
         )
 
         return result, file_path
+
+    # ===== SIMPLE ARGUMENT TEST CASES =====
+
+    @action
+    def test_simple_string(self, message: str) -> str:
+        """Test action with a simple string argument"""
+        self.logger.log(f"Received string: {message}")
+        return f"Processed: {message}"
+
+    @action
+    def test_simple_int(self, number: int) -> int:
+        """Test action with a simple integer argument"""
+        self.logger.log(f"Received int: {number}")
+        return number * 2
+
+    @action
+    def test_simple_float(self, value: float) -> float:
+        """Test action with a simple float argument"""
+        self.logger.log(f"Received float: {value}")
+        return round(value * 3.14, 2)
+
+    @action
+    def test_simple_bool(self, flag: bool) -> bool:
+        """Test action with a simple boolean argument"""
+        self.logger.log(f"Received bool: {flag}")
+        return not flag
+
+    @action
+    def test_multiple_simple_args(
+        self, name: str, age: int, height: float, active: bool
+    ) -> dict:
+        """Test action with multiple simple arguments of different types"""
+        self.logger.log(
+            f"Received: name={name}, age={age}, height={height}, active={active}"
+        )
+        return {
+            "name": name.upper(),
+            "age_doubled": age * 2,
+            "height_cm": height * 100,
+            "status": "active" if active else "inactive",
+        }
+
+    # ===== OPTIONAL ARGUMENT TEST CASES =====
+
+    @action
+    def test_optional_string(self, message: str, prefix: Optional[str] = None) -> str:
+        """Test action with optional string argument"""
+        result = f"{prefix}: {message}" if prefix else message
+        self.logger.log(f"Result: {result}")
+        return result
+
+    @action
+    def test_optional_with_defaults(
+        self,
+        required_param: str,
+        optional_int: Optional[int] = None,
+        default_string: str = "default_value",
+        default_float: float = 1.0,
+        default_bool: bool = False,
+    ) -> dict:
+        """Test action with various optional parameters and defaults"""
+        self.logger.log(f"Processing with required_param={required_param}")
+        return {
+            "required": required_param,
+            "optional_int": optional_int
+            if optional_int is not None
+            else "not_provided",
[…]
