+++ src/madsci_node_module/tests/test_openapi_schema_validation.py
@@ -5,11 +5,12 @@ import tempfile
 import time
 from datetime import datetime
 from pathlib import Path
-from typing import Any, Dict, List, Optional
+from typing import Annotated, Any, Dict, List, Optional, Union
 
 import pytest
 from fastapi.testclient import TestClient
 from madsci.common.types.action_types import ActionFiles
+from madsci.common.types.location_types import LocationArgument
 from madsci.common.types.node_types import NodeDefinition
 from madsci.node_module.helpers import action
 from pydantic import BaseModel, Field
@@ -140,6 +141,43 @@ class TestFiles(ActionFiles):
 
     data_file: Path
     config_file: Path
+
+
+# ============================================================================
+# Pydantic Models for Argument Testing
+# ============================================================================
+
+
+class SampleProcessingRequest(BaseModel):
+    """Example complex input model for testing pydantic argument handling"""
+
+    sample_ids: list[str] = Field(description="List of sample identifiers to process")
+    processing_type: str = Field(description="Type of processing to perform")
+    parameters: dict[str, Union[str, int, float]] = Field(
+        description="Processing parameters"
+    )
+    priority: int = Field(description="Processing priority (1-10)", ge=1, le=10)
+    notify_on_completion: bool = Field(
+        default=False, description="Whether to send notifications"
+    )
+
+
+class AnalysisResult(BaseModel):
+    """Example custom pydantic model for analysis results"""
+
+    sample_id: str = Field(description="Unique identifier for the sample")
+    concentration: float = Field(description="Measured concentration in mg/mL")
+    ph_level: float = Field(description="pH level of the sample")
+    temperature: float = Field(description="Temperature in Celsius during measurement")
+    quality_score: int = Field(description="Quality score from 0-100", ge=0, le=100)
+    notes: str = Field(default="", description="Additional notes about the analysis")
+
+
+class TestFileOutput(ActionFiles):
+    """Test file output model for argument testing"""
+
+    log_file_1: Path
+    log_file_2: Path
 
 
 class OpenAPISchemaTestNode(TestNode):
@@ -290,6 +328,322 @@ class OpenAPISchemaTestNode(TestNode):
         return SimpleTestResult(value=file_size, message="file_processed")
 
 
+class OpenAPIArgumentTestNode(TestNode):
+    """Test node specifically for testing OpenAPI schema generation for action arguments."""
+
+    # ============================================================================
+    # Simple Argument Type Tests
+    # ============================================================================
+
+    @action
+    def test_simple_string_arg(self, message: str) -> str:
+        """Test action with a simple string argument"""
+        return f"Processed: {message}"
+
+    @action
+    def test_simple_int_arg(self, number: int) -> int:
+        """Test action with a simple integer argument"""
+        return number * 2
+
+    @action
+    def test_simple_float_arg(self, value: float) -> float:
+        """Test action with a simple float argument"""
+        return round(value * 3.14, 2)
+
+    @action
+    def test_simple_bool_arg(self, flag: bool) -> bool:
+        """Test action with a simple boolean argument"""
+        return not flag
+
+    @action
+    def test_multiple_simple_args(
+        self, name: str, age: int, height: float, active: bool
+    ) -> dict:
+        """Test action with multiple simple arguments of different types"""
+        return {
+            "name": name.upper(),
+            "age_doubled": age * 2,
+            "height_cm": height * 100,
+            "status": "active" if active else "inactive",
+        }
+
+    # ============================================================================
+    # Optional Argument Tests
+    # ============================================================================
+
+    @action
+    def test_optional_string_arg(
+        self, message: str, prefix: Optional[str] = None
+    ) -> str:
+        """Test action with optional string argument"""
+        return f"{prefix}: {message}" if prefix else message
+
+    @action
[…]

[…]
