### 2025-10-03 4e1e399
Handle args/kwargs
