+++ src/madsci_node_module/tests/test_openapi_schema_validation.py
@@ -1,6 +1,5 @@
 """Test-driven development approach to verify OpenAPI schema generation for json_result field."""
 
-import contextlib
 import tempfile
 import time
 from datetime import datetime
@@ -16,79 +15,7 @@ from madsci.node_module.helpers import action
 from pydantic import BaseModel, Field
 
 from madsci_node_module.tests.test_node import TestNode, TestNodeConfig
-
-
-def wait_for_node_ready(client: TestClient, timeout: float = 5.0) -> bool:
-    """Wait for node to be ready before executing actions.
-
-    Args:
-        client: FastAPI test client
-        timeout: Maximum time to wait in seconds
-
-    Returns:
-        True if node is ready, False if timeout
-    """
-    start_time = time.time()
-    while time.time() - start_time < timeout:
-        with contextlib.suppress(Exception):
-            response = client.get("/status")
-            if response.status_code == 200:
-                status_data = response.json()
-                if status_data.get("ready", False):
-                    return True
-        time.sleep(0.1)  # Check every 100ms
-    return False
-
-
-def execute_action_and_wait(
-    client: TestClient,
-    action_name: str,
-    parameters: Optional[dict] = None,
-    timeout: float = 10.0,
-) -> dict:
-    """Execute an action and wait for it to complete.
-
-    Args:
-        client: FastAPI test client
-        action_name: Name of the action to execute
-        parameters: Optional parameters for the action
-        timeout: Maximum time to wait for completion
-
-    Returns:
-        Final action result dict
-
-    Raises:
-        AssertionError: If action fails or times out
-    """
-    if parameters is None:
-        parameters = {}
-
-    # Create action
-    response = client.post(f"/action/{action_name}", json=parameters)
-    assert response.status_code == 200, f"Failed to create action: {response.text}"
-    action_id = response.json()["action_id"]
-
-    # Start action
-    response = client.post(f"/action/{action_name}/{action_id}/start")
-    assert response.status_code == 200, f"Failed to start action: {response.text}"
-
-    # Wait for completion
-    start_time = time.time()
-    while time.time() - start_time < timeout:
-        response = client.get(f"/action/{action_id}/result")
-        if response.status_code == 200:
-            result = response.json()
-            status = result.get("status")
-            if status in ["succeeded", "failed", "error"]:
-                return result
-        time.sleep(0.1)  # Check every 100ms
-
-    # If we get here, the action timed out
-    response = client.get(f"/action/{action_id}/result")
-    result = response.json() if response.status_code == 200 else {"status": "timeout"}
-    raise AssertionError(
-        f"Action {action_name} timed out after {timeout}s. Last status: {result.get('status', 'unknown')}"
-    )
+from madsci_node_module.tests.test_rest_utils import execute_action_and_wait
 
 
 # Test result models for various type scenarios
@@ -326,6 +253,76 @@ class OpenAPISchemaTestNode(TestNode):
         """Action that takes a file and returns a model."""
         file_size = input_file.stat().st_size if input_file.exists() else 0
         return SimpleTestResult(value=file_size, message="file_processed")
+
+    # ============================================================================
+    # Variable Arguments Tests (*args, **kwargs)
+    # ============================================================================
+
+    @action
+    def test_var_args_only(self, required_param: str, *args) -> dict:
+        """Action that accepts additional positional arguments."""
+        return {
+            "required_param": required_param,
+            "var_args": list(args),
+            "var_args_count": len(args),
+        }
+
+    @action
+    def test_var_kwargs_only(self, required_param: str, **kwargs) -> dict:
+        """Action that accepts additional keyword arguments."""
+        return {
+            "required_param": required_param,
[…]
