+++ src/madsci_node_module/tests/test_rest_node_client.py
@@ -823,3 +823,206 @@ def test_send_action_with_list_files(
 
         # Verify all three calls were made
         assert mock_post.call_count == 3
+
+
+@patch("requests.post")
+def test_send_action_with_var_args(
+    mock_post: MagicMock, rest_node_client: RestNodeClient
+) -> None:
+    """Test sending an action request with variable arguments (*args)."""
+    action_id = new_ulid_str()
+
+    # Mock create action response
+    create_response = MagicMock()
+    create_response.ok = True
+    create_response.json.return_value = {"action_id": action_id}
+
+    # Mock start action response
+    start_response = MagicMock()
+    start_response.ok = True
+    start_response.json.return_value = ActionSucceeded(
+        action_id=action_id, json_result={"var_args": ["arg1", "arg2", 123]}
+    ).model_dump(mode="json")
+
+    mock_post.side_effect = [create_response, start_response]
+
+    # Create action request with var_args
+    action_request = ActionRequest(
+        action_name="test_var_args",
+        args={"required_param": "value"},
+        var_args=["arg1", "arg2", 123],
+    )
+
+    result = rest_node_client.send_action(action_request, await_result=False)
+
+    # Verify the result
+    assert isinstance(result, ActionResult)
+    assert result.action_id == action_id
+    assert result.status == ActionStatus.SUCCEEDED
+    assert result.json_result["var_args"] == ["arg1", "arg2", 123]
+
+    # Verify the calls
+    assert mock_post.call_count == 2
+
+    # Check that var_args was included in the request payload
+    create_call_kwargs = mock_post.call_args_list[0][1]
+    request_data = create_call_kwargs["json"]
+    assert request_data["var_args"] == ["arg1", "arg2", 123]
+
+
+@patch("requests.post")
+def test_send_action_with_var_kwargs(
+    mock_post: MagicMock, rest_node_client: RestNodeClient
+) -> None:
+    """Test sending an action request with variable keyword arguments (**kwargs)."""
+    action_id = new_ulid_str()
+
+    # Mock create action response
+    create_response = MagicMock()
+    create_response.ok = True
+    create_response.json.return_value = {"action_id": action_id}
+
+    # Mock start action response
+    start_response = MagicMock()
+    start_response.ok = True
+    start_response.json.return_value = ActionSucceeded(
+        action_id=action_id,
+        json_result={"var_kwargs": {"extra1": "value1", "extra2": 42}},
+    ).model_dump(mode="json")
+
+    mock_post.side_effect = [create_response, start_response]
+
+    # Create action request with var_kwargs
+    action_request = ActionRequest(
+        action_name="test_var_kwargs",
+        args={"required_param": "value"},
+        var_kwargs={"extra1": "value1", "extra2": 42},
+    )
+
+    result = rest_node_client.send_action(action_request, await_result=False)
+
+    # Verify the result
+    assert isinstance(result, ActionResult)
+    assert result.action_id == action_id
+    assert result.status == ActionStatus.SUCCEEDED
+    assert result.json_result["var_kwargs"] == {"extra1": "value1", "extra2": 42}
+
+    # Verify the calls
+    assert mock_post.call_count == 2
+
+    # Check that var_kwargs was included in the request payload
+    create_call_kwargs = mock_post.call_args_list[0][1]
+    request_data = create_call_kwargs["json"]
+    assert request_data["var_kwargs"] == {"extra1": "value1", "extra2": 42}
+
+
+@patch("requests.post")
+def test_send_action_with_var_args_and_kwargs(
+    mock_post: MagicMock, rest_node_client: RestNodeClient
+) -> None:
+    """Test sending an action request with both *args and **kwargs."""
+    action_id = new_ulid_str()
+
[…]
