### 2025-10-06 3624508
Fix: bad pipe type operator
