### 2025-10-06 b3b1047
Fix: more bad pipe type operators
