### 2025-10-09 4c7e120
Update src/madsci_workcell_manager/madsci/workcell_manager/workcell_engine.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
