### 2025-10-09 393a4fe
Fix datapoint result validation
