+++ example_lab/node_definitions/advanced_example_node.node.info.yaml
@@ -35,6 +35,10 @@ actions:
     results: {}
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_json:
     name: return_json
     description: Return a JSON object
@@ -57,6 +61,10 @@ actions:
           type: object
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_dict:
     name: return_dict
     description: Return a JSON object
@@ -80,6 +88,10 @@ actions:
           type: object
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_file:
     name: return_file
     description: Return a file
@@ -93,6 +105,10 @@ actions:
         result_type: file
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_file_and_json:
     name: return_file_and_json
     description: Return data and a file
@@ -119,6 +135,10 @@ actions:
         result_type: file
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_labeled_json_values:
     name: return_labeled_json_values
     description: Return labeled JSON values
@@ -142,6 +162,10 @@ actions:
           type: object
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_labeled_file_values:
     name: return_labeled_file_values
     description: Return labeled file values
@@ -159,6 +183,10 @@ actions:
         result_type: file
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_labeled_file_and_json_values:
     name: return_labeled_file_and_json_values
     description: Return labeled file and json values
@@ -190,6 +218,10 @@ actions:
         result_type: file
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_analysis_result:
     name: return_analysis_result
     description: Return a custom pydantic model with analysis results
@@ -247,6 +279,10 @@ actions:
           type: object
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_experiment_metadata:
     name: return_experiment_metadata
     description: Return complex experiment metadata as a custom pydantic model
@@ -308,6 +344,10 @@ actions:
           type: object
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   return_mixed_custom_and_files:
     name: return_mixed_custom_and_files
     description: Return a custom pydantic model and a file
@@ -369,6 +409,10 @@ actions:
         result_type: file
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   test_simple_string:
     name: test_simple_string
     description: Test action with a simple string argument
@@ -397,6 +441,10 @@ actions:
           type: object
     blocking: false
     asynchronous: true
[…]
