+++ example_lab/node_definitions/platereader_1.info.yaml
@@ -47,6 +47,10 @@ actions:
           type: object
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   create_plate_file:
     name: create_plate_file
     description: Run a command on the plate reader.
@@ -60,6 +64,10 @@ actions:
         result_type: file
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
   create_plate_files:
     name: create_plate_files
     description: Run a command on the plate reader.
@@ -77,6 +85,10 @@ actions:
         result_type: file
     blocking: false
     asynchronous: true
+    accepts_var_args: false
+    accepts_var_kwargs: false
+    var_args_schema: null
+    var_kwargs_schema: null
 config:
   node_definition: node_definitions/platereader_1.node.yaml
   node_info_path: null
