+++ src/madsci_data_manager/tests/test_data_server.py
@@ -179,7 +179,7 @@ def is_docker_available():  # noqa
             ["docker", "version"],  # noqa
             capture_output=True,
             text=True,
-            timeout=10,
+            timeout=60,
             check=False,
         )
         return result.returncode == 0

[…]
