### 2025-10-13 d995dcc
Added init template with create resource updates
