### 2025-10-13 d6819f9
Utilize resource templates
