### 2025-10-13 8150436
Utilize resource templates
