### 2025-10-13 2874d92
Utilize resource templates
