+++ .github/workflows/integration_test.yml
@@ -27,4 +27,4 @@ jobs:
       - name: Run Docker Compose
         uses: hoverkraft-tech/compose-action@v2.0.1
         with:
-        compose-file: "./compose.yaml"
+          compose-file: "./compose.yaml"
