+++ src/madsci_node_module/tests/test_openapi_schemas.py
@@ -1,0 +1,484 @@
+"""Streamlined OpenAPI schema validation tests.
+
+Consolidates and parametrizes tests from test_openapi_schema_validation.py
+to reduce redundancy while maintaining comprehensive coverage.
+"""
+
+import tempfile
+import time
+from datetime import datetime
+from pathlib import Path
+from typing import Any, Dict, List
+
+import pytest
+from fastapi.testclient import TestClient
+from madsci.common.types.action_types import ActionFiles
+from madsci.common.types.node_types import NodeDefinition
+from madsci.node_module.helpers import action
+from pydantic import BaseModel, Field
+
+from madsci_node_module.tests.test_node import TestNode, TestNodeConfig
+
+
+# Test result models for schema validation
+class SimpleTestResult(BaseModel):
+    """Simple test result model."""
+
+    value: int = Field(description="Simple integer value")
+    message: str = Field(description="Status message")
+
+
+class ComplexTestResult(BaseModel):
+    """Complex test result model with nested data."""
+
+    id: str = Field(description="Unique identifier")
+    measurements: List[float] = Field(description="List of measurement values")
+    metadata: Dict[str, Any] = Field(description="Additional metadata")
+    timestamp: datetime = Field(description="Processing timestamp")
+
+
+class TestFiles(ActionFiles):
+    """Test file result model."""
+
+    data_file: Path
+    config_file: Path
+
+
+class OpenAPISchemaTestNode(TestNode):
+    """Test node for OpenAPI schema validation."""
+
+    @action
+    def return_int(self) -> int:
+        """Action that returns a simple integer."""
+        return 42
+
+    @action
+    def return_string(self) -> str:
+        """Action that returns a simple string."""
+        return "test_string"
+
+    @action
+    def return_float(self) -> float:
+        """Action that returns a simple float."""
+        return 3.14159
+
+    @action
+    def return_bool(self) -> bool:
+        """Action that returns a simple boolean."""
+        return True
+
+    @action
+    def return_dict(self) -> dict:
+        """Action that returns a dictionary."""
+        return {"key": "value", "number": 123, "flag": True}
+
+    @action
+    def return_list(self) -> list:
+        """Action that returns a list."""
+        return [1, 2, 3, "four", 5.0]
+
+    @action
+    def return_simple_model(self) -> SimpleTestResult:
+        """Action that returns a simple Pydantic model."""
+        return SimpleTestResult(value=42, message="success")
+
+    @action
+    def return_complex_model(self) -> ComplexTestResult:
+        """Action that returns a complex Pydantic model."""
+        return ComplexTestResult(
+            id="test_001",
+            measurements=[1.0, 2.5, 3.7],
+            metadata={"instrument": "test", "operator": "user"},
+            timestamp=datetime.now(),
+        )
+
+    @action
+    def return_single_file(self) -> Path:
+        """Action that returns a single file."""
+        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".txt") as f:
+            f.write("test file content")
+            return Path(f.name)
+
+    @action
+    def return_multiple_files(self) -> TestFiles:
+        """Action that returns multiple files."""
+        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".dat") as f1:
+            f1.write("data content")
+            data_path = Path(f1.name)
+
+        with tempfile.NamedTemporaryFile(mode="w", delete=False, suffix=".cfg") as f2:
+            f2.write("config content")
+            config_path = Path(f2.name)
+
+        return TestFiles(data_file=data_path, config_file=config_path)
+
+    @action
+    def mixed_return(self) -> tuple[SimpleTestResult, Path]:
[…]
