+++ src/madsci_node_module/tests/test_node_infrastructure.py
@@ -1,0 +1,358 @@
+"""Test node infrastructure including lifecycle, configuration, and basic actions.
+
+Consolidates tests from:
+- test_node.py (TestNode implementation)
+- test_rest_utils.py (utilities)
+- Basic node lifecycle tests from test_rest_node_module.py
+"""
+
+import time
+
+from fastapi.testclient import TestClient
+from madsci.common.types.action_types import ActionResult, ActionStatus
+from madsci.common.types.admin_command_types import AdminCommandResponse
+from madsci.common.types.node_types import NodeStatus
+from ulid import ULID
+
+from madsci_node_module.tests.test_node import TestNode
+from madsci_node_module.tests.test_rest_utils import (
+    execute_action_with_validation,
+    parametrize_admin_commands,
+    validate_admin_command_response,
+)
+
+
+class TestNodeLifecycle:
+    """Test node startup, shutdown, state management."""
+
+    def test_startup_and_shutdown_handlers(self, basic_test_node: TestNode) -> None:
+        """Test the startup_handler and shutdown_handler methods."""
+        assert not hasattr(basic_test_node, "startup_has_run")
+        assert not hasattr(basic_test_node, "shutdown_has_run")
+        assert basic_test_node.test_interface is None
+
+        basic_test_node.start_node(testing=True)
+
+        with TestClient(basic_test_node.rest_api) as client:
+            time.sleep(0.5)
+            assert basic_test_node.startup_has_run
+            assert not hasattr(basic_test_node, "shutdown_has_run")
+            assert basic_test_node.test_interface is not None
+
+            response = client.get("/status")
+            assert response.status_code == 200
+
+        time.sleep(0.5)
+
+        assert basic_test_node.startup_has_run
+        assert basic_test_node.shutdown_has_run
+        assert basic_test_node.test_interface is None
+
+    def test_node_status_endpoint(self, test_client: TestClient) -> None:
+        """Test the /status endpoint."""
+        with test_client as client:
+            time.sleep(0.1)
+            response = client.get("/status")
+            assert response.status_code == 200
+            status = NodeStatus.model_validate(response.json())
+            assert isinstance(status.ready, bool)
+
+    def test_node_state_endpoint(self, test_client: TestClient) -> None:
+        """Test the /state endpoint."""
+        with test_client as client:
+            time.sleep(0.1)
+            response = client.get("/state")
+            assert response.status_code == 200
+            state_data = response.json()
+            assert isinstance(state_data, dict)
+
+    def test_node_info_endpoint(self, test_client: TestClient) -> None:
+        """Test the /info endpoint."""
+        with test_client as client:
+            time.sleep(0.1)
+            response = client.get("/info")
+            assert response.status_code == 200
+            info_data = response.json()
+            assert isinstance(info_data, dict)
+            # Should contain basic node information
+            assert isinstance(info_data, dict) and len(info_data) > 0
+
+    def test_shutdown_command(self, basic_test_node: TestNode) -> None:
+        """Test the shutdown admin command."""
+        basic_test_node.start_node(testing=True)
+
+        with TestClient(basic_test_node.rest_api) as client:
+            time.sleep(0.5)
+            response = client.post("/admin/shutdown")
+            assert response.status_code == 200
+            validated_response = AdminCommandResponse.model_validate(response.json())
+            assert validated_response.success is True
+            assert not validated_response.errors
+            assert basic_test_node.shutdown_has_run
+
+
+class TestNodeConfiguration:
+    """Test node configuration and settings."""
+
+    def test_node_factory_basic_config(self, test_node_factory) -> None:
+        """Test basic node configuration through factory."""
+        node = test_node_factory(
+            node_name="Config Test Node",
[…]

[…]
