+++ pyproject.toml
@@ -14,6 +14,7 @@ dev = [
     "pre-commit>=4.3.0",
     "pydantic-settings-export>=1.0.2",
     "ruff>=0.13.0",
+    "nbconvert"
 ]
 
 [tool.pydantic_settings_export]
