+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "madsci"
 version = "0.4.7"
-dependencies = ["pre-commit>=4.3.0"]
+dependencies = ["pre-commit>=4.3.0", "nbconvert>=7.16.6"]
 readme="README.md"
 requires-python = ">=3.9.1"
 
