### 2025-10-15 38e4c7b
Adding in fake color picker nodes for WF testing purposes.
