+++ src/madsci_common/pyproject.toml
@@ -29,7 +29,8 @@ dependencies = [
     "pydantic-settings>=2.9.1",
     "pydantic-settings-export>=1.0.2",
     "regex>=2025.7.34",
-    "classy-fastapi>=0.6.1"
+    "classy-fastapi>=0.6.1",
+    "nbconvert>=7.0.0"
 ]
 
 
