+++ .github/workflows/integration_test.yml
@@ -1,40 +1,0 @@
-name: Docker Image
-
-on:
-  push: # * Run for every push
-    branches: [ "*" ]
-    tags: [ '*' ]
-  schedule: # Run on Tuesday's at 12:00
-    - cron: '0 12 * * 2'
-  workflow_dispatch: # Run manually
-
-env:
-  # Use docker.io for Docker Hub if empty
-  REGISTRY: ghcr.io
-  # github.repository as <account>/<repo>
-  IMAGE_NAME: ${{ github.repository }}
-  DASHBOARD_IMAGE_NAME: ${{ github.repository }}_dashboard
-
-
-jobs:
-  docker-compose-test:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Copy Example Env
-        run: cp .env.example .env
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v3
-
-      - name: Build Docker Compose services
-        run: docker compose build
-
-      - name: Start services to verify build
-        run: docker compose up -d
-
-      - name: Run a command in the lab manager container
-        run: docker compose exec -T lab_manager jupyter nbconvert --to script ./MADSci/example_lab/notebooks/experiment_notebook.ipynb
