+++ .github/workflows/docker.yml
@@ -77,7 +77,7 @@ jobs:
         run: docker compose up -d
 
       - name: Run a command in the lab manager container
-        run: docker compose exec -T lab_manager source ~/MADSci/.venv/bin/activate & jupyter nbconvert --to script ./MADSci/example_lab/notebooks/experiment_notebook.ipynb
+        run: docker compose exec lab_manager jupyter nbconvert --to script ./notebooks/experiment_notebook.ipynb
 
     outputs:
       image_tags: ${{ steps.meta.outputs.tags }}
