+++ .github/workflows/docker.yml
@@ -73,6 +73,8 @@ jobs:
           cache-to: type=gha,mode=max
       - name: example.env to .env
         run: cp .env.example .env
+      - name: docker pull
+        run: docker compose pull
       - name: Start services to verify build
         run: docker compose up -d
 
