+++ .github/workflows/docker.yml
@@ -77,9 +77,10 @@ jobs:
         run: docker compose pull
       - name: Start services to verify build
         run: docker compose up -d
-
+      - name: sleep 10
+        run: sleep 10
       - name: Run a command in the lab manager container
-        run: docker compose exec lab_manager jupyter nbconvert --to script ./notebooks/experiment_notebook.ipynb
+        run: docker compose run --rm --no-deps lab_manager jupyter nbconvert --to script ./notebooks/experiment_notebook.ipynb
 
     outputs:
       image_tags: ${{ steps.meta.outputs.tags }}
