### 2025-10-15 ec6c62d
Slightly improved error handling for workflows
