+++ .github/workflows/docker.yml
@@ -73,14 +73,12 @@ jobs:
           cache-to: type=gha,mode=max
       - name: example.env to .env
         run: cp .env.example .env
-      - name: docker pull
-        run: docker compose pull
       - name: Start services to verify build
         run: docker compose up -d
       - name: sleep 10
         run: sleep 10
       - name: Run a command in the lab manager container
-        run: docker compose run --rm --no-deps lab_manager python -m nbconvert --to script ./notebooks/experiment_notebook.ipynb
+        run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to script ./notebooks/experiment_notebook.ipynb
 
     outputs:
       image_tags: ${{ steps.meta.outputs.tags }}
