+++ src/madsci_node_module/tests/test_openapi_schema_validation.py
@@ -1384,9 +1384,13 @@ class TestActionArgumentSchemaGeneration:
             "schema"
         ]
 
+        # With RestActionRequest structure, parameters are in the args field
         string_props = self._get_schema_props(string_schema, components)
-        assert "message" in string_props
-        message_prop = string_props["message"]
+        assert "args" in string_props
+        args_schema = string_props["args"]
+        args_props = self._get_schema_props(args_schema, components)
+        assert "message" in args_props
+        message_prop = args_props["message"]
         assert message_prop["type"] == "string"
 
         # Check int argument schema
@@ -1396,8 +1400,11 @@ class TestActionArgumentSchemaGeneration:
         int_schema = int_post["requestBody"]["content"]["application/json"]["schema"]
 
         int_props = self._get_schema_props(int_schema, components)
-        assert "number" in int_props
-        number_prop = int_props["number"]
+        assert "args" in int_props
+        int_args_schema = int_props["args"]
+        int_args_props = self._get_schema_props(int_args_schema, components)
+        assert "number" in int_args_props
+        number_prop = int_args_props["number"]
         assert number_prop["type"] == "integer"
 
         # Check float argument schema
@@ -1409,8 +1416,11 @@ class TestActionArgumentSchemaGeneration:
         ]
 
         float_props = self._get_schema_props(float_schema, components)
-        assert "value" in float_props
-        value_prop = float_props["value"]
+        assert "args" in float_props
+        float_args_schema = float_props["args"]
+        float_args_props = self._get_schema_props(float_args_schema, components)
+        assert "value" in float_args_props
+        value_prop = float_args_props["value"]
         assert value_prop["type"] == "number"
 
         # Check bool argument schema
@@ -1420,8 +1430,11 @@ class TestActionArgumentSchemaGeneration:
         bool_schema = bool_post["requestBody"]["content"]["application/json"]["schema"]
 
         bool_props = self._get_schema_props(bool_schema, components)
-        assert "flag" in bool_props
-        flag_prop = bool_props["flag"]
+        assert "args" in bool_props
+        bool_args_schema = bool_props["args"]
+        bool_args_props = self._get_schema_props(bool_args_schema, components)
+        assert "flag" in bool_args_props
+        flag_prop = bool_args_props["flag"]
         assert flag_prop["type"] == "boolean"
 
     def test_multiple_arguments_schema_generation(self, argument_test_client):
@@ -1439,24 +1452,29 @@ class TestActionArgumentSchemaGeneration:
         ]["schema"]
         props = self._get_schema_props(multi_schema, components)
 
+        # With RestActionRequest structure, parameters are in the args field
+        assert "args" in props
+        args_schema = props["args"]
+        args_props = self._get_schema_props(args_schema, components)
+
         # Should have all four arguments
-        assert "name" in props
-        assert "age" in props
-        assert "height" in props
-        assert "active" in props
+        assert "name" in args_props
+        assert "age" in args_props
+        assert "height" in args_props
+        assert "active" in args_props
 
         # Check types
-        assert props["name"]["type"] == "string"
-        assert props["age"]["type"] == "integer"
-        assert props["height"]["type"] == "number"
-        assert props["active"]["type"] == "boolean"
+        assert args_props["name"]["type"] == "string"
+        assert args_props["age"]["type"] == "integer"
+        assert args_props["height"]["type"] == "number"
+        assert args_props["active"]["type"] == "boolean"
 
-        # Should mark all as required
-        required = self._get_schema_required(multi_schema, components)
-        assert "name" in required
-        assert "age" in required
-        assert "height" in required
-        assert "active" in required
+        # Should mark all as required - check in the args schema
[…]
