+++ .github/workflows/docker.yml
@@ -71,14 +71,6 @@ jobs:
           platforms: linux/amd64,linux/arm64
           cache-from: type=gha
           cache-to: type=gha,mode=max
-      - name: example.env to .env
-        run: cp .env.example .env
-      - name: Start services to verify build
-        run: docker compose up -d
-      - name: sleep 10
-        run: sleep 10
-      - name: Run a command in the lab manager container
-        run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to script ./notebooks/experiment_notebook.ipynb
 
     outputs:
       image_tags: ${{ steps.meta.outputs.tags }}
