+++ example_lab/example_modules/integration_test.yml
@@ -25,6 +25,8 @@ jobs:
         uses: actions/checkout@v4
       - name: example.env to .env
         run: cp .env.example .env
+      - name: docker build
+        run: docker compose build
       - name: Start services to verify build
         run: docker compose up -d
       - name: sleep 10
