+++ .github/workflows/integration_test.yml
@@ -29,9 +29,10 @@ jobs:
         run: docker compose build
       - name: Start services to verify build
         run: docker compose up -d
-      - name: sleep 10
-        run: sleep 10
       - name: fix permissions
         run: sudo chown -R $USER:$USER .madsci
+      - name: fix other permissions
+        run: sudo chown -R $USER:$USER example_lab
+
       - name: Run a command in the lab manager container
         run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace ./notebooks/experiment_notebook.ipynb
