+++ src/madsci_node_module/tests/test_openapi_schemas.py
@@ -1,23 +1,21 @@
-"""Streamlined OpenAPI schema validation tests.
-
-Consolidates and parametrizes tests from test_openapi_schema_validation.py
-to reduce redundancy while maintaining comprehensive coverage.
-"""
+"""Comprehensive OpenAPI schema validation tests."""
 
 import tempfile
 import time
 from datetime import datetime
 from pathlib import Path
-from typing import Any, Dict, List
+from typing import Annotated, Any, Dict, List, Optional, Union
 
 import pytest
 from fastapi.testclient import TestClient
 from madsci.common.types.action_types import ActionFiles
+from madsci.common.types.location_types import LocationArgument
 from madsci.common.types.node_types import NodeDefinition
 from madsci.node_module.helpers import action
 from pydantic import BaseModel, Field
 
 from madsci_node_module.tests.test_node import TestNode, TestNodeConfig
+from madsci_node_module.tests.test_rest_utils import execute_action_and_wait
 
 
 # Test result models for schema validation
@@ -35,6 +33,66 @@ class ComplexTestResult(BaseModel):
     measurements: List[float] = Field(description="List of measurement values")
     metadata: Dict[str, Any] = Field(description="Additional metadata")
     timestamp: datetime = Field(description="Processing timestamp")
+
+
+class NestedTestResult(BaseModel):
+    """Test result with nested models."""
+
+    primary: SimpleTestResult = Field(description="Primary result data")
+    secondary: Optional[SimpleTestResult] = Field(
+        default=None, description="Optional secondary data"
+    )
+    status: str = Field(description="Overall status")
+
+
+class OptionalFieldsResult(BaseModel):
+    """Test result with various optional field types."""
+
+    required_field: str = Field(description="Always present field")
+    optional_string: Optional[str] = Field(
+        default=None, description="Optional string field"
+    )
+    optional_int: Optional[int] = Field(
+        default=None, description="Optional integer field"
+    )
+    optional_list: Optional[List[str]] = Field(
+        default=None, description="Optional list field"
+    )
+    optional_dict: Optional[Dict[str, Any]] = Field(
+        default=None, description="Optional dict field"
+    )
+
+
+class SampleProcessingRequest(BaseModel):
+    """Example complex input model for testing pydantic argument handling"""
+
+    sample_ids: list[str] = Field(description="List of sample identifiers to process")
+    processing_type: str = Field(description="Type of processing to perform")
+    parameters: dict[str, Union[str, int, float]] = Field(
+        description="Processing parameters"
+    )
+    priority: int = Field(description="Processing priority (1-10)", ge=1, le=10)
+    notify_on_completion: bool = Field(
+        default=False, description="Whether to send notifications"
+    )
+
+
+class AnalysisResult(BaseModel):
+    """Example custom pydantic model for analysis results"""
+
+    sample_id: str = Field(description="Unique identifier for the sample")
+    concentration: float = Field(description="Measured concentration in mg/mL")
+    ph_level: float = Field(description="pH level of the sample")
+    temperature: float = Field(description="Temperature in Celsius during measurement")
+    quality_score: int = Field(description="Quality score from 0-100", ge=0, le=100)
+    notes: str = Field(default="", description="Additional notes about the analysis")
+
+
+class TestFileOutput(ActionFiles):
+    """Test file output model for argument testing"""
+
+    log_file_1: Path
+    log_file_2: Path
 
 
 class TestFiles(ActionFiles):
@@ -123,6 +181,26 @@ class OpenAPISchemaTestNode(TestNode):
         return TestFiles(data_file=data_path, config_file=config_path)
 
     @action
+    def return_nested_model(self) -> NestedTestResult:
[…]
