### 2025-10-16 f25e114
Removed old resources logic
