### 2025-10-16 1543156
Removed old resources logic
