### 2025-10-16 e18c799
Removed old resources logic
