### 2025-10-16 b98a2d8
Check template version for updates
