+++ example_lab/example_modules/platereader.py
@@ -26,7 +26,6 @@ class PlateReaderInterface:
 
     status_code: int = 0
     device_number: int = 0
-    config: PlateReaderConfig = PlateReaderConfig()
     config_model = PlateReaderConfig
 
     def __init__(
