### 2025-10-16 f6e9f59
Add get_datapoint and get_datapoint_id helper methods to Workflow model
