+++ example_lab/example_modules/robotarm.py
@@ -4,7 +4,9 @@ import time
 from typing import Annotated, Any, Optional
 
 from madsci.client.event_client import EventClient
+from madsci.common.types.action_types import ActionFailed
 from madsci.common.types.admin_command_types import AdminCommandResponse
+from madsci.common.types.base_types import Error
 from madsci.common.types.location_types import LocationArgument
 from madsci.common.types.node_types import RestNodeConfig
 from madsci.common.types.resource_types.definitions import SlotResourceDefinition
@@ -86,10 +88,17 @@ class RobotArmNode(RestNode):
             speed = self.config.speed
         speed = max(1.0, min(100.0, speed))  # Clamp speed to 1-100 mm/s
         if self.resource_client:
+            target_resource = self.resource_client.get_resource(target.resource_id)
+            if len(target_resource.children) + 1 > target_resource.capacity:
+                return ActionFailed(
+                    errors=[
+                        Error(message=f"Target location {target.resource_id} is full")
+                    ]
+                )
             try:
                 popped_plate, _ = self.resource_client.pop(resource=source.resource_id)
             except Exception as e:
-                raise ValueError("No plate in source!") from e
+                return ActionFailed(errors=[Error(message=str(e))])
             self.resource_client.push(
                 resource=self.gripper.resource_id, child=popped_plate
             )
@@ -100,6 +109,7 @@ class RobotArmNode(RestNode):
                 resource=self.gripper.resource_id
             )
             self.resource_client.push(resource=target.resource_id, child=popped_plate)
+        return None
 
     def get_location(self) -> AdminCommandResponse:
         """Get location for the robot arm"""
