+++ .github/workflows/integration_test.yml
@@ -32,4 +32,4 @@ jobs:
       - name: fix  permissions
         run: sudo chmod -R 777 .
       - name: Run a command in the lab manager container
-        run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
+        run: docker compose run --rm --no-deps workcell_manager sudo chmod -R 777 . & python -m nbconvert --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
