### 2025-10-16 488cf3a
Add unit tests for datapoint helper methods
