+++ src/madsci_workcell_manager/tests/test_workcell_engine.py
@@ -34,6 +34,7 @@ from madsci.common.types.workflow_types import (
     WorkflowParameters,
     WorkflowStatus,
 )
+from madsci.common.utils import new_ulid_str
 from madsci.workcell_manager.state_handler import WorkcellStateHandler
 from madsci.workcell_manager.workcell_engine import Engine
 from pytest_mock_resources import RedisConfig, create_redis_fixture
@@ -897,6 +898,301 @@ def test_feed_data_forward_no_matching_parameters(engine: Engine) -> None:
     # Should not modify existing values
     assert updated_wf.parameter_values == {"existing": "value"}
     assert updated_wf.file_input_ids == {"existing_file": "file_id"}
+
+
+# Workflow get_datapoint_id and get_datapoint method tests
+def test_get_datapoint_id_single_datapoint_no_label() -> None:
+    """Test get_datapoint_id with single datapoint and no label specified."""
+    datapoint_id = new_ulid_str()
+    step = Step(
+        name="Test Step",
+        key="step1",
+        action="test_action",
+        node="node1",
+        result=ActionResult(
+            status=ActionStatus.SUCCEEDED,
+            datapoints={"result": datapoint_id},
+        ),
+    )
+
+    workflow = Workflow(
+        name="Test Workflow",
+        steps=[step],
+    )
+
+    # Should return the single datapoint ID when no label specified
+    result_id = workflow.get_datapoint_id(step_key="step1")
+    assert result_id == datapoint_id
+
+
+def test_get_datapoint_id_single_datapoint_with_label() -> None:
+    """Test get_datapoint_id with single datapoint and specific label."""
+    datapoint_id = new_ulid_str()
+    step = Step(
+        name="Test Step",
+        key="step1",
+        action="test_action",
+        node="node1",
+        result=ActionResult(
+            status=ActionStatus.SUCCEEDED,
+            datapoints={"output_file": datapoint_id},
+        ),
+    )
+
+    workflow = Workflow(
+        name="Test Workflow",
+        steps=[step],
+    )
+
+    # Should return the datapoint ID for the specific label
+    result_id = workflow.get_datapoint_id(step_key="step1", label="output_file")
+    assert result_id == datapoint_id
+
+
+def test_get_datapoint_id_multiple_datapoints_with_label() -> None:
+    """Test get_datapoint_id with multiple datapoints and specific label."""
+    datapoint_id_1 = new_ulid_str()
+    datapoint_id_2 = new_ulid_str()
+    step = Step(
+        name="Test Step",
+        key="step1",
+        action="test_action",
+        node="node1",
+        result=ActionResult(
+            status=ActionStatus.SUCCEEDED,
+            datapoints={
+                "result_data": datapoint_id_1,
+                "log_file": datapoint_id_2,
+            },
+        ),
+    )
+
+    workflow = Workflow(
+        name="Test Workflow",
+        steps=[step],
+    )
+
+    # Should return the correct datapoint ID for the specified label
+    result_id = workflow.get_datapoint_id(step_key="step1", label="log_file")
+    assert result_id == datapoint_id_2
+
+
+def test_get_datapoint_id_multiple_datapoints_no_label_raises_error() -> None:
+    """Test get_datapoint_id with multiple datapoints but no label raises ValueError."""
+    step = Step(
+        name="Test Step",
+        key="step1",
+        action="test_action",
+        node="node1",
+        result=ActionResult(
+            status=ActionStatus.SUCCEEDED,
+            datapoints={
+                "result_data": new_ulid_str(),
+                "log_file": new_ulid_str(),
+            },
+        ),
+    )
+
+    workflow = Workflow(
+        name="Test Workflow",
+        steps=[step],
+    )
+
[…]
