### 2025-10-16 befab4c
Add: check_node_capability()
