### 2025-10-16 584c123
Fix: typo in capabilites check
