### 2025-10-16 ccfb4ae
fix permissions again
