### 2025-10-16 6242f3f
fix permissions again
