+++ .github/workflows/integration_test.yml
@@ -31,5 +31,7 @@ jobs:
         run: docker compose up -d
       - name: fix  permissions
         run: sudo chmod -R 777 .
+      - name: fix other permissions
+        run: sudo chmod -R 777 .madsci
       - name: Run a command in the lab manager container
-        run: docker compose run --rm --user "$(id -u):$(id -g)" --no-deps workcell_manager python -m nbconvert --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
+        run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
