### 2025-10-16 eb6c7b7
fix permissions again
