### 2025-10-16 eca2c9c
fix permissions again
