### 2025-10-16 6189a7e
fix permissions again
