+++ .github/workflows/integration_test.yml
@@ -27,13 +27,13 @@ jobs:
         run: cp .env.example .env
       - name: check ownership
         run: ls -ld .madsci
-      - name: docker build
-        run: docker compose build
-      - name: Start services to verify build
-        run: docker compose up -d
-      - name: fix  permissions
-        run: sudo chmod -R 777 .
-      - name: fix other permissions
-        run: sudo chown $USER_ID:$GROUP_ID .madsci && sudo chmod -R 777 .madsci
-      - name: Run a command in the lab manager container
-        run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
+      # - name: docker build
+      #   run: docker compose build
+      # - name: Start services to verify build
+      #   run: docker compose up -d
+      # - name: fix  permissions
+      #   run: sudo chmod -R 777 .
+      # - name: fix other permissions
+      #   run: sudo chown $USER_ID:$GROUP_ID .madsci && sudo chmod -R 777 .madsci
+      # - name: Run a command in the lab manager container
+      #   run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
