### 2025-10-16 7807b20
fix permissions again
