+++ .github/workflows/integration_test.yml
@@ -25,14 +25,13 @@ jobs:
         uses: actions/checkout@v4
       - name: example.env to .env
         run: cp .env.example .env
-      - name: check permissions
-        run: sudo chmod -R 777 .madsci
+      - name: docker build
+        run: docker compose build
+      - name: Start services to verify build
+        run: docker compose up -d
       - name: check ownership
         run: ls -ld .madsci
-      # - name: docker build
-      #   run: docker compose build
-      # - name: Start services to verify build
-      #   run: docker compose up -d
+
       # - name: fix  permissions
       #   run: sudo chmod -R 777 .
       # - name: fix other permissions
