+++ .github/workflows/integration_test.yml
@@ -34,11 +34,9 @@ jobs:
       - name: check ownership
         run: ls -ld .madsci
       - name: fix  permissions
-        run: sudo chown -R 1000:1000 .madsci && sudo chmod -R 777 .
+        run: sudo chown -R madsci:madsci .madsci && sudo chmod -R 777 .
       - name: check ownership
         run: ls -ld .madsci
 
-      # - name: fix other permissions
-      #   run: sudo chown $USER_ID:$GROUP_ID .madsci && sudo chmod -R 777 .madsci
       # - name: Run a command in the lab manager container
       #   run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
