+++ README.md
@@ -12,7 +12,7 @@
 | src/madsci\_client/madsci/client/node/\_\_init\_\_.py                                         |        4 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/node/abstract\_node\_client.py                               |       29 |        4 |     86% |26, 29, 95-96 |
 | src/madsci\_client/madsci/client/node/rest\_node\_client.py                                   |      115 |        2 |     98% |   33, 131 |
-| src/madsci\_client/madsci/client/resource\_client.py                                          |      613 |      186 |     70% |100, 125, 147, 174-177, 185, 227-230, 247, 309-313, 326-341, 367, 399-402, 451-458, 464, 477, 488-489, 525, 551, 596-597, 632-634, 674-675, 712-713, 745, 777-778, 810-811, 843-844, 876-877, 906-907, 935-940, 968-970, 1019-1029, 1055-1058, 1106-1116, 1133, 1136-1147, 1175-1182, 1202-1205, 1248-1267, 1282-1288, 1335-1361, 1403-1441, 1470-1487, 1571-1575, 1589, 1612-1613, 1621-1625, 1641-1642, 1653, 1687, 1700, 1713-1723 |
+| src/madsci\_client/madsci/client/resource\_client.py                                          |      627 |      199 |     68% |100, 125, 147, 174-177, 185, 227-230, 247, 309-313, 326-341, 367, 399-402, 451-458, 464, 477, 488-489, 525, 551, 596-597, 632-634, 674-675, 712-713, 745, 777-778, 810-811, 843-844, 876-877, 906-907, 935-940, 968-970, 1001-1054, 1102-1112, 1138-1141, 1189-1199, 1216, 1219-1230, 1258-1265, 1285-1288, 1331-1350, 1365-1371, 1418-1444, 1486-1524, 1553-1570, 1654-1658, 1672, 1695-1696, 1704-1708, 1724-1725, 1736, 1770, 1783, 1796-1806 |
 | src/madsci\_client/madsci/client/workcell\_client.py                                          |      252 |       48 |     81% |122, 151, 179, 194, 243-247, 253, 275, 279, 303-308, 416-437, 464-492, 531, 539, 547, 552-553, 657, 686 |
 | src/madsci\_common/madsci/common/context.py                                                   |       20 |        1 |     95% |        11 |
 | src/madsci\_common/madsci/common/data\_manipulation.py                                        |       43 |        0 |    100% |           |
@@ -64,8 +64,8 @@
 | src/madsci\_node\_module/madsci/node\_module/helpers.py                                       |      108 |       61 |     44% |44, 71-114, 123-135, 146, 148, 150, 156-158, 174-194 |
 | src/madsci\_node\_module/madsci/node\_module/rest\_node\_module.py                            |      141 |       50 |     65% |46, 50, 52, 54-58, 70-84, 95-97, 110, 116-121, 135, 148, 177, 183-185, 192-194, 205-226, 238-240, 276 |
 | src/madsci\_resource\_manager/madsci/resource\_manager/\_\_init\_\_.py                        |        0 |        0 |    100% |           |
[…]
