+++ .github/workflows/integration_test.yml
@@ -31,11 +31,7 @@ jobs:
         run: docker compose up -d
       - name: sleep 10
         run: sleep 10
-      - name: check ownership
-        run: ls -ld .madsci
       - name: fix  permissions
         run: sudo chown -R 1000:1000 .madsci && sudo chmod -R 777 .
-      - name: check ownership
-        run: ls -ld .madsci
       - name: Run a command in the lab manager container
-        run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
+        run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --stdout --to notebook --inplace --execute ./notebooks/experiment_notebook.ipynb
