+++ .github/workflows/integration_test.yml
@@ -27,15 +27,10 @@ jobs:
         run: cp .env.example .env
       - name: docker build
         run: docker compose build
+      - name: Set permissions for madsci log directory
+        run: sudo mkdir -p /home/madsci/.madsci/logs && sudo chown -R $USER:$USER /home/madsci/.madsci
       - name: Start services to verify build
         run: docker compose up -d
-      - name: sleep 10
-        run: sleep 10
-      - name: check ownership
-        run: ls -ld .madsci
-      - name: fix  permissions
-        run: sudo chown -R 1000:1000 .madsci && sudo chmod -R 777 .
-      - name: check ownership
-        run: ls -ld .madsci
+
       - name: Run a command in the lab manager container
         run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --stdout --execute ./notebooks/experiment_notebook.ipynb
