### 2025-10-17 521533e
Bump to v0.5.0-rc.1
