+++ src/madsci_squid/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.squid"
-version = "0.4.7"
+version = "0.5.0-rc.1"
 description = "The Modular Autonomous Discovery for Science (MADSci) Control Server and Scheduler, aka Squid."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
