+++ src/madsci_experiment_application/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.experiment_application"
-version = "0.4.7"
+version = "0.5.0-rc.1"
 description = "The Modular Autonomous Discovery for Science (MADSci) Experiment Application framework."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
