+++ src/madsci_data_manager/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.data_manager"
-version = "0.4.7"
+version = "0.5.0-rc.1"
 description = "The Modular Autonomous Discovery for Science (MADSci) Data Manager."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
