+++ src/madsci_workcell_manager/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.workcell_manager"
-version = "0.4.7"
+version = "0.5.0-rc.1"
 description = "The Modular Autonomous Discovery for Science (MADSci) Workcell Manager."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
