+++ src/madsci_node_module/pyproject.toml
@@ -1,7 +1,7 @@
 
 [project]
 name = "madsci.node_module"
-version = "0.4.7"
+version = "0.5.0-rc.1"
 description = "The Modular Autonomous Discovery for Science (MADSci) Node Module Helper Classes."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
