### 2025-10-17 427900e
Fix: add support back for Annotated return types
