+++ src/madsci_location_manager/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.location_manager"
-version = "0.5.0-rc.1"
+version = "0.5.0-rc.2"
 description = "The Modular Autonomous Discovery for Science (MADSci) Location Manager."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
