+++ src/madsci_client/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.client"
-version = "0.5.0-rc.1"
+version = "0.5.0-rc.2"
 description = "The Modular Autonomous Discovery for Science (MADSci) Python Client and CLI."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
