+++ src/madsci_experiment_application/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.experiment_application"
-version = "0.5.0-rc.1"
+version = "0.5.0-rc.2"
 description = "The Modular Autonomous Discovery for Science (MADSci) Experiment Application framework."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
