+++ src/madsci_experiment_manager/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.experiment_manager"
-version = "0.5.0-rc.1"
+version = "0.5.0-rc.2"
 description = "The Modular Autonomous Discovery for Science (MADSci) Experiment Manager."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
