+++ README.md
@@ -62,8 +62,8 @@
 | src/madsci\_location\_manager/madsci/location\_manager/transfer\_planner.py                   |      175 |        2 |     99% |  282, 315 |
 | src/madsci\_node\_module/madsci/node\_module/\_\_init\_\_.py                                  |        4 |        0 |    100% |           |
 | src/madsci\_node\_module/madsci/node\_module/abstract\_node\_module.py                        |      474 |      119 |     75% |78, 82, 84, 86, 88, 90, 92, 96, 98, 100, 102, 113, 129, 140, 165-169, 198-200, 209, 230-235, 240-248, 258-265, 293-305, 316-317, 328-338, 360-362, 376-378, 486, 564, 591, 616, 638, 663, 669, 689, 692, 739-740, 771-772, 820, 832, 836, 846-851, 872, 875-887, 904, 917, 923-924, 939, 949-950, 956-957, 984-992, 1008, 1024-1028, 1042-1046, 1060-1061, 1075-1076, 1097-1100 |
-| src/madsci\_node\_module/madsci/node\_module/helpers.py                                       |      147 |       59 |     60% |42, 69-112, 117-134, 141, 162-167, 213, 232 |
-| src/madsci\_node\_module/madsci/node\_module/rest\_node\_module.py                            |      344 |       68 |     80% |53, 57, 59, 61, 71-85, 96-110, 134, 139-150, 189-193, 219, 230-232, 243-257, 272, 280-294, 298, 316-318, 352, 355, 378, 381, 445, 471, 529, 566, 589-590, 688, 776 |
+| src/madsci\_node\_module/madsci/node\_module/helpers.py                                       |      152 |       59 |     61% |42, 69-112, 117-134, 141, 162-167, 224, 243 |
+| src/madsci\_node\_module/madsci/node\_module/rest\_node\_module.py                            |      348 |       70 |     80% |62, 66, 68, 70, 80-94, 105-119, 143, 148-159, 198-202, 228, 239-241, 252-266, 281, 289-303, 307, 325-327, 361, 364, 387, 390, 454, 480, 538, 565, 579, 595, 606-607, 705, 793 |
 | src/madsci\_resource\_manager/madsci/resource\_manager/\_\_init\_\_.py                        |        0 |        0 |    100% |           |
 | src/madsci\_resource\_manager/madsci/resource\_manager/resource\_interface.py                 |      643 |      125 |     81% |76, 80, 82, 88-98, 110, 171-173, 305-309, 367, 369, 371, 373, 375, 456, 460-462, 479, 570, 788, 793-794, 825, 843-888, 927-928, 956-960, 989-991, 1023, 1051-1053, 1090-1092, 1119, 1123, 1141-1143, 1155-1156, 1201-1205, 1225, 1281, 1295-1299, 1340-1358, 1373-1375, 1378-1382, 1394-1401, 1444, 1460-1462, 1497-1500, 1511-1513, 1551-1553, 1600 |
 | src/madsci\_resource\_manager/madsci/resource\_manager/resource\_server.py                    |      400 |       93 |     77% |82-86, 134-135, 175-177, 182-185, 197, 223-225, 236-238, 247-251, 277-279, 332-334, 376-380, 391-393, 400-402, 410-412, 426-428, 436, 440-444, 455-459, 473-475, 521-524, 535-537, 776-781, 796-801, 827-829, 834, 858-860, 865, 890-892, 919-921, 926-927 |
@@ -81,7 +81,7 @@
[…]
