+++ .github/workflows/integration_test.yml
@@ -25,13 +25,13 @@ jobs:
         uses: actions/checkout@v4
       - name: example.env to .env
         run: cp .env.example .env
-      - name: docker build
-        run: docker compose build
-      - name: Start services to verify build
-        run: docker compose up -d
-      - name: Sleep for 30 seconds to allow services to start
-        run: sleep 30
-      - name: Change ownership of /home/madsci/.madsci
-        run: sudo chown -R $USER:$USER .madsci
+      - name: Set USER_ID and GROUP_ID in .env
+        run: |
+          echo "USER_ID=$(id -u)" >> .env
+          echo "GROUP_ID=$(id -g)" >> .env
+      - name: Ensure .madsci directory exists
+        run: mkdir -p ./.madsci
+      - name: Build and start example lab services, and wait for startup
+        run: docker compose up --build -d && sleep 30
       - name: Run a command in the lab manager container
-        run: docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --stdout --execute ./notebooks/experiment_notebook.ipynb
+        run: timeout 10m docker compose run --rm --no-deps workcell_manager python -m nbconvert --to notebook --inplace --stdout --execute ./notebooks/experiment_notebook.ipynb
