### 2025-10-21 3b03680
Initial exploration and test creation for Optional[LocationArgument] issue
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
