+++ src/madsci_node_module/tests/test_optional_location_argument.py
@@ -1,0 +1,244 @@
+"""Test for Optional[LocationArgument] and other complex type handling in node actions."""
+
+from typing import Optional, Union
+from unittest.mock import MagicMock, patch
+import pytest
+from pydantic import BaseModel
+
+from madsci.common.types.location_types import LocationArgument
+from madsci.common.types.action_types import ActionRequest
+from madsci.node_module.abstract_node_module import AbstractNode
+from madsci.node_module.helpers import action
+from madsci.common.types.node_types import NodeConfig
+
+
+class CustomModel(BaseModel):
+    """A custom pydantic model for testing."""
+    value: str
+    count: int = 0
+
+
+class TestNodeConfig(NodeConfig):
+    """Test configuration."""
+    update_node_files: bool = False
+
+
+class MockNode(AbstractNode):
+    """Mock node for testing."""
+    
+    __test__ = False
+    config_model = TestNodeConfig
+    
+    def __init__(self):
+        """Initialize without calling parent __init__ to avoid file dependencies."""
+        # Minimal initialization needed for testing
+        from madsci.common.types.node_types import NodeDefinition, NodeInfo
+        from madsci.common.utils import new_ulid_str
+        
+        self.config = TestNodeConfig()
+        self.node_definition = NodeDefinition(
+            node_name="test_node",
+            node_id=new_ulid_str(),
+            module_name="test_module",
+            module_version="0.0.1"
+        )
+        self.node_info = NodeInfo.from_node_def_and_config(
+            self.node_definition, self.config
+        )
+        self.action_handlers = {}
+        self.action_history = {}
+        self.node_state = {}
+        
+        # Configure logger and clients
+        from madsci.client.event_client import EventClient
+        from madsci.client.resource_client import ResourceClient
+        from madsci.client.data_client import DataClient
+        
+        self.logger = self.event_client = EventClient()
+        self.resource_client = ResourceClient(event_client=self.event_client)
+        self.data_client = DataClient()
+        
+        # Initialize node status
+        from madsci.common.types.node_types import NodeStatus
+        self.node_status = NodeStatus(ready=True)
+        
+        # Add actions
+        for action_callable in self.__class__.__dict__.values():
+            if hasattr(action_callable, "__is_madsci_action__"):
+                self._add_action(
+                    func=action_callable,
+                    action_name=action_callable.__madsci_action_name__,
+                    description=action_callable.__madsci_action_description__,
+                    blocking=action_callable.__madsci_action_blocking__,
+                    result_definitions=action_callable.__madsci_action_result_definitions__,
+                )
+    
+    @action
+    def action_with_optional_location(
+        self,
+        target: Optional[LocationArgument] = None,
+        speed: int = 100,
+    ) -> str:
+        """Action with optional location argument."""
+        if target is not None:
+            assert isinstance(target, LocationArgument), f"Expected LocationArgument, got {type(target)}"
+            return f"Moving to {target.representation} at speed {speed}"
+        return f"No movement at speed {speed}"
+    
+    @action
+    def action_with_union_location(
+        self,
+        target: Union[LocationArgument, str],
+        speed: int = 100,
+    ) -> str:
+        """Action with union type including location argument."""
+        if isinstance(target, LocationArgument):
+            return f"Moving to location {target.representation} at speed {speed}"
+        return f"Moving to string {target} at speed {speed}"
+    
+    @action
+    def action_with_optional_custom_model(
+        self,
+        data: Optional[CustomModel] = None,
+    ) -> str:
+        """Action with optional custom BaseModel."""
+        if data is not None:
+            assert isinstance(data, CustomModel), f"Expected CustomModel, got {type(data)}"
+            return f"Data: {data.value}, count: {data.count}"
+        return "No data"
+    
+    @action
+    def action_with_union_custom_model(
+        self,
[…]
