### 2025-10-21 2ac3b68
Fix Optional[LocationArgument] and Optional[BaseModel] handling in node actions
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
