+++ src/madsci_node_module/madsci/node_module/abstract_node_module.py
@@ -63,7 +63,7 @@ from madsci.common.utils import (
     threaded_daemon,
     to_snake_case,
 )
-from pydantic import ValidationError
+from pydantic import BaseModel, ValidationError
 from semver import Version
 
 
@@ -554,8 +554,8 @@ class AbstractNode:
             arg_dict, action_request, param_analysis, parameters
         )
 
-        # Validate any arguments that expect a LocationArgument
-        return self._validate_location_arguments(action_callable, arg_dict)
+        # Validate any arguments that expect a Pydantic BaseModel (LocationArgument, etc.)
+        return self._validate_pydantic_arguments(action_callable, arg_dict)
 
     def _get_action_callable(self, action_name: str) -> Callable:
         """Get the callable for an action, raising an error if not found."""
@@ -743,35 +743,115 @@ class AbstractNode:
             if var_args:
                 arg_dict["__madsci_var_args__"] = var_args
 
-    def _validate_location_arguments(
+    def _extract_pydantic_types_from_hint(
+        self, type_hint: Any
+    ) -> list[type[BaseModel]]:
+        """
+        Extract all Pydantic BaseModel types from a type hint.
+
+        Handles:
+        - Direct BaseModel subclasses
+        - Optional[BaseModel] (Union[BaseModel, None])
+        - Union[BaseModel, OtherType, ...]
+        - Annotated[BaseModel, ...]
+
+        Args:
+            type_hint: The type hint to analyze
+
+        Returns:
+            List of BaseModel subclass types found in the hint
+        """
+        pydantic_types = []
+
+        # Handle Annotated types - extract the actual type
+        origin = get_origin(type_hint)
+        if origin is Annotated:
+            type_hint = get_args(type_hint)[0]
+            origin = get_origin(type_hint)
+
+        # Handle Union types (including Optional)
+        if origin is Union:
+            for arg in get_args(type_hint):
+                # Skip None type
+                if arg is type(None):
+                    continue
+                # Check if it's a BaseModel subclass
+                try:
+                    if isinstance(arg, type) and issubclass(arg, BaseModel):
+                        pydantic_types.append(arg)
+                except TypeError:
+                    # issubclass raises TypeError if arg is not a class
+                    pass
+        else:
+            # Direct type - check if it's a BaseModel subclass
+            try:
+                if isinstance(type_hint, type) and issubclass(type_hint, BaseModel):
+                    pydantic_types.append(type_hint)
+            except TypeError:
+                pass
+
+        return pydantic_types
+
+    def _validate_pydantic_arguments(
         self,
         action_callable: Callable,
         arg_dict: dict[str, Any],
     ) -> dict[str, Any]:
         """
-        Validate and convert any arguments expected as LocationArgument.
+        Validate and convert any arguments expected as Pydantic BaseModel instances.
 
-        If the action function declares a parameter with type LocationArgument and the
-        corresponding value in arg_dict is a dictionary (e.g., from a deserialized JSON payload),
-        this function uses Pydantic's model_validate to reconstruct a valid LocationArgument instance.
+        This handles LocationArgument and any other BaseModel subclasses that may be
+        passed as dictionaries from deserialized JSON payloads. It properly handles:
+        - Direct BaseModel types: LocationArgument, CustomModel, etc.
+        - Optional[BaseModel]: Optional[LocationArgument], etc.
+        - Union[BaseModel, OtherType]: Union[LocationArgument, str], etc.
+
+        If the action function declares a parameter with a BaseModel type (directly or
+        within Optional/Union) and the corresponding value in arg_dict is a dictionary,
+        this function uses Pydantic's model_validate to reconstruct the proper model instance.
 
         Raises:
[…]
