### 2025-10-21 4cd0f90
Move Optional[LocationArgument] tests to test_node_infrastructure.py
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
