### 2025-10-21 86a078f
Get current owner to query on the resource
