### 2025-10-21 176600b
Use node ID to query on resource
