### 2025-10-21 6057247
Renamed list templates to query templates
