+++ src/madsci_node_module/tests/test_node_infrastructure.py
@@ -1,17 +1,25 @@
-"""Test node infrastructure including lifecycle, configuration, and basic actions.
-
-Consolidates tests from:
-- test_node.py (TestNode implementation)
-- test_rest_utils.py (utilities)
-- Basic node lifecycle tests from test_rest_node_module.py
-"""
+"""Test node infrastructure including lifecycle, configuration, and basic actions."""
 
 import time
+from typing import Optional
 
 from fastapi.testclient import TestClient
-from madsci.common.types.action_types import ActionResult, ActionStatus
+from madsci.client.data_client import DataClient
+from madsci.client.event_client import EventClient
+from madsci.client.resource_client import ResourceClient
+from madsci.common.types.action_types import ActionRequest, ActionResult, ActionStatus
 from madsci.common.types.admin_command_types import AdminCommandResponse
-from madsci.common.types.node_types import NodeStatus
+from madsci.common.types.location_types import LocationArgument
+from madsci.common.types.node_types import (
+    NodeConfig,
+    NodeDefinition,
+    NodeInfo,
+    NodeStatus,
+)
+from madsci.common.utils import new_ulid_str
+from madsci.node_module.abstract_node_module import AbstractNode
+from madsci.node_module.helpers import action
+from pydantic import BaseModel
 from ulid import ULID
 
 from madsci_node_module.tests.test_node import TestNode
@@ -370,191 +378,129 @@ class TestAdminCommands:
             assert validated_response.success is True
 
 
+class CustomModel(BaseModel):
+    """A custom pydantic model for testing."""
+
+    value: str
+    count: int = 0
+
+
+class TestNodeConfig(NodeConfig):
+    """Test configuration."""
+
+    update_node_files: bool = False
+
+
+class MockNode(AbstractNode):
+    """Mock node for testing."""
+
+    __test__ = False
+    config_model = TestNodeConfig
+
+    def __init__(self):
+        """Initialize without calling parent __init__ to avoid file dependencies."""
+        self.config = TestNodeConfig()
+        self.node_definition = NodeDefinition(
+            node_name="test_node",
+            node_id=new_ulid_str(),
+            module_name="test_module",
+            module_version="0.0.1",
+        )
+        self.node_info = NodeInfo.from_node_def_and_config(
+            self.node_definition, self.config
+        )
+        self.action_handlers = {}
+        self.action_history = {}
+        self.node_state = {}
+        self.logger = self.event_client = EventClient()
+        self.resource_client = ResourceClient(event_client=self.event_client)
+        self.data_client = DataClient()
+        self.node_status = NodeStatus(ready=True)
+
+        # Add actions
+        self._register_actions()
+
+    def _register_actions(self):
+        """Register all actions for this mock node."""
+        for action_callable in self.__class__.__dict__.values():
+            if hasattr(action_callable, "__is_madsci_action__"):
+                self._add_action(
+                    func=action_callable,
+                    action_name=action_callable.__madsci_action_name__,
+                    description=action_callable.__madsci_action_description__,
+                    blocking=action_callable.__madsci_action_blocking__,
+                    result_definitions=action_callable.__madsci_action_result_definitions__,
+                )
+
+    @action
+    def action_with_optional_location(
+        self,
+        target: Optional[LocationArgument] = None,
+        speed: int = 100,
+    ) -> str:
+        """Action with optional location argument."""
+        if target is not None:
+            assert isinstance(target, LocationArgument), (
+                f"Expected LocationArgument, got {type(target)}"
+            )
+            return f"Moving to {target.representation} at speed {speed}"
+        return f"No movement at speed {speed}"
+
+    @action
+    def action_with_union_location(
+        self,
[…]
