### 2025-10-21 0b16d20
Bug fix for query_templates
