### 2025-10-21 cb831fe
Bug fix for query_templates
