### 2025-10-21 2050320
Fix pipe vs. union issue
