### 2025-10-21 f766bbb
Fix github actions workflows to run checks on pull request
