### 2025-10-22 3ef3c9d
Update to classy-fastapi v0.7.0 and centralize common routes
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
