### 2025-10-22 db3ffd0
Add comprehensive tests for inherited routes in AbstractManagerBase
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
