### 2025-10-22 6276df1
Fix linting issues in manager_base and test_manager_base
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
