### 2025-10-22 891f270
Fix some outdated tests and route shenanigans
