### 2025-10-22 4fd0be6
Updated docstring
