+++ README.md
@@ -17,7 +17,7 @@
 | src/madsci\_common/madsci/common/context.py                                                   |       20 |        1 |     95% |        11 |
 | src/madsci\_common/madsci/common/data\_manipulation.py                                        |       43 |        0 |    100% |           |
 | src/madsci\_common/madsci/common/exceptions.py                                                |       23 |        2 |     91% |     30-31 |
-| src/madsci\_common/madsci/common/manager\_base.py                                             |       85 |       24 |     72% |120-124, 163-164, 171-184, 246-257, 281-282 |
+| src/madsci\_common/madsci/common/manager\_base.py                                             |      108 |       26 |     76% |80-82, 146-150, 189-190, 235-248, 310-321, 345-346 |
 | src/madsci\_common/madsci/common/nodes.py                                                     |        7 |        0 |    100% |           |
 | src/madsci\_common/madsci/common/object\_storage\_helpers.py                                  |      107 |       35 |     67% |45-52, 62-68, 86-92, 130, 199-204, 211-216, 226, 237-243, 294-299, 312-318, 334-355 |
 | src/madsci\_common/madsci/common/ownership.py                                                 |       20 |        1 |     95% |        11 |
@@ -47,17 +47,17 @@
 | src/madsci\_common/madsci/common/validators.py                                                |       27 |        0 |    100% |           |
 | src/madsci\_common/madsci/common/warnings.py                                                  |        4 |        2 |     50% |       4-7 |
 | src/madsci\_common/madsci/common/workflows.py                                                 |       28 |       15 |     46% |14-29, 39-48, 64, 68 |
-| src/madsci\_data\_manager/madsci/data\_manager/data\_server.py                                |      125 |       15 |     88% |59, 94-102, 119, 131, 195, 224, 231, 269-270 |
-| src/madsci\_event\_manager/madsci/event\_manager/event\_server.py                             |      119 |       45 |     62% |59-60, 79-82, 99, 108-115, 132-133, 171-210, 214-218, 224-232, 237-238 |
+| src/madsci\_data\_manager/madsci/data\_manager/data\_server.py                                |      116 |       14 |     88% |59, 94-102, 116, 180, 209, 216, 254-255 |
+| src/madsci\_event\_manager/madsci/event\_manager/event\_server.py                             |      110 |       44 |     60% |59-60, 79-82, 93-100, 117-118, 156-195, 199-203, 209-217, 222-223 |
 | src/madsci\_event\_manager/madsci/event\_manager/events\_csv\_exporter.py                     |      275 |      231 |     16% |29-57, 62-66, 71-72, 77-85, 90-111, 116-144, 149-180, 185-193, 200-233, 240-260, 265-267, 280-282, 297-335, 356-507, 525-638, 645-657, 674-710, 727-753, 770-796 |
 | src/madsci\_event\_manager/madsci/event\_manager/notifications.py                             |       52 |        5 |     90% |31-32, 43, 84-85 |
[…]
