### 2025-10-22 b1fa5ed
query_templates_simple->list_templates
