+++ src/madsci_resource_manager/madsci/resource_manager/resource_server.py
@@ -395,8 +395,8 @@ class ResourceManager(
             raise HTTPException(status_code=500, detail=str(e)) from e
 
     @get("/templates")
-    async def query_templates_simple(self) -> list[ResourceDataModels]:
-        """Query all templates (simple endpoint without filtering)."""
+    async def list_templates(self) -> list[ResourceDataModels]:
+        """List all templates."""
         try:
             return self._resource_interface.query_templates()
         except Exception as e:
