### 2025-10-23 dbe8abd
Add toggle for root_definition endpoint and update lab server to disable
