### 2025-10-23 5395468
Fix _analyze_file_parameter_type to handle Annotated[Path] types
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
