### 2025-10-23 6bd346a
Add _is_file_type helper to handle list[Path] in abstract_node_module
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
