### 2025-10-23 40e27fa
Add comprehensive tests for Annotated[Path] in node module
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
