### 2025-10-23 d842311
Fix linting issues and handle Optional[Annotated[Path]]
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
