### 2025-10-23 cfac6c7
Refactor to pass ruff/pre-commit checks - reduce complexity and fix formatting
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
