+++ tests/test_annotated_path_bug.py
@@ -19,7 +19,9 @@ def test_annotated_path_recognized_as_file_parameter():
 
     # Test Annotated[Path, "description"] - this should also be a file parameter
     result = _analyze_file_parameter_type(Annotated[Path, "A file path description"])
-    assert result["is_file_param"] is True, "Annotated[Path, ...] should be recognized as a file parameter"
+    assert result["is_file_param"] is True, (
+        "Annotated[Path, ...] should be recognized as a file parameter"
+    )
     assert result["is_list"] is False
     assert result["is_optional"] is False
 
@@ -34,7 +36,9 @@ def test_annotated_path_list_recognized_as_file_parameter():
 
     # Test Annotated[list[Path], "description"] - this should also be a file list parameter
     result = _analyze_file_parameter_type(Annotated[list[Path], "Multiple file paths"])
-    assert result["is_file_param"] is True, "Annotated[list[Path], ...] should be recognized as a file list parameter"
+    assert result["is_file_param"] is True, (
+        "Annotated[list[Path], ...] should be recognized as a file list parameter"
+    )
     assert result["is_list"] is True
     assert result["is_optional"] is False
 
@@ -52,14 +56,20 @@ def test_extract_file_parameters_with_annotated_path():
     file_params = extract_file_parameters(action_with_annotated_path)
 
     # Check that all path parameters are recognized
-    assert "config_file" in file_params, "config_file should be recognized as a file parameter"
+    assert "config_file" in file_params, (
+        "config_file should be recognized as a file parameter"
+    )
     assert file_params["config_file"]["required"] is True
     assert file_params["config_file"]["is_list"] is False
 
-    assert "data_files" in file_params, "data_files should be recognized as a file parameter"
+    assert "data_files" in file_params, (
+        "data_files should be recognized as a file parameter"
+    )
     assert file_params["data_files"]["required"] is True
     assert file_params["data_files"]["is_list"] is True
 
-    assert "optional_file" in file_params, "optional_file should be recognized as a file parameter"
+    assert "optional_file" in file_params, (
+        "optional_file should be recognized as a file parameter"
+    )
     assert file_params["optional_file"]["required"] is False
     assert file_params["optional_file"]["is_list"] is False
