### 2025-10-24 82453c1
Tweak location manager settings fields
