+++ Configuration.md
@@ -160,12 +160,12 @@ Settings for the LocationManager.
 
 **Environment Prefix**: `LOCATION_`
 
-| Name                          | Type                   | Default                    | Description                                      | Example                    |
-|-------------------------------|------------------------|----------------------------|--------------------------------------------------|----------------------------|
-| `LOCATION_SERVER_URL`         | `AnyUrl`               | `"http://localhost:8006/"` | The URL where this manager's server runs.        | `"http://localhost:8006/"` |
-| `LOCATION_MANAGER_DEFINITION` | `string` \| `Path`     | `"manager.yaml"`           | Path to the manager definition file to use.      | `"manager.yaml"`           |
-| `LOCATION_SERVER_HOST`        | `string`               | `"localhost"`              | The host to run the server on.                   | `"localhost"`              |
-| `LOCATION_SERVER_PORT`        | `integer`              | `8006`                     | The port to run the server on.                   | `8006`                     |
-| `LOCATION_REDIS_HOST`         | `string`               | `"localhost"`              | The host of the Redis server for state storage.  | `"localhost"`              |
-| `LOCATION_REDIS_PORT`         | `integer`              | `6379`                     | The port of the Redis server for state storage.  | `6379`                     |
-| `LOCATION_REDIS_PASSWORD`     | `string` \| `NoneType` | `null`                     | The password for the Redis server (if required). | `null`                     |
+| Name                          | Type                   | Default                    | Description                                          | Example                    |
+|-------------------------------|------------------------|----------------------------|------------------------------------------------------|----------------------------|
+| `LOCATION_SERVER_URL`         | `AnyUrl`               | `"http://localhost:8006/"` | The URL where this manager's server runs.            | `"http://localhost:8006/"` |
+| `LOCATION_MANAGER_DEFINITION` | `string` \| `Path`     | `"location.manager.yaml"`  | Path to the location manager definition file to use. | `"location.manager.yaml"`  |
+| `LOCATION_SERVER_HOST`        | `string`               | `"localhost"`              | The host to run the server on.                       | `"localhost"`              |
+| `LOCATION_SERVER_PORT`        | `integer`              | `8006`                     | The port to run the server on.                       | `8006`                     |
+| `LOCATION_REDIS_HOST`         | `string`               | `"localhost"`              | The host of the Redis server for state storage.      | `"localhost"`              |
+| `LOCATION_REDIS_PORT`         | `integer`              | `6379`                     | The port of the Redis server for state storage.      | `6379`                     |
+| `LOCATION_REDIS_PASSWORD`     | `string` \| `NoneType` | `null`                     | The password for the Redis server (if required).     | `null`                     |
