### 2025-10-24 51749bf
Remove some vestigial transfer manager definitions
