+++ example_lab/managers/transfer_config.yaml
@@ -1,152 +1,0 @@
-robots:
-  platecrane:
-    robot_name: platecrane
-    default_step_template:
-      name: "Transfer via {robot_name}"
-      node: platecrane
-      action: transfer
-      args:
-        plate_type: "flat_bottom_96well"
-      locations:
-        source: "{source}"
-        target: "{target}"
-    default_args:
-      plate_type: "flat_bottom_96well"
-
-  sciclops:
-    robot_name: sciclops
-    default_step_template:
-      name: "Get plate via {robot_name}"
-      node: sciclops
-      action: get_plate
-      locations:
-        source: "{source}"
-        target: "{target}"
-    default_args: {}
-
-  pf400:
-    robot_name: pf400
-    default_step_template:
-      name: "Transfer via {robot_name}"
-      node: pf400
-      action: transfer
-      locations:
-        source: "{source}"
-        target: "{target}"
-    default_args: {}
-
-  robotarm_1:
-    robot_name: robotarm_1
-    default_step_template:
-      name: "Transfer via {robot_name}"
-      node: robotarm_1
-      action: transfer
-      locations:
-        source: "{source}"
-        target: "{target}"
-    default_args: {}
-
-locations:
-  tower1:
-    location_name: tower1
-    description: "Storage tower 1"
-    accessible_by: ["sciclops"]
-    capacity: 20
-    default_args: {}
-    lookup:
-      sciclops: [100, 200, 300, 400, 500, 600]
-
-  exchange:
-    location_name: exchange
-    description: "Central exchange location - hub for transfers"
-    accessible_by: ["sciclops", "pf400"]
-    capacity: 2
-    default_args: {}
-    lookup:
-      sciclops: [150, 250, 350, 450, 550, 650]
-      pf400: [75.5, 25.3, 45.7, 600.2, 75.8, 900.1]
-
-  camera_station:
-    location_name: camera_station
-    description: "Camera imaging station"
-    accessible_by: ["pf400"]
-    capacity: 1
-    default_args: {}
-    robot_overrides:
-      pf400:
-        source_plate_rotation: "narrow"
-        target_plate_rotation: "narrow"
-    lookup:
-      pf400: [94.597, 26.416, 66.422, 714.811, 81.916, 995.074]
-
-  ot2_gamma_deck_2:
-    location_name: ot2_gamma_deck_2
-    description: "OT2 Gamma robot deck position 2"
-    accessible_by: ["pf400"]
-    capacity: 1
-    default_args: {}
-    robot_overrides:
-      pf400:
-        target_plate_rotation: "wide"
-    lookup:
-      pf400: [85.2, 30.1, 70.5, 720.3, 85.4, 1000.2]
-
-  hidex_nest:
-    location_name: hidex_nest
-    description: "Hidex plate reader nest position"
-    accessible_by: ["platecrane"]
-    capacity: 1
-    default_args:
-      height_offset: 8
-    robot_overrides:
-      platecrane:
-        approach_speed: "slow"
-    lookup:
-      platecrane: [45.0, 15.0, 30.0, 500.0, 40.0, 750.0]
-
-  sealer_nest:
-    location_name: sealer_nest
-    description: "Plate sealer nest position"
-    accessible_by: ["platecrane"]
-    capacity: 1
-    default_args: {}
-    lookup:
-      platecrane: [50.0, 20.0, 35.0, 520.0, 45.0, 780.0]
-
-  liconic_nest:
-    location_name: liconic_nest
-    description: "Liconic incubator loading position"
-    accessible_by: ["platecrane"]
-    capacity: 1
[…]
