### 2025-10-24 8069920
Remove more vestigial transfer manager definitions
