+++ example_lab/node_definitions/transfer_manager.info.yaml
@@ -1,178 +1,0 @@
-node_name: transfer_manager
-node_id: 01K1HJKYJN6S4HWX2P5K4B1R4J
-node_description: This node manages the transfer operations.
-node_type: transfer_manager
-module_name: transfer_manager
-module_version: 0.0.1
-capabilities:
-  get_info: true
-  get_state: true
-  get_status: true
-  send_action: true
-  get_action_result: true
-  get_action_history: true
-  action_files: true
-  send_admin_commands: true
-  set_config: true
-  get_resources: false
-  get_log: true
-  admin_commands:
-  - lock
-  - reset
-  - shutdown
-  - unlock
-node_url: null
-actions:
-  transfer:
-    name: transfer
-    description: Transfer between two locations by submitting child workflow
-    args: {}
-    locations:
-      source:
-        name: source
-        description: Location to transfer from
-        argument_type: location
-        required: true
-        default: null
-      target:
-        name: target
-        description: Location to transfer to
-        argument_type: location
-        required: true
-        default: null
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-  transfer_resource:
-    name: transfer_resource
-    description: Transfer a resource by ID to a target location
-    args:
-      source_resource:
-        name: source_resource
-        description: ID of the resource to transfer
-        argument_type: Union[LocationArgument, str]
-        required: true
-        default: null
-    locations:
-      target:
-        name: target
-        description: Target location to transfer to
-        argument_type: location
-        required: true
-        default: null
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-  get_transfer_options:
-    name: get_transfer_options
-    description: Get available transfer path options between two locations
-    args:
-      source:
-        name: source
-        description: Source location name
-        argument_type: str
-        required: true
-        default: null
-      target:
-        name: target
-        description: Target location name
-        argument_type: str
-        required: true
-        default: null
-    locations: {}
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-  validate_transfer:
-    name: validate_transfer
-    description: Validate if a transfer between two locations is possible
-    args:
-      source:
-        name: source
-        description: Source location name
-        argument_type: str
-        required: true
-        default: null
-      target:
-        name: target
-        description: Target location name
-        argument_type: str
-        required: true
-        default: null
-    locations: {}
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-config:
-  node_definition: node_definitions/transfer_manager.node.yaml
-  node_info_path: null
-  update_node_files: true
-  status_update_interval: 2.0
-  state_update_interval: 2.0
-  node_url: http://localhost:8006
-  uvicorn_kwargs: {}
-  workcell_server_url: http://localhost:8005
-config_schema:
-  description: Configuration for the Transfer Node.
-  properties:
-    node_definition:
-      anyOf:
-      - type: string
-      - format: path
-        type: string
-      - type: 'null'
-      default: default.node.yaml
-      description: Path to the node definition file to use. If set, the node will
-        load the definition from this file on startup. Otherwise, a default configuration
-        will be created.
-      title: Node Definition File
-    node_info_path:
-      anyOf:
-      - type: string
-      - format: path
-        type: string
-      - type: 'null'
-      default: null
-      description: Path to export the generated node info file. If not set, will use
-        the node name and the node_definition's path.
-      title: Node Info Path
-    update_node_files:
-      default: true
-      description: Whether to update the node definition and info files on startup.
-        If set to False, the node will not update the files even if they are out of
-        date.
-      title: Update Node Files
-      type: boolean
-    status_update_interval:
-      anyOf:
[…]
