### 2025-10-24 699090a
More cleanup and dependency fix
