+++ managers/example_event.manager.yaml
@@ -1,4 +1,0 @@
-name: Event Manager
-description: null
-manager_type: event_manager
-event_manager_id: 01K4XCDBH131Q3ZSRSFM7TP5GH
