+++ managers/example_data.manager.yaml
@@ -1,4 +1,0 @@
-name: Data Manager
-description: null
-manager_type: data_manager
-data_manager_id: 01K4XCH44YXVEZ0GSZZW78Y83A
