+++ managers/example_experiment.manager.yaml
@@ -1,4 +1,0 @@
-name: Experiment Manager
-description: null
-manager_type: experiment_manager
-experiment_manager_id: 01K5A16JMRQNPA25Y7NCH30J79
