+++ managers/example_resource.manager.yaml
@@ -1,5 +1,0 @@
-name: Resource Manager
-description: null
-manager_type: resource_manager
-resource_manager_id: 01K4E01QHXBDN1544Y5F6X6645
-custom_types: {}
