### 2025-10-24 cf3db94
Fix pre-commit checks
