+++ example_lab/lab.manager.yaml
@@ -1,4 +1,0 @@
-name: MADSci Lab Manager
-description: null
-manager_id: 01K5AN9H004WHKA4590K63NEVS
-manager_type: lab_manager
