### 2025-10-24 9f12453
Support semantic version checks
