### 2025-10-24 2b25a2a
Remove default node artifacts
