+++ default.node.info.yaml
@@ -1,65 +1,0 @@
-node_name: default.node
-node_id: 01K895ZXJD88DGCEWQX7ABJ74G
-node_description: null
-node_type: device
-module_name: dummy_node
-module_version: 0.0.1
-capabilities:
-  admin_commands:
-  - lock
-  - unlock
-node_url: null
-actions: {}
-config:
-  node_definition: default.node.yaml
-  node_info_path: null
-  update_node_files: true
-  status_update_interval: 2.0
-  state_update_interval: 2.0
-config_schema:
-  description: Basic Configuration for a MADSci Node.
-  properties:
-    node_definition:
-      anyOf:
-      - type: string
-      - format: path
-        type: string
-      - type: 'null'
-      default: default.node.yaml
-      description: Path to the node definition file to use. If set, the node will
-        load the definition from this file on startup. Otherwise, a default configuration
-        will be created.
-      title: Node Definition File
-    node_info_path:
-      anyOf:
-      - type: string
-      - format: path
-        type: string
-      - type: 'null'
-      default: null
-      description: Path to export the generated node info file. If not set, will use
-        the node name and the node_definition's path.
-      title: Node Info Path
-    update_node_files:
-      default: true
-      description: Whether to update the node definition and info files on startup.
-        If set to False, the node will not update the files even if they are out of
-        date.
-      title: Update Node Files
-      type: boolean
-    status_update_interval:
-      anyOf:
-      - type: number
-      - type: 'null'
-      default: 2.0
-      description: The interval in seconds at which the node should update its status.
-      title: Status Update Interval
-    state_update_interval:
-      anyOf:
-      - type: number
-      - type: 'null'
-      default: 2.0
-      description: The interval in seconds at which the node should update its state.
-      title: State Update Interval
-  title: NodeConfig
-  type: object
