### 2025-10-24 8066fbd
Switch to __del__ deconstructors
