+++ src/madsci_data_manager/madsci/data_manager/data_server.py
@@ -97,10 +97,6 @@ class DataManager(AbstractManagerBase[DataManagerSettings, DataManagerDefinition
                 "\nTo resolve this issue, run the migration tool and restart the server."
             )
             raise
-        finally:
-            # Always dispose of the version checker
-            if version_checker:
-                version_checker.dispose()
 
     def _setup_database(self) -> None:
         """Setup database connection and collections."""
