+++ src/madsci_event_manager/madsci/event_manager/event_server.py
@@ -91,10 +91,6 @@ class EventManager(AbstractManagerBase[EventManagerSettings, EventManagerDefinit
                 "\nTo resolve this issue, run the migration tool and restart the server."
             )
             raise
-        finally:
-            # Always dispose of the version checker
-            if version_checker:
-                version_checker.dispose()
 
     def setup_logging(self) -> None:
         """Setup logging for the event manager. Prevent recursive logging."""
