### 2025-10-24 aa12f41
Add comprehensive unit tests for save_workflow_files function
Added two new unit tests to directly test the save_workflow_files function:
1. test_save_workflow_files_flushes_buffer - Tests that file buffer is properly flushed
2. test_save_workflow_files_with_multiple_files - Tests handling of multiple file inputs

These tests provide better coverage and serve as regression tests for the file emptying bug.

Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
