+++ src/madsci_workcell_manager/tests/test_workcell_utils.py
@@ -1,9 +1,15 @@
 """Unit tests for madsci.workcell_manager.workcell_utils module."""
 
-from unittest.mock import MagicMock, patch
+from io import BytesIO
+from unittest.mock import MagicMock, Mock, patch
 
+from fastapi import UploadFile
 from madsci.client.node import AbstractNodeClient
+from madsci.common.types.datapoint_types import FileDataPoint
+from madsci.common.types.parameter_types import ParameterInputFile
+from madsci.common.types.workflow_types import Workflow, WorkflowParameters
 from madsci.workcell_manager.workcell_utils import find_node_client
+from madsci.workcell_manager.workflow_utils import save_workflow_files
 
 
 class MockNodeClient(AbstractNodeClient):
@@ -145,3 +151,126 @@ def test_find_node_client_validation_fails():
 
         assert result is None
         mock_client.validate_url.assert_called_once_with("http://test.com")
+
+
+def test_save_workflow_files_flushes_buffer():
+    """Test that save_workflow_files properly flushes file buffer before reading.
+
+    This is a regression test for the bug where files were being emptied because
+    the buffer wasn't flushed after writing to the temp file, resulting in the
+    data_client reading an empty file.
+    """
+    # Create a mock workflow with a file input parameter
+    workflow = Workflow(
+        name="Test Workflow",
+        parameters=WorkflowParameters(
+            file_inputs=[ParameterInputFile(key="test_file")]
+        ),
+        file_input_paths={"test_file": "test.txt"},
+    )
+
+    # Create test file content
+    test_content = b"This is test file content that should not be empty!"
+    file_handle = BytesIO(test_content)
+
+    # Create UploadFile with the test content
+    upload_file = UploadFile(filename="test_file", file=file_handle)
+
+    # Mock the data_client to capture what file is being submitted
+    mock_data_client = Mock()
+    submitted_file_contents = []
+
+    def mock_submit_datapoint(datapoint):
+        """Mock submit that reads the file to verify its content."""
+        if isinstance(datapoint, FileDataPoint):
+            file_path = datapoint.path
+            # Read the file content to verify it's not empty
+            content = file_path.read_bytes()
+            submitted_file_contents.append(content)
+            # Clean up the temp file
+            file_path.unlink(missing_ok=True)
+
+        # Return a mock datapoint with an ID
+        mock_result = Mock()
+        mock_result.datapoint_id = "mock_datapoint_id_123"
+        return mock_result
+
+    mock_data_client.submit_datapoint = mock_submit_datapoint
+
+    # Call the function
+    result = save_workflow_files(workflow, [upload_file], mock_data_client)
+
+    # Verify that the file was submitted with non-empty content
+    assert len(submitted_file_contents) == 1, "Expected one file to be submitted"
+    assert (
+        len(submitted_file_contents[0]) == len(test_content)
+    ), f"File content should be {len(test_content)} bytes, but got {len(submitted_file_contents[0])} bytes"
+    assert (
+        submitted_file_contents[0] == test_content
+    ), "File content does not match expected content"
+
+    # Verify that the workflow was updated with the file input ID
+    assert "test_file" in result.file_input_ids
+    assert result.file_input_ids["test_file"] == "mock_datapoint_id_123"
+
+
+def test_save_workflow_files_with_multiple_files():
+    """Test that save_workflow_files correctly handles multiple file inputs."""
+    # Create a mock workflow with multiple file input parameters
+    workflow = Workflow(
+        name="Test Workflow",
+        parameters=WorkflowParameters(
+            file_inputs=[
+                ParameterInputFile(key="file1"),
+                ParameterInputFile(key="file2"),
+                ParameterInputFile(key="file3"),
[…]
