### 2025-10-24 402946b
Run ruff formatter on all files - automatic formatting fixes
Applied ruff format to entire codebase as requested. Reformatted assert statements
in test_workcell_utils.py for better readability. All tests still passing.

Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
