### 2025-10-24 713e487
Refactor to use pydantic settings for config
