+++ src/madsci_workcell_manager/tests/test_workcell_migration.py
@@ -8,9 +8,9 @@ from unittest.mock import Mock, patch
 
 import pytest
 from madsci.common.mongodb_migration_tool import (
+    MongoDBMigrationSettings,
     MongoDBMigrator,
     main,
-    resolve_schema_file_path,
 )
 from madsci.common.mongodb_version_checker import MongoDBVersionChecker
 
@@ -175,23 +175,40 @@ def test_workcell_version_checker_still_requires_migration_for_existing_db():
             schema_file.unlink()
 
 
-@patch("madsci.common.mongodb_migration_tool.resolve_schema_file_path")
 @patch("madsci.common.mongodb_migration_tool.MongoDBMigrator")
 @patch(
-    "sys.argv", ["migration_tool.py", "--database", "madsci_workcells", "--backup-only"]
+    "sys.argv",
+    [
+        "migration_tool.py",
+        "--database",
+        "madsci_workcells",
+        "--backup_only",
+        "true",
+    ],
 )
-def test_workcell_backup_only_command(mock_migrator_class, mock_resolve_schema):
+def test_workcell_backup_only_command(mock_migrator_class):
     """Test backup only command for workcells database"""
-    # Mock the schema file resolution
-    mock_resolve_schema.return_value = Path("fake_schema.json")
-
     mock_migrator = Mock()
     backup_dir = Path(tempfile.gettempdir()) / "madsci_workcells_backup"
     mock_migrator.create_backup.return_value = backup_dir
     mock_migrator_class.return_value = mock_migrator
 
     with patch.dict(os.environ, {"MONGODB_URL": "mongodb://localhost:27017"}):
-        main()
+        # Create temp schema structure for auto-detection
+        temp_dir = Path(tempfile.gettempdir()) / "test_backup_command"
+        temp_dir.mkdir(exist_ok=True)
+        schema_dir = temp_dir / "madsci" / "workcell_manager"
+        schema_dir.mkdir(parents=True, exist_ok=True)
+        temp_schema = schema_dir / "schema.json"
+        temp_schema.write_text(json.dumps({"version": "1.0.0"}))
+
+        original_cwd = Path.cwd()
+        os.chdir(temp_dir)
+
+        try:
+            main()
+        finally:
+            os.chdir(original_cwd)
 
     # Verify only backup was called
     mock_migrator.create_backup.assert_called_once()
@@ -209,9 +226,12 @@ def test_workcell_backup_creation(mock_subprocess):
     schema_file.write_text(json.dumps({"version": "1.0.0"}))
 
     try:
-        migrator = MongoDBMigrator(
-            "mongodb://localhost:27017", "madsci_workcells", str(schema_file)
+        settings = MongoDBMigrationSettings(
+            mongo_db_url="mongodb://localhost:27017",
+            database="madsci_workcells",
+            schema_file=str(schema_file),
         )
+        migrator = MongoDBMigrator(settings)
 
         backup_path = migrator.create_backup()
 
@@ -241,9 +261,12 @@ def test_workcell_collection_creation(mock_mongo_client_workcells):  # noqa
     schema_file.write_text(json.dumps(schema_content))
 
     try:
-        migrator = MongoDBMigrator(
-            "mongodb://localhost:27017", "madsci_workcells", str(schema_file)
+        settings = MongoDBMigrationSettings(
+            mongo_db_url="mongodb://localhost:27017",
+            database="madsci_workcells",
+            schema_file=str(schema_file),
         )
+        migrator = MongoDBMigrator(settings)
 
         # Access the database from the migrator instance, not the mock client
         migrator.database.list_collection_names.return_value = []
@@ -256,34 +279,47 @@ def test_workcell_collection_creation(mock_mongo_client_workcells):  # noqa
         schema_file.unlink()
[…]
