+++ src/madsci_resource_manager/madsci/resource_manager/migration_tool.py
@@ -1,6 +1,5 @@
 """Database migration tool for MADSci resources using Alembic with automatic type conversion handling."""
 
-import argparse
 import os
 import re
 import subprocess
@@ -16,28 +15,83 @@ from alembic.autogenerate import compare_metadata
 from alembic.config import Config
 from alembic.runtime.migration import MigrationContext
 from madsci.client.event_client import EventClient
+from madsci.common.types.base_types import MadsciBaseSettings, PathLike
 from madsci.resource_manager.database_version_checker import DatabaseVersionChecker
 from madsci.resource_manager.resource_tables import ResourceTable, SchemaVersionTable
 from psycopg2 import sql
+from pydantic import Field
 from sqlalchemy import inspect
 from sqlmodel import Session, create_engine, select
+
+
+class DatabaseMigrationSettings(
+    MadsciBaseSettings,
+    env_file=(".env", "resources.env", "migration.env"),
+    toml_file=("settings.toml", "resources.settings.toml", "migration.settings.toml"),
+    yaml_file=("settings.yaml", "resources.settings.yaml", "migration.settings.yaml"),
+    json_file=("settings.json", "resources.settings.json", "migration.settings.json"),
+    env_prefix="RESOURCES_MIGRATION_",
+):
+    """Configuration settings for PostgreSQL database migration operations."""
+
+    db_url: Optional[str] = Field(
+        default=None,
+        title="Database URL",
+        description="PostgreSQL connection URL (e.g., postgresql://user:pass@localhost:5432/resources). "
+        "If not provided, will try RESOURCES_DB_URL environment variable.",
+    )
+    target_version: Optional[str] = Field(
+        default=None,
+        title="Target Version",
+        description="Target version to migrate to (defaults to current MADSci version)",
+    )
+    backup_only: bool = Field(
+        default=False,
+        title="Backup Only",
+        description="Only create a backup, do not run migration",
+    )
+    restore_from: Optional[PathLike] = Field(
+        default=None,
+        title="Restore From",
+        description="Restore from specified backup file instead of migrating",
+    )
+    generate_migration: Optional[str] = Field(
+        default=None,
+        title="Generate Migration",
+        description="Generate a new migration file with the given message",
+    )
+
+    def get_effective_db_url(self) -> str:
+        """Get the effective database URL, trying fallback environment variables if needed."""
+        if self.db_url:
+            return self.db_url
+
+        # Try environment variable
+        db_url = os.getenv("RESOURCES_DB_URL")
+        if db_url:
+            return db_url
+
+        raise RuntimeError(
+            "No database URL provided and RESOURCES_DB_URL environment variable not found. "
+            "Please either:\n"
+            "1. Provide --db-url argument, or\n"
+            "2. Set RESOURCES_DB_URL in your environment/.env file\n"
+            "Example: RESOURCES_DB_URL=postgresql://user:pass@localhost:5432/resources"
+        )
 
 
 class DatabaseMigrator:
     """Handles database schema migrations for MADSci using Alembic with automatic type conversion handling."""
 
     def __init__(
-        self, db_url: Optional[str] = None, logger: Optional[EventClient] = None
+        self, settings: DatabaseMigrationSettings, logger: Optional[EventClient] = None
     ) -> None:
-        """Initialize the migrator with database URL and logger."""
-        # If no db_url provided, try to get from environment
-        if db_url is None:
-            db_url = self._get_database_url_from_env()
-
-        self.db_url = db_url
+        """Initialize the migrator with settings and logger."""
+        self.settings = settings
+        self.db_url = settings.get_effective_db_url()
         self.logger = logger or EventClient()
-        self.engine = create_engine(db_url)
[…]
