+++ src/madsci_common/madsci/common/types/experiment_types.py
@@ -42,10 +42,11 @@ class ExperimentManagerSettings(
         description="Path to the experiment manager definition file to use.",
         default="experiment.manager.yaml",
     )
-    db_url: str = Field(
-        title="Database URL",
-        description="The URL of the database for the experiment manager.",
+    mongo_db_url: str = Field(
+        title="MongoDB URL",
+        description="The URL of the MongoDB database for the experiment manager.",
         default="mongodb://localhost:27017",
+        validation_alias=AliasChoices("mongo_db_url", "EXPERIMENT_DB_URL", "db_url"),
     )
     database_name: str = Field(
         default="madsci_experiments",
