### 2025-10-27 3f85349
WIP: Standardize mongodb url field; pydantify validation logic; explore schema-version vs. madsci version
