+++ src/madsci_data_manager/tests/test_data_migrations.py
@@ -13,6 +13,7 @@ from madsci.common.mongodb_migration_tool import (
     main,
 )
 from madsci.common.mongodb_version_checker import MongoDBVersionChecker
+from pydantic_extra_types.semantic_version import SemanticVersion
 
 
 @pytest.fixture
@@ -26,7 +27,7 @@ def temp_data_schema():
         data_manager_dir.mkdir(parents=True)
         data_schema = {
             "database": "madsci_data",
-            "version": "1.0.0",
+            "schema_version": "1.0.0",
             "description": "Schema definition for MADSci Data Manager MongoDB",
             "collections": {
                 "datapoints": {
@@ -96,7 +97,7 @@ def test_data_version_checker_detects_missing_version_tracking():
         mock_client.return_value = mock_client_instance
 
         schema_file = Path("data_schema.json")
-        schema_file.write_text(json.dumps({"version": "1.0.0"}))
+        schema_file.write_text(json.dumps({"schema_version": "1.0.0"}))
 
         try:
             checker = MongoDBVersionChecker(
@@ -104,7 +105,7 @@ def test_data_version_checker_detects_missing_version_tracking():
             )
 
             db_version = checker.get_database_version()
-            assert db_version == "NO_VERSION_TRACKING"
+            assert db_version == SemanticVersion(0, 0, 0)
 
             needs_migration, _, _ = checker.is_migration_needed()
             assert needs_migration is True
@@ -132,19 +133,15 @@ def test_data_version_checker_auto_initializes_fresh_database():
         mock_collection.find_one.return_value = None
 
         schema_file = Path("data_schema.json")
-        schema_file.write_text(json.dumps({"version": "1.0.0"}))
+        schema_file.write_text(json.dumps({"schema_version": "1.0.0"}))
 
         try:
             checker = MongoDBVersionChecker(
                 "mongodb://localhost:27017", "madsci_data", str(schema_file)
             )
 
-            # Mock the current MADSci version
-            with patch.object(
-                checker, "get_current_madsci_version", return_value="0.5.0rc3"
-            ):
-                # This should auto-initialize without raising an error
-                checker.validate_or_fail()
+            # This should auto-initialize without raising an error
+            checker.validate_or_fail()
 
             # Verify that create_schema_versions_collection was called
             mock_collection.create_index.assert_called()
@@ -166,7 +163,7 @@ def test_data_version_checker_still_requires_migration_for_existing_db():
         mock_client.return_value = mock_client_instance
 
         schema_file = Path("data_schema.json")
-        schema_file.write_text(json.dumps({"version": "1.0.0"}))
+        schema_file.write_text(json.dumps({"schema_version": "1.0.0"}))
 
         try:
             checker = MongoDBVersionChecker(
@@ -207,7 +204,7 @@ def test_data_backup_only_command(mock_migrator_class):
         schema_dir = temp_dir / "madsci" / "data_manager"
         schema_dir.mkdir(parents=True, exist_ok=True)
         temp_schema = schema_dir / "schema.json"
-        temp_schema.write_text(json.dumps({"version": "1.0.0"}))
+        temp_schema.write_text(json.dumps({"schema_version": "1.0.0"}))
 
         original_cwd = Path.cwd()
         os.chdir(temp_dir)
@@ -230,7 +227,7 @@ def test_data_backup_creation(mock_subprocess):
     mock_subprocess.return_value = mock_result
 
     schema_file = Path("data_schema.json")
-    schema_file.write_text(json.dumps({"version": "1.0.0"}))
[…]
