+++ src/madsci_common/madsci/common/types/datapoint_types.py
@@ -265,8 +265,8 @@ class DataManagerSettings(
         title="Collection Name",
         description="The name of the MongoDB collection where data are stored.",
     )
-    mongo_db_url: str = Field(
-        default="mongodb://localhost:27017",
+    mongo_db_url: AnyUrl = Field(
+        default=AnyUrl("mongodb://localhost:27017"),
         title="MongoDB URL",
         description="The URL of the MongoDB database used by the Data Manager.",
         validation_alias=AliasChoices("mongo_db_url", "DATA_DB_URL", "db_url"),
