+++ src/madsci_workcell_manager/tests/test_workcell_migration.py
@@ -13,6 +13,7 @@ from madsci.common.mongodb_migration_tool import (
     main,
 )
 from madsci.common.mongodb_version_checker import MongoDBVersionChecker
+from pydantic_extra_types.semantic_version import SemanticVersion
 
 
 @pytest.fixture
@@ -26,7 +27,7 @@ def temp_workcell_schema():
         workcell_manager_dir.mkdir(parents=True)
         workcell_schema = {
             "database": "madsci_workcells",
-            "version": "1.0.0",
+            "schema_version": "1.0.0",
             "description": "Schema definition for MADSci Workcell Manager MongoDB",
             "collections": {
                 "archived_workflows": {
@@ -89,7 +90,7 @@ def test_workcell_version_checker_detects_missing_version_tracking():
         mock_client.return_value = mock_client_instance
 
         schema_file = Path("workcell_schema.json")
-        schema_file.write_text(json.dumps({"version": "1.0.0"}))
+        schema_file.write_text(json.dumps({"schema_version": "1.0.0"}))
 
         try:
             checker = MongoDBVersionChecker(
@@ -97,7 +98,7 @@ def test_workcell_version_checker_detects_missing_version_tracking():
             )
 
             db_version = checker.get_database_version()
-            assert db_version == "NO_VERSION_TRACKING"
+            assert db_version == SemanticVersion(0, 0, 0)
 
             needs_migration, _, _ = checker.is_migration_needed()
             assert needs_migration is True
@@ -125,19 +126,15 @@ def test_workcell_version_checker_auto_initializes_fresh_database():
         mock_collection.find_one.return_value = None
 
         schema_file = Path("workcell_schema.json")
-        schema_file.write_text(json.dumps({"version": "1.0.0"}))
+        schema_file.write_text(json.dumps({"schema_version": "1.0.0"}))
 
         try:
             checker = MongoDBVersionChecker(
                 "mongodb://localhost:27017", "madsci_workcells", str(schema_file)
             )
 
-            # Mock the current MADSci version
-            with patch.object(
-                checker, "get_current_madsci_version", return_value="0.5.0rc3"
-            ):
-                # This should auto-initialize without raising an error
-                checker.validate_or_fail()
+            # This should auto-initialize without raising an error
+            checker.validate_or_fail()
 
             # Verify that create_schema_versions_collection was called
             mock_collection.create_index.assert_called()
@@ -159,7 +156,7 @@ def test_workcell_version_checker_still_requires_migration_for_existing_db():
         mock_client.return_value = mock_client_instance
 
         schema_file = Path("workcell_schema.json")
-        schema_file.write_text(json.dumps({"version": "1.0.0"}))
+        schema_file.write_text(json.dumps({"schema_version": "1.0.0"}))
 
         try:
             checker = MongoDBVersionChecker(
@@ -200,7 +197,7 @@ def test_workcell_backup_only_command(mock_migrator_class):
         schema_dir = temp_dir / "madsci" / "workcell_manager"
         schema_dir.mkdir(parents=True, exist_ok=True)
         temp_schema = schema_dir / "schema.json"
-        temp_schema.write_text(json.dumps({"version": "1.0.0"}))
+        temp_schema.write_text(json.dumps({"schema_version": "1.0.0"}))
 
         original_cwd = Path.cwd()
         os.chdir(temp_dir)
@@ -223,7 +220,7 @@ def test_workcell_backup_creation(mock_subprocess):
     mock_subprocess.return_value = mock_result
 
     schema_file = Path("workcell_schema.json")
-    schema_file.write_text(json.dumps({"version": "1.0.0"}))
+    schema_file.write_text(json.dumps({"schema_version": "1.0.0"}))
 
     try:
[…]
