### 2025-10-27 83bf617
Bump to v0.5.0-rc.3
