+++ src/madsci_event_manager/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.event_manager"
-version = "0.5.0-rc.2"
+version = "0.5.0-rc.3"
 description = "The Modular Autonomous Discovery for Science (MADSci) Event and Logging Manager."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
