+++ src/madsci_squid/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.squid"
-version = "0.5.0-rc.2"
+version = "0.5.0-rc.3"
 description = "The Modular Autonomous Discovery for Science (MADSci) Control Server and Scheduler, aka Squid."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
