+++ src/madsci_common/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.common"
-version = "0.5.0-rc.2"
+version = "0.5.0-rc.3"
 description = "The Modular Autonomous Discovery for Science (MADSci) Common Definitions and Utilities."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
