### 2025-10-27 160fb4c
Bump to v0.5.0
