+++ example_lab/example_modules/pf400.py
@@ -1,13 +1,0 @@
-""" A fake PF400 module for testing."""
-
-
-import time
-from pathlib import Path
-from typing import Any, Optional
-
-from madsci.client.event_client import EventClient
-from madsci.common.types.action_types import ActionFailed, ActionResult, ActionSucceeded
-from madsci.common.types.admin_command_types import AdminCommandResponse
-from madsci.common.types.node_types import RestNodeConfig
-from madsci.node_module.helpers import action
-from madsci.node_module.rest_node_module import RestNode
