+++ example_lab/node_definitions/ot2_gamma.info.yaml
@@ -1,123 +1,0 @@
-node_name: ot2_gamma
-node_id: 01JQF8FYR6N7FXT2QC2EJWQX94
-node_description: A (fake) OT2 for testing purposes.
-node_type: device
-module_name: ot2_module
-module_version: 2.1.1
-capabilities:
-  get_info: true
-  get_state: true
-  get_status: true
-  send_action: true
-  get_action_result: true
-  get_action_history: true
-  action_files: true
-  send_admin_commands: true
-  set_config: true
-  get_resources: false
-  get_log: true
-  admin_commands:
-  - cancel
-  - lock
-  - pause
-  - reset
-  - resume
-  - safety_stop
-  - shutdown
-  - unlock
-node_url: null
-actions:
-  run_protocol:
-    name: run_protocol
-    description: Run a OT2 protocol.
-    args:
-      parameters:
-        name: parameters
-        description: Parameters for insertion into the protocol.
-        argument_type: dict[str, Any]
-        required: false
-        default: {}
-    locations: {}
-    files:
-      protocol:
-        name: protocol
-        description: Protocol file.
-        argument_type: file
-        required: true
-        default: null
-    results: {}
-    blocking: false
-    asynchronous: true
-  config:
-    node_definition: node_definitions/ot2_gamma.node.yaml
-    node_info_path: null
-    update_node_files: true
-    status_update_interval: 2.0
-    state_update_interval: 2.0
-    node_url: http://localhost:2003/
-    uvicorn_kwargs: {}
-    device_number: 0
-  config_schema:
-    description: Configuration for the OT2 gamma node module.
-    properties:
-    node_definition:
-      anyOf:
-      - type: string
-      - format: path
-        type: string
-      - type: 'null'
-      default: default.node.yaml
-      description: Path to the node definition file to use. If set, the node will
-        load the definition from this file on startup. Otherwise, a default configuration
-        will be created.
-      title: Node Definition File
-    node_info_path:
-      anyOf:
-      - type: string
-      - format: path
-        type: string
-      - type: 'null'
-      default: null
-      description: Path to export the generated node info file. If not set, will use
-        the node name and the node_definition's path.
-      title: Node Info Path
-    update_node_files:
-      default: true
-      description: Whether to update the node definition and info files on startup.
-        If set to False, the node will not update the files even if they are out of
-        date.
-      title: Update Node Files
-      type: boolean
-    status_update_interval:
-      anyOf:
-      - type: number
-      - type: 'null'
-      default: 2.0
-      description: The interval in seconds at which the node should update its status.
-      title: Status Update Interval
-    state_update_interval:
-      anyOf:
-      - type: number
-      - type: 'null'
-      default: 2.0
-      description: The interval in seconds at which the node should update its state.
-      title: State Update Interval
-    node_url:
-      default: http://127.0.0.1:2000/
-      description: The URL used to communicate with the node. This is the base URL
-        for the REST API.
-      format: uri
-      minLength: 1
-      title: Node URL
-      type: string
-    uvicorn_kwargs:
-      additionalProperties: true
-      description: Configuration for the Uvicorn server that runs the REST API.
-      title: Uvicorn Configuration
-      type: object
-    device_number:
-      default: 0
-      title: Device Number
-      type: integer
-  title: OT2GammaConfig
-  type: object
