+++ example_lab/node_definitions/pf400_1.info.yaml
@@ -1,301 +1,0 @@
-node_name: pf400_1
-node_id: 01JQ9WC5N1CE6EVBNAJBE0WW1K
-node_description: A (fake) PF400 for testing purposes.
-node_type: device
-module_name: pf400
-module_version: 2.1.1
-capabilities:
-  get_info: true
-  get_state: true
-  get_status: true
-  send_action: true
-  get_action_result: true
-  get_action_history: true
-  action_files: true
-  send_admin_commands: true
-  set_config: true
-  get_resources: false
-  get_log: true
-  admin_commands:
-  - cancel
-  - lock
-  - pause
-  - reset
-  - resume
-  - safety_stop
-  - shutdown
-  - unlock
-node_url: null
-actions:
-  transfer:
-    name: transfer
-    description: Transfer a plate from one location to another
-    args:
-      source_plate_rotation:
-        name: source_plate_rotation
-        description: Orientation of the plate at the source, wide or narrow
-        argument_type: str
-        required: false
-        default: ''
-      target_plate_rotation:
-        name: target_plate_rotation
-        description: Final orientation of the plate at the target, wide or narrow
-        argument_type: str
-        required: false
-        default: ''
-    locations:
-      source:
-        name: source
-        description: Location to pick a plate from
-        argument_type: location
-        required: true
-        default: null
-      target:
-        name: target
-        description: Location to place a plate to
-        argument_type: location
-        required: true
-        default: null
-      source_approach:
-        name: source_approach
-        description: Location to approach from
-        argument_type: location
-        required: false
-        default: null
-      target_approach:
-        name: target_approach
-        description: Location to approach from
-        argument_type: location
-        required: false
-        default: null
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-  pick_plate:
-    name: pick_plate
-    description: Pick a plate from a source location
-    args: {}
-    locations:
-      source:
-        name: source
-        description: Location to pick a plate from
-        argument_type: location
-        required: true
-        default: null
-      source_approach:
-        name: source_approach
-        description: Location to approach from
-        argument_type: location
-        required: false
-        default: null
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-  place_plate:
-    name: place_plate
-    description: Place a plate in a target location, optionally moving first to target_approach
-    args: {}
-    locations:
-      target:
-        name: target
-        description: Location to place a plate to
-        argument_type: location
-        required: true
-        default: null
-      target_approach:
-        name: target_approach
-        description: Location to approach from
-        argument_type: location
-        required: false
-        default: null
-    files: {}
-    results: {}
-    blocking: false
-    asynchronous: true
-  remove_lid:
-    name: remove_lid
-    description: Remove a lid from a plate
-    args:
-      source_plate_rotation:
-        name: source_plate_rotation
-        description: Orientation of the plate at the source, wide or narrow
-        argument_type: str
-        required: false
-        default: ''
-      target_plate_rotation:
-        name: target_plate_rotation
-        description: Final orientation of the plate at the target, wide or narrow
-        argument_type: str
-        required: false
-        default: ''
-      lid_height:
-        name: lid_height
-        description: height of the lid, in steps
-        argument_type: float
-        required: false
-        default: 7.0
-    locations:
-      source:
-        name: source
-        description: Location to pick a plate from
-        argument_type: location
-        required: true
-        default: null
-      target:
-        name: target
-        description: Location to place a plate to
-        argument_type: location
-        required: true
-        default: null
-      source_approach:
-        name: source_approach
-        description: Location to approach from
-        argument_type: location
-        required: false
-        default: null
-      target_approach:
[…]
