+++ example_lab/managers/example_workcell.yaml
@@ -1,0 +1,5 @@
+name: example_workcell
+manager_type: workcell_manager
+manager_id: 01K871R2HDSWXABFE99MB1SKB9
+description: null
+nodes: {}
