### 2025-11-05 55ef8c7
Handle major and minor version updates only
