### 2025-11-05 eed659d
Avoid running migration tool if version has not been tracked before
