### 2025-11-06 32d3b82
Fix #159, bump to v0.5.1
