+++ src/madsci_node_module/tests/test_node_infrastructure.py
@@ -14,7 +14,9 @@ from madsci.common.types.action_types import (
     ActionRequest,
     ActionResult,
     ActionStatus,
+    ArgumentDefinition,
     FileArgumentDefinition,
+    LocationArgumentDefinition,
 )
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.location_types import LocationArgument
@@ -869,3 +871,225 @@ class TestAnnotatedPathInNode:
         )
         assert isinstance(action_def.files["optional_file"], FileArgumentDefinition)
         assert action_def.files["optional_file"].description == "An optional file"
+
+
+class TestLocationArgumentDetection:
+    """Test that Union types containing LocationArgument are properly detected as location arguments."""
+
+    def test_union_location_argument_detection(self) -> None:
+        """Test that Union[LocationArgument, str] is properly detected as a location argument."""
+        # Create test action definition
+        action_def = ActionDefinition(
+            name="test_action",
+            description="Test action",
+            blocking=False,
+        )
+
+        # Create a dummy node
+        class DummyNode(AbstractNode):
+            __test__ = False
+            config_model = TestNodeConfig
+
+            def __init__(self):
+                """Initialize without calling parent __init__ to avoid file dependencies."""
+                self.config = TestNodeConfig(test_required_param=42)
+                self.node_definition = NodeDefinition(
+                    node_name="test_node",
+                    node_id=new_ulid_str(),
+                    module_name="test_module",
+                    module_version="0.0.1",
+                )
+                self.node_info = NodeInfo.from_node_def_and_config(
+                    self.node_definition, self.config
+                )
+                self.action_handlers = {}
+                self.action_history = {}
+                self.node_state = {}
+                self.logger = self.event_client = EventClient()
+                self.resource_client = ResourceClient(event_client=self.event_client)
+                self.data_client = DataClient()
+                self.node_status = NodeStatus(ready=True)
+
+        node = DummyNode()
+
+        # Simulate a function signature with Union[LocationArgument, str]
+        def test_func(target: Union[LocationArgument, str] = "default"):
+            pass
+
+        signature = inspect.signature(test_func)
+        type_hints = get_type_hints(test_func, include_extras=True)
+
+        node._parse_action_arg(action_def, signature, "target", type_hints["target"])
+
+        # BUG: This currently fails because Union[LocationArgument, str] is not detected as location
+        # The parameter gets added to args instead of locations
+        assert "target" in action_def.locations, (
+            "target should be detected as a location argument because Union contains LocationArgument"
+        )
+        assert "target" not in action_def.args, (
+            "target should not be in regular args when Union contains LocationArgument"
+        )
+        assert isinstance(action_def.locations["target"], LocationArgumentDefinition)
+
+    def test_optional_location_argument_detection(self) -> None:
+        """Test that Optional[LocationArgument] is properly detected as a location argument."""
+        # Create test action definition
+        action_def = ActionDefinition(
+            name="test_action",
+            description="Test action",
+            blocking=False,
+        )
+
+        # Create a dummy node
+        class DummyNode(AbstractNode):
+            __test__ = False
+            config_model = TestNodeConfig
+
+            def __init__(self):
+                """Initialize without calling parent __init__ to avoid file dependencies."""
+                self.config = TestNodeConfig(test_required_param=42)
+                self.node_definition = NodeDefinition(
+                    node_name="test_node",
+                    node_id=new_ulid_str(),
+                    module_name="test_module",
+                    module_version="0.0.1",
+                )
+                self.node_info = NodeInfo.from_node_def_and_config(
+                    self.node_definition, self.config
+                )
+                self.action_handlers = {}
+                self.action_history = {}
+                self.node_state = {}
[…]
