### 2025-11-07 8b730bb
Batter handle db version tracking
