### 2025-11-07 39817a7
Handle version updates based on major and minor updates to MADSci
