### 2025-11-07 292c841
Track schema versions for migration rather than MADSci version
