### 2025-11-07 3123ac8
Track schema versions for migration rather than MADSci version
