+++ src/madsci_common/madsci/common/mongodb_version_checker.py
@@ -1,7 +1,6 @@
 """MongoDB version checking and validation for MADSci."""
 
 import json
-import os
 import traceback
 from datetime import datetime, timezone
 from pathlib import Path
@@ -76,29 +75,31 @@ class MongoDBVersionChecker:
 
         Returns:
             SemanticVersion if a valid semantic version is found
-            0.0.0 if no version tracking exists
-            None if database/collection doesn't exist or an error occurs
+            None if database/collection doesn't exist or no version records
         """
         try:
-            # Try to access the database directly instead of checking list_database_names()
-            # MongoDB only shows databases in list_database_names() if they have collections with data
             collection_names = self.database.list_collection_names()
             if not collection_names:
-                # If database has no collections at all, it doesn't really exist
-                raise Exception("No collections found")
+                # Database has no collections
+                return None
 
             # Check if schema_versions collection exists
             if "schema_versions" not in collection_names:
-                # Database exists with collections but no schema_versions collection
-                return SemanticVersion(0, 0, 0)
+                # Database exists but no schema_versions collection
+                return None
+
+            # Check if collection has any records
+            version_record = self.database["schema_versions"].find_one(
+                {},
+                sort=[("applied_at", -1)],  # Most recent first
+            )
+
+            if not version_record:
+                # Collection exists but is empty
+                return None
 
             # Get the latest version entry
-            return SemanticVersion.parse(
-                self.database["schema_versions"].find_one(
-                    {},
-                    sort=[("applied_at", -1)],  # Most recent first
-                )["version"]
-            ) or SemanticVersion(0, 0, 0)
+            return SemanticVersion.parse(version_record["version"])
 
         except Exception:
             self.logger.error(
@@ -106,15 +107,37 @@ class MongoDBVersionChecker:
             )
             return None
 
+    def is_version_tracked(self) -> bool:
+        """
+        Check if version tracking exists in the database.
+
+        Returns True if the schema_versions collection exists AND has at least one version record.
+        Returns False if the collection doesn't exist or is empty.
+        """
+        try:
+            collection_names = self.database.list_collection_names()
+
+            if "schema_versions" not in collection_names:
+                return False
+
+            # Check if collection has any records
+            version_record = self.database["schema_versions"].find_one({})
+            return version_record is not None
+
+        except Exception:
+            return False
+
     def is_migration_needed(
         self,
-    ) -> tuple[bool, SemanticVersion, Optional[Union[SemanticVersion, str]]]:
+    ) -> tuple[bool, SemanticVersion, Optional[SemanticVersion]]:
         """
         Check if database migration is needed.
 
-        Migration is only needed for major or minor schema version mismatches,
-        not for patch versions or pre-release versions. Uses SemanticVersion
-        for semantic version comparison.
+        Migration is only needed if:
+        1. Version IS tracked in the database, AND
+        2. Schema version mismatch between expected and database
+
+        If version is NOT tracked, no migration is needed (server will start normally).
 
         Returns:
             tuple: (needs_migration, expected_schema_version, database_version)
@@ -122,115 +145,93 @@ class MongoDBVersionChecker:
         expected_schema_version = self.get_expected_schema_version()
         db_version = self.get_database_version()
 
-        if db_version is None:
-            # Database doesn't exist at all - fresh install
-            self.logger.info(
-                f"Database {self.database_name} does not exist - migration needed for initial setup"
-            )
-            return True, expected_schema_version, None
+        # If version tracking doesn't exist, no migration needed
+        # User can manually run migration if they want to start tracking
+        if not self.is_version_tracked():
+            is_docker = self._is_running_in_docker()
 
-        if db_version == SemanticVersion(0, 0, 0):
-            # Database exists but no version tracking - needs initialization
-            self.logger.info(
[…]
