### 2025-11-07 3d7fedd
fixed parameterized locations
