+++ src/madsci_workcell_manager/madsci/workcell_manager/workflow_utils.py
@@ -231,9 +231,7 @@ def create_workflow(
                 state_handler,
                 step,
                 wf,
-                data_client,
-                location_client,
-                running=False,
+                location_client=location_client
             )
         )
 
@@ -265,9 +263,8 @@ def prepare_workflow_step(
     state_handler: WorkcellStateHandler,
     step: Step,
     workflow: Workflow,
-    data_client: DataClient,
+    data_client: Optional[DataClient] = None,
     location_client: Optional[LocationClient] = None,
-    running: bool = True,
 ) -> Step:
     """Prepares a step for execution by replacing locations and validating it"""
     parameter_values = workflow.parameter_values
@@ -280,7 +277,7 @@ def prepare_workflow_step(
         state_handler=state_handler,
         feedforward_parameters=workflow.parameters.feed_forward,
     )
-    if running:
+    if data_client is not None:
         working_step = prepare_workflow_files(working_step, workflow, data_client)
     EventClient().info(validation_string)
     if not valid:
