### 2025-11-07 76ee93d
Fix pre-commit and unit tests
