### 2025-11-07 a8f7f6b
Fix test_get_log
