### 2025-11-07 7578cc8
Better handle db migrations
