### 2025-11-07 26c8505
Better handle db migrations
