### 2025-11-07 230cce9
Better handle db migrations
