+++ .github/workflows/integration_tests.yml
@@ -1,8 +1,9 @@
 name: Integration Tests
 
 on:
-  push: # * Run for every push
-    branches: [ "*" ]
+  pull_request:
+  push:
+    branches: [main]
     tags: [ '*' ]
   schedule: # Run on Tuesday's at 12:00
     - cron: '0 12 * * 2'
