+++ madsci-entrypoint.sh
@@ -40,7 +40,15 @@ mkdir -p /home/${CONTAINER_USER}/.madsci/logs
 chown $USER_ID:$GROUP_ID /home/${CONTAINER_USER} || true
 chown $USER_ID:$GROUP_ID /home/${CONTAINER_USER}/.madsci || true
 chown $USER_ID:$GROUP_ID /home/${CONTAINER_USER}/.madsci/logs || true
-chown -R $USER_ID:$GROUP_ID /home/${CONTAINER_USER}/MADSci || true
+# Prepare writable dirs used at runtime (no recursive chown)
+# Backups
+install -d -m 0775 -o "${USER_ID}" -g "${GROUP_ID}" /home/${CONTAINER_USER}/.madsci/mongodb/backups || true
+install -d -m 0775 -o "${USER_ID}" -g "${GROUP_ID}" /home/${CONTAINER_USER}/.madsci/postgres/backups || true
+
+# Alembic revisions (repo location) — make ONLY this dir writable
+ALEMBIC_VERSIONS_DIR="${ALEMBIC_VERSIONS_DIR:-/home/${CONTAINER_USER}/MADSci/src/madsci_resource_manager/madsci/resource_manager/alembic/versions}"
+install -d -m 0775 -o "${USER_ID}" -g "${GROUP_ID}" "${ALEMBIC_VERSIONS_DIR}" || true
+chown "${USER_ID}:${GROUP_ID}" "${ALEMBIC_VERSIONS_DIR}" || true
 
 
 # *Run the container command as the specified user
