+++ src/madsci_common/madsci/common/mongodb_migration_tool.py
@@ -11,120 +11,9 @@ from typing import Any, Dict, List, Optional
 
 from madsci.client.event_client import EventClient
 from madsci.common.mongodb_version_checker import MongoDBVersionChecker
-from madsci.common.types.base_types import MadsciBaseSettings, PathLike
-from pydantic import AliasChoices, AnyUrl, Field
+from madsci.common.types.mongodb_migration_types import MongoDBMigrationSettings
+from pydantic import AnyUrl
 from pymongo import MongoClient
-
-
-class MongoDBMigrationSettings(
-    MadsciBaseSettings,
-    env_file=(".env", "mongodb.env", "migration.env"),
-    toml_file=("settings.toml", "mongodb.settings.toml", "migration.settings.toml"),
-    yaml_file=("settings.yaml", "mongodb.settings.yaml", "migration.settings.yaml"),
-    json_file=("settings.json", "mongodb.settings.json", "migration.settings.json"),
-    env_prefix="MONGODB_MIGRATION_",
-):
-    """Configuration settings for MongoDB migration operations."""
-
-    mongo_db_url: AnyUrl = Field(
-        default=AnyUrl("mongodb://localhost:27017"),
-        title="MongoDB URL",
-        description="MongoDB connection URL (e.g., mongodb://localhost:27017). "
-        "Defaults to localhost MongoDB instance.",
-        validation_alias=AliasChoices(
-            "mongo_db_url", "MONGODB_URL", "MONGO_URL", "DATABASE_URL", "db_url"
-        ),
-    )
-    database: str = Field(
-        title="Database Name",
-        description="Database name to migrate (e.g., madsci_events, madsci_data)",
-    )
-    schema_file: Optional[PathLike] = Field(
-        default=None,
-        title="Schema File Path",
-        description="Path to schema.json file. If not provided, will auto-detect based on database name.",
-    )
-    target_version: Optional[str] = Field(
-        default=None,
-        title="Target Version",
-        description="Target version to migrate to (defaults to current MADSci version)",
-    )
-    backup_only: bool = Field(
-        default=False,
-        title="Backup Only",
-        description="Only create a backup, do not run migration",
-    )
-    restore_from: Optional[PathLike] = Field(
-        default=None,
-        title="Restore From",
-        description="Restore from specified backup directory instead of migrating",
-    )
-    check_version: bool = Field(
-        default=False,
-        title="Check Version Only",
-        description="Only check version compatibility, do not migrate",
-    )
-
-    def get_effective_schema_file_path(self) -> Path:
-        """Get the effective schema file path, auto-detecting if needed."""
-        if self.schema_file:
-            schema_path = Path(self.schema_file)
-            if not schema_path.exists():
-                raise FileNotFoundError(f"Schema file not found: {schema_path}")
-            return schema_path
-
-        # Auto-detect schema file based on database name
-        current_dir = Path.cwd()
-
-        # Common schema file locations based on database name
-        possible_paths = []
-
-        if self.database == "madsci_events":
-            possible_paths = [
-                current_dir / "madsci" / "event_manager" / "schema.json",
-                current_dir / "event_manager" / "schema.json",
-                current_dir / "schema" / "event_manager.json",
-            ]
-        elif self.database == "madsci_data":
-            possible_paths = [
-                current_dir / "madsci" / "data_manager" / "schema.json",
-                current_dir / "data_manager" / "schema.json",
-                current_dir / "schema" / "data_manager.json",
-            ]
-        elif self.database == "madsci_workcells":
-            possible_paths = [
-                current_dir / "madsci" / "workcell_manager" / "schema.json",
-                current_dir / "workcell_manager" / "schema.json",
-                current_dir / "schema" / "workcell_manager.json",
-            ]
-        elif self.database == "madsci_experiments":
[…]

[…]
