+++ ui/README.md
@@ -1,6 +1,8 @@
-# MADSci Dashboard
+# MADSci Dashboard (squid_dashboard)
 
 Vue 3 + Vuetify web dashboard for monitoring and controlling MADSci-powered laboratories.
+
+> **Note**: This package is internally named `squid_dashboard` but serves as the primary MADSci Dashboard UI component.
 
 ## Features
 
@@ -14,8 +16,11 @@ Vue 3 + Vuetify web dashboard for monitoring and controlling MADSci-powered labo
 ## Development Setup
 
 ### Prerequisites
-- Node.js 16+ and yarn
-- MADSci lab services running (for API integration)
+- **Node.js**: Version 16.0 or higher (LTS recommended)
+- **Yarn**: Version 1.22+ for package management (do NOT use npm)
+- **TypeScript**: Automatically installed with dependencies
+- **System Requirements**: 2GB+ RAM, modern browser with ES2020 support
+- **MADSci Services**: Lab Manager (SQUID) and manager services running for full functionality
 
 ### Quick Start
 
@@ -44,10 +49,13 @@ yarn preview    # Preview production build locally
 
 ### Technology Stack
 - **Vue 3**: Progressive JavaScript framework with Composition API
-- **Vuetify 3**: Material Design component library
+- **Vuetify 3**: Primary Material Design component library
+- **PrimeVue**: Secondary component library (for specialized components)
 - **TypeScript**: Type-safe JavaScript development
 - **Vite**: Fast build tool and dev server
-- **Pinia**: State management (via `store.ts`)
+- **Vue Reactivity**: State management using Vue's built-in reactivity system (via `store.ts`)
+
+> **UI Framework Note**: The dashboard primarily uses Vuetify for the main interface, with PrimeVue providing additional specialized components where needed. Both frameworks are configured in `vite.config.mts` with auto-import resolvers.
 
 ### Project Structure
 ```
@@ -82,21 +90,27 @@ The main dashboard uses a tabbed interface:
 ```
 
 ### State Management
-Global state is managed via Pinia store:
+Global state is managed using Vue 3's Composition API and reactive references:
 
 ```typescript
-// store.ts patterns
-interface LabState {
-  context: MadsciContext
-  workcells: Workcell[]
-  workflows: Workflow[]
-  resources: Resource[]
-  // ... other lab state
-}
+// store.ts actual implementation
+import { ref, watchEffect } from 'vue'
 
-// Reactive updates from MADSci services
-const store = useLabStore()
-store.updateWorkcellStatus(workcellId, newStatus)
+// Reactive state variables
+const workcell_state = ref<WorkcellState>()
+const active_workflows = ref<Record<string, Workflow>>({})
+const resources = ref()
+const locations = ref<Record<string, any>>({})
+const labContext = ref<any>(null)
+
+// Real-time updates with automatic refresh intervals
+watchEffect(async () => {
+  // Initialize API URLs from lab context
+  const urls = await (await fetch(main_url.value + "/context")).json()
+  setInterval(updateWorkcellState, 1000)
+  setInterval(updateWorkflows, 1000)
+  setInterval(updateResources, 1000)
+})
 ```
 
 ### Component Patterns
@@ -138,23 +152,51 @@ store.updateWorkcellStatus(workcellId, newStatus)
 
 The dashboard communicates with MADSci services via REST APIs:
 
-### Service Connections
+### Service Discovery
 ```typescript
-// Connecting to MADSci services
-const labContext = await fetch('http://localhost:8000/context')
-const workcells = await fetch(`${context.workcell_server_url}/workcells`)
-const resources = await fetch(`${context.resource_server_url}/resources`)
+// Fetch lab context to discover all service URLs
[…]
