### 2025-11-08 6bbd2ac
Example lab audit
