+++ example_lab/example_modules/liquidhandler.py
@@ -1,4 +1,24 @@
-"""A fake liquid handler module for testing."""
+"""
+Example Virtual Liquid Handler Node Module for MADSci
+
+This module demonstrates how to implement a virtual laboratory instrument node in MADSci.
+It simulates a liquid handling robot with deck positions and pipettes, showing:
+
+1. Node configuration with custom settings
+2. Resource template creation and management
+3. Action method implementation with proper decorators
+4. Resource tracking and location management
+5. Event logging and error handling
+
+Key Concepts Demonstrated:
+- RestNode inheritance for HTTP API nodes
+- @action decorator for exposing callable methods
+- Resource templates for reusable lab components
+- ULID-based resource identification
+- Location-based resource transfers
+
+This serves as a template for creating real instrument drivers.
+"""
 
 import time
 from pathlib import Path
@@ -14,16 +34,39 @@ from madsci.node_module.rest_node_module import RestNode
 
 
 class LiquidHandlerConfig(RestNodeConfig):
-    """Configuration for the liquid handler node module."""
+    """
+    Configuration for the liquid handler node module.
+
+    Extends RestNodeConfig to add liquid handler-specific settings.
+    These settings can be overridden via environment variables with
+    the node name prefix (e.g., LIQUIDHANDLER_1_DEVICE_NUMBER=1).
+
+    Attributes:
+        device_number: Identifier for the physical/virtual liquid handler device
+        wait_time: Simulation delay for action execution (seconds)
+    """
 
     device_number: int = 0
-    """The device number of the liquid handler."""
+    """The device number of the liquid handler (for multi-device setups)."""
+
     wait_time: float = 2.0
-    """Time to wait while running an action, in seconds."""
+    """Time to wait while running an action, in seconds (simulates real hardware timing)."""
 
 
 class LiquidHandlerInterface:
-    """A fake liquid handler interface for testing."""
+    """
+    Simulated liquid handler hardware interface.
+
+    This class represents the low-level hardware communication layer that
+    would typically interface with real liquid handler firmware/drivers.
+    In a real implementation, this would contain actual device communication
+    protocols (e.g., serial, TCP/IP, vendor SDK calls).
+
+    Attributes:
+        status_code: Current device status (0=OK, >0=error conditions)
+        device_number: Hardware device identifier
+        logger: Event client for logging device operations
+    """
 
     status_code: int = 0
     device_number: int = 0
@@ -31,30 +74,92 @@ class LiquidHandlerInterface:
     def __init__(
         self, device_number: int = 0, logger: Optional[EventClient] = None
     ) -> "LiquidHandlerInterface":
-        """Initialize the liquid handler."""
+        """
+        Initialize the liquid handler hardware interface.
+
+        Args:
+            device_number: Unique identifier for this liquid handler device
+            logger: Event client for logging (creates default if None)
+        """
         self.logger = logger or EventClient()
         self.device_number = device_number
 
     def run_command(self, command: str) -> None:
-        """Run a command on the liquid handler."""
+        """
+        Execute a command on the liquid handler hardware.
+
+        In a real implementation, this would send commands to the actual
+        liquid handler via the appropriate communication protocol.
+
+        Args:
+            command: Hardware command string to execute
+        """
         self.logger.log(
             f"Running command {command} on device number {self.device_number}."
         )
-        time.sleep(2)  # Simulate command execution
+        time.sleep(2)  # Simulate hardware command execution time
 
 
 class LiquidHandlerNode(RestNode):
-    """A fake liquid handler node module for testing."""
+    """
+    Virtual Liquid Handler Node for MADSci Laboratory Automation.
 
+    This node simulates a multi-deck liquid handling robot with the following capabilities:
+    - Two deck positions (deck1, deck2) for placing labware
+    - Multi-channel pipette system for liquid transfers
[…]
