+++ example_lab/NODE_DEVELOPMENT_GUIDE.md
@@ -1,0 +1,226 @@
+# Node Development Quick Reference
+
+> For comprehensive, interactive node development tutorials, see **[node_notebook.ipynb](notebooks/node_notebook.ipynb)**
+
+This guide provides quick reference information for node development patterns not covered in the interactive tutorial.
+
+## Quick Start Checklist
+
+1. **📓 Start with the interactive tutorial**: Run `jupyter lab notebooks/node_notebook.ipynb` for hands-on learning
+2. **Choose a template**: Copy from `example_modules/` that matches your instrument type
+3. **Define configuration**: Create your config class inheriting from `RestNodeConfig`
+4. **Implement hardware interface**: Separate device communication logic
+5. **Add action methods**: Use `@action` decorator on your methods
+6. **Create resource templates**: Define lab materials your instrument handles
+7. **Test thoroughly**: Use both unit tests and integration tests
+8. **Deploy**: Add to `compose.yaml` and configure YAML files
+
+## Production Deployment Patterns
+
+### Docker Configuration Template
+```yaml
+# Add to compose.yaml
+my_instrument:
+  <<: *madsci-service
+  container_name: my_instrument_1
+  environment:
+    - NODE_DEFINITION=node_definitions/my_instrument.node.yaml
+    - NODE_URL=http://localhost:2010
+  command: python example_modules/my_instrument.py
+```
+
+### Node Definition Files
+**`my_instrument.node.yaml`**:
+```yaml
+node_name: my_instrument_1
+node_id: 01234567890123456789012345  # Use new_ulid_str()
+node_type: measurement_device
+module_path: example_modules/my_instrument.py
+node_class: MyInstrumentNode
+```
+
+**`my_instrument.info.yaml`**:
+```yaml
+name: "My Custom Instrument"
+description: "Custom laboratory measurement device"
+manufacturer: "Lab Equipment Inc."
+model: "Model-X1000"
+version: "1.0.0"
+capabilities: [optical_density_measurement, sample_transfer]
+resources: [sample_holder_slot, reagent_reservoir_pool]
+```
+
+## Testing Patterns
+
+### Unit Testing Template
+```python
+import pytest
+from unittest.mock import Mock, patch
+from my_instrument import MyInstrumentNode
+
+class TestMyInstrumentNode:
+    @pytest.fixture
+    def node(self):
+        node = MyInstrumentNode()
+        node.logger = Mock()
+        node.resource_client = Mock()
+        return node
+
+    def test_startup_handler(self, node):
+        with patch.object(MyInstrumentInterface, 'connect'):
+            node.startup_handler()
+            assert node.hardware is not None
+```
+
+### Integration Testing
+```python
+def test_node_integration():
+    base_url = "http://localhost:2010"
+    response = requests.get(f"{base_url}/health")
+    assert response.status_code == 200
+```
+
+## Advanced Patterns
+
+### Concurrent Operations
+```python
+from concurrent.futures import ThreadPoolExecutor
+
+class MyInstrumentNode(RestNode):
+    def __init__(self):
+        super().__init__()
+        self.executor = ThreadPoolExecutor(max_workers=4)
+
+    @action
+    async def parallel_measurements(self, samples: list[str]) -> list[dict]:
+        futures = [self.executor.submit(self._measure_single, s) for s in samples]
+        return [f.result() for f in futures]
+```
+
+### State Machine Implementation
+```python
+from enum import Enum
+
+class InstrumentState(Enum):
+    IDLE = "idle"
+    MEASURING = "measuring"
+    ERROR = "error"
+
+class MyInstrumentNode(RestNode):
+    def _transition_state(self, new_state: InstrumentState):
+        # Validation logic here
+        self.state = new_state
+```
+
+## Security Best Practices
+
[…]

[…]
