### 2025-11-08 ad48b6e
Fix node double logging by using self.logger instead of creating new EventClient instances
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
