### 2025-11-10 4c4564e
Updated template endpoints
