### 2025-11-10 e0ddc23
Updated template endpoints
