### 2025-11-10 d4e7a62
Updated query endpoint
