### 2025-11-10 c95d5e2
Updated query endpoint
