### 2025-11-11 64baa36
Bug fix for action file return; bump to v0.5.2
