+++ src/madsci_client/madsci/client/node/rest_node_client.py
@@ -16,6 +16,7 @@ from madsci.common.types.action_types import (
     ActionRequest,
     ActionResult,
     ActionStatus,
+    RestActionResult,
 )
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.event_types import Event
@@ -196,7 +197,9 @@ class RestNodeClient(AbstractNodeClient):
             timeout=60,
         )
         rest_response.raise_for_status()
-        return ActionResult.model_validate(rest_response.json())
+        return self._convert_rest_result_to_action_result(
+            rest_response.json(), action_name, action_id
+        )
 
     def get_action_status_by_name(
         self, action_name: str, action_id: str
@@ -209,53 +212,89 @@ class RestNodeClient(AbstractNodeClient):
         rest_response.raise_for_status()
         return ActionStatus(rest_response.json())
 
-    def _get_result_files(
-        self, action_name: str, action_id: str, result: ActionResult
-    ) -> None:
-        """Return the files associated with an action result. REST-implementation specific."""
-        if result.files is None:
-            return
+    def _find_file_by_key(self, temp_dir: Path, file_key: str) -> Path | None:
+        """Find a file in the temp directory by its key."""
+        # Look for file with matching label (with any extension)
+        extracted_files = list(temp_dir.glob(f"{file_key}*"))
+        if extracted_files:
+            return extracted_files[0]
 
-        # Always download as ZIP and extract files
-        zip_path = self._get_action_files_zip(action_name, action_id)
+        # Fallback: try exact name
+        exact_file = temp_dir / file_key
+        if exact_file.exists():
+            return exact_file
 
-        if isinstance(result.files, ActionFiles):
-            # Multiple files case - extract from ZIP
-            files_dict = result.files.model_dump()
-            downloaded_files = {}
+        # Last fallback: use any available file
+        all_files = list(temp_dir.iterdir())
+        return all_files[0] if all_files else None
+
+    def _convert_rest_result_to_action_result(
+        self, rest_result_data: dict, action_name: str, action_id: str
+    ) -> ActionResult:
+        """Convert a REST API result (RestActionResult format) to ActionResult format.
+
+        The REST API returns files as a list of strings (file keys), but ActionResult
+        expects files to be Path objects or ActionFiles. This method handles the conversion.
+        """
+        # First validate as RestActionResult to ensure proper format
+        rest_result = RestActionResult.model_validate(rest_result_data)
+
+        # Convert to ActionResult format
+        result_data = rest_result.model_dump()
+
+        # Handle files field - fetch actual files if file keys are present
+        files_list = result_data.get("files")
+        if files_list:
+            # Fetch the actual files using the file keys
+            action_files = self._fetch_files_from_keys(
+                action_name, action_id, files_list
+            )
+            result_data["files"] = action_files
+        else:
+            result_data["files"] = None
+
+        return ActionResult.model_validate(result_data)
+
+    def _fetch_files_from_keys(
+        self, action_name: str, action_id: str, file_keys: list[str]
+    ) -> Path | ActionFiles | None:
+        """Fetch actual files from the server using the provided file keys."""
+        try:
+            # Download the ZIP file containing all files
+            zip_path = self._get_action_files_zip(action_name, action_id)
 
             # Extract files from ZIP
             with zipfile.ZipFile(zip_path, "r") as zip_file:
                 temp_dir = Path(tempfile.mkdtemp())
                 zip_file.extractall(temp_dir)
 
-                # Map extracted files back to their labels
-                for file_key in files_dict:
-                    # Look for file with matching label (with any extension)
-                    extracted_files = list(temp_dir.glob(f"{file_key}*"))
-                    if extracted_files:
-                        downloaded_files[file_key] = extracted_files[0]
-                    else:
[…]
