### 2025-11-11 ea83915
Fix 3.9 backcompat issue
