### 2025-11-13 b462920
Added pydantic validator for schema
