+++ src/madsci_common/madsci/common/types/mongodb_migration_types.py
@@ -1,10 +1,11 @@
 """ "MongoDB migration configuration types."""
 
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any, Dict, List, Optional
 
 from madsci.common.types.base_types import MadsciBaseSettings, PathLike
-from pydantic import AliasChoices, AnyUrl, Field, field_validator
+from pydantic import AliasChoices, AnyUrl, BaseModel, Field, field_validator
+from pydantic_extra_types.semantic_version import SemanticVersion
 
 
 class MongoDBMigrationSettings(
@@ -74,3 +75,363 @@ class MongoDBMigrationSettings(
         if not p.exists():
             raise FileNotFoundError(f"Schema file not found: {p}")
         return p
+
+
+class IndexKey(BaseModel):
+    """Represents a single key in a MongoDB index."""
+
+    field: str = Field(description="Field name to index")
+    direction: int = Field(
+        description="Sort direction: 1 for ascending, -1 for descending"
+    )
+
+    @field_validator("direction")
+    @classmethod
+    def validate_direction(cls, v: int) -> int:
+        """Validate index direction is 1 or -1."""
+        if v not in (1, -1):
+            raise ValueError("Index direction must be 1 (ascending) or -1 (descending)")
+        return v
+
+    def to_tuple(self) -> tuple[str, int]:
+        """Convert to tuple format for MongoDB operations."""
+        return (self.field, self.direction)
+
+    def to_list(self) -> list:
+        """Convert to list format for schema.json."""
+        return [self.field, self.direction]
+
+    @classmethod
+    def from_list(cls, key_list: List[Any]) -> "IndexKey":
+        """Create IndexKey from list format [field, direction]."""
+        if len(key_list) != 2:
+            raise ValueError(
+                f"Index key must have exactly 2 elements, got {len(key_list)}"
+            )
+        return cls(field=key_list[0], direction=key_list[1])
+
+
+class IndexDefinition(BaseModel):
+    """MongoDB index definition."""
+
+    keys: List[IndexKey] = Field(description="List of index keys with directions")
+    name: str = Field(description="Name of the index")
+    unique: bool = Field(
+        default=False, description="Whether the index enforces uniqueness"
+    )
+    background: bool = Field(
+        default=True, description="Whether to build index in background"
+    )
+    description: Optional[str] = Field(
+        default=None, description="Human-readable description of the index"
+    )
+
+    @field_validator("keys", mode="before")
+    @classmethod
+    def validate_keys(cls, v: Any) -> List[IndexKey]:
+        """Validate and convert keys from various formats."""
+        if not v:
+            raise ValueError("Index must have at least one key")
+
+        if isinstance(v, list):
+            result = []
+            for item in v:
+                if isinstance(item, IndexKey):
+                    result.append(item)
+                elif isinstance(item, (list, tuple)) and len(item) == 2:
+                    result.append(IndexKey.from_list(item))
+                elif isinstance(item, dict):
+                    result.append(IndexKey(**item))
+                else:
+                    raise ValueError(f"Invalid key format: {item}")
+            return result
+
+        raise ValueError(f"Keys must be a list, got {type(v)}")
+
+    def to_mongo_format(self) -> Dict[str, Any]:
+        """Convert to MongoDB index creation format."""
+        return {
+            "name": self.name,
+            "unique": self.unique,
+            "background": self.background,
+        }
+
+    def to_schema_dict(self) -> Dict[str, Any]:
+        """Convert to schema.json format."""
+        result = {
+            "keys": [key.to_list() for key in self.keys],
+            "name": self.name,
+            "background": self.background,
+        }
+        if self.unique:
+            result["unique"] = True
+        if self.description:
[…]
