### 2025-11-13 5a0ed69
Enable system logs
