+++ src/madsci_resource_manager/madsci/resource_manager/alembic/env.py
@@ -63,7 +63,7 @@ config = context.config
 # Interpret the config file for Python logging.
 # This line sets up loggers basically.
 if config.config_file_name is not None:
-    fileConfig(config.config_file_name)
+    fileConfig(config.config_file_name, disable_existing_loggers=False)
 
 # Set target metadata from your models
 target_metadata = ResourceTable.metadata
