+++ src/madsci_data_manager/madsci/data_manager/data_server.py
@@ -25,7 +25,6 @@ from madsci.common.types.datapoint_types import (
     DataPoint,
     ObjectStorageSettings,
 )
-from madsci.common.types.mongodb_migration_types import MongoDBMigrationSettings
 from minio import Minio
 from pymongo import MongoClient
 
@@ -67,12 +66,10 @@ class DataManager(AbstractManagerBase[DataManagerSettings, DataManagerDefinition
         self.logger.info("Validating MongoDB schema version...")
 
         schema_file_path = Path(__file__).parent / "schema.json"
-        mig_cfg = MongoDBMigrationSettings(database=self.settings.database_name)
         version_checker = MongoDBVersionChecker(
             db_url=str(self.settings.mongo_db_url),
             database_name=self.settings.database_name,
             schema_file_path=str(schema_file_path),
-            backup_dir=str(mig_cfg.backup_dir),
             logger=self.logger,
         )
 
