### 2025-11-13 a85000f
Use pydentic args
