+++ src/madsci_common/madsci/common/mongodb_migration_tool.py
@@ -1,7 +1,5 @@
 """MongoDB migration tool for MADSci databases with backup, schema management, and CLI."""
 
-import argparse
-import json
 import subprocess
 import sys
 import traceback
@@ -11,7 +9,10 @@ from typing import Any, Dict, List, Optional
 
 from madsci.client.event_client import EventClient
 from madsci.common.mongodb_version_checker import MongoDBVersionChecker
-from madsci.common.types.mongodb_migration_types import MongoDBMigrationSettings
+from madsci.common.types.mongodb_migration_types import (
+    MongoDBMigrationSettings,
+    MongoDBSchema,
+)
 from pydantic import AnyUrl
 from pymongo import MongoClient
 
@@ -72,7 +73,7 @@ class MongoDBMigrator:
             if hasattr(self, "logger") and self.logger:
                 self.logger.debug("MongoDB migrator client disposed")
 
-    def load_expected_schema(self) -> Dict[str, Any]:
+    def load_expected_schema(self) -> MongoDBSchema:
         """Load the expected schema from the schema.json file."""
         try:
             if not self.schema_file_path.exists():
@@ -80,15 +81,29 @@ class MongoDBMigrator:
                     f"Schema file not found: {self.schema_file_path}"
                 )
 
-            with open(self.schema_file_path) as f:  # noqa
-                schema = json.load(f)
-
+            schema = MongoDBSchema.from_file(str(self.schema_file_path))
             self.logger.info(f"Loaded schema from {self.schema_file_path}")
             return schema
 
         except Exception as e:
             self.logger.error(f"Error loading schema file: {e}")
             raise RuntimeError(f"Cannot load schema file: {e}") from e
+
+    def get_current_database_schema(self) -> MongoDBSchema:
+        """Get the current database schema using Pydantic models."""
+        try:
+            current_version = self.version_checker.get_database_version()
+            version_str = str(current_version) if current_version else "0.0.0"
+
+            return MongoDBSchema.from_mongodb_database(
+                database_name=self.database_name,
+                mongo_client=self.client,
+                schema_version=version_str,
+            )
+
+        except Exception as e:
+            self.logger.error(f"Error getting current database schema: {e}")
+            raise
 
     def create_backup(self) -> Path:
         """Create a backup of the current database using mongodump."""
@@ -183,45 +198,6 @@ class MongoDBMigrator:
             self.logger.error(f"Restore failed: {e.stderr}")
             raise RuntimeError(f"Database restore failed: {e}") from e
 
-    def get_current_database_schema(self) -> Dict[str, Any]:
-        """Get the current database schema (collections and indexes)."""
-        current_schema = {"database": self.database_name, "collections": {}}
-
-        try:
-            # Get all collections except system collections
-            collection_names = [
-                name
-                for name in self.database.list_collection_names()
-                if not name.startswith("system.")
-            ]
-
-            for collection_name in collection_names:
-                collection = self.database[collection_name]
-                indexes = list(collection.list_indexes())
-
-                # Filter out the default _id index for comparison
-                filtered_indexes = []
-                for index in indexes:
-                    if index["name"] != "_id_":
-                        filtered_indexes.append(
-                            {
-                                "keys": list(index["key"].items()),
-                                "name": index["name"],
-                                "unique": index.get("unique", False),
-                                "background": index.get("background", False),
-                            }
-                        )
-
-                current_schema["collections"][collection_name] = {
-                    "indexes": filtered_indexes
-                }
-
-            return current_schema
-
-        except Exception as e:
-            self.logger.error(f"Error getting current database schema: {e}")
-            raise
-
     def apply_schema_migrations(self) -> None:
[…]
