### 2025-11-13 b2f4d17
Updated arguments when commands are generated
