### 2025-11-13 7e06dba
Updated arguments when commands are generated
