### 2025-11-13 e89a44e
Added validate schema field to use with paydentic args
