### 2025-11-13 3ca2d02
Fixed bug with validate_schema arg
