### 2025-11-15 #9416 @editorialbot
[REVIEW]: MADSci: A modular Python-based framework to enable autonomous science
**Submitting author:** @RyanTheRobothead  (<a href="http://orcid.org/0000-0002-3000-2811">Ryan D. Lewis</a>)
**Repository:** https://github.com/AD-SDL/MADSci
**Branch with paper.md** (empty if default branch): madsci_paper
**Version:** v0.4.7
