### 2025-11-15 5a835b6
Standardize requests client configuration with retry and timeout support
This commit addresses issue #171 by implementing a standardized configuration
system for all MADSci HTTP clients with support for configurable retry
strategies and timeout values.

## Changes

### New Infrastructure

1. **Created MadsciClientConfig base class** (client_types.py)
   - Centralized Pydantic settings for HTTP client configuration
   - Configurable retry strategy (enabled by default with 3 retries)
   - Configurable backoff factor and status codes for retry
   - Three timeout tiers: default (10s), data operati
[…]
