+++ src/madsci_common/madsci/common/types/client_types.py
@@ -1,0 +1,202 @@
+"""
+Client configuration types for MADSci.
+
+This module provides Pydantic settings models for configuring HTTP clients
+across the MADSci ecosystem, including retry strategies, timeout values,
+and backoff algorithms.
+"""
+
+from typing import Optional
+
+from madsci.common.types.base_types import MadsciBaseSettings
+from pydantic import Field
+from pydantic_settings import SettingsConfigDict
+
+
+class MadsciClientConfig(MadsciBaseSettings):
+    """
+    Base configuration for MADSci HTTP clients.
+
+    This class provides standardized configuration for requests library usage,
+    including retry strategies, timeout values, and backoff algorithms.
+    All MADSci clients should use this configuration to ensure consistency.
+
+    Attributes
+    ----------
+    retry_enabled : bool
+        Whether to enable automatic retries for failed requests. Default: True.
+    retry_total : int
+        Total number of retry attempts. Default: 3.
+    retry_backoff_factor : float
+        Backoff factor between retries (in seconds). The actual delay is calculated
+        as {backoff factor} * (2 ** ({retry number} - 1)). Default: 0.3.
+    retry_status_forcelist : list[int]
+        HTTP status codes that should trigger a retry. Default: [429, 500, 502, 503, 504].
+    retry_allowed_methods : Optional[list[str]]
+        HTTP methods that are allowed to be retried. If None, uses urllib3 defaults
+        (HEAD, GET, PUT, DELETE, OPTIONS, TRACE). Default: None.
+    timeout_default : float
+        Default timeout in seconds for standard requests. Default: 10.
+    timeout_data_operations : float
+        Timeout in seconds for data-heavy operations (e.g., uploads, downloads). Default: 60.
+    timeout_long_operations : float
+        Timeout in seconds for long-running operations (e.g., workflow queries, utilization). Default: 100.
+    pool_connections : int
+        Number of connection pool entries for the session. Default: 10.
+    pool_maxsize : int
+        Maximum size of the connection pool. Default: 10.
+    """
+
+    model_config = SettingsConfigDict(
+        env_prefix="MADSCI_CLIENT_",
+        env_file=None,
+        env_file_encoding="utf-8",
+        validate_assignment=True,
+        validate_default=True,
+        env_nested_delimiter="__",
+        case_sensitive=False,
+        extra="ignore",
+    )
+
+    # Retry configuration
+    retry_enabled: bool = Field(
+        default=True,
+        description="Whether to enable automatic retries for failed requests",
+    )
+    retry_total: int = Field(
+        default=3,
+        ge=0,
+        description="Total number of retry attempts",
+    )
+    retry_backoff_factor: float = Field(
+        default=0.3,
+        ge=0.0,
+        description="Backoff factor between retries in seconds",
+    )
+    retry_status_forcelist: list[int] = Field(
+        default=[429, 500, 502, 503, 504],
+        description="HTTP status codes that should trigger a retry",
+    )
+    retry_allowed_methods: Optional[list[str]] = Field(
+        default=None,
+        description="HTTP methods allowed to be retried (None uses urllib3 defaults)",
+    )
+
+    # Timeout configuration
+    timeout_default: float = Field(
+        default=10.0,
+        gt=0.0,
+        description="Default timeout in seconds for standard requests",
+    )
+    timeout_data_operations: float = Field(
+        default=60.0,
+        gt=0.0,
+        description="Timeout in seconds for data-heavy operations",
+    )
+    timeout_long_operations: float = Field(
+        default=100.0,
+        gt=0.0,
[…]

[…]
