### 2025-11-16 b3a4e3c
Fix shutdown admin action missing BackgroundTasks parameter
Fixes #174

The dashboard's shutdown admin action was throwing a TypeError because
the RestNode.shutdown() method requires a BackgroundTasks parameter,
but the admin command route handler wasn't providing it.

Changes:
- Modified RestNode.run_admin_command() to accept optional BackgroundTasks parameter
- Added special handling for SHUTDOWN command to pass BackgroundTasks to shutdown method
- Added error handling for cases where BackgroundTasks is unexpectedly None
- FastAPI will automatically inject BackgroundTasks dependency when pre
[…]
