### 2025-11-16 4da3ff9
Add REST server resilience features for issue #149
This commit implements resilience improvements for MADSci REST servers
to handle request storms, high demand, and security scanning.

Changes include:
- Rate limiting middleware with configurable requests/time window
- Request tracking middleware for performance monitoring
- Configurable uvicorn resource constraints (workers, concurrency, max requests)
- Settings in ManagerSettings for all resilience features
- Comprehensive stress tests and middleware tests
- Added 'slow' pytest marker for extended stress tests

All managers inheriting from A
[…]
