+++ src/madsci_common/tests/test_stress.py
@@ -1,0 +1,292 @@
+"""
+Stress tests for MADSci REST servers.
+
+These tests simulate high-load scenarios to ensure servers remain
+resilient under stress conditions like request storms and high concurrency.
+"""
+
+import asyncio
+import time
+from concurrent.futures import ThreadPoolExecutor, as_completed
+from typing import ClassVar
+
+import pytest
+from madsci.common.manager_base import AbstractManagerBase
+from madsci.common.types.manager_types import (
+    ManagerDefinition,
+    ManagerSettings,
+    ManagerType,
+)
+from pydantic import Field
+from starlette.testclient import TestClient
+
+
+class TestManagerSettings(ManagerSettings):
+    """Test settings for the manager."""
+
+    model_config: ClassVar[dict] = {"env_prefix": "TEST_"}
+
+
+class TestManagerDefinition(ManagerDefinition):
+    """Test definition for the manager."""
+
+    manager_type: ManagerType = Field(default=ManagerType.EVENT_MANAGER)
+
+
+class TestManager(AbstractManagerBase[TestManagerSettings, TestManagerDefinition]):
+    """Test manager implementation."""
+
+    SETTINGS_CLASS = TestManagerSettings
+    DEFINITION_CLASS = TestManagerDefinition
+
+
+@pytest.fixture
+def stress_test_manager() -> TestManager:
+    """Create a manager configured for stress testing."""
+    settings = TestManagerSettings(
+        rate_limit_enabled=True,
+        rate_limit_requests=100,
+        rate_limit_window=10,
+    )
+    definition = TestManagerDefinition(name="Stress Test Manager")
+    return TestManager(settings=settings, definition=definition)
+
+
+@pytest.fixture
+def stress_test_client(stress_test_manager: TestManager) -> TestClient:
+    """Create a test client for stress testing."""
+    app = stress_test_manager.create_server()
+    return TestClient(app)
+
+
+def make_request(client: TestClient, endpoint: str = "/health") -> dict:
+    """
+    Make a single request and return status and timing info.
+
+    Args:
+        client: The test client
+        endpoint: The endpoint to request
+
+    Returns:
+        Dictionary with status_code, success, and response_time
+    """
+    start_time = time.time()
+    try:
+        response = client.get(endpoint)
+        response_time = time.time() - start_time
+        return {
+            "status_code": response.status_code,
+            "success": response.status_code == 200,
+            "response_time": response_time,
+            "rate_limited": response.status_code == 429,
+        }
+    except Exception as e:
+        response_time = time.time() - start_time
+        return {
+            "status_code": 0,
+            "success": False,
+            "response_time": response_time,
+            "error": str(e),
+            "rate_limited": False,
+        }
+
+
+def test_concurrent_requests(stress_test_client: TestClient) -> None:
+    """Test server handles concurrent requests without crashing."""
+    num_requests = 50
+    results = []
+
+    # Use ThreadPoolExecutor for concurrent requests
+    with ThreadPoolExecutor(max_workers=10) as executor:
+        futures = [
+            executor.submit(make_request, stress_test_client)
+            for _ in range(num_requests)
+        ]
+
+        for future in as_completed(futures):
+            results.append(future.result())
+
+    # Check that we got responses for all requests
+    assert len(results) == num_requests
+
+    # Count successes and rate limited responses
+    successes = sum(1 for r in results if r["success"])
+    rate_limited = sum(1 for r in results if r["rate_limited"])
+
+    # At least some requests should succeed
+    assert successes > 0, "No requests succeeded"
+
+    # Some may be rate limited, but server should handle it gracefully
+    if rate_limited > 0:
[…]
