+++ src/madsci_common/madsci/common/types/manager_types.py
@@ -54,6 +54,45 @@ class ManagerSettings(MadsciBaseSettings):
         default=Path("manager.yaml"),
     )
 
+    # Rate limiting settings
+    rate_limit_enabled: bool = Field(
+        title="Rate Limiting Enabled",
+        description="Enable rate limiting for API endpoints.",
+        default=True,
+    )
+    rate_limit_requests: int = Field(
+        title="Rate Limit Requests",
+        description="Maximum number of requests allowed per time window.",
+        default=100,
+        ge=1,
+    )
+    rate_limit_window: int = Field(
+        title="Rate Limit Window",
+        description="Time window for rate limiting in seconds.",
+        default=60,
+        ge=1,
+    )
+
+    # Server resource constraints
+    uvicorn_workers: Optional[int] = Field(
+        title="Uvicorn Workers",
+        description="Number of uvicorn worker processes. If None, uses uvicorn default (1).",
+        default=None,
+        ge=1,
+    )
+    uvicorn_limit_concurrency: Optional[int] = Field(
+        title="Uvicorn Limit Concurrency",
+        description="Maximum number of concurrent connections. If None, no limit is enforced.",
+        default=None,
+        ge=1,
+    )
+    uvicorn_limit_max_requests: Optional[int] = Field(
+        title="Uvicorn Limit Max Requests",
+        description="Maximum number of requests a worker will process before restarting. Helps prevent memory leaks.",
+        default=None,
+        ge=1,
+    )
+
 
 class ManagerHealth(MadsciBaseModel):
     """Base health status for MADSci Manager services.

[…]
