### 2025-11-16 f283c60
Add comprehensive automated testing infrastructure for dashboard
Implements automated testing framework to address issue #145 by adding
Vitest for unit/component testing and Playwright for E2E testing.

Testing Infrastructure:
- Vitest: Fast unit and component testing with native Vite integration
- Vue Test Utils: Official Vue component testing utilities
- Testing Library: User-centric testing patterns
- Playwright: Cross-browser end-to-end testing
- happy-dom: Lightweight DOM implementation for tests

New Files:
- vitest.config.ts: Vitest configuration with coverage settings
- playwright.conf
[…]
