+++ ui/vitest.config.ts
@@ -1,0 +1,27 @@
+import { fileURLToPath } from 'node:url'
+import { configDefaults, defineConfig, mergeConfig } from 'vitest/config'
+import viteConfig from './vite.config.mts'
+
+export default mergeConfig(
+  viteConfig,
+  defineConfig({
+    test: {
+      environment: 'happy-dom',
+      exclude: [...configDefaults.exclude, 'tests/e2e/*'],
+      root: fileURLToPath(new URL('./', import.meta.url)),
+      globals: true,
+      setupFiles: ['./tests/setup.ts'],
+      coverage: {
+        provider: 'v8',
+        reporter: ['text', 'json', 'html', 'lcov'],
+        exclude: [
+          'node_modules/',
+          'tests/',
+          '*.config.*',
+          '**/*.d.ts',
+          'dist/',
+        ],
+      },
+    },
+  })
+)
