### 2025-11-17 7e29206
Fix rate limiting middleware and stress test
- Replace HTTPException with JSONResponse in rate limiting middleware
  to ensure proper handling in TestClient and production environments
- Fix test_burst_traffic to wait for rate limit window to expire
  before verifying server health
- All tests now pass successfully
