### 2025-11-17 1eb89e3
Update ExperimentClient tests to use new session-based mocking
- Replace direct requests.get/post mocking with create_http_session mocking
- Update all test methods to mock the session object returned by create_http_session
- Change timeout assertions from integer 10 to float 10.0
- All 22 tests passing successfully
