+++ src/madsci_client/tests/test_experiment_client.py
@@ -16,6 +16,13 @@ from ulid import ULID
 
 
 @pytest.fixture
+def mock_session():
+    """Create a mock session object."""
+    session = Mock()
+    return session
+
+
+@pytest.fixture
 def mock_response():
     """Create a mock response object."""
     response = Mock()
@@ -76,48 +83,57 @@ class TestExperimentClientInit:
 class TestExperimentClientGetExperiment:
     """Test ExperimentClient get_experiment method."""
 
-    @patch("madsci.client.experiment_client.requests.get")
-    def test_get_experiment_success(self, mock_get, experiment):
+    @patch("madsci.client.experiment_client.create_http_session")
+    def test_get_experiment_success(self, mock_create_session, experiment):
         """Test successful get_experiment call."""
         mock_response = Mock()
         mock_response.ok = True
         mock_response.json.return_value = experiment.model_dump()
-        mock_get.return_value = mock_response
+
+        mock_session = Mock()
+        mock_session.get.return_value = mock_response
+        mock_create_session.return_value = mock_session
 
         client = ExperimentClient(experiment_server_url="http://localhost:8002")
         result = client.get_experiment(experiment.experiment_id)
 
-        mock_get.assert_called_once_with(
-            f"http://localhost:8002/experiment/{experiment.experiment_id}", timeout=10
+        mock_session.get.assert_called_once_with(
+            f"http://localhost:8002/experiment/{experiment.experiment_id}", timeout=10.0
         )
         assert isinstance(result, Experiment)
         assert result.run_name == experiment.run_name
         assert result.status == experiment.status
 
-    @patch("madsci.client.experiment_client.requests.get")
-    def test_get_experiment_with_ulid(self, mock_get, experiment):
+    @patch("madsci.client.experiment_client.create_http_session")
+    def test_get_experiment_with_ulid(self, mock_create_session, experiment):
         """Test get_experiment with ULID object."""
         mock_response = Mock()
         mock_response.ok = True
         mock_response.json.return_value = experiment.model_dump()
-        mock_get.return_value = mock_response
+
+        mock_session = Mock()
+        mock_session.get.return_value = mock_response
+        mock_create_session.return_value = mock_session
 
         client = ExperimentClient(experiment_server_url="http://localhost:8002")
         ulid_obj = ULID.from_str(experiment.experiment_id)
         result = client.get_experiment(ulid_obj)
 
-        mock_get.assert_called_once_with(
-            f"http://localhost:8002/experiment/{ulid_obj}", timeout=10
+        mock_session.get.assert_called_once_with(
+            f"http://localhost:8002/experiment/{ulid_obj}", timeout=10.0
         )
         assert isinstance(result, Experiment)
 
-    @patch("madsci.client.experiment_client.requests.get")
-    def test_get_experiment_http_error(self, mock_get):
+    @patch("madsci.client.experiment_client.create_http_session")
+    def test_get_experiment_http_error(self, mock_create_session):
         """Test get_experiment with HTTP error."""
         mock_response = Mock()
         mock_response.ok = False
         mock_response.raise_for_status.side_effect = requests.HTTPError("404 Not Found")
-        mock_get.return_value = mock_response
+
+        mock_session = Mock()
+        mock_session.get.return_value = mock_response
+        mock_create_session.return_value = mock_session
 
         client = ExperimentClient(experiment_server_url="http://localhost:8002")
 
@@ -128,55 +144,64 @@ class TestExperimentClientGetExperiment:
 class TestExperimentClientGetExperiments:
     """Test ExperimentClient get_experiments method."""
 
-    @patch("madsci.client.experiment_client.requests.get")
-    def test_get_experiments_success(self, mock_get, experiment):
[…]
