### 2025-11-17 1e32904
Update EventClient tests to use new session-based mocking
- Replace direct requests.get/post mocking with create_http_session mocking
- Update all test methods to mock the session object
- Use appropriate timeout values (10.0s, 30.0s, 60.0s) for different operations
- All 36 EventClient tests passing successfully
