+++ src/madsci_client/tests/test_event_client.py
@@ -108,9 +108,9 @@ class TestEventClientInit:
 class TestEventClientLogging:
     """Test EventClient logging methods."""
 
-    @patch("madsci.client.event_client.requests.post")
+    @patch("madsci.client.event_client.create_http_session")
     def test_log_event_object(
-        self, mock_post, config_with_server, temp_log_dir, sample_event
+        self, mock_create_session, config_with_server, temp_log_dir, sample_event
     ):
         """Test logging an Event object."""
         config_with_server.log_dir = temp_log_dir
@@ -118,19 +118,22 @@ class TestEventClientLogging:
         # Mock successful POST to event server
         mock_response = Mock()
         mock_response.ok = True
-        mock_post.return_value = mock_response
+
+        mock_session = Mock()
+        mock_session.post.return_value = mock_response
+        mock_create_session.return_value = mock_session
 
         client = EventClient(config=config_with_server)
 
         # Reset mock after initialization (which logs debug messages)
-        mock_post.reset_mock()
+        mock_session.post.reset_mock()
 
         client.log(sample_event)
 
         # Verify event was sent to server
-        mock_post.assert_called_once()
-        call_args = mock_post.call_args
-        assert call_args[1]["timeout"] == 10
+        mock_session.post.assert_called_once()
+        call_args = mock_session.post.call_args
+        assert call_args[1]["timeout"] == 10.0
         # Verify the JSON payload contains our event data
         json_data = call_args[1]["json"]
         assert json_data["event_type"] == "test"
@@ -251,15 +254,17 @@ class TestEventClientLogging:
 
         assert client.logfile.exists()
 
-    @patch("madsci.client.event_client.requests.post")
+    @patch("madsci.client.event_client.create_http_session")
     def test_event_server_error_handling(
-        self, mock_post, config_with_server, temp_log_dir
+        self, mock_create_session, config_with_server, temp_log_dir
     ):
         """Test handling of event server errors."""
         config_with_server.log_dir = temp_log_dir
 
         # Mock failed POST to event server
-        mock_post.side_effect = requests.RequestException("Server error")
+        mock_session = Mock()
+        mock_session.post.side_effect = requests.RequestException("Server error")
+        mock_create_session.return_value = mock_session
 
         client = EventClient(config=config_with_server)
 
@@ -278,25 +283,29 @@ class TestEventClientLogging:
 class TestEventClientEventRetrieval:
     """Test EventClient event retrieval methods."""
 
-    @patch("madsci.client.event_client.requests.get")
+    @patch("madsci.client.event_client.create_http_session")
     def test_get_event_with_server(
-        self, mock_get, config_with_server, temp_log_dir, sample_event
+        self, mock_create_session, config_with_server, temp_log_dir, sample_event
     ):
         """Test get_event with event server configured."""
         config_with_server.log_dir = temp_log_dir
-        client = EventClient(config=config_with_server)
 
         # Mock successful GET from event server
         mock_response = Mock()
         mock_response.ok = True
         mock_response.json.return_value = sample_event.model_dump()
-        mock_get.return_value = mock_response
+
+        mock_session = Mock()
+        mock_session.get.return_value = mock_response
+        mock_create_session.return_value = mock_session
+
+        client = EventClient(config=config_with_server)
 
         result = client.get_event(sample_event.event_id)
 
-        mock_get.assert_called_once_with(
+        mock_session.get.assert_called_once_with(
             f"http://localhost:8001/event/{sample_event.event_id}",
-            timeout=10,
+            timeout=10.0,
         )
         assert isinstance(result, Event)
         assert result.event_id == sample_event.event_id
@@ -325,27 +334,30 @@ class TestEventClientEventRetrieval:
 
[…]
