### 2025-11-17 44225a3
Apply pre-commit fixes and update generated config
- Fix ruff linting issues (unused loop variable, code formatting)
- Update comment in middleware.py to avoid ERA001 false positive
- Auto-generate .env.example and Configuration.md with new settings
- All pre-commit hooks now pass successfully
