+++ src/madsci_common/tests/test_stress.py
@@ -5,7 +5,6 @@ These tests simulate high-load scenarios to ensure servers remain
 resilient under stress conditions like request storms and high concurrency.
 """
 
-import asyncio
 import time
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from typing import ClassVar
@@ -125,14 +124,8 @@ def test_concurrent_requests(stress_test_client: TestClient) -> None:
     # Calculate statistics
     response_times = [r["response_time"] for r in results if r["success"]]
     if response_times:
-        avg_response_time = sum(response_times) / len(response_times)
-        max_response_time = max(response_times)
-        print(f"\nConcurrent Request Stats:")
-        print(f"  Total requests: {num_requests}")
-        print(f"  Successful: {successes}")
-        print(f"  Rate limited: {rate_limited}")
-        print(f"  Average response time: {avg_response_time:.4f}s")
-        print(f"  Max response time: {max_response_time:.4f}s")
+        sum(response_times) / len(response_times)
+        max(response_times)
 
 
 def test_sustained_load(stress_test_client: TestClient) -> None:
@@ -163,12 +156,6 @@ def test_sustained_load(stress_test_client: TestClient) -> None:
     final_health = make_request(stress_test_client, "/health")
     assert final_health["success"], "Server became unhealthy after sustained load"
 
-    print(f"\nSustained Load Stats ({duration_seconds}s):")
-    print(f"  Total requests: {len(results)}")
-    print(f"  Successful: {successes}")
-    print(f"  Rate limited: {rate_limited}")
-    print(f"  Errors: {errors}")
-
 
 def test_burst_traffic(stress_test_client: TestClient) -> None:
     """Test server handles sudden traffic bursts."""
@@ -179,23 +166,22 @@ def test_burst_traffic(stress_test_client: TestClient) -> None:
     start_time = time.time()
     with ThreadPoolExecutor(max_workers=20) as executor:
         futures = [
-            executor.submit(make_request, stress_test_client)
-            for _ in range(burst_size)
+            executor.submit(make_request, stress_test_client) for _ in range(burst_size)
         ]
 
         for future in as_completed(futures):
             results.append(future.result())
-    total_time = time.time() - start_time
+    time.time() - start_time
 
     # Analyze results
-    successes = sum(1 for r in results if r["success"])
-    rate_limited = sum(1 for r in results if r["rate_limited"])
+    sum(1 for r in results if r["success"])
+    sum(1 for r in results if r["rate_limited"])
 
     # With rate limiting, we expect some requests to be limited
     # but none should cause the server to crash
-    assert all(r["status_code"] in [200, 429] for r in results if r["status_code"] != 0), (
-        "Server returned unexpected status codes during burst"
-    )
+    assert all(
+        r["status_code"] in [200, 429] for r in results if r["status_code"] != 0
+    ), "Server returned unexpected status codes during burst"
 
     # Server should still be responsive after burst
     # Wait for rate limit window to expire so we can verify server is still functioning
@@ -204,13 +190,6 @@ def test_burst_traffic(stress_test_client: TestClient) -> None:
     assert health_check["success"] or health_check["rate_limited"], (
         "Server became completely unresponsive after burst traffic"
     )
-
-    print(f"\nBurst Traffic Stats:")
-    print(f"  Burst size: {burst_size}")
-    print(f"  Total time: {total_time:.2f}s")
-    print(f"  Requests per second: {burst_size/total_time:.2f}")
-    print(f"  Successful: {successes}")
-    print(f"  Rate limited: {rate_limited}")
 
 
 def test_multiple_endpoints_under_load(stress_test_client: TestClient) -> None:
[…]
