### 2025-11-17 efe603a
Fix CI/CD checks: pre-commit, linting, and test issues
- Fix RET504 linter error in test_experiment_client.py (unnecessary assignment)
- Add noqa comments for intentional local imports in create_http_session to avoid circular dependencies
- Add sleep in EventClient test to wait for threaded event posting
- Auto-generated .env.example and Configuration.md updates from pydantic-settings-export

All pre-commit checks now pass:
- ruff check ✓
- ruff format ✓
- pydantic-settings-export ✓

All client tests pass:
- ExperimentClient: 22/22 ✓
- EventClient: 36/36 ✓
