### 2025-11-17 d852f49
Update src/madsci_node_module/madsci/node_module/rest_node_module.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
