### 2025-11-17 30c74d3
Modify global context to prevent --help arg capture
