+++ src/madsci_common/tests/test_context.py
@@ -3,47 +3,61 @@
 import threading
 
 from madsci.common.context import (
+    GlobalMadsciContext,
     get_current_madsci_context,
-    global_madsci_context,
     madsci_context,
 )
 from madsci.common.types.context_types import MadsciContext
 
 
 def test_global_madsci_context_default() -> None:
-    """Test that global_madsci_context is an instance of MadsciContext by default."""
-    assert isinstance(global_madsci_context, MadsciContext)
+    """Test that GlobalMadsciContext._context is an instance of MadsciContext by default."""
+    assert isinstance(GlobalMadsciContext._context, MadsciContext)
 
 
 def test_global_context_across_threads() -> None:
-    """Tests that changes to global_madsci_context are consistent across threads."""
-    original_url = global_madsci_context.lab_server_url
+    """Tests that changes to GlobalMadsciContext are consistent across threads."""
+    original_url = GlobalMadsciContext._context.lab_server_url
     test_url = "http://test-lab:8000"
-    global_madsci_context.lab_server_url = test_url
-    assert str(global_madsci_context.lab_server_url) == "http://test-lab:8000/"
+
+    # Create a new context with modified URL
+    new_context = GlobalMadsciContext._context.model_copy()
+    new_context.lab_server_url = test_url
+    GlobalMadsciContext._context = new_context
+    assert str(GlobalMadsciContext._context.lab_server_url) == "http://test-lab:8000/"
 
     def check_context() -> None:
         """Function to check context in a separate thread."""
-        assert str(global_madsci_context.lab_server_url) == "http://test-lab:8000/"
-        global_madsci_context.lab_server_url = original_url
+        assert (
+            str(GlobalMadsciContext._context.lab_server_url) == "http://test-lab:8000/"
+        )
+        # Restore original context
+        restore_context = GlobalMadsciContext._context.model_copy()
+        restore_context.lab_server_url = original_url
+        GlobalMadsciContext._context = restore_context
 
     # Run the check in a separate thread
     thread = threading.Thread(target=check_context)
     thread.start()
     thread.join()
     # Ensure the original state is restored
-    global_madsci_context.lab_server_url = original_url
+    final_context = GlobalMadsciContext._context.model_copy()
+    final_context.lab_server_url = original_url
+    GlobalMadsciContext._context = final_context
 
 
 def test_madsci_context_temporary_override() -> None:
     """Test that madsci_context temporarily overrides and restores context."""
-    original_lab_url = global_madsci_context.lab_server_url
-    original_event_url = global_madsci_context.event_server_url
+    original_lab_url = GlobalMadsciContext._context.lab_server_url
+    original_event_url = GlobalMadsciContext._context.event_server_url
     test_lab_url = "http://test-lab:8000"
     test_event_url = "http://test-event:8001"
 
-    global_madsci_context.lab_server_url = original_lab_url
-    global_madsci_context.event_server_url = original_event_url
+    # Ensure we start with known state
+    base_context = GlobalMadsciContext._context.model_copy()
+    base_context.lab_server_url = original_lab_url
+    base_context.event_server_url = original_event_url
+    GlobalMadsciContext._context = base_context
 
     with madsci_context(
         lab_server_url=test_lab_url, event_server_url=test_event_url
@@ -65,12 +79,15 @@ def test_madsci_context_temporary_override() -> None:
 
 def test_madsci_context_partial_override() -> None:
     """Test that madsci_context only overrides specified fields."""
-    original_lab_url = global_madsci_context.lab_server_url
-    original_event_url = global_madsci_context.event_server_url
+    original_lab_url = GlobalMadsciContext._context.lab_server_url
[…]
