### 2025-11-17 ca23f3b
Fix PR #180 review issues: decorator stacking, lazy loading, and docstrings
Addresses all review comments on PR #180:

1. Fixed invalid @property/@classmethod decorator stacking (Comments #1 & #2)
   - Removed incompatible property/classmethod combinations
   - Replaced with proper @classmethod get_context() and set_context() methods
   - Maintains clean API while fixing Python decorator compatibility issues

2. Implemented true lazy loading to prevent --help arg capture
   - Changed _context from MadsciContext() to None as initial value
   - Context is now only created on first access via get
[…]
