### 2025-11-17 1ee2dfd
Fix Python 3.9 compatibility: replace pipe operator with Optional
Co-authored-by: LuckierDodge <15006408+LuckierDodge@users.noreply.github.com>
