+++ src/madsci_location_manager/tests/test_location_client.py
@@ -64,7 +64,7 @@ def test_transfer_method_signatures(location_client):
     # Check get_transfer_graph signature
     sig = inspect.signature(location_client.get_transfer_graph)
     params = list(sig.parameters.keys())
-    assert "retry" in params
+    assert "timeout" in params
 
     # Check plan_transfer signature
     sig = inspect.signature(location_client.plan_transfer)
@@ -72,13 +72,13 @@ def test_transfer_method_signatures(location_client):
     assert "source_location_id" in params
     assert "target_location_id" in params
     assert "resource_id" in params
-    assert "retry" in params
+    assert "timeout" in params
 
     # Check get_location_resources signature
     sig = inspect.signature(location_client.get_location_resources)
     params = list(sig.parameters.keys())
     assert "location_id" in params
-    assert "retry" in params
+    assert "timeout" in params
 
 
 def test_get_location_by_name_method_exists(location_client):
@@ -92,14 +92,14 @@ def test_get_location_by_name_method_signature(location_client):
     sig = inspect.signature(location_client.get_location_by_name)
     params = list(sig.parameters.keys())
     assert "location_name" in params
-    assert "retry" in params
+    assert "timeout" in params
 
     # Check return type annotation
     assert sig.return_annotation.__name__ == "Location"
 
 
-@patch("madsci.client.location_client.requests.Session.get")
-def test_get_location_by_name_method_request(mock_get, location_client):
+@patch("madsci.client.location_client.create_http_session")
+def test_get_location_by_name_method_request(mock_create_session):
     """Test that get_location_by_name makes correct HTTP request."""
 
     # Mock successful response
@@ -107,30 +107,36 @@ def test_get_location_by_name_method_request(mock_get, location_client):
     test_location_id = new_ulid_str()  # Generate valid ULID
     mock_location_data = {
         "location_id": test_location_id,
-        "name": "test_location_name",
+        "location_name": "test_location_name",
         "description": "Test location description",
     }
     mock_response.json.return_value = mock_location_data
     mock_response.raise_for_status.return_value = None
-    mock_get.return_value = mock_response
+
+    mock_session = Mock()
+    mock_session.get.return_value = mock_response
+    mock_create_session.return_value = mock_session
+
+    # Create a new client to use mocked session
+    client = LocationClient(location_server_url="http://test/")
 
     # Call the method
-    result = location_client.get_location_by_name("test_location_name")
+    result = client.get_location_by_name("test_location_name")
 
     # Verify the request was made correctly
-    mock_get.assert_called_once()
-    call_args = mock_get.call_args
+    mock_session.get.assert_called_once()
+    call_args = mock_session.get.call_args
     # Check that the URL is correct and query parameters are used
     assert call_args[0][0].endswith("/location")
     assert call_args[1]["params"] == {"name": "test_location_name"}
 
     # Verify the result is a Location object
     assert isinstance(result, Location)
-    assert result.name == "test_location_name"
+    assert result.location_name == "test_location_name"
     assert result.location_id == test_location_id
 
 
-@patch("madsci.client.location_client.requests.Session.get")
+@patch("madsci.client.location_client.create_http_session")
 def test_get_location_by_name_method_error_handling(mock_get, location_client):
     """Test that get_location_by_name handles errors correctly."""
 
@@ -146,8 +152,8 @@ def test_get_location_by_name_method_error_handling(mock_get, location_client):
         location_client.get_location_by_name("nonexistent_location")
 
 
-@patch("madsci.client.location_client.requests.Session.get")
-def test_get_location_resources_empty_hierarchy(mock_get, location_client):
+@patch("madsci.client.location_client.create_http_session")
[…]
