### 2025-11-18 1e0bbb7
feat: Add MadsciClientMixin for centralized client lifecycle management
Implements issue #162 by creating a reusable mixin class that manages
MADSci client initialization and lifecycle across all components.

## New Features

- **MadsciClientMixin**: Central mixin class for managing all 7 MADSci clients
  - EventClient, ResourceClient, DataClient, ExperimentClient
  - WorkcellClient, LocationClient, LabClient
- Lazy initialization with property-based access
- Flexible configuration via context, explicit URLs, or direct injection
- EventClient sharing across dependent clients
- Comprehensive un
[…]
