+++ src/madsci_client/tests/test_client_mixin.py
@@ -1,0 +1,507 @@
+"""Unit tests for MadsciClientMixin."""
+
+import tempfile
+from pathlib import Path
+from unittest.mock import Mock, patch
+
+import pytest
+from madsci.client.client_mixin import MadsciClientMixin
+from madsci.client.data_client import DataClient
+from madsci.client.event_client import EventClient
+from madsci.client.experiment_client import ExperimentClient
+from madsci.client.lab_client import LabClient
+from madsci.client.location_client import LocationClient
+from madsci.client.resource_client import ResourceClient
+from madsci.client.workcell_client import WorkcellClient
+from madsci.common.types.context_types import MadsciContext
+from madsci.common.types.event_types import EventClientConfig
+from pydantic import AnyUrl
+
+
+@pytest.fixture
+def temp_log_dir():
+    """Create a temporary directory for log files."""
+    with tempfile.TemporaryDirectory() as temp_dir:
+        yield Path(temp_dir)
+
+
+@pytest.fixture
+def mock_context():
+    """Create a mock MADSci context."""
+    return MadsciContext(
+        event_server_url=AnyUrl("http://localhost:8001"),
+        experiment_server_url=AnyUrl("http://localhost:8002"),
+        resource_server_url=AnyUrl("http://localhost:8003"),
+        data_server_url=AnyUrl("http://localhost:8004"),
+        workcell_server_url=AnyUrl("http://localhost:8005"),
+        location_server_url=AnyUrl("http://localhost:8006"),
+        lab_server_url=AnyUrl("http://localhost:8000"),
+    )
+
+
+class TestMixinBasicUsage:
+    """Test basic usage patterns of MadsciClientMixin."""
+
+    def test_mixin_can_be_inherited(self):
+        """Test that mixin can be inherited by a class."""
+
+        class MyComponent(MadsciClientMixin):
+            pass
+
+        component = MyComponent()
+        assert isinstance(component, MadsciClientMixin)
+
+    def test_mixin_with_required_clients(self, temp_log_dir, mock_context):
+        """Test mixin with REQUIRED_CLIENTS class attribute."""
+
+        class MyComponent(MadsciClientMixin):
+            REQUIRED_CLIENTS = ["event", "resource"]
+
+        with patch(
+            "madsci.client.client_mixin.get_current_madsci_context",
+            return_value=mock_context,
+        ):
+            with patch("madsci.client.event_client.EventClientConfig") as mock_config:
+                mock_config.return_value.name = None
+                mock_config.return_value.log_dir = temp_log_dir
+                mock_config.return_value.log_level = "INFO"
+                mock_config.return_value.event_server_url = None
+
+                with patch(
+                    "madsci.client.event_client.get_current_madsci_context",
+                    return_value=mock_context,
+                ):
+                    with patch("madsci.client.resource_client.requests.get"):
+                        component = MyComponent()
+                        component.setup_clients()
+
+                        assert component._event_client is not None
+                        assert component._resource_client is not None
+                        assert component._data_client is None
+
+    def test_lazy_initialization(self, temp_log_dir, mock_context):
+        """Test that clients are lazily initialized."""
+
+        class MyComponent(MadsciClientMixin):
+            pass
+
+        with patch(
+            "madsci.client.client_mixin.get_current_madsci_context",
+            return_value=mock_context,
+        ):
+            with patch("madsci.client.event_client.EventClientConfig") as mock_config:
+                mock_config.return_value.name = None
+                mock_config.return_value.log_dir = temp_log_dir
+                mock_config.return_value.log_level = "INFO"
+                mock_config.return_value.event_server_url = None
+
+                with patch(
+                    "madsci.client.event_client.get_current_madsci_context",
+                    return_value=mock_context,
+                ):
+                    component = MyComponent()
+
[…]

[…]
