### 2025-11-18 80e1037
fix: Simplify client mixin tests and fix _init_client_batch
- Rewrote tests to be simpler and more focused on actual functionality
- Fixed _init_client_batch to avoid eager property access
- Tests now use direct mock injection instead of complex nested patches
- All 22 mixin tests pass successfully
- Removed dependency on context mocking in mixin itself
