### 2025-11-18 af3a009
Fixing tests post-merge
