+++ src/madsci_resource_manager/tests/test_fresh_database_initialization.py
@@ -10,7 +10,7 @@ from madsci.resource_manager.resource_interface import ResourceInterface
 from madsci.resource_manager.resource_server import ResourceManager
 from madsci.resource_manager.resource_tables import create_session
 from pytest_mock_resources import PostgresConfig, create_postgres_fixture
-from sqlalchemy import Engine
+from sqlalchemy import Engine, inspect, text
 from starlette.testclient import TestClient
 
 
@@ -78,3 +78,111 @@ def test_auto_initialization_methods_exist():
     assert hasattr(ResourceManager, "_auto_initialize_fresh_database")
     assert callable(ResourceManager._is_fresh_database)
     assert callable(ResourceManager._auto_initialize_fresh_database)
+
+
+def test_fresh_database_detection(fresh_postgres_engine: Engine) -> None:
+    """Test that is_fresh_database correctly detects fresh databases."""
+    version_checker = DatabaseVersionChecker(str(fresh_postgres_engine.url))
+    version_checker.engine = fresh_postgres_engine
+
+    # Fresh database should be detected as fresh
+    assert version_checker.is_fresh_database() is True, (
+        "Fresh database should be detected as fresh"
+    )
+
+
+def test_fresh_database_auto_initialization(fresh_postgres_engine: Engine) -> None:
+    """Test that validate_or_fail auto-initializes fresh databases."""
+    # Create a version checker that uses the existing engine directly
+    version_checker = DatabaseVersionChecker(str(fresh_postgres_engine.url))
+    # Replace the engine with the fresh one to avoid connection issues
+    version_checker.engine = fresh_postgres_engine
+
+    # This should auto-initialize without raising an error
+    version_checker.validate_or_fail()
+
+    # After auto-initialization, version should be tracked
+    assert version_checker.is_version_tracked() is True, (
+        "Version tracking should be enabled after auto-initialization"
+    )
+
+    # Should have a database version now
+    db_version = version_checker.get_database_version()
+    assert db_version is not None, (
+        "Database should have a version after auto-initialization"
+    )
+
+    # Should match current MADSci version (accepting that it might be 0.0.0 in test)
+    madsci_version = version_checker.get_current_madsci_version()
+    # Using major.minor comparison like the PostgreSQL implementation does
+    assert version_checker.versions_match(db_version, madsci_version), (
+        "Database version should match MADSci version after auto-initialization"
+    )
+
+
+def test_fresh_database_auto_init_does_not_require_migration_after(
+    fresh_postgres_engine: Engine,
+) -> None:
+    """Test that after auto-initialization, no further migration is needed."""
+    version_checker = DatabaseVersionChecker(str(fresh_postgres_engine.url))
+    version_checker.engine = fresh_postgres_engine
+
+    # Auto-initialize
+    version_checker.validate_or_fail()
+
+    # Should not need migration after auto-initialization
+    needs_migration, madsci_version, db_version = version_checker.is_migration_needed()
+    assert needs_migration is False, (
+        "Should not need migration after auto-initialization"
+    )
+    assert madsci_version is not None, "MADSci version should be available"
+    assert db_version is not None, (
+        "Database version should be available after auto-initialization"
+    )
+
+
+def test_fresh_database_auto_init_creates_version_table(
+    fresh_postgres_engine: Engine,
+) -> None:
+    """Test that auto-initialization creates the schema version table."""
+    version_checker = DatabaseVersionChecker(str(fresh_postgres_engine.url))
+    version_checker.engine = fresh_postgres_engine
+
+    # Initially no schema version table
+    inspector = inspect(fresh_postgres_engine)
+    initial_tables = inspector.get_table_names()
+    assert "madsci_schema_version" not in initial_tables, (
[…]
