### 2025-11-19 8c3b605
Backing up utilization reporting fixes on rpl_sdl
