+++ src/madsci_event_manager/madsci/event_manager/event_server.py
@@ -1,7 +1,7 @@
 """Example Event Manager implementation using the new AbstractManagerBase class."""
 
 from datetime import datetime
-from typing import Any, Dict, Optional, Tuple, Union
+from typing import Any, Dict, Optional, Tuple, Union, Annotated
 
 import pymongo
 from classy_fastapi import get, post
@@ -20,6 +20,7 @@ from madsci.common.types.event_types import (
 )
 from madsci.event_manager.events_csv_exporter import CSVExporter
 from madsci.event_manager.notifications import EmailAlerts
+from madsci.event_manager.time_series_analyzer import TimeSeriesAnalyzer
 from madsci.event_manager.utilization_analyzer import UtilizationAnalyzer
 from pymongo import MongoClient, errors
 from pymongo.synchronous.database import Database
@@ -141,18 +142,35 @@ class EventManager(AbstractManagerBase[EventManagerSettings, EventManagerDefinit
 
     # Utilization endpoints (examples of more complex endpoints)
 
-    @get("/utilization/sessions")
+    @get("/utilization/sessions", response_model=None)
     async def get_session_utilization(
         self,
         start_time: Optional[str] = None,
         end_time: Optional[str] = None,
-        csv_format: bool = Query(False, description="Return data in CSV format"),
-        save_to_file: bool = Query(False, description="Save CSV to server filesystem"),
-        output_path: Optional[str] = Query(
-            None, description="Server path to save CSV files"
-        ),
+        csv_format: Annotated[
+            bool, Query(description="Return data in CSV format")
+        ] = False,
+        save_to_file: Annotated[
+            bool, Query(description="Save CSV to server filesystem")
+        ] = False,
+        output_path: Annotated[
+            Optional[str], Query(description="Server path to save CSV files")
+        ] = None,
     ) -> Union[Dict[str, Any], Response]:
-        """Generate comprehensive session-based utilization report."""
+        """
+        Generate comprehensive session-based utilization report.
+
+        PARAMETERS:
+        - csv_format: If True, returns CSV data instead of JSON
+        - save_to_file: If True, saves CSV to server filesystem (requires output_path)
+        - output_path: Server path where CSV files should be saved
+
+        RESPONSE TYPES:
+        - csv_format=False: JSON dict (default behavior)
+        - csv_format=True, save_to_file=False: CSV download
+        - csv_format=True, save_to_file=True: JSON with file path info
+        """
+
         analyzer = self._get_session_analyzer()
         if analyzer is None:
             return {"error": "Failed to create session analyzer"}
@@ -191,19 +209,157 @@ class EventManager(AbstractManagerBase[EventManagerSettings, EventManagerDefinit
             return report
 
         except Exception as e:
-            self.logger.error(f"Error generating session utilization: {e}")
+            self.logger.log_error(f"Error generating session utilization: {e}")
             return {"error": f"Failed to generate report: {e!s}"}
+
+    @get("/utilization/periods", response_model=None)
+    async def get_utilization_periods(
+        self,
+        start_time: Optional[str] = None,
+        end_time: Optional[str] = None,
+        analysis_type: str = "daily",
+        user_timezone: str = "America/Chicago",
+        include_users: bool = True,
+        csv_format: Annotated[
+            bool, Query(description="Return data in CSV format")
+        ] = False,
+        save_to_file: Annotated[
+            bool, Query(description="Save CSV to server filesystem")
+        ] = False,
+        output_path: Annotated[
+            Optional[str], Query(description="Server path to save CSV files")
+        ] = None,
+    ) -> Union[Dict[str, Any], Response]:
+        """Get time-series utilization analysis with periodic breakdowns."""
+
+        # Setup analyzer
[…]
