+++ tests/manual_test_utilization.py
@@ -8,10 +8,16 @@ client = EventClient(event_server_url="http://parker.cels.anl.gov:8001/")
 
 # Test 1: Basic utilization periods CSV export
 saved_files = client.get_utilization_periods(
-    analysis_type="daily", csv_export=True, save_to_file=True, output_path="~/.madsci/reports"
+    analysis_type="daily",
+    csv_export=True,
+    save_to_file=True,
+    output_path="~/.madsci/reports",
 )
 saved_files = client.get_utilization_periods(
-    analysis_type="weekly", csv_export=True, save_to_file=True, output_path="~/.madsci/reports"
+    analysis_type="weekly",
+    csv_export=True,
+    save_to_file=True,
+    output_path="~/.madsci/reports",
 )
 saved_files = client.get_utilization_periods(
     analysis_type="monthly",
