+++ src/madsci_workcell_manager/madsci/workcell_manager/workflow_utils.py
@@ -23,9 +23,9 @@ from madsci.common.types.workflow_types import (
     Workflow,
     WorkflowDefinition,
 )
+from madsci.common.types.action_types import ActionStatus, ActionResult
 from madsci.workcell_manager.state_handler import WorkcellStateHandler
 from madsci.workcell_manager.workcell_actions import workcell_action_dict
-from sqlalchemy import Null, null
 
 
 def validate_node_names(
@@ -442,15 +442,24 @@ def get_workflow_inputs_directory(
     return Path(working_directory).expanduser() / "Workflows" / workflow_id / "Inputs"
 
 
-def cancel_workflow(wf: Workflow, state_handler: WorkcellStateHandler) -> None:
+def cancel_workflow(wf: Workflow) -> Workflow:
     """Cancels the workflow run"""
+    # Handle WF status.
     wf.status.cancelled = True
-    # if not wf.start_time:
-    #     wf.start_time = datetime.now()
-    # if not wf.end_time:
-    #     wf.end_time = datetime.now()
-    with state_handler.wc_state_lock():
-        state_handler.set_active_workflow(wf)
+    wf.status.running = False
+
+    # Handle WF current action/step.
+    current_step = wf.steps[wf.step_index]
+    current_step.status = ActionStatus.CANCELLED
+    current_step.result = ActionResult(status=ActionStatus.CANCELLED)
+    if not current_step.end_time:
+        current_step.end_time = datetime.now()
+    if current_step.start_time and current_step.end_time:
+        current_step.duration = current_step.end_time - current_step.start_time
+
+    # handle start/end/duration.
+    if not wf.end_time:
+        wf.end_time = datetime.now()
     return wf
 
 
