+++ src/madsci_data_manager/tests/test_data_migrations.py
@@ -237,17 +237,21 @@ def test_data_backup_creation(mock_subprocess):
         )
         migrator = MongoDBMigrator(settings)
 
-        backup_path = migrator.create_backup()
+        # Mock the validation methods to bypass filesystem checks
+        with (
+            patch.object(migrator, "_validate_backup_integrity", return_value=True),
+        ):
+            backup_path = migrator.create_backup()
 
-        # Verify backup path format
-        assert "madsci_data_backup_" in backup_path.name
+            # Verify backup path format
+            assert "madsci_data_backup_" in backup_path.name
 
-        # Verify mongodump was called with data database
-        mock_subprocess.assert_called_once()
-        call_args = mock_subprocess.call_args[0][0]
-        assert "mongodump" in call_args
-        assert "--db" in call_args
-        assert "madsci_data" in call_args
+            # Verify mongodump was called with data database
+            mock_subprocess.assert_called_once()
+            call_args = mock_subprocess.call_args[0][0]
+            assert "mongodump" in call_args
+            assert "--db" in call_args
+            assert "madsci_data" in call_args
     finally:
         schema_file.unlink()
 
