+++ src/madsci_resource_manager/madsci/resource_manager/migration_tool.py
@@ -1,14 +1,19 @@
 """Database migration tool for MADSci resources using Alembic with automatic type conversion handling."""
 
+import fcntl
+import hashlib
+import json
 import os
 import re
 import subprocess
 import sys
+import tempfile
+import time
 import traceback
 import urllib.parse as urlparse
 from datetime import datetime
 from pathlib import Path
-from typing import Any, Optional
+from typing import Any, List, Optional
 
 from alembic import command
 from alembic.autogenerate import compare_metadata
@@ -22,7 +27,7 @@ from madsci.resource_manager.database_version_checker import DatabaseVersionChec
 from madsci.resource_manager.resource_tables import ResourceTable, SchemaVersionTable
 from psycopg2 import sql
 from pydantic import Field
-from sqlalchemy import inspect
+from sqlalchemy import inspect, text
 from sqlmodel import Session, create_engine, select
 
 
@@ -114,6 +119,10 @@ class DatabaseMigrator:
         # Setup Alembic configuration
         self.alembic_cfg = self._setup_alembic_config()
 
+        # Migration locking attributes
+        self.lock_file: Optional[Path] = None
+        self.lock_fd: Optional[int] = None
+
     def _get_package_root(self) -> Path:
         """Get the root directory of the madsci.resource_manager package."""
         try:
@@ -187,6 +196,100 @@ class DatabaseMigrator:
             "database": parsed.path.lstrip("/") if parsed.path else "resources",
         }
 
+    def _acquire_migration_lock(self) -> None:
+        """Acquire exclusive lock for migration operations."""
+        db_info = self._parse_db_url()
+        database_name = db_info["database"]
+
+        # Create lock file in system temp directory
+        temp_dir = Path(tempfile.gettempdir())
+        self.lock_file = temp_dir / f"madsci_migration_{database_name}.lock"
+
+        try:
+            # Open lock file for writing
+            self.lock_fd = os.open(
+                str(self.lock_file), os.O_CREAT | os.O_WRONLY | os.O_TRUNC, 0o644
+            )
+
+            # Try to acquire exclusive lock (non-blocking)
+            fcntl.flock(self.lock_fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
+
+            # Write process info to lock file
+            lock_info = f"pid={os.getpid()}\ntimestamp={time.time()}\ndatabase={database_name}\n"
+            os.write(self.lock_fd, lock_info.encode())
+
+            self.logger.info(f"Acquired migration lock: {self.lock_file}")
+
+        except BlockingIOError:
+            if self.lock_fd is not None:
+                os.close(self.lock_fd)
+                self.lock_fd = None
+            raise RuntimeError(
+                f"Another migration is already in progress for database '{database_name}'. "
+                f"Lock file: {self.lock_file}"
+            ) from None
+        except Exception as e:
+            if self.lock_fd is not None:
+                os.close(self.lock_fd)
+                self.lock_fd = None
+            raise RuntimeError(f"Failed to acquire migration lock: {e}") from e
+
+    def _release_migration_lock(self) -> None:
+        """Release migration lock."""
+        if self.lock_fd is not None:
+            try:
+                fcntl.flock(self.lock_fd, fcntl.LOCK_UN)
+                os.close(self.lock_fd)
+                self.logger.info("Released migration lock")
+            except Exception as e:
+                self.logger.warning(f"Error releasing lock: {e}")
+            finally:
+                self.lock_fd = None
+
+        if self.lock_file and self.lock_file.exists():
+            try:
+                self.lock_file.unlink()
+                self.logger.info(f"Removed lock file: {self.lock_file}")
+            except Exception as e:
[…]
