+++ src/madsci_workcell_manager/tests/test_workcell_migration.py
@@ -230,17 +230,21 @@ def test_workcell_backup_creation(mock_subprocess):
         )
         migrator = MongoDBMigrator(settings)
 
-        backup_path = migrator.create_backup()
+        # Mock the validation methods to bypass filesystem checks
+        with (
+            patch.object(migrator, "_validate_backup_integrity", return_value=True),
+        ):
+            backup_path = migrator.create_backup()
 
-        # Verify backup path format
-        assert "madsci_workcells_backup_" in backup_path.name
+            # Verify backup path format
+            assert "madsci_workcells_backup_" in backup_path.name
 
-        # Verify mongodump was called with workcells database
-        mock_subprocess.assert_called_once()
-        call_args = mock_subprocess.call_args[0][0]
-        assert "mongodump" in call_args
-        assert "--db" in call_args
-        assert "madsci_workcells" in call_args
+            # Verify mongodump was called with workcells database
+            mock_subprocess.assert_called_once()
+            call_args = mock_subprocess.call_args[0][0]
+            assert "mongodump" in call_args
+            assert "--db" in call_args
+            assert "madsci_workcells" in call_args
     finally:
         schema_file.unlink()
 
