### 2025-11-19 de596c5
Bump to v0.5.4; add 'step_key' alias for step's key field
