### 2025-11-19 85b326d
Fix location manager resource attachment and updating from definition file
