### 2025-11-19 85c12b8
Cleanup location manager settings
