### 2025-11-20 e42e377
Fix up e2e tests, address context issue
