### 2025-11-24 1d10887
WIP: standalone backup tooling phase 1
