+++ src/madsci_common/madsci/common/types/backup_types.py
@@ -1,0 +1,66 @@
+"""Configuration types for MADSci backup operations."""
+
+from pathlib import Path
+from typing import List, Optional, Union
+
+from madsci.common.types.base_types import MadsciBaseSettings
+from pydantic import AnyUrl, Field, field_validator
+
+
+class BaseBackupSettings(MadsciBaseSettings):
+    """Common backup configuration settings."""
+
+    backup_dir: Path = Field(
+        default=Path(".madsci/backups"),
+        title="Backup Directory",
+        description="Directory for storing backups",
+    )
+    max_backups: int = Field(
+        default=10,
+        ge=0,
+        title="Maximum Backups",
+        description="Maximum number of backups to retain",
+    )
+    validate_integrity: bool = Field(
+        default=True,
+        title="Validate Integrity",
+        description="Perform integrity validation after backup",
+    )
+    compression: bool = Field(
+        default=True,
+        title="Enable Compression",
+        description="Enable backup compression",
+    )
+
+    @field_validator("backup_dir", mode="before")
+    @classmethod
+    def convert_backup_dir_to_path(cls, v: Union[str, Path]) -> Path:
+        """Convert backup_dir to Path object."""
+        if isinstance(v, str):
+            return Path(v)
+        return v
+
+
+class PostgreSQLBackupSettings(BaseBackupSettings):
+    """PostgreSQL-specific backup settings."""
+
+    db_url: str = Field(title="Database URL", description="PostgreSQL connection URL")
+    backup_format: str = Field(
+        default="custom",
+        title="Backup Format",
+        description="pg_dump format: custom, plain, directory, tar",
+    )
+
+
+class MongoDBBackupSettings(BaseBackupSettings):
+    """MongoDB-specific backup settings."""
+
+    mongo_db_url: AnyUrl = Field(
+        title="MongoDB URL", description="MongoDB connection URL"
+    )
+    database: str = Field(title="Database Name", description="Database name to backup")
+    collections: Optional[List[str]] = Field(
+        default=None,
+        title="Collections",
+        description="Specific collections to backup (all if None)",
+    )
