+++ src/madsci_common/madsci/common/backup_tools/backup_manager.py
@@ -1,0 +1,395 @@
+"""Backup management operations and utilities."""
+
+import shutil
+from datetime import datetime
+from pathlib import Path
+from typing import Any, Dict, List, Optional
+
+from madsci.client.event_client import EventClient
+
+from .backup_validator import BackupValidator
+from .base_backup import BackupInfo
+
+
+class BackupManager:
+    """Manages backup operations like rotation, listing, and cleanup."""
+
+    def __init__(self, logger: Optional[EventClient] = None) -> None:
+        """
+        Initialize the backup manager.
+
+        Args:
+            logger: Optional logger instance
+        """
+        self.logger = logger or EventClient()
+        self.validator = BackupValidator(logger)
+
+    def list_backups(self, backup_dir: Path) -> List[BackupInfo]:
+        """
+        List all available backups in a directory with metadata.
+
+        Args:
+            backup_dir: Directory containing backups
+
+        Returns:
+            List of BackupInfo objects sorted by creation time (newest first)
+        """
+        if not backup_dir.exists():
+            self.logger.warning(f"Backup directory does not exist: {backup_dir}")
+            return []
+
+        backups = []
+
+        # Look for backup files (SQL files and MongoDB backup directories)
+        backup_files = []
+        backup_files.extend(backup_dir.glob("*.sql"))  # PostgreSQL backups
+        backup_files.extend(
+            [
+                d
+                for d in backup_dir.iterdir()
+                if d.is_dir() and not d.name.startswith(".")
+            ]
+        )  # MongoDB backups
+
+        for backup_path in backup_files:
+            try:
+                backup_info = self._create_backup_info(backup_path)
+                if backup_info:
+                    backups.append(backup_info)
+            except Exception as e:
+                self.logger.warning(f"Error processing backup {backup_path}: {e}")
+
+        # Sort by creation time (newest first)
+        backups.sort(key=lambda x: x.created_at, reverse=True)
+
+        return backups
+
+    def _create_backup_info(self, backup_path: Path) -> Optional[BackupInfo]:
+        """
+        Create BackupInfo from a backup file/directory.
+
+        Args:
+            backup_path: Path to backup file or directory
+
+        Returns:
+            BackupInfo object or None if unable to create
+        """
+        try:
+            # Load metadata if available
+            metadata = self.validator.load_metadata(backup_path)
+
+            if metadata:
+                return BackupInfo(
+                    backup_path=backup_path,
+                    created_at=datetime.fromisoformat(metadata["timestamp"]),
+                    database_version=metadata.get("database_version"),
+                    backup_size=metadata.get("backup_size", 0),
+                    checksum=metadata.get("checksum", ""),
+                    backup_type=metadata.get("backup_type", "unknown"),
+                    is_valid=metadata.get("is_valid", False),
+                )
+            # Fallback to file system metadata
+            stat = backup_path.stat()
+            return BackupInfo(
+                backup_path=backup_path,
+                created_at=datetime.fromtimestamp(stat.st_mtime),
+                database_version=None,
+                backup_size=stat.st_size
+                if backup_path.is_file()
+                else self._get_directory_size(backup_path),
+                checksum="",
+                backup_type="unknown",
+                is_valid=False,
+            )
+
+        except Exception as e:
+            self.logger.error(f"Error creating BackupInfo for {backup_path}: {e}")
+            return None
+
+    def _get_directory_size(self, directory: Path) -> int:
+        """Get total size of all files in a directory."""
+        total = 0
+        for file_path in directory.rglob("*"):
+            if file_path.is_file():
+                total += file_path.stat().st_size
+        return total
+
[…]
