### 2025-11-24 ba8b583
WIP: standalone backup tooling phase 2
