+++ src/madsci_resource_manager/madsci/resource_manager/backup_tools/postgres_backup.py
@@ -1,0 +1,503 @@
+"""Standalone PostgreSQL backup and restore tool for MADSci."""
+
+import os
+import subprocess
+import threading
+import urllib.parse as urlparse
+from datetime import datetime
+from pathlib import Path
+from typing import TYPE_CHECKING, Dict, List, Optional
+
+if TYPE_CHECKING:
+    from types import TracebackType
+
+from madsci.client.event_client import EventClient
+from madsci.common.backup_tools.backup_manager import BackupManager
+from madsci.common.backup_tools.backup_validator import BackupValidator
+from madsci.common.backup_tools.base_backup import AbstractBackupTool, BackupInfo
+from madsci.common.types.backup_types import PostgreSQLBackupSettings
+
+
+class BackupLockManager:
+    """Manages exclusive locks for backup operations."""
+
+    def __init__(self) -> None:
+        """Initialize the backup lock manager."""
+        self._locks: Dict[str, threading.Lock] = {}
+        self._lock_creation_lock = threading.Lock()
+
+    def acquire_lock(self, db_url: str) -> "BackupLockContext":
+        """Acquire lock for a specific database URL."""
+        return BackupLockContext(self, db_url)
+
+    def _get_lock(self, db_url: str) -> threading.Lock:
+        """Get or create lock for database URL."""
+        with self._lock_creation_lock:
+            if db_url not in self._locks:
+                self._locks[db_url] = threading.Lock()
+            return self._locks[db_url]
+
+
+class BackupLockContext:
+    """Context manager for database backup locks."""
+
+    def __init__(self, lock_manager: BackupLockManager, db_url: str) -> None:
+        """Initialize the backup lock context."""
+        self.lock_manager = lock_manager
+        self.db_url = db_url
+        self.lock = lock_manager._get_lock(db_url)
+
+    def __enter__(self) -> "BackupLockContext":
+        """Acquire the lock."""
+        self.lock.acquire()
+        return self
+
+    def __exit__(
+        self,
+        exc_type: Optional[type],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional["TracebackType"],
+    ) -> None:
+        """Release the lock."""
+        self.lock.release()
+
+
+class PostgreSQLBackupTool(AbstractBackupTool):
+    """Standalone PostgreSQL backup and restore tool."""
+
+    def __init__(
+        self, settings: PostgreSQLBackupSettings, logger: Optional[EventClient] = None
+    ) -> None:
+        """
+        Initialize PostgreSQL backup tool.
+
+        Args:
+            settings: PostgreSQL backup configuration
+            logger: Optional event client for logging
+        """
+        self.settings = settings
+        self.logger = logger or EventClient()
+        self.backup_dir = self._setup_backup_dir()
+        self.validator = BackupValidator()
+        self.manager = BackupManager()
+        self._lock_manager = BackupLockManager()
+
+    def _setup_backup_dir(self) -> Path:
+        """Create and return the backup directory."""
+        backup_dir = Path(self.settings.backup_dir)
+        backup_dir.mkdir(parents=True, exist_ok=True)
+        return backup_dir
+
+    def _parse_db_url(self) -> Dict[str, str]:
+        """Parse database URL into components."""
+        parsed = urlparse.urlparse(self.settings.db_url)
+
+        return {
+            "host": parsed.hostname or "localhost",
+            "port": str(parsed.port or 5432),
+            "user": parsed.username or "postgres",
+            "password": parsed.password or "",
+            "database": parsed.path.lstrip("/") or "postgres",
+        }
+
+    def _build_pg_dump_command(
+        self, backup_path: Path, backup_format: str
+    ) -> List[str]:
+        """Build pg_dump command for backup creation."""
[…]
