### 2025-11-24 29fd16a
WIP: standalone backup tooling phase 3
