+++ src/madsci_common/madsci/common/backup_tools/mongodb_backup.py
@@ -1,0 +1,513 @@
+"""Standalone MongoDB backup and restore tool."""
+
+import hashlib
+import json
+import shutil
+import subprocess
+import time
+from datetime import datetime
+from pathlib import Path
+from typing import List, Optional
+
+from madsci.client.event_client import EventClient
+from madsci.common.backup_tools.backup_manager import BackupManager
+from madsci.common.backup_tools.backup_validator import BackupValidator
+from madsci.common.backup_tools.base_backup import AbstractBackupTool, BackupInfo
+from madsci.common.types.backup_types import MongoDBBackupSettings
+from pymongo import MongoClient
+
+
+class MongoDBBackupTool(AbstractBackupTool):
+    """Standalone MongoDB backup and restore tool."""
+
+    def __init__(
+        self, settings: MongoDBBackupSettings, logger: Optional[EventClient] = None
+    ) -> None:
+        """
+        Initialize MongoDB backup tool.
+
+        Args:
+            settings: MongoDB backup configuration settings
+            logger: Optional logger instance
+        """
+        self.settings = settings
+        self.logger = logger or EventClient()
+        self.backup_dir = self._setup_backup_dir()
+        self.validator = BackupValidator()
+        self.manager = BackupManager()
+
+        # Initialize MongoDB connection
+        self.client = MongoClient(str(settings.mongo_db_url))
+        self.database = self.client[settings.database]
+
+    def _setup_backup_dir(self) -> Path:
+        """Set up backup directory, creating it if it doesn't exist."""
+        backup_dir = self.settings.backup_dir
+        if not backup_dir.is_absolute():
+            backup_dir = Path.cwd() / backup_dir
+
+        backup_dir.mkdir(parents=True, exist_ok=True)
+        return backup_dir
+
+    def create_backup(self, name_suffix: Optional[str] = None) -> Path:
+        """
+        Create a MongoDB backup using mongodump.
+
+        Args:
+            name_suffix: Optional suffix to add to backup name
+
+        Returns:
+            Path to the created backup directory
+
+        Raises:
+            RuntimeError: If backup creation fails
+        """
+        backup_path = self._generate_backup_path(name_suffix)
+        mongodump_cmd = self._build_mongodump_command(backup_path)
+
+        try:
+            self._execute_backup_command(mongodump_cmd, backup_path)
+            self._post_backup_processing(backup_path)
+            return backup_path
+
+        except subprocess.CalledProcessError as e:
+            self._cleanup_failed_backup(backup_path, e)
+            raise RuntimeError(f"Database backup failed: {e}") from e
+        except FileNotFoundError as fe:
+            raise RuntimeError(
+                f"mongodump command not found. Please ensure MongoDB tools are installed. {fe}"
+            ) from fe
+
+    def _generate_backup_path(self, name_suffix: Optional[str]) -> Path:
+        """Generate backup path with timestamp and optional suffix."""
+        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S_%f")
+        suffix = f"_{name_suffix}" if name_suffix else ""
+        backup_filename = f"{self.settings.database}_backup_{timestamp}{suffix}"
+        return self.backup_dir / backup_filename
+
+    def _build_mongodump_command(self, backup_path: Path) -> List[str]:
+        """Build the mongodump command with all necessary parameters."""
+        mongodump_cmd = ["mongodump"]
+
+        # Add connection parameters
+        parsed_url = self.settings.mongo_db_url
+        if parsed_url.host:
+            port = parsed_url.port or 27017
+            mongodump_cmd.extend(["--host", f"{parsed_url.host}:{port}"])
+
+        if parsed_url.username:
[…]
