+++ src/madsci_common/tests/test_mongodb_backup_tools/test_mongodb_backup_cli.py
@@ -1,0 +1,629 @@
+"""Test MongoDB backup CLI interface."""
+
+import json
+import tempfile
+from datetime import datetime
+from pathlib import Path
+from unittest.mock import Mock, patch
+
+import pytest
+from click.testing import CliRunner
+from madsci.common.backup_tools.base_backup import BackupInfo
+from madsci.common.backup_tools.cli import mongodb_backup
+
+
+class TestMongoDBBackupCLI:
+    """Test MongoDB backup CLI interface."""
+
+    @pytest.fixture
+    def temp_backup_dir(self):
+        """Create temporary backup directory for testing."""
+        with tempfile.TemporaryDirectory() as temp_dir:
+            yield Path(temp_dir)
+
+    @pytest.fixture
+    def cli_runner(self):
+        """Create Click CLI runner."""
+        return CliRunner()
+
+    @pytest.fixture
+    def mock_backup_tool(self, temp_backup_dir):
+        """Create mock backup tool for CLI testing."""
+        mock_tool = Mock()
+        mock_tool.create_backup.return_value = temp_backup_dir / "test_backup"
+        mock_tool.list_available_backups.return_value = []
+        mock_tool.validate_backup_integrity.return_value = True
+        mock_tool.restore_from_backup.return_value = None
+        mock_tool.delete_backup.return_value = None
+        return mock_tool
+
+    def test_cli_create_backup(self, cli_runner, temp_backup_dir, mock_backup_tool):
+        """Test CLI backup creation command."""
+
+        with patch(
+            "madsci.common.backup_tools.cli.MongoDBBackupTool"
+        ) as mock_tool_class:
+            mock_tool_class.return_value = mock_backup_tool
+            mock_backup_tool.create_backup.return_value = Path(
+                str(temp_backup_dir / "test_backup")
+            )
+
+            result = cli_runner.invoke(
+                mongodb_backup,
+                [
+                    "--mongo-url",
+                    "mongodb://localhost:27017/",
+                    "--database",
+                    "test_db",
+                    "--backup-dir",
+                    str(temp_backup_dir),
+                    "create",
+                ],
+            )
+
+            assert result.exit_code == 0
+            assert "Backup created successfully" in result.output
+            mock_backup_tool.create_backup.assert_called_once_with(None)
+
+    def test_cli_create_backup_with_name_suffix(self, cli_runner, mock_backup_tool):
+        """Test CLI backup creation with custom name suffix."""
+
+        with patch(
+            "madsci.common.backup_tools.cli.MongoDBBackupTool"
+        ) as mock_tool_class:
+            mock_tool_class.return_value = mock_backup_tool
+
+            result = cli_runner.invoke(
+                mongodb_backup,
+                [
+                    "--mongo-url",
+                    "mongodb://localhost:27017/",
+                    "--database",
+                    "test_db",
+                    "create",
+                    "--name-suffix",
+                    "pre_migration",
+                ],
+            )
+
+            assert result.exit_code == 0
+            mock_backup_tool.create_backup.assert_called_once_with("pre_migration")
+
+    def test_cli_create_backup_no_validation(self, cli_runner, mock_backup_tool):
+        """Test CLI backup creation with validation disabled."""
+
+        with (
+            patch(
+                "madsci.common.backup_tools.cli.MongoDBBackupTool"
+            ) as mock_tool_class,
+            patch(
+                "madsci.common.backup_tools.cli.MongoDBBackupSettings"
+            ) as mock_settings,
+        ):
+            mock_tool_class.return_value = mock_backup_tool
+
+            result = cli_runner.invoke(
+                mongodb_backup,
+                [
+                    "--mongo-url",
+                    "mongodb://localhost:27017/",
+                    "--database",
+                    "test_db",
+                    "create",
+                    "--no-validate",
+                ],
[…]
