+++ src/madsci_common/madsci/common/backup_tools/__init__.py
@@ -3,5 +3,14 @@
 from .backup_manager import BackupManager
 from .backup_validator import BackupValidator
 from .base_backup import AbstractBackupTool, BackupInfo
+from .cli import main_mongodb_backup
+from .mongodb_backup import MongoDBBackupTool
 
-__all__ = ["AbstractBackupTool", "BackupInfo", "BackupManager", "BackupValidator"]
+__all__ = [
+    "AbstractBackupTool",
+    "BackupInfo",
+    "BackupManager",
+    "BackupValidator",
+    "MongoDBBackupTool",
+    "main_mongodb_backup",
+]

[…]
