### 2025-11-25 fac29ca
WIP: standalone backup tooling phase 4
