+++ src/madsci_workcell_manager/tests/test_workcell_migration.py
@@ -187,7 +187,7 @@ def test_workcell_backup_only_command(mock_migrator_class):
     """Test backup only command for workcells database"""
     mock_migrator = Mock()
     backup_dir = Path(tempfile.gettempdir()) / "madsci_workcells_backup"
-    mock_migrator.create_backup.return_value = backup_dir
+    mock_migrator.backup_tool.create_backup.return_value = backup_dir
     mock_migrator_class.return_value = mock_migrator
 
     with patch.dict(os.environ, {"MONGODB_URL": "mongodb://localhost:27017"}):
@@ -208,7 +208,7 @@ def test_workcell_backup_only_command(mock_migrator_class):
             os.chdir(original_cwd)
 
     # Verify only backup was called
-    mock_migrator.create_backup.assert_called_once()
+    mock_migrator.backup_tool.create_backup.assert_called_once()
     mock_migrator.run_migration.assert_not_called()
 
 
@@ -230,11 +230,9 @@ def test_workcell_backup_creation(mock_subprocess):
         )
         migrator = MongoDBMigrator(settings)
 
-        # Mock the validation methods to bypass filesystem checks
-        with (
-            patch.object(migrator, "_validate_backup_integrity", return_value=True),
-        ):
-            backup_path = migrator.create_backup()
+        # Mock the post-backup processing to bypass filesystem checks
+        with patch.object(migrator.backup_tool, "_post_backup_processing"):
+            backup_path = migrator.backup_tool.create_backup()
 
             # Verify backup path format
             assert "madsci_workcells_backup_" in backup_path.name
@@ -427,7 +425,9 @@ def test_workcell_restore_command(mock_migrator_class):
             os.chdir(original_cwd)
 
     # Verify restore was called with correct path
-    mock_migrator.restore_from_backup.assert_called_once_with(Path("workcells_backup"))
+    mock_migrator.backup_tool.restore_from_backup.assert_called_once_with(
+        Path("workcells_backup")
+    )
     mock_migrator.run_migration.assert_not_called()
 
 
