### 2025-11-25 1c3881b
Unify CLI and add documentation
