+++ src/madsci_common/madsci/common/backup_tools/unified_cli.py
@@ -1,0 +1,385 @@
+"""Unified CLI for MADSci backup operations across all database types."""
+
+import json
+import sys
+from datetime import datetime
+from pathlib import Path
+from typing import Literal, Optional
+
+import click
+from pydantic import AnyUrl
+
+# Import backup tools - will be created in subsequent tasks
+try:
+    from .encryption import BackupEncryption  # type: ignore[attr-defined]
+except ImportError:
+    BackupEncryption = None  # type: ignore[assignment, misc]
+
+try:
+    from .registry import BackupRegistry  # type: ignore[attr-defined]
+except ImportError:
+    BackupRegistry = None  # type: ignore[assignment, misc]
+
+from .mongodb_backup import MongoDBBackupSettings, MongoDBBackupTool
+
+# Import PostgreSQL tools from resource_manager package
+try:
+    from madsci.resource_manager.backup_tools.postgres_backup import (
+        PostgreSQLBackupSettings,
+        PostgreSQLBackupTool,
+    )
+except ImportError:
+    # Fallback for testing or when resource_manager not installed
+    PostgreSQLBackupSettings = None  # type: ignore[assignment, misc]
+    PostgreSQLBackupTool = None  # type: ignore[assignment, misc]
+
+
+def detect_database_type(db_url: str) -> Literal["postgresql", "mongodb"]:
+    """Auto-detect database type from connection URL.
+
+    Args:
+        db_url: Database connection URL
+
+    Returns:
+        Database type: "postgresql" or "mongodb"
+
+    Raises:
+        ValueError: If database type cannot be detected
+    """
+    db_url_lower = db_url.lower()
+
+    if db_url_lower.startswith(("postgresql://", "postgres://")):
+        return "postgresql"
+    if db_url_lower.startswith(("mongodb://", "mongodb+srv://")):
+        return "mongodb"
+    raise ValueError(
+        f"Unable to detect database type from URL: {db_url}. "
+        "Supported prefixes: postgresql://, postgres://, mongodb://, mongodb+srv://"
+    )
+
+
+def load_config(config_path: str) -> dict:
+    """Load configuration from JSON file.
+
+    Args:
+        config_path: Path to configuration file
+
+    Returns:
+        Configuration dictionary
+    """
+    with Path(config_path).open() as f:
+        return json.load(f)
+
+
+@click.group()
+@click.option(
+    "--config", type=click.Path(exists=True), help="Configuration file (JSON)"
+)
+@click.pass_context
+def madsci_backup(ctx: click.Context, config: Optional[str]) -> None:
+    """MADSci unified backup management tool.
+
+    Supports PostgreSQL and MongoDB backups with automatic database type detection.
+    """
+    ctx.ensure_object(dict)
+    if config:
+        ctx.obj["config"] = load_config(config)
+
+
+@madsci_backup.command()
+@click.option("--db-url", required=True, help="Database connection URL")
+@click.option("--backup-dir", default=".madsci/backups", help="Backup directory")
+@click.option(
+    "--type",
+    "db_type",
+    type=click.Choice(["postgresql", "mongodb"]),
+    help="Database type (auto-detected if omitted)",
+)
+@click.option("--name", help="Backup name suffix")
+@click.option("--encrypt/--no-encrypt", default=False, help="Enable encryption")
+@click.option("--encrypt-key", type=click.Path(exists=True), help="Encryption key file")
+def create(
+    db_url: str,
+    backup_dir: str,
+    db_type: Optional[str],
+    name: Optional[str],
+    encrypt: bool,
+    encrypt_key: Optional[str],
+) -> None:
+    """Create a new database backup.
+
+    Examples:
+        madsci-backup create --db-url postgresql://localhost/mydb
+
+        madsci-backup create --db-url mongodb://localhost/mydb --name pre-deploy
+
[…]
