+++ src/madsci_common/tests/test_backup_tools/test_unified_cli.py
@@ -1,0 +1,476 @@
+"""Tests for unified backup CLI."""
+
+import json
+import tempfile
+from pathlib import Path
+from unittest.mock import MagicMock, patch
+
+import pytest
+from click.testing import CliRunner
+from madsci.common.backup_tools.unified_cli import (
+    detect_database_type,
+    madsci_backup,
+)
+
+
+class TestDatabaseTypeDetection:
+    """Test database type auto-detection."""
+
+    def test_detect_postgresql_standard_url(self) -> None:
+        """Test detection of PostgreSQL from standard URL."""
+        assert (
+            detect_database_type("postgresql://user:pass@localhost/db") == "postgresql"
+        )
+
+    def test_detect_postgresql_short_url(self) -> None:
+        """Test detection of PostgreSQL from postgres:// URL."""
+        assert detect_database_type("postgres://user:pass@localhost/db") == "postgresql"
+
+    def test_detect_mongodb_standard_url(self) -> None:
+        """Test detection of MongoDB from standard URL."""
+        assert detect_database_type("mongodb://localhost:27017/db") == "mongodb"
+
+    def test_detect_mongodb_srv_url(self) -> None:
+        """Test detection of MongoDB from SRV URL."""
+        assert detect_database_type("mongodb+srv://cluster.mongodb.net/db") == "mongodb"
+
+    def test_detect_invalid_url_raises_error(self) -> None:
+        """Test invalid URL raises ValueError."""
+        with pytest.raises(ValueError, match="Unable to detect database type"):
+            detect_database_type("mysql://localhost/db")
+
+
+class TestUnifiedCLICreate:
+    """Test unified CLI create command."""
+
+    @pytest.fixture
+    def runner(self) -> CliRunner:
+        """Create Click CLI test runner."""
+        return CliRunner()
+
+    def test_create_postgres_backup_auto_detect(self, runner: CliRunner) -> None:
+        """Test creating PostgreSQL backup with auto-detection."""
+        with patch(
+            "madsci.common.backup_tools.unified_cli.PostgreSQLBackupTool"
+        ) as mock_tool:
+            mock_instance = MagicMock()
+            mock_instance.create_backup.return_value = Path("/test/backup.dump")
+            mock_tool.return_value = mock_instance
+
+            result = runner.invoke(
+                madsci_backup, ["create", "--db-url", "postgresql://localhost/test"]
+            )
+
+            assert result.exit_code == 0
+            assert "Backup created:" in result.output
+            mock_tool.assert_called_once()
+            mock_instance.create_backup.assert_called_once()
+
+    def test_create_mongodb_backup_auto_detect(self, runner: CliRunner) -> None:
+        """Test creating MongoDB backup with auto-detection."""
+        with patch(
+            "madsci.common.backup_tools.unified_cli.MongoDBBackupTool"
+        ) as mock_tool:
+            mock_instance = MagicMock()
+            mock_instance.create_backup.return_value = Path("/test/backup")
+            mock_tool.return_value = mock_instance
+
+            result = runner.invoke(
+                madsci_backup, ["create", "--db-url", "mongodb://localhost/testdb"]
+            )
+
+            assert result.exit_code == 0
+            assert "Backup created:" in result.output
+            mock_tool.assert_called_once()
+            mock_instance.create_backup.assert_called_once()
+
+    def test_create_with_explicit_type(self, runner: CliRunner) -> None:
+        """Test creating backup with explicit database type."""
+        with patch(
+            "madsci.common.backup_tools.unified_cli.PostgreSQLBackupTool"
+        ) as mock_tool:
+            mock_instance = MagicMock()
+            mock_instance.create_backup.return_value = Path("/test/backup.dump")
+            mock_tool.return_value = mock_instance
+
[…]
