+++ src/madsci_resource_manager/tests/test_resource_backup_tools/__init__.py
@@ -1,1 +1,0 @@
-"""Test package for backup tools."""
