+++ src/madsci_common/madsci/common/backup_tools/__init__.py
@@ -5,6 +5,8 @@ from .backup_validator import BackupValidator
 from .base_backup import AbstractBackupTool, BackupInfo
 from .cli import main_mongodb_backup
 from .mongodb_backup import MongoDBBackupTool
+from .postgres_backup import PostgreSQLBackupTool
+from .postgres_cli import main_postgres_backup
 
 __all__ = [
     "AbstractBackupTool",
@@ -12,5 +14,7 @@ __all__ = [
     "BackupManager",
     "BackupValidator",
     "MongoDBBackupTool",
+    "PostgreSQLBackupTool",
     "main_mongodb_backup",
+    "main_postgres_backup",
 ]
