+++ src/madsci_resource_manager/madsci/resource_manager/backup_tools/__init__.py
@@ -1,6 +1,0 @@
-"""Database backup tools for MADSci Resource Manager."""
-
-from .cli import main_postgres_backup, postgres_backup
-from .postgres_backup import PostgreSQLBackupTool
-
-__all__ = ["PostgreSQLBackupTool", "main_postgres_backup", "postgres_backup"]
