+++ src/madsci_common/tests/test_postgres_backup_tools/test_postgres_backup_cli.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 from click.testing import CliRunner
-from madsci.resource_manager.backup_tools.cli import (
+from madsci.common.backup_tools.postgres_cli import (
     main_postgres_backup,
     postgres_backup,
 )
@@ -38,7 +38,7 @@ class TestPostgreSQLBackupCLI:
         assert result.exit_code == 0
         assert "PostgreSQL backup management commands" in result.output
 
-    @patch("madsci.resource_manager.backup_tools.cli.PostgreSQLBackupTool")
+    @patch("madsci.common.backup_tools.postgres_cli.PostgreSQLBackupTool")
     def test_cli_create_backup(
         self,
         mock_tool_class: Mock,
@@ -72,7 +72,7 @@ class TestPostgreSQLBackupCLI:
         # Verify backup creation was called
         mock_tool.create_backup.assert_called_once_with(None)
 
-    @patch("madsci.resource_manager.backup_tools.cli.PostgreSQLBackupTool")
+    @patch("madsci.common.backup_tools.postgres_cli.PostgreSQLBackupTool")
     def test_cli_create_backup_with_name_suffix(
         self,
         mock_tool_class: Mock,
@@ -107,7 +107,7 @@ class TestPostgreSQLBackupCLI:
         # Verify backup creation was called with suffix
         mock_tool.create_backup.assert_called_once_with("pre_migration")
 
-    @patch("madsci.resource_manager.backup_tools.cli.PostgreSQLBackupTool")
+    @patch("madsci.common.backup_tools.postgres_cli.PostgreSQLBackupTool")
     def test_cli_create_backup_no_validate(
         self,
         mock_tool_class: Mock,
@@ -141,7 +141,7 @@ class TestPostgreSQLBackupCLI:
         settings = mock_tool_class.call_args[0][0]
         assert settings.validate_integrity is False
 
-    @patch("madsci.resource_manager.backup_tools.cli.PostgreSQLBackupTool")
+    @patch("madsci.common.backup_tools.postgres_cli.PostgreSQLBackupTool")
     def test_cli_create_backup_different_format(
         self,
         mock_tool_class: Mock,
@@ -176,7 +176,7 @@ class TestPostgreSQLBackupCLI:
         settings = mock_tool_class.call_args[0][0]
         assert settings.backup_format == "plain"
 
-    @patch("madsci.resource_manager.backup_tools.cli.PostgreSQLBackupTool")
+    @patch("madsci.common.backup_tools.postgres_cli.PostgreSQLBackupTool")
     def test_cli_create_backup_failure(
         self,
         mock_tool_class: Mock,
@@ -199,7 +199,7 @@ class TestPostgreSQLBackupCLI:
         assert result.exit_code == 1
         assert "Backup failed: Backup failed" in result.output
 
-    @patch("madsci.resource_manager.backup_tools.cli.PostgreSQLBackupTool")
+    @patch("madsci.common.backup_tools.postgres_cli.PostgreSQLBackupTool")
     def test_cli_list_backups(
         self,
         mock_tool_class: Mock,
@@ -228,7 +228,7 @@ class TestPostgreSQLBackupCLI:
         assert "2024-01-01 12:00:00" in result.output
         assert "1024" in result.output
 
-    @patch("madsci.resource_manager.backup_tools.cli.PostgreSQLBackupTool")
+    @patch("madsci.common.backup_tools.postgres_cli.PostgreSQLBackupTool")
     def test_cli_list_backups_empty(
         self,
         mock_tool_class: Mock,
@@ -251,7 +251,7 @@ class TestPostgreSQLBackupCLI:
[…]
