### 2025-11-29 e06c413
Minor edits/fixes to node status display/cancel display. WIP liquidhandler cancel
