+++ example_lab/default.node.yaml
@@ -1,5 +1,5 @@
 node_name: example_app
-node_id: 01KAEKF82701HZJCGXQQESA81G
+node_id: 01KB0X8HCE8H9Y9H5FQKY1TKZ7
 node_description: null
 node_type: device
 module_name: example_app
