### 2025-11-30 71f807b
Final: backup cli tooling separated from migration tool
