+++ src/madsci_common/tests/test_mongodb_backup_tools/test_mongodb_backup_tool.py
@@ -77,29 +77,42 @@ class TestMongoDBBackupTool:
 
     def test_create_backup_all_collections_success(self, backup_tool, temp_backup_dir):
         """Test successful backup creation for all collections."""
+
+        def mock_mongodump(*args, **_kwargs):
+            """Mock mongodump by creating expected directory structure."""
+            # Extract backup path from mongodump command
+            cmd = args[0]
+            out_index = cmd.index("--out") + 1
+            backup_path = Path(cmd[out_index])
+            db_name = cmd[cmd.index("--db") + 1]
+
+            # Create the directory structure that mongodump would create
+            db_backup_path = backup_path / db_name
+            db_backup_path.mkdir(parents=True, exist_ok=True)
+
+            # Create a mock collection file
+            (db_backup_path / "collection1.bson").touch()
+            (db_backup_path / "collection1.metadata.json").touch()
+
+            mock_result = Mock()
+            mock_result.returncode = 0
+            mock_result.stderr = ""
+            return mock_result
+
         # Mock subprocess.run for mongodump
         with (
-            patch("subprocess.run") as mock_run,
+            patch("subprocess.run", side_effect=mock_mongodump) as mock_run,
             patch(
                 "madsci.common.backup_tools.mongodb_backup.datetime"
             ) as mock_datetime,
         ):
-            mock_run.return_value.returncode = 0
-            mock_run.return_value.stderr = ""
-
             # Mock datetime to get predictable backup path
             mock_now = datetime(2024, 1, 1, 12, 0, 0, 0)
             mock_datetime.now.return_value = mock_now
             mock_datetime.strftime = mock_now.strftime
 
-            # Expected backup path based on mocked timestamp
-            (temp_backup_dir / "test_events_backup_20240101_120000_000000")
-
             # Mock the backup tool's internal methods
             with (
-                patch.object(
-                    backup_tool, "_verify_backup_completion", return_value=True
-                ),
                 patch.object(
                     backup_tool,
                     "_generate_backup_checksum",
@@ -122,30 +135,51 @@ class TestMongoDBBackupTool:
                 assert result_path.parent == temp_backup_dir
 
     def test_create_backup_specific_collections(
-        self, mongodb_backup_settings, temp_backup_dir
+        self,
+        mongodb_backup_settings,
+        temp_backup_dir,  # noqa: ARG002
     ):
         """Test backup creation for specific collections only."""
         # Update settings to specify collections
         mongodb_backup_settings.collections = ["collection1"]
 
-        with (
-            patch("madsci.common.backup_tools.mongodb_backup.MongoClient"),
-            patch("subprocess.run") as mock_run,
-        ):
-            mock_run.return_value.returncode = 0
-            mock_run.return_value.stderr = ""
+        def mock_mongodump(*args, **_kwargs):
+            """Mock mongodump by creating expected directory structure."""
+            # Extract backup path from mongodump command
+            cmd = args[0]
+            out_index = cmd.index("--out") + 1
+            backup_path = Path(cmd[out_index])
+            db_name = cmd[cmd.index("--db") + 1]
 
+            # Create the directory structure that mongodump would create
+            db_backup_path = backup_path / db_name
+            db_backup_path.mkdir(parents=True, exist_ok=True)
+
+            # Create a mock collection file for the specified collection
+            (db_backup_path / "collection1.bson").touch()
+            (db_backup_path / "collection1.metadata.json").touch()
+
+            mock_result = Mock()
+            mock_result.returncode = 0
+            mock_result.stderr = ""
+            return mock_result
+
+        # Create a mock MongoDB client
+        mock_client = Mock()
+        mock_database = Mock()
+        mock_client.__getitem__ = Mock(return_value=mock_database)
[…]
