### 2025-12-02 585be52
Remove references to not-implemented features
