+++ src/madsci_common/madsci/common/backup_tools/unified_cli.py
@@ -2,25 +2,12 @@
 
 import json
 import sys
-from datetime import datetime
 from pathlib import Path
 from typing import Literal, Optional
 
 import click
-from pydantic import AnyUrl
-
-# Import backup tools - will be created in subsequent tasks
-try:
-    from .encryption import BackupEncryption  # type: ignore[attr-defined]
-except ImportError:
-    BackupEncryption = None  # type: ignore[assignment, misc]
-
-try:
-    from .registry import BackupRegistry  # type: ignore[attr-defined]
-except ImportError:
-    BackupRegistry = None  # type: ignore[assignment, misc]
-
 from madsci.common.types.backup_types import PostgreSQLBackupSettings
+from pydantic import AnyUrl
 
 from .mongodb_backup import MongoDBBackupSettings, MongoDBBackupTool
 from .postgres_backup import PostgreSQLBackupTool
@@ -88,15 +75,11 @@ def madsci_backup(ctx: click.Context, config: Optional[str]) -> None:
     help="Database type (auto-detected if omitted)",
 )
 @click.option("--name", help="Backup name suffix")
-@click.option("--encrypt/--no-encrypt", default=False, help="Enable encryption")
-@click.option("--encrypt-key", type=click.Path(exists=True), help="Encryption key file")
 def create(
     db_url: str,
     backup_dir: str,
     db_type: Optional[str],
     name: Optional[str],
-    encrypt: bool,
-    encrypt_key: Optional[str],
 ) -> None:
     """Create a new database backup.
 
@@ -104,8 +87,6 @@ def create(
         madsci-backup create --db-url postgresql://localhost/mydb
 
         madsci-backup create --db-url mongodb://localhost/mydb --name pre-deploy
-
-        madsci-backup create --db-url postgresql://localhost/mydb --encrypt --encrypt-key ~/.madsci/key
     """
     try:
         # Auto-detect database type if not specified
@@ -144,89 +125,10 @@ def create(
             tool = MongoDBBackupTool(settings)
             backup_path = tool.create_backup(name)
 
-        # Apply encryption if requested
-        if encrypt:
-            if not BackupEncryption:
-                click.echo(
-                    "ERROR: Encryption module not available. "
-                    "Please install with: pip install madsci-common[encryption]",
-                    err=True,
-                )
-                sys.exit(1)
-
-            if not encrypt_key:
-                click.echo(
-                    "ERROR: --encrypt-key required when --encrypt is enabled", err=True
-                )
-                sys.exit(1)
-
-            encryptor = BackupEncryption(Path(encrypt_key))
-            encrypted_path = encryptor.encrypt_backup(backup_path)
-            click.echo(f"✓ Backup created and encrypted: {encrypted_path}")
-        else:
-            click.echo(f"✓ Backup created: {backup_path}")
+        click.echo(f"✓ Backup created: {backup_path}")
 
     except Exception as e:
         click.echo(f"✗ Backup failed: {e}", err=True)
-        sys.exit(1)
-
-
-@madsci_backup.command()
-@click.option(
-    "--type",
-    "db_type",
-    type=click.Choice(["postgresql", "mongodb"]),
-    help="Filter by database type",
-)
-@click.option("--from", "date_from", help="Filter from date (YYYY-MM-DD)")
-@click.option("--to", "date_to", help="Filter to date (YYYY-MM-DD)")
-def list(
-    db_type: Optional[str], date_from: Optional[str], date_to: Optional[str]
-) -> None:
-    """List available backups.
-
-    Examples:
-        madsci-backup list
-
-        madsci-backup list --type postgresql
-
-        madsci-backup list --from 2024-01-01 --to 2024-01-31
-    """
-    try:
-        if not BackupRegistry:
-            click.echo(
-                "ERROR: Backup registry not available. "
[…]
