### 2025-12-02 6ae09e5
Remove unneccessary check
