### 2025-12-02 f4403c7
Fix default values for db urls
