### 2025-12-02 44c8c99
Address review comments; add rate limits to rest node
