+++ src/madsci_common/madsci/common/middleware.py
@@ -5,6 +5,7 @@ This module provides middleware for enhancing server resilience,
 including rate limiting and request tracking.
 """
 
+import asyncio
 import time
 from collections import defaultdict
 from typing import Callable
@@ -22,10 +23,18 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
     rate limits based on a sliding window algorithm. When a client
     exceeds the rate limit, a 429 Too Many Requests response is returned.
 
+    Thread-safe implementation using asyncio locks to prevent race conditions
+    in concurrent request handling. Includes periodic cleanup to prevent
+    memory leaks from inactive client IPs.
+
     Attributes:
         requests_limit: Maximum number of requests allowed per time window
         time_window: Time window in seconds for rate limiting
         storage: Dictionary tracking request timestamps per client IP
+        locks: Dictionary of locks for thread-safe access per client IP
+        global_lock: Lock for managing the locks dictionary itself
+        cleanup_interval: Interval in seconds between cleanup operations
+        last_cleanup: Timestamp of last cleanup operation
     """
 
     def __init__(
@@ -33,6 +42,7 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
         app: Callable,
         requests_limit: int = 100,
         time_window: int = 60,
+        cleanup_interval: int = 300,
     ) -> None:
         """
         Initialize the rate limiting middleware.
@@ -41,16 +51,68 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
             app: The ASGI application
             requests_limit: Maximum number of requests allowed per time window
             time_window: Time window in seconds for rate limiting
+            cleanup_interval: Interval in seconds between cleanup operations (default: 300)
         """
         super().__init__(app)
         self.requests_limit = requests_limit
         self.time_window = time_window
+        self.cleanup_interval = cleanup_interval
         # Storage maps client IP addresses to lists of request timestamps
         self.storage: defaultdict[str, list[float]] = defaultdict(list)
+        # Per-client locks for thread-safe access
+        self.locks: dict[str, asyncio.Lock] = {}
+        # Global lock for managing the locks dictionary
+        self.global_lock = asyncio.Lock()
+        # Track last cleanup time
+        self.last_cleanup = time.time()
+
+    async def _get_client_lock(self, client_ip: str) -> asyncio.Lock:
+        """
+        Get or create a lock for a specific client IP.
+
+        Args:
+            client_ip: The client IP address
+
+        Returns:
+            asyncio.Lock: The lock for this client
+        """
+        # Use global lock to safely manage the locks dictionary
+        async with self.global_lock:
+            if client_ip not in self.locks:
+                self.locks[client_ip] = asyncio.Lock()
+            return self.locks[client_ip]
+
+    async def _cleanup_inactive_clients(self, now: float) -> None:
+        """
+        Remove inactive client IPs from storage to prevent memory leaks.
+
+        This method removes clients whose last request was more than
+        time_window seconds ago, as they no longer need rate limiting tracking.
+
+        Args:
+            now: Current timestamp
+        """
+        async with self.global_lock:
+            cutoff = now - self.time_window
+            inactive_clients = [
+                client_ip
+                for client_ip, timestamps in self.storage.items()
+                if not timestamps or timestamps[-1] < cutoff
+            ]
+
+            for client_ip in inactive_clients:
+                del self.storage[client_ip]
+                if client_ip in self.locks:
+                    del self.locks[client_ip]
+
+            self.last_cleanup = now
 
     async def dispatch(self, request: Request, call_next: Callable) -> Response:
         """
         Process each request and enforce rate limiting.
+
+        Thread-safe implementation that uses per-client locks to prevent
[…]
