### 2025-12-02 d225965
Remove vestigial test assertions
