### 2025-12-02 52ac75a
Switch to async and update docs/tests to match
