### 2025-12-03 a8694de
Cancellable decorator added to liquidhandler. @cancellable actions can now be cancelled via node modal
