### 2025-12-03 f08759d
Test WF and experiment to test admin commands in example lab. Can cancel WF successfully by cancelling action via node modal!
