+++ example_lab/test_admin_commands.py
@@ -1,0 +1,33 @@
+"""Test Experiment for Workflow Admin Commands"""
+
+from madsci.common.types.experiment_types import ExperimentDesign
+from madsci.common.types.node_types import NodeDefinition
+from madsci.experiment_application import (
+    ExperimentApplication,
+    ExperimentApplicationConfig,
+)
+from pydantic import AnyUrl
+
+class ExampleApp(ExperimentApplication):
+    """An Example Application"""
+
+    experiment_design = ExperimentDesign(
+        experiment_name="Example_App",
+    )
+    config = ExperimentApplicationConfig(node_url=AnyUrl("http://localhost:6000"))
+
+    def run_experiment(self) -> str:
+        """main experiment function"""
+
+        workflow = self.workcell_client.submit_workflow(
+            "./workflows/admin_commands_test.workflow.yaml",
+        )
+        return "test"
+
+if __name__ == "__main__":
+    app = ExampleApp(
+        node_definition=NodeDefinition(
+            node_name="example_app", module_name="example_app"
+        )
+    )
+    app.start_app()
