### 2025-12-03 f725268
Cancel button added to WF modal, seems to work in simplest case (running test admin command WF/experiment).
