### 2025-12-03 c4b12ce
Cancelling a WF now retries until reaches a cancellable node, or timeout. Seems to work!
