### 2025-12-04 4c87e19
Fix some cli naming issues for clarity
