+++ src/madsci_common/pyproject.toml
@@ -36,7 +36,7 @@ dependencies = [
 ]
 
 [project.scripts]
-madsci-backup = "madsci.common.backup_tools.unified_cli:main"
+madsci-backup = "madsci.common.backup_tools.cli:main"
 madsci-mongodb-backup = "madsci.common.backup_tools.cli:main_mongodb_backup"
 madsci-postgres-backup = "madsci.common.backup_tools.postgres_cli:main_postgres_backup"
 

[…]
