### 2025-12-04 9a06761
Add burst protection via rate limits
