+++ src/madsci_common/madsci/common/middleware.py
@@ -23,13 +23,18 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
     rate limits based on a sliding window algorithm. When a client
     exceeds the rate limit, a 429 Too Many Requests response is returned.
 
+    Supports dual rate limiting with both short (burst) and long (sustained)
+    windows. Both limits must be satisfied for a request to proceed.
+
     Async-safe implementation using asyncio locks to prevent race conditions
     in concurrent coroutine handling. Includes periodic cleanup to prevent
     memory leaks from inactive client IPs.
 
     Attributes:
-        requests_limit: Maximum number of requests allowed per time window
-        time_window: Time window in seconds for rate limiting
+        requests_limit: Maximum number of requests allowed per long time window
+        time_window: Long time window in seconds for rate limiting
+        short_requests_limit: Maximum number of requests allowed per short time window (optional)
+        short_time_window: Short time window in seconds for burst protection (optional)
         storage: Dictionary tracking request timestamps per client IP
         locks: Dictionary of locks for thread-safe access per client IP
         global_lock: Lock for managing the locks dictionary itself
@@ -42,6 +47,8 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
         app: Callable,
         requests_limit: int = 100,
         time_window: int = 60,
+        short_requests_limit: int | None = None,
+        short_time_window: int | None = None,
         cleanup_interval: int = 300,
     ) -> None:
         """
@@ -49,13 +56,17 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
 
         Args:
             app: The ASGI application
-            requests_limit: Maximum number of requests allowed per time window
-            time_window: Time window in seconds for rate limiting
+            requests_limit: Maximum number of requests allowed per long time window
+            time_window: Long time window in seconds for rate limiting
+            short_requests_limit: Maximum number of requests per short window (optional, for burst protection)
+            short_time_window: Short time window in seconds (optional, typically 1 second)
             cleanup_interval: Interval in seconds between cleanup operations (default: 300)
         """
         super().__init__(app)
         self.requests_limit = requests_limit
         self.time_window = time_window
+        self.short_requests_limit = short_requests_limit
+        self.short_time_window = short_time_window
         self.cleanup_interval = cleanup_interval
         # Storage maps client IP addresses to lists of request timestamps
         self.storage: defaultdict[str, list[float]] = defaultdict(list)
@@ -111,8 +122,11 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
         """
         Process each request and enforce rate limiting.
 
-        Thread-safe implementation that uses per-client locks to prevent
+        Async-safe implementation that uses per-client locks to prevent
         race conditions in concurrent request handling.
+
+        For dual rate limiting, both short and long window limits must be
+        satisfied for a request to proceed.
 
         Args:
             request: The incoming HTTP request
@@ -134,23 +148,61 @@ class RateLimitMiddleware(BaseHTTPMiddleware):
         # Get the lock for this client IP
         client_lock = await self._get_client_lock(client_ip)
 
-        # Use client-specific lock to ensure thread-safe access
+        # Use client-specific lock to ensure async-safe access
         async with client_lock:
-            # Clean up old timestamps outside the current window
+            # Clean up old timestamps outside the longest window
             cutoff = now - self.time_window
             self.storage[client_ip] = [
                 ts for ts in self.storage[client_ip] if ts > cutoff
             ]
 
-            # Check if client has exceeded rate limit
+            # Check short window limit (burst protection) if configured
+            if (
+                self.short_requests_limit is not None
+                and self.short_time_window is not None
+            ):
+                short_cutoff = now - self.short_time_window
+                short_window_requests = [
[…]
