+++ src/madsci_common/tests/test_middleware.py
@@ -57,6 +57,20 @@ def test_manager_without_rate_limiting() -> TestManager:
 
 
 @pytest.fixture
+def test_manager_with_dual_rate_limiting() -> TestManager:
+    """Create a test manager instance with dual rate limiting enabled."""
+    settings = TestManagerSettings(
+        rate_limit_enabled=True,
+        rate_limit_requests=20,  # Long window: 20 requests per 10 seconds
+        rate_limit_window=10,
+        rate_limit_short_requests=5,  # Short window: 5 requests per 1 second (burst) - intentionally low for testing
+        rate_limit_short_window=1,
+    )
+    definition = TestManagerDefinition(name="Dual Rate Limited Manager")
+    return TestManager(settings=settings, definition=definition)
+
+
+@pytest.fixture
 def rate_limited_client(test_manager_with_rate_limiting: TestManager) -> TestClient:
     """Create a test client with rate limiting."""
     app = test_manager_with_rate_limiting.create_server()
@@ -67,6 +81,15 @@ def rate_limited_client(test_manager_with_rate_limiting: TestManager) -> TestCli
 def unlimited_client(test_manager_without_rate_limiting: TestManager) -> TestClient:
     """Create a test client without rate limiting."""
     app = test_manager_without_rate_limiting.create_server()
+    return TestClient(app)
+
+
+@pytest.fixture
+def dual_rate_limited_client(
+    test_manager_with_dual_rate_limiting: TestManager,
+) -> TestClient:
+    """Create a test client with dual rate limiting."""
+    app = test_manager_with_dual_rate_limiting.create_server()
     return TestClient(app)
 
 
@@ -225,3 +248,138 @@ def test_memory_leak_prevention(test_manager_with_rate_limiting: TestManager) ->
     assert final_size == 0, (
         f"Memory leak detected: storage should be 0 after cleanup, but is {final_size}"
     )
+
+
+def test_dual_rate_limiting_burst_protection(
+    dual_rate_limited_client: TestClient,
+) -> None:
+    """Test that burst limit (short window) prevents rapid request bursts."""
+    # Make 5 requests rapidly (at the burst limit: 5 per second)
+    for i in range(5):
+        response = dual_rate_limited_client.get("/health")
+        assert response.status_code == 200, f"Request {i + 1} should succeed"
+        # Check that burst headers are present
+        assert "X-RateLimit-Burst-Limit" in response.headers
+        assert response.headers["X-RateLimit-Burst-Limit"] == "5"
+        assert "X-RateLimit-Burst-Remaining" in response.headers
+
+    # The 6th rapid request should be rejected by burst limit
+    response = dual_rate_limited_client.get("/health")
+    assert response.status_code == 429
+    assert "burst limit" in response.json()["detail"]
+    assert "X-RateLimit-Burst-Limit" in response.headers
+    assert response.headers["X-RateLimit-Burst-Remaining"] == "0"
+
+
+def test_dual_rate_limiting_long_window_protection(
+    dual_rate_limited_client: TestClient,
+) -> None:
+    """Test that long window limit prevents sustained high load."""
+    # Make requests slowly to avoid burst limit (20 requests over ~4 seconds)
+    # Burst limit is 5/second, so we'll space them out
+    for i in range(20):
+        response = dual_rate_limited_client.get("/health")
+        assert response.status_code == 200, f"Request {i + 1} should succeed"
+        # Small delay to avoid burst limit (every 5th request, wait a bit longer)
+        if (i + 1) % 5 == 0:
+            time.sleep(1.1)  # Wait for burst window to reset
+
+    # The 21st request should be rejected by long window limit
+    response = dual_rate_limited_client.get("/health")
+    assert response.status_code == 429
+    # Should mention the long window (10 seconds), not burst
[…]
