### 2025-12-04 a5a0ed5
Remove | operator from type hint
