+++ README.md
@@ -6,7 +6,7 @@
 |---------------------------------------------------------------------------------------------- | -------: | -------: | ------: | --------: |
 | src/madsci\_client/madsci/client/\_\_init\_\_.py                                              |       15 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/data\_client.py                                              |      155 |       41 |     74% |72, 88-96, 106-107, 116-126, 175-183, 190, 233-234, 253-254, 314, 331, 347-354, 382-384, 431-441 |
-| src/madsci\_client/madsci/client/event\_client.py                                             |      266 |       30 |     89% |64, 88-89, 270-273, 281-284, 289, 347, 395-398, 407-410, 415, 440, 445-446, 455-456, 469, 496, 500-504 |
+| src/madsci\_client/madsci/client/event\_client.py                                             |      266 |       30 |     89% |64, 88-89, 270-273, 280-283, 288, 345, 393-396, 404-407, 412, 437, 442-443, 452-453, 466, 493, 497-501 |
 | src/madsci\_client/madsci/client/experiment\_client.py                                        |       56 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/lab\_client.py                                               |       32 |       19 |     41% |28-34, 40-43, 47-50, 54-57, 61-64 |
 | src/madsci\_client/madsci/client/location\_client.py                                          |      143 |       66 |     54% |57, 74, 91-93, 103, 113, 131-142, 160-171, 223-235, 255-266, 294-306, 364-376, 424-435, 463-482 |
@@ -15,41 +15,55 @@
 | src/madsci\_client/madsci/client/node/rest\_node\_client.py                                   |      205 |       44 |     79% |62, 118, 219-230, 250-260, 270-305, 320-323, 394, 412, 414, 428 |
 | src/madsci\_client/madsci/client/resource\_client.py                                          |      631 |      198 |     69% |101, 126, 148, 175-178, 186, 228-231, 248, 310-314, 327-342, 368, 400-403, 452-459, 465, 478, 489-490, 526, 552, 597-598, 633-635, 675-676, 713-714, 746, 778-779, 811-812, 844-845, 877-878, 907-908, 936-941, 969-971, 1002-1055, 1103-1113, 1139-1142, 1190-1200, 1217, 1220-1231, 1259-1266, 1286-1289, 1345-1363, 1378-1384, 1431-1457, 1499-1537, 1566-1583, 1667-1671, 1685, 1708-1709, 1717-1721, 1737-1738, 1749, 1783, 1796, 1809-1819 |
 | src/madsci\_client/madsci/client/workcell\_client.py                                          |      251 |       48 |     81% |122, 151, 179, 194, 243-247, 253, 275, 279, 303-308, 416-437, 464-492, 531, 539, 547, 552-553, 657, 686 |
+| src/madsci\_common/madsci/common/backup\_tools/\_\_init\_\_.py                                |        8 |        0 |    100% |           |
+| src/madsci\_common/madsci/common/backup\_tools/backup\_manager.py                             |      172 |       26 |     85% |63-64, 109-111, 133, 149-150, 159-160, 190, 231, 247, 284, 287-288, 304-305, 321-322, 351-356 |
+| src/madsci\_common/madsci/common/backup\_tools/backup\_validator.py                           |       90 |       11 |     88% |105-107, 176-178, 256, 269-273, 287 |
+| src/madsci\_common/madsci/common/backup\_tools/base\_backup.py                                |       25 |        0 |    100% |           |
[…]
