### 2025-12-05 8cb380c
Fix pre-commit; extend e2e test timeout
