### 2025-12-05 d7c7f6b
Update autodocs, rename for clarity
