+++ README.md
@@ -5,16 +5,16 @@
 | Name                                                                                          |    Stmts |     Miss |   Cover |   Missing |
 |---------------------------------------------------------------------------------------------- | -------: | -------: | ------: | --------: |
 | src/madsci\_client/madsci/client/\_\_init\_\_.py                                              |       15 |        0 |    100% |           |
-| src/madsci\_client/madsci/client/data\_client.py                                              |      155 |       41 |     74% |72, 88-96, 106-107, 116-126, 175-183, 190, 233-234, 253-254, 314, 331, 347-354, 382-384, 431-441 |
-| src/madsci\_client/madsci/client/event\_client.py                                             |      266 |       30 |     89% |64, 88-89, 270-273, 280-283, 288, 345, 393-396, 404-407, 412, 437, 442-443, 452-453, 466, 493, 497-501 |
-| src/madsci\_client/madsci/client/experiment\_client.py                                        |       56 |        0 |    100% |           |
-| src/madsci\_client/madsci/client/lab\_client.py                                               |       32 |       19 |     41% |28-34, 40-43, 47-50, 54-57, 61-64 |
-| src/madsci\_client/madsci/client/location\_client.py                                          |      143 |       66 |     54% |57, 74, 91-93, 103, 113, 131-142, 160-171, 223-235, 255-266, 294-306, 364-376, 424-435, 463-482 |
+| src/madsci\_client/madsci/client/data\_client.py                                              |      157 |       41 |     74% |85, 109-117, 127-128, 137-148, 214-224, 238, 287-288, 308-309, 371, 388, 404-411, 439-441, 488-498 |
+| src/madsci\_client/madsci/client/event\_client.py                                             |      267 |       30 |     89% |64, 91-92, 302-305, 313-316, 321, 379, 427-430, 439-442, 447, 472, 477-478, 487-488, 501, 528, 532-536 |
+| src/madsci\_client/madsci/client/experiment\_client.py                                        |       59 |        0 |    100% |           |
+| src/madsci\_client/madsci/client/lab\_client.py                                               |       35 |       21 |     40% |31-43, 52-58, 67-73, 82-88, 97-103 |
+| src/madsci\_client/madsci/client/location\_client.py                                          |       98 |       39 |     60% |45, 62, 76, 86, 104-112, 132-140, 189-198, 218-226, 254-263, 318-327, 374-382, 410-426 |
 | src/madsci\_client/madsci/client/node/\_\_init\_\_.py                                         |        4 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/node/abstract\_node\_client.py                               |       30 |        4 |     87% |27, 30, 100-101 |
-| src/madsci\_client/madsci/client/node/rest\_node\_client.py                                   |      205 |       44 |     79% |62, 118, 219-230, 250-260, 270-305, 320-323, 394, 412, 414, 428 |
[…]
