### 2025-12-05 #9416 @ktahar
Hi @LuckierDodge , thanks for response!
I was not aware of madsci_paper branch and was looking at main branch. I guess it's okay to review main branch.

Permission issue has been solved by setting `USER_ID` and `GROUP_ID`. Thanks and sorry for being a docker noob...
So I was able to go over the example readme and notebooks.

I've got a question regarding installation with pip.
11 packages are listed in [readme](https://github.com/AD-SDL/MADSci?tab=readme-ov-file#python-packages), but I can find only 9 in [PyPI](https://pypi.org/search/?q=madsci) - The location_manager and squid don't exist on PyPI.
Is the readme wrong?

Regarding "Community guidelines" bullet point, is [Developer Guide](https://github.com/AD-SDL/MADSci?tab=readme-ov-file#developer-guide) in README the corresponding document? I can guess we may use GitHub issue/PR to contribute, report problem, or seek support. But I think it can be stated clearly.
