### 2025-12-05 417dd12
Publish madsci.location_manager
