### 2025-12-05 3bfd922
Bump to v0.6.0-rc.1
