+++ src/madsci_client/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "madsci.client"
-version = "0.5.5"
+version = "0.6.0-rc.1"
 description = "The Modular Autonomous Discovery for Science (MADSci) Python Clients."
 authors = [
     {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
