### 2025-12-05 edb9818
Publish madsci.squid
