+++ .github/workflows/pypi.yml
@@ -33,3 +33,4 @@ jobs:
       - run: pdm publish -p src/madsci_data_manager
       - run: pdm publish -p src/madsci_experiment_application
       - run: pdm publish -p src/madsci_location_manager
+      - run: pdm publish -p src/madsci_squid
