### 2025-12-05 aa9004a
Breakout developer guide to Contributions.md; add better guidelines for external contributors
