+++ CONTRIBUTING.md
@@ -1,0 +1,211 @@
+# Contributing to MADSci
+
+Thank you for your interest in contributing to MADSci! This guide will help you get started with development and understand how to submit contributions.
+
+## Table of Contents
+
+- [Developer Guide](#developer-guide)
+- [Submitting Feedback and Contributions](#submitting-feedback-and-contributions)
+- [Configuration](#configuration)
+
+## Developer Guide
+
+### Prerequisites
+
+- **Python 3.9+**: Required for all MADSci components
+- **[PDM](https://pdm-project.org/)**: For dependency management and virtual environments
+- **[Docker](https://docs.docker.com/engine/install/)**: Required for services and integration tests
+  - Alternatives: [Rancher Desktop](https://rancherdesktop.io/), [Podman](https://podman.io/)
+- **[just](https://github.com/casey/just)**: Task runner for development commands
+   - Commands are defined in the hidden `.justfile` in the repository root, and can be used without `just` if desired
+   - Run `just` without a subcommand for a list of all available commands
+- **Node.js/yarn**: Only needed for dashboard development
+
+### Quick Setup
+
+```bash
+# Clone and initialize
+git clone https://github.com/AD-SDL/MADSci.git
+cd MADSci
+just init  # Installs all dependencies and sets up pre-commit hooks
+
+# See all available commands
+just list
+
+# Start example lab for testing
+just up
+```
+
+### Development Commands
+
+```bash
+# Testing
+pytest                    # Run all tests
+just test                 # Alternative test runner
+pytest -k workcell        # Run specific component tests
+
+# Code Quality
+just checks               # Run all pre-commit checks (ruff, formatting, etc.)
+ruff check               # Manual linting
+ruff format              # Manual formatting
+
+# Services
+just build               # Build Docker images
+just up                  # Start example lab
+just down               # Stop services
+
+# Dashboard Development
+cd ui/
+yarn dev                # Start Vue dev server
+yarn build              # Build for production
+```
+
+### Development Patterns
+
+**Manager Implementation:**
+Each manager service follows this structure:
+- Settings class inheriting from `MadsciBaseSettings`
+- FastAPI server with REST endpoints
+- Client class for programmatic interaction
+- Database models (SQLModel/Pydantic)
+
+**Testing:**
+- Integration tests use Docker containers via pytest-mock-resources
+- Component tests are in each package's `tests/` directory
+- Use `pytest -k EXPRESSION` to filter tests
+
+**Configuration:**
+- Environment variables with hierarchical precedence
+- Each manager has unique prefix (e.g., `WORKCELL_`, `EVENT_`)
+- See [Configuration.md](./Configuration.md) for full details
+
+### Dev Container Support
+
+For VS Code users, use the included [.devcontainer](./.devcontainer) for instant setup:
+- Automatic dependency installation
+- Pre-configured development environment
+- Docker services ready to run
+
+## Submitting Feedback and Contributions
+
+We welcome all forms of contribution to MADSci! Whether you've found a bug, have a feature request, or want to contribute code, here's how to get involved:
+
+### Reporting Bugs
+
+If you encounter a bug or issue:
+
+1. **Check existing issues**: Search the [issue tracker](https://github.com/AD-SDL/MADSci/issues) to see if it's already been reported
+2. **Create a new issue**: If not found, [open a new issue](https://github.com/AD-SDL/MADSci/issues/new) with:
+   - A clear, descriptive title
+   - Steps to reproduce the problem
+   - Expected vs. actual behavior
+   - Your environment details (OS, Python version, MADSci version)
+   - Relevant logs or error messages
+   - Minimal code example if applicable
+
+### Feature Requests
+
+Have an idea for improving MADSci?
+
[…]
