### 2025-12-05 99eb8b1
Another, more logical attempt to fix the flaky tests
