+++ src/madsci_node_module/tests/test_rest_node_client.py
@@ -1,6 +1,7 @@
 """Comprehensive unit tests for the RestNodeClient class."""
 
 import tempfile
+import time
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
@@ -578,13 +579,11 @@ def test_send_action_file_response(mock_create_session: MagicMock) -> None:
     assert result.action_id == action_id
 
 
-@patch("time.sleep")
 @patch("madsci.client.node.rest_node_client.EventClient")
 @patch("madsci.client.node.rest_node_client.create_http_session")
 def test_await_action_result_timeout(
     mock_create_session: MagicMock,
-    mock_event_client: MagicMock,  # noqa: ARG001
-    mock_sleep: MagicMock,  # noqa: ARG001
+    mock_event_client: MagicMock,  # noqa: ARG001 Need to patch event client to avoid real calls
 ) -> None:
     """Test await_action_result with timeout."""
     action_id = new_ulid_str()
@@ -603,20 +602,26 @@ def test_await_action_result_timeout(
     # Create a new client to use mocked session
     client = RestNodeClient(url="http://localhost:2000")
 
+    # Use a very short timeout to make the test fast
     with pytest.raises(TimeoutError, match="Timed out waiting for action to complete"):
-        client.await_action_result(action_id, timeout=0.1)
+        client.await_action_result(action_id, timeout=0.5)
 
 
-@patch("time.sleep")
 @patch("madsci.client.node.rest_node_client.EventClient")
 @patch("madsci.client.node.rest_node_client.create_http_session")
 def test_await_action_result_exponential_backoff(
     mock_create_session: MagicMock,
     mock_event_client: MagicMock,  # noqa: ARG001 Need to patch event client to avoid real calls
-    mock_sleep: MagicMock,
 ) -> None:
-    """Test await_action_result with exponential backoff."""
+    """Test await_action_result with exponential backoff.
+
+    This test verifies that the polling interval increases with each retry
+    by measuring the actual time between status checks.
+    """
     action_id = new_ulid_str()
+
+    # Track when each GET call happens to verify exponential backoff
+    call_times = []
 
     # Mock responses: running status -> running status -> succeeded status -> result
     running_status_response = MagicMock()
@@ -640,12 +645,26 @@ def test_await_action_result_exponential_backoff(
     result_response.raise_for_status.return_value = None
 
     mock_session = MagicMock()
-    mock_session.get.side_effect = [
-        running_status_response,
-        running_status_response,
-        succeeded_status_response,
-        result_response,
-    ]
+    # Use a side_effect function to track call timing
+    get_call_count = {"count": 0}
+
+    def mock_get(*args, **kwargs):  # noqa: ARG001 kwargs needed for mock signature compatibility
+        get_call_count["count"] += 1
+        call_times.append(time.time())
+
+        if get_call_count["count"] == 1 or get_call_count["count"] == 2:
+            return running_status_response
+        if get_call_count["count"] == 3:
+            return succeeded_status_response
+        if get_call_count["count"] == 4:
+            return result_response
+        # This should not happen - fail with a clear error
+        raise AssertionError(
+            f"Unexpected GET call #{get_call_count['count']}. "
+            f"Test expected only 4 GET calls. URL: {args[0] if args else 'unknown'}"
+        )
+
+    mock_session.get.side_effect = mock_get
     mock_create_session.return_value = mock_session
 
     # Create a new client to use mocked session
@@ -655,11 +674,30 @@ def test_await_action_result_exponential_backoff(
[…]
