+++ README.md
@@ -6,7 +6,7 @@
 |---------------------------------------------------------------------------------------------- | -------: | -------: | ------: | --------: |
 | src/madsci\_client/madsci/client/\_\_init\_\_.py                                              |       15 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/data\_client.py                                              |      157 |       41 |     74% |85, 109-117, 127-128, 137-148, 214-224, 238, 287-288, 308-309, 371, 388, 404-411, 439-441, 488-498 |
-| src/madsci\_client/madsci/client/event\_client.py                                             |      267 |       30 |     89% |64, 91-92, 302-305, 313-316, 321, 379, 427-430, 439-442, 447, 472, 477-478, 487-488, 501, 528, 532-536 |
+| src/madsci\_client/madsci/client/event\_client.py                                             |      275 |       33 |     88% |65, 93-96, 107-108, 318-321, 329-332, 337, 395, 443-446, 455-458, 463, 488, 493-494, 503-504, 517, 544, 548-552 |
 | src/madsci\_client/madsci/client/experiment\_client.py                                        |       59 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/lab\_client.py                                               |       35 |       21 |     40% |31-43, 52-58, 67-73, 82-88, 97-103 |
 | src/madsci\_client/madsci/client/location\_client.py                                          |       98 |       39 |     60% |45, 62, 76, 86, 104-112, 132-140, 189-198, 218-226, 254-263, 318-327, 374-382, 410-426 |
@@ -50,7 +50,7 @@
 | src/madsci\_common/madsci/common/types/location\_types.py                                     |       98 |        9 |     91% |30, 32, 34, 36, 41, 46, 51, 98, 170 |
 | src/madsci\_common/madsci/common/types/manager\_types.py                                      |       45 |        5 |     89% |     27-31 |
 | src/madsci\_common/madsci/common/types/mongodb\_migration\_types.py                           |      214 |      116 |     46% |84, 105, 129-130, 150, 159, 165, 191, 197, 200-203, 206, 218-227, 248-262, 266-271, 293-297, 303-319, 323-333, 338-348, 352-359, 363, 367, 371, 385-441, 458-490 |
-| src/madsci\_common/madsci/common/types/node\_types.py                                         |      144 |        5 |     97% |398, 410, 419, 429, 459 |
+| src/madsci\_common/madsci/common/types/node\_types.py                                         |      144 |        7 |     95% |398, 406, 410, 419, 425, 429, 459 |
 | src/madsci\_common/madsci/common/types/parameter\_types.py                                    |       37 |        8 |     78% |14, 23, 47, 49, 51, 68, 70, 72 |
 | src/madsci\_common/madsci/common/types/resource\_types/\_\_init\_\_.py                        |      296 |       46 |     84% |119, 163, 240, 250, 271, 276, 280, 284, 325, 338, 344, 366, 370, 404, 426, 430, 478-482, 512, 555, 560-561, 572-588, 591-600, 607-609, 640, 655, 684-687, 724-727, 764, 769, 779, 783 |
 | src/madsci\_common/madsci/common/types/resource\_types/definitions.py                         |      118 |       10 |     92% |32-34, 219-227, 348, 424 |
@@ -80,7 +80,7 @@
 | src/madsci\_node\_module/madsci/node\_module/\_\_init\_\_.py                                  |        4 |        0 |    100% |           |
[…]
