### 2025-12-05 85fa7a4
Address e2e test failures
