+++ README.md
@@ -4,7 +4,8 @@
 
 | Name                                                                                          |    Stmts |     Miss |   Cover |   Missing |
 |---------------------------------------------------------------------------------------------- | -------: | -------: | ------: | --------: |
-| src/madsci\_client/madsci/client/\_\_init\_\_.py                                              |       15 |        0 |    100% |           |
+| src/madsci\_client/madsci/client/\_\_init\_\_.py                                              |       17 |        0 |    100% |           |
+| src/madsci\_client/madsci/client/client\_mixin.py                                             |      183 |       33 |     82% |235, 242, 283, 320, 327, 363, 377, 392-430, 464, 482, 500, 515-521 |
 | src/madsci\_client/madsci/client/data\_client.py                                              |      157 |       41 |     74% |85, 109-117, 127-128, 137-148, 214-224, 238, 287-288, 308-309, 371, 388, 404-411, 439-441, 488-498 |
 | src/madsci\_client/madsci/client/event\_client.py                                             |      275 |       33 |     88% |65, 93-96, 107-108, 318-321, 329-332, 337, 395, 443-446, 455-458, 463, 488, 493-494, 503-504, 517, 544, 548-552 |
 | src/madsci\_client/madsci/client/experiment\_client.py                                        |       59 |        0 |    100% |           |
@@ -13,7 +14,7 @@
 | src/madsci\_client/madsci/client/node/\_\_init\_\_.py                                         |        4 |        0 |    100% |           |
 | src/madsci\_client/madsci/client/node/abstract\_node\_client.py                               |       30 |        4 |     87% |27, 30, 100-101 |
 | src/madsci\_client/madsci/client/node/rest\_node\_client.py                                   |      208 |       44 |     79% |63, 146, 282-293, 324-334, 356-391, 418-421, 528, 558, 560, 578 |
-| src/madsci\_client/madsci/client/resource\_client.py                                          |      634 |      198 |     69% |102, 127, 149, 176-179, 187, 244-247, 264, 330-334, 350-365, 394, 428-431, 484-491, 497, 514, 526-527, 576, 607, 654-655, 692-694, 736-737, 776-777, 813, 849-850, 886-887, 923-924, 960-961, 992-993, 1024-1029, 1060-1062, 1093-1146, 1198-1208, 1238-1241, 1292-1302, 1323, 1326-1337, 1369-1376, 1400-1403, 1461-1479, 1500-1506, 1555-1581, 1625-1663, 1694-1711, 1795-1799, 1813, 1836-1837, 1845-1849, 1865-1866, 1877, 1914, 1927, 1940-1950 |
+| src/madsci\_client/madsci/client/resource\_client.py                                          |      636 |      200 |     69% |102, 127, 149, 176-179, 187, 244-247, 264, 330-334, 350-365, 394, 428-431, 484-491, 497, 514, 526-527, 576, 607, 654-657, 694-696, 738-739, 778-779, 815, 851-852, 888-889, 925-926, 962-963, 994-995, 1026-1031, 1062-1064, 1095-1148, 1200-1210, 1240-1243, 1294-1304, 1325, 1328-1339, 1371-1378, 1402-1405, 1463-1481, 1502-1508, 1557-1583, 1627-1665, 1696-1713, 1797-1801, 1815, 1838-1839, 1847-1851, 1867-1868, 1879, 1916, 1929, 1942-1952 |
[…]
