### 2025-12-05 fec8499
Add: location manager definition <-> redis sync functionality
