+++ src/madsci_location_manager/tests/test_location_persistence.py
@@ -1,0 +1,351 @@
+"""Tests for location persistence to YAML definition files."""
+
+from pathlib import Path
+from tempfile import TemporaryDirectory
+
+import pytest
+from fastapi.testclient import TestClient
+from madsci.common.types.location_types import (
+    Location,
+    LocationDefinition,
+    LocationManagerDefinition,
+    LocationManagerSettings,
+    LocationTransferCapabilities,
+    TransferStepTemplate,
+)
+from madsci.common.utils import new_ulid_str
+from madsci.location_manager.location_server import LocationManager
+from pytest_mock_resources import RedisConfig, create_redis_fixture
+from redis import Redis
+
+
+# Create a Redis server fixture for testing
+@pytest.fixture(scope="session")
+def pmr_redis_config() -> RedisConfig:
+    """Configure the Redis server."""
+    return RedisConfig(image="redis:7.4")
+
+
+redis_server = create_redis_fixture()
+
+
+@pytest.fixture
+def temp_dir():
+    """Create a temporary directory for test files."""
+    with TemporaryDirectory() as tmpdir:
+        yield Path(tmpdir)
+
+
+@pytest.fixture
+def definition_file(temp_dir):
+    """Create a test definition file."""
+    def_path = temp_dir / "location.manager.yaml"
+
+    # Create initial definition with some locations
+    initial_definition = LocationManagerDefinition(
+        name="Test Location Manager",
+        description="A test location manager instance",
+        locations=[
+            LocationDefinition(
+                location_name="Initial Location 1",
+                location_id=new_ulid_str(),
+                description="First initial location",
+            ),
+            LocationDefinition(
+                location_name="Initial Location 2",
+                location_id=new_ulid_str(),
+                description="Second initial location",
+            ),
+        ],
+    )
+
+    # Write to file
+    initial_definition.to_yaml(def_path)
+    return def_path
+
+
+@pytest.fixture
+def app(redis_server: Redis, definition_file: Path):
+    """Create a test app with test settings, Redis server, and definition file."""
+    settings = LocationManagerSettings(
+        redis_host=redis_server.connection_pool.connection_kwargs["host"],
+        redis_port=redis_server.connection_pool.connection_kwargs["port"],
+        manager_definition=definition_file,
+    )
+
+    # Load definition from file
+    definition = LocationManagerDefinition.from_yaml(definition_file)
+
+    # Create the app with the Redis connection passed to the LocationManager
+    manager = LocationManager(settings=settings, definition=definition)
+    # Override the state handler's Redis connection to use the test Redis instance
+    manager.state_handler._redis_connection = redis_server
+
+    return manager.create_server(version="0.1.0")
+
+
+@pytest.fixture
+def client(app):
+    """Create a test client."""
+    return TestClient(app)
+
+
+@pytest.fixture
+def sample_location():
+    """Create a sample location for testing."""
+    return Location(
+        location_id=new_ulid_str(),
+        location_name="Test Location",
+        description="A test location for persistence testing",
+    )
+
+
+# RED PHASE: Write failing tests
+
+
+def test_add_location_persists_to_yaml(client, sample_location, definition_file):
+    """Test that adding a location persists it to the YAML definition file.
+
+    This test verifies that when a location is added via the API,
+    it is written back to the definition file so it persists across restarts.
+    """
+    # Add a new location via API
+    response = client.post("/location", json=sample_location.model_dump())
+    assert response.status_code == 200
+
+    # Reload the definition from the file
+    reloaded_definition = LocationManagerDefinition.from_yaml(definition_file)
+
+    # Check that the new location is in the definition
+    location_ids = [loc.location_id for loc in reloaded_definition.locations]
[…]
