### 2025-12-05 d39ffde
Fix: exempt localhost from rate limits by default
