+++ src/madsci_common/tests/test_stress.py
@@ -48,6 +48,7 @@ def stress_test_manager() -> TestManager:
         rate_limit_enabled=True,
         rate_limit_requests=100,
         rate_limit_window=10,
+        rate_limit_exempt_ips=[],  # Disable localhost exemption for stress testing
     )
     definition = TestManagerDefinition(name="Stress Test Manager")
     return TestManager(settings=settings, definition=definition)
