### 2025-12-05 30f54fd
Update src/madsci_location_manager/tests/test_location_persistence.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
