### 2025-12-05 852efb5
Update src/madsci_location_manager/tests/test_location_persistence.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
