### 2025-12-05 b0db710
Address PR review comments
