### 2025-12-05 eafa298
Bump to v0.6.0-rc.3
