+++ README.md
@@ -25,7 +25,7 @@
 | src/madsci\_common/madsci/common/backup\_tools/mongodb\_backup.py                             |      280 |       90 |     68% |47, 82-84, 140, 149-168, 204, 209, 249-256, 284, 304-305, 326-328, 336, 358, 363, 368-369, 403, 428-430, 434-482, 515-540, 544-553, 559-560 |
 | src/madsci\_common/madsci/common/backup\_tools/postgres\_backup.py                            |      217 |       55 |     75% |12, 128-134, 196, 204-205, 207, 262, 275, 277, 279, 290-296, 303, 330, 356, 360, 373-374, 407-419, 441-442, 456-458, 473-475, 488-494, 506-508 |
 | src/madsci\_common/madsci/common/backup\_tools/postgres\_cli.py                               |      104 |        7 |     93% |109-111, 162-163, 179-180 |
-| src/madsci\_common/madsci/common/context.py                                                   |       22 |        2 |     91% |    11, 44 |
+| src/madsci\_common/madsci/common/context.py                                                   |       36 |        2 |     94% |    58, 80 |
 | src/madsci\_common/madsci/common/data\_manipulation.py                                        |       43 |        0 |    100% |           |
 | src/madsci\_common/madsci/common/exceptions.py                                                |       23 |        2 |     91% |     30-31 |
 | src/madsci\_common/madsci/common/manager\_base.py                                             |      136 |       35 |     74% |94, 162-166, 216-217, 262-275, 381-414, 438-439 |
@@ -51,7 +51,7 @@
 | src/madsci\_common/madsci/common/types/location\_types.py                                     |       98 |        9 |     91% |30, 32, 34, 36, 41, 46, 51, 98, 170 |
 | src/madsci\_common/madsci/common/types/manager\_types.py                                      |       46 |        5 |     89% |     27-31 |
 | src/madsci\_common/madsci/common/types/mongodb\_migration\_types.py                           |      214 |      116 |     46% |84, 105, 129-130, 150, 159, 165, 191, 197, 200-203, 206, 218-227, 248-262, 266-271, 293-297, 303-319, 323-333, 338-348, 352-359, 363, 367, 371, 385-441, 458-490 |
-| src/madsci\_common/madsci/common/types/node\_types.py                                         |      144 |        7 |     95% |398, 406, 410, 419, 425, 429, 459 |
+| src/madsci\_common/madsci/common/types/node\_types.py                                         |      144 |        5 |     97% |398, 410, 419, 429, 459 |
 | src/madsci\_common/madsci/common/types/parameter\_types.py                                    |       37 |        8 |     78% |14, 23, 47, 49, 51, 68, 70, 72 |
 | src/madsci\_common/madsci/common/types/resource\_types/\_\_init\_\_.py                        |      296 |       46 |     84% |119, 163, 240, 250, 271, 276, 280, 284, 325, 338, 344, 366, 370, 404, 426, 430, 478-482, 512, 555, 560-561, 572-588, 591-600, 607-609, 640, 655, 684-687, 724-727, 764, 769, 779, 783 |
 | src/madsci\_common/madsci/common/types/resource\_types/definitions.py                         |      118 |       10 |     92% |32-34, 219-227, 348, 424 |
@@ -101,7 +101,7 @@
 | src/madsci\_workcell\_manager/madsci/workcell\_manager/workcell\_server.py                    |      264 |       67 |     75% |86-87, 123-158, 183-184, 231, 243-247, 271-272, 284, 292, 302-305, 313-317, 348, 363, 377, 391, 397, 426-428, 432-437, 462-464, 501-503, 518, 529, 560-565, 572-573 |
[…]
