### 2025-12-07 #9416 @ktahar
Yo @LuckierDodge ,
Thanks and both points look good to me!

Now I have completed my checklist.
Overall I think this project should be a significant contribution to research area in lab automation.
It provides quite high-level and versatile framework for lab, experiment, and data managements in automated labs.
The code is written neatly and well-organized, leveraging modern web-technology. It also comes with docs and tests.
I do recommend the publication in JOSS.

As final note, I would recommend compiling the document into HTML using a tool like sphinx and publish it in Read the Docs or somewhere else. README.md in each subdirectory is okay but it would be easier to navigate if they are structured as a document with single entry point. However, it would not be a requirement to publish this paper in JOSS.
