### 2025-12-08 ccebd38
Example Location manager definition tweaks
