+++ example_lab/managers/example_location.manager.yaml
@@ -14,12 +14,48 @@ locations:
       deck_type: standard
       max_plates: 1
     robotarm_1:
+      gripper_config: standard
+      max_payload: 2.0
       position:
       - 10
       - 15
       - 5
+  resource_template_name: plate_nest
+  resource_template_overrides:
+    resource_class: deck_slot
+    resource_description: Deck position accessible by liquid handler and robot arm
+  allow_transfers: true
+- location_name: platereader_1.plate_carriage
+  location_id: 01K5HDZZCF27YHD2WDGSXFPPKQ
+  description: Plate carriage - accessible by robotarm_1 for plate transport
+  representations:
+    robotarm_1:
+      gripper_config: wide
+      max_payload: 3.0
+      position:
+      - 15
+      - 10
+      - 8
+  resource_template_name: plate_nest
+  resource_template_overrides:
+    resource_class: plate_carriage
+    resource_description: Plate carriage for platereader_1
+  allow_transfers: true
+- location_name: liquidhandler_2.deck_3
+  location_id: 01K5HDZZCF27YHD2WDGSXFPPKN
+  description: Deck 3 on liquidhandler_2 - accessible by robotarm_1
+  representations:
+    liquidhandler_2:
+      deck_position: 3
+      deck_type: standard
+      max_plates: 1
+    robotarm_1:
       gripper_config: standard
       max_payload: 2.0
+      position:
+      - 25
+      - 15
+      - 5
   resource_template_name: plate_nest
   resource_template_overrides:
     resource_class: deck_slot
@@ -38,6 +74,19 @@ locations:
     resource_class: deck_slot
     resource_description: Deck position accessible by liquid handler only
   allow_transfers: true
+- location_name: liquidhandler_2.deck_4
+  location_id: 01K5HDZZCF27YHD2WDGSXFPPKP
+  description: Deck 4 on liquidhandler_2 - liquid handler only
+  representations:
+    liquidhandler_2:
+      deck_position: 4
+      deck_type: standard
+      max_plates: 1
+  resource_template_name: plate_nest
+  resource_template_overrides:
+    resource_class: deck_slot
+    resource_description: Deck position accessible by liquid handler only
+  allow_transfers: true
 - location_name: liquidhandler_1.deck_3
   location_id: 01K5HDZZCF27YHD2WDGSXFPPKH
   description: Deck 3 on liquidhandler_1 - liquid handler only
@@ -47,45 +96,12 @@ locations:
       deck_type: standard
       max_plates: 1
     robotarm_1:
+      gripper_config: standard
+      max_payload: 2.0
       position:
       - 20
       - 15
       - 5
-      gripper_config: standard
-      max_payload: 2.0
-  resource_template_name: plate_nest
-  resource_template_overrides:
-    resource_class: deck_slot
-    resource_description: Deck position accessible by liquid handler and robot arm
-  allow_transfers: true
-- location_name: liquidhandler_1.deck_4
-  location_id: 01K5HDZZCF27YHD2WDGSXFPPKJ
-  description: Deck 4 on liquidhandler_1 - liquid handler only
-  representations:
-    liquidhandler_1:
-      deck_position: 4
-      deck_type: standard
-      max_plates: 1
-  resource_template_name: plate_nest
-  resource_template_overrides:
-    resource_class: deck_slot
-    resource_description: Deck position accessible by liquid handler only
-  allow_transfers: true
-- location_name: liquidhandler_2.deck_1
-  location_id: 01K5HDZZCF27YHD2WDGSXFPPKK
-  description: Deck 1 on liquidhandler_2 - accessible by robotarm_1
-  representations:
-    liquidhandler_2:
-      deck_position: 1
[…]
