### 2025-12-08 266e0c9
Update rate limits based on RPL testing
