+++ Configuration.md
@@ -37,7 +37,7 @@ Settings for the MADSci Data Manager.
 | `DATA_SERVER_URL`                           | `AnyUrl`                | `"http://localhost:8004/"`    | The URL of the data manager server.                                                                                        | `"http://localhost:8004/"`    |
 | `DATA_MANAGER_DEFINITION`                   | `string` \| `Path`      | `"data.manager.yaml"`         | Path to the data manager definition file to use.                                                                           | `"data.manager.yaml"`         |
 | `DATA_RATE_LIMIT_ENABLED`                   | `boolean`               | `true`                        | Enable rate limiting for API endpoints.                                                                                    | `true`                        |
-| `DATA_RATE_LIMIT_REQUESTS`                  | `integer`               | `100`                         | Maximum number of requests allowed per long time window.                                                                   | `100`                         |
+| `DATA_RATE_LIMIT_REQUESTS`                  | `integer`               | `300`                         | Maximum number of requests allowed per long time window.                                                                   | `300`                         |
 | `DATA_RATE_LIMIT_WINDOW`                    | `integer`               | `60`                          | Long time window for rate limiting in seconds.                                                                             | `60`                          |
 | `DATA_RATE_LIMIT_SHORT_REQUESTS`            | `integer` \| `NoneType` | `50`                          | Maximum number of requests allowed per short time window for burst protection. If None, short window limiting is disabled. | `50`                          |
 | `DATA_RATE_LIMIT_SHORT_WINDOW`              | `integer` \| `NoneType` | `1`                           | Short time window for burst protection in seconds. If None, short window limiting is disabled.                             | `1`                           |
@@ -77,7 +77,7 @@ Handles settings and configuration for the Event Manager.
 | `EVENT_SERVER_URL`                           | `AnyUrl` \| `NoneType`            | `"http://localhost:8001"`     | The URL of the Event Manager server.                                                                                       | `"http://localhost:8001"`     |
 | `EVENT_MANAGER_DEFINITION`                   | `string` \| `Path`                | `"event.manager.yaml"`        | Path to the event manager definition file to use.                                                                          | `"event.manager.yaml"`        |
 | `EVENT_RATE_LIMIT_ENABLED`                   | `boolean`                         | `true`                        | Enable rate limiting for API endpoints.                                                                                    | `true`                        |
-| `EVENT_RATE_LIMIT_REQUESTS`                  | `integer`                         | `100`                         | Maximum number of requests allowed per long time window.                                                                   | `100`                         |
+| `EVENT_RATE_LIMIT_REQUESTS`                  | `integer`                         | `300`                         | Maximum number of requests allowed per long time window.                                                                   | `300`                         |
 | `EVENT_RATE_LIMIT_WINDOW`                    | `integer`                         | `60`                          | Long time window for rate limiting in seconds.                                                                             | `60`                          |
 | `EVENT_RATE_LIMIT_SHORT_REQUESTS`            | `integer` \| `NoneType`           | `50`                          | Maximum number of requests allowed per short time window for burst protection. If None, short window limiting is disabled. | `50`                          |
 | `EVENT_RATE_LIMIT_SHORT_WINDOW`              | `integer` \| `NoneType`           | `1`                           | Short time window for burst protection in seconds. If None, short window limiting is disabled.                             | `1`                           |
@@ -103,7 +103,7 @@ Settings for the MADSci Workcell Manager.
 | `WORKCELL_SERVER_URL`                                                          | `AnyUrl`                         | `"http://localhost:8005/"`                               | The URL of the workcell manager server.                                                                                                                                                  | `"http://localhost:8005/"`                               |
[…]
