### 2025-12-08 cc2b75d
Bump to v0.6.0-rc.4
