+++ README.md
@@ -15,7 +15,7 @@
 | src/madsci\_client/madsci/client/node/abstract\_node\_client.py                               |       30 |        4 |     87% |27, 30, 100-101 |
 | src/madsci\_client/madsci/client/node/rest\_node\_client.py                                   |      208 |       44 |     79% |63, 146, 282-293, 324-334, 356-391, 418-421, 528, 558, 560, 578 |
 | src/madsci\_client/madsci/client/resource\_client.py                                          |      636 |      200 |     69% |102, 127, 149, 176-179, 187, 244-247, 264, 330-334, 350-365, 394, 428-431, 484-491, 497, 514, 526-527, 576, 607, 654-657, 694-696, 738-739, 778-779, 815, 851-852, 888-889, 925-926, 962-963, 994-995, 1026-1031, 1062-1064, 1095-1148, 1200-1210, 1240-1243, 1294-1304, 1325, 1328-1339, 1371-1378, 1402-1405, 1463-1481, 1502-1508, 1557-1583, 1627-1665, 1696-1713, 1797-1801, 1815, 1838-1839, 1847-1851, 1867-1868, 1879, 1916, 1929, 1942-1952 |
-| src/madsci\_client/madsci/client/workcell\_client.py                                          |      198 |       45 |     77% |94, 120, 146, 160, 209-213, 219, 245, 249, 273-278, 388-406, 433-461, 500, 508, 516, 521-522, 635, 668 |
+| src/madsci\_client/madsci/client/workcell\_client.py                                          |      198 |       45 |     77% |94, 120, 146, 160, 209-213, 219, 245, 249, 273-278, 388-406, 433-461, 503, 511, 519, 524-525, 638, 671 |
 | src/madsci\_common/madsci/common/backup\_tools/\_\_init\_\_.py                                |        8 |        0 |    100% |           |
 | src/madsci\_common/madsci/common/backup\_tools/backup\_manager.py                             |      172 |       26 |     85% |63-64, 109-111, 133, 149-150, 159-160, 190, 231, 247, 284, 287-288, 304-305, 321-322, 351-356 |
 | src/madsci\_common/madsci/common/backup\_tools/backup\_validator.py                           |       90 |       11 |     88% |105-107, 176-178, 256, 269-273, 287 |
