### 2025-12-08 8b1373b
Set default concurrency limit for nodes; bump to v0.6.0-rc.5
