+++ example_lab/node_definitions/advanced_example_node.node.info.yaml
@@ -1278,7 +1278,8 @@ config:
   status_update_interval: 2.0
   state_update_interval: 2.0
   node_url: http://localhost:2004/
-  uvicorn_kwargs: {}
+  uvicorn_kwargs:
+    limit_concurrency: 10
   enable_rate_limiting: true
   rate_limit_requests: 100
   rate_limit_window: 60
@@ -1341,7 +1342,9 @@ config_schema:
       type: string
     uvicorn_kwargs:
       additionalProperties: true
-      description: Configuration for the Uvicorn server that runs the REST API.
+      description: Configuration for the Uvicorn server that runs the REST API. By
+        default, sets limit_concurrency=10 to protect against connection exhaustion
+        attacks.
       title: Uvicorn Configuration
       type: object
     enable_rate_limiting:
