### 2025-12-08 bce3366
Update UI dependencies, sort locations in location manager definition
