### 2025-12-11 9965d4e
Initial sketch of labware types
