### 2025-12-11 8589633
WIP: phase 1, type parser implemented
