+++ src/madsci_node_module/tests/test_type_analyzer.py
@@ -1,0 +1,864 @@
+"""Tests for the type analyzer module.
+
+This test suite comprehensively tests the type analysis functionality,
+covering all aspects of recursive type unwrapping, special type detection,
+and error handling.
+"""
+
+from pathlib import Path, PosixPath, PurePath, WindowsPath
+from typing import Annotated, Any, Optional, TypeVar, Union
+
+import pytest
+from madsci.common.types.action_types import (
+    ActionCancelled,
+    ActionFailed,
+    ActionFiles,
+    ActionJSON,
+    ActionNotReady,
+    ActionPaused,
+    ActionResult,
+    ActionRunning,
+    ActionSucceeded,
+    ArgumentDefinition,
+    FileArgumentDefinition,
+)
+from madsci.common.types.base_types import MadsciBaseModel
+from madsci.common.types.location_types import LocationArgument
+from madsci.node_module.type_analyzer import (
+    analyze_type,
+    extract_metadata_from_annotated,
+    is_action_result_type,
+    is_file_type,
+    is_location_argument_type,
+    is_optional_type,
+)
+
+# =============================================================================
+# BASIC TYPES (10 tests)
+# =============================================================================
+
+
+def test_analyze_basic_types():
+    """Test analysis of basic Python types."""
+    # String
+    info = analyze_type(str)
+    assert info.base_type is str
+    assert not info.is_optional
+    assert info.special_type is None
+
+    # Integer
+    info = analyze_type(int)
+    assert info.base_type is int
+    assert not info.is_optional
+
+    # Float
+    info = analyze_type(float)
+    assert info.base_type is float
+
+    # Bool
+    info = analyze_type(bool)
+    assert info.base_type is bool
+
+    # Dict
+    info = analyze_type(dict)
+    assert info.base_type is dict
+
+    # List
+    info = analyze_type(list)
+    assert info.base_type is list
+
+
+def test_analyze_path():
+    """Test analysis of Path type."""
+    info = analyze_type(Path)
+    assert info.base_type is Path
+    assert info.special_type == "file"
+    assert info.special_type_class == Path
+    assert not info.is_optional
+
+
+def test_analyze_location_argument():
+    """Test analysis of LocationArgument type."""
+    info = analyze_type(LocationArgument)
+    assert info.base_type is LocationArgument
+    assert info.special_type == "location"
+    assert info.special_type_class == LocationArgument
+    assert not info.is_optional
+
+
+def test_analyze_action_result():
+    """Test analysis of ActionResult and subclasses."""
+    info = analyze_type(ActionResult)
+    assert info.base_type is ActionResult
+    assert info.special_type == "action_result"
+    assert info.special_type_class == ActionResult
+
+
+def test_analyze_action_files():
+    """Test analysis of ActionFiles type."""
+    info = analyze_type(ActionFiles)
+    assert info.base_type is ActionFiles
+    assert info.special_type == "action_files"
+    assert info.special_type_class == ActionFiles
+
+
+def test_analyze_action_json():
+    """Test analysis of ActionJSON type."""
+    info = analyze_type(ActionJSON)
+    assert info.base_type is ActionJSON
+    assert info.special_type == "action_json"
+    assert info.special_type_class == ActionJSON
+
+
+def test_analyze_pydantic_model():
+    """Test analysis of custom BaseModel subclass."""
+
+    class CustomModel(MadsciBaseModel):
+        field: str
+
+    info = analyze_type(CustomModel)
+    assert info.base_type is CustomModel
+    # Custom pydantic models are not special types
+    assert info.special_type is None
+
+
+def test_analyze_none_type():
+    """Test analysis of None/NoneType."""
+    info = analyze_type(type(None))
+    assert info.base_type is type(None)
[…]
