+++ src/madsci_node_module/madsci/node_module/type_analyzer.py
@@ -1,0 +1,516 @@
+"""Type analysis utilities for MADSci action parsing.
+
+This module provides centralized, recursive type analysis for parsing action
+arguments and return types, handling arbitrary nesting of type hints like
+Optional, Union, Annotated, list, dict, etc.
+
+## Overview
+
+The type analyzer solves the problem of parsing complex, deeply-nested Python type
+hints commonly found in MADSci action signatures. It recursively unwraps type
+wrappers and identifies special MADSci types like LocationArgument, Path,
+ActionResult, etc., at any nesting level.
+
+## Key Features
+
+- **Recursive Type Unwrapping**: Handles arbitrary nesting of Optional, Union,
+  Annotated, list, dict, and tuple types
+- **Special Type Detection**: Identifies MADSci-specific types (LocationArgument,
+  Path, ActionResult, ActionFiles, ActionJSON) regardless of nesting depth
+- **Metadata Preservation**: Collects and preserves metadata from Annotated types
+  throughout the type hierarchy
+- **Union Validation**: Detects and reports conflicting special types in unions
+- **Python 3.10+ Support**: Handles both typing.Union and types.UnionType (| operator)
+- **Depth Protection**: Prevents infinite recursion with configurable depth limits
+
+## Usage Examples
+
+### Basic Type Analysis
+
+```python
+from madsci.node_module.type_analyzer import analyze_type
+from madsci.common.types.location_types import LocationArgument
+
+# Simple type
+info = analyze_type(int)
+assert info.base_type == int
+assert not info.is_optional
+
+# Optional type
+info = analyze_type(Optional[LocationArgument])
+assert info.base_type == LocationArgument
+assert info.is_optional
+assert info.special_type == "location"
+```
+
+### Complex Nested Types
+
+```python
+from typing import Annotated, Optional
+
+# Deeply nested type (the motivating use case for this module)
+info = analyze_type(Optional[Annotated[LocationArgument, "Transfer source"]])
+assert info.base_type == LocationArgument
+assert info.is_optional
+assert info.special_type == "location"
+assert "Transfer source" in info.metadata
+
+# Container with special types
+info = analyze_type(list[Path])
+assert info.is_list
+assert info.special_type == "file"
+```
+
+### Helper Functions
+
+```python
+from madsci.node_module.type_analyzer import (
+    is_optional_type,
+    is_location_argument_type,
+    is_file_type,
+    is_action_result_type,
+)
+
+# Quick type checks
+assert is_optional_type(Optional[int])
+assert is_location_argument_type(LocationArgument)
+assert is_file_type(Path)
+assert is_action_result_type(ActionFailed)
+```
+
+## Architecture
+
+The module uses a two-phase approach:
+
+1. **Type Unwrapping**: Recursively unwraps type wrappers (Annotated, Union, etc.)
+   to extract the base type and characteristics
+2. **Classification**: Identifies special MADSci types and validates type combinations
+
+This allows the action parser to correctly handle types like:
+- `Optional[Annotated[LocationArgument, "description"]]`
+- `list[Annotated[Path, FileArgumentDefinition(...)]]`
+- `Union[str, int] | None` (Python 3.10+)
+
+## Error Handling
+
+The analyzer raises `ValueError` for:
+- Unions with conflicting special types (e.g., `Union[Path, LocationArgument]`)
+- Excessive recursion depth (configurable, default 20 levels)
+- Other invalid type combinations
+
+All errors include detailed messages with the problematic type hint for debugging.
+"""
+
+from __future__ import annotations
+
+import types
+from dataclasses import dataclass, field
+from pathlib import Path, PurePath
[…]

[…]
