### 2025-12-11 ede8477
WIP: phase 2, update action result parsing
