+++ src/madsci_node_module/tests/test_helpers.py
@@ -5,8 +5,15 @@ from typing import Annotated, Dict, List, Optional, Union
 
 import pytest
 from madsci.common.types.action_types import (
+    ActionCancelled,
+    ActionFailed,
     ActionFiles,
     ActionJSON,
+    ActionNotReady,
+    ActionPaused,
+    ActionResult,
+    ActionRunning,
+    ActionSucceeded,
     FileActionResultDefinition,
     JSONActionResultDefinition,
     _analyze_file_parameter_type,
@@ -588,3 +595,247 @@ def test_extract_file_parameters_with_annotated_path():
     )
     assert file_params["optional_file"]["required"] is False
     assert file_params["optional_file"]["is_list"] is False
+
+
+# =============================================================================
+# PHASE 2: ActionResult Return Types (Issue #199)
+# =============================================================================
+
+
+def test_parse_result_action_result_base():
+    """Test parse_result with ActionResult base class."""
+    result = parse_result(ActionResult)
+    # ActionResult types should return empty list (handled by framework)
+    assert result == []
+
+
+def test_parse_result_action_failed():
+    """Test parse_result with ActionFailed."""
+    result = parse_result(ActionFailed)
+    assert result == []
+
+
+def test_parse_result_action_succeeded():
+    """Test parse_result with ActionSucceeded."""
+    result = parse_result(ActionSucceeded)
+    assert result == []
+
+
+def test_parse_result_action_running():
+    """Test parse_result with ActionRunning."""
+    result = parse_result(ActionRunning)
+    assert result == []
+
+
+def test_parse_result_action_not_ready():
+    """Test parse_result with ActionNotReady."""
+    result = parse_result(ActionNotReady)
+    assert result == []
+
+
+def test_parse_result_action_cancelled():
+    """Test parse_result with ActionCancelled."""
+    result = parse_result(ActionCancelled)
+    assert result == []
+
+
+def test_parse_result_action_paused():
+    """Test parse_result with ActionPaused."""
+    result = parse_result(ActionPaused)
+    assert result == []
+
+
+def test_parse_result_optional_action_failed():
+    """Test parse_result with Optional[ActionFailed]."""
+    result = parse_result(Optional[ActionFailed])
+    # Should still return empty list - ActionResult wrapped in Optional
+    assert result == []
+
+
+def test_parse_result_union_action_failed_none():
+    """Test parse_result with Union[ActionFailed, None]."""
+    result = parse_result(Union[ActionFailed, None])
+    assert result == []
+
+
+def test_parse_result_annotated_action_failed():
+    """Test parse_result with Annotated[ActionFailed, 'description']."""
+    result = parse_result(Annotated[ActionFailed, "Action failure result"])
+    assert result == []
+
+
+def test_parse_result_annotated_optional_action_failed():
+    """Test parse_result with Annotated[Optional[ActionFailed], ...]."""
+    result = parse_result(
+        Annotated[Optional[ActionFailed], "Optional action failure result"]
+    )
+    assert result == []
+
+
+def test_parse_result_optional_annotated_action_failed():
+    """Test parse_result with Optional[Annotated[ActionFailed, ...]]."""
+    result = parse_result(Optional[Annotated[ActionFailed, "Action failure result"]])
+    assert result == []
+
+
+def test_parse_result_returns_empty_for_action_result():
+    """Verify that all ActionResult subclasses return empty list."""
+    action_result_types = [
+        ActionResult,
+        ActionFailed,
+        ActionSucceeded,
+        ActionRunning,
+        ActionNotReady,
+        ActionCancelled,
+        ActionPaused,
+    ]
+
+    for action_type in action_result_types:
+        result = parse_result(action_type)
+        assert result == [], f"{action_type.__name__} should return empty list"
+
+
[…]
