### 2025-12-11 4ed7e10
WIP: phase 3, argument parsing complete
