+++ src/madsci_node_module/tests/test_argument_parsing.py
@@ -1,0 +1,1198 @@
+"""
+Comprehensive tests for action argument parsing with complex type hints.
+
+Phase 3 of action parsing fix (Issue #184).
+Tests handling of complex nested type hints in action arguments.
+"""
+
+from pathlib import Path, PosixPath, PurePath
+from typing import Annotated, Optional, Union
+
+import pytest
+from madsci.common.types.action_types import (
+    FileArgumentDefinition,
+    LocationArgumentDefinition,
+)
+from madsci.common.types.location_types import LocationArgument
+from madsci.common.types.node_types import RestNodeConfig
+from madsci.node_module.helpers import action
+from madsci.node_module.rest_node_module import RestNode
+
+
+class TestConfig(RestNodeConfig):
+    """Test configuration for argument parsing tests."""
+
+    __test__ = False
+    test_param: int = 1
+
+
+class TestArgumentParsingNode(RestNode):
+    """Test node for argument parsing tests."""
+
+    __test__ = False
+    config_model = TestConfig
+
+    def startup_handler(self) -> None:
+        """Initialize the node."""
+
+    def shutdown_handler(self) -> None:
+        """Shutdown the node."""
+
+    def state_handler(self) -> None:
+        """Update node state."""
+
+
+# ============================================================================
+# LocationArgument Parameters (Tests 110-124)
+# ============================================================================
+
+
+def test_parse_location_direct():
+    """Test 110: Parse direct LocationArgument parameter."""
+
+    class Node(TestArgumentParsingNode):
+        @action
+        def test_action(self, loc: LocationArgument) -> None:
+            pass
+
+    node = Node()
+    action_def = node.node_info.actions["test_action"]
+
+    assert "loc" in action_def.locations
+    assert action_def.locations["loc"].name == "loc"
+    assert action_def.locations["loc"].required is True
+    assert "loc" not in action_def.args
+    assert "loc" not in action_def.files
+
+
+def test_parse_location_optional():
+    """Test 111: Parse Optional[LocationArgument] parameter."""
+
+    class Node(TestArgumentParsingNode):
+        @action
+        def test_action(self, loc: Optional[LocationArgument] = None) -> None:
+            pass
+
+    node = Node()
+    action_def = node.node_info.actions["test_action"]
+
+    assert "loc" in action_def.locations
+    assert action_def.locations["loc"].required is False
+    assert "loc" not in action_def.args
+    assert "loc" not in action_def.files
+
+
+def test_parse_location_union_none():
+    """Test 112: Parse Union[LocationArgument, None] parameter."""
+
+    class Node(TestArgumentParsingNode):
+        @action
+        def test_action(self, loc: Union[LocationArgument, None] = None) -> None:
+            pass
+
+    node = Node()
+    action_def = node.node_info.actions["test_action"]
+
+    assert "loc" in action_def.locations
+    assert action_def.locations["loc"].required is False
+
+
+def test_parse_location_annotated():
+    """Test 113: Parse Annotated[LocationArgument, 'desc'] parameter."""
+
+    class Node(TestArgumentParsingNode):
+        @action
+        def test_action(
+            self, loc: Annotated[LocationArgument, "Location description"]
+        ) -> None:
+            pass
+
+    node = Node()
+    action_def = node.node_info.actions["test_action"]
+
+    assert "loc" in action_def.locations
+    assert action_def.locations["loc"].required is True
+    assert action_def.locations["loc"].description == "Location description"
+
+
+def test_parse_location_optional_annotated():
+    """Test 114: Parse Optional[Annotated[LocationArgument, 'desc']] parameter."""
+
+    class Node(TestArgumentParsingNode):
+        @action
+        def test_action(
+            self,
[…]
