+++ src/madsci_node_module/madsci/node_module/abstract_node_module.py
@@ -42,9 +42,6 @@ from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.base_types import Error
 from madsci.common.types.datapoint_types import DataPoint, FileDataPoint, ValueDataPoint
 from madsci.common.types.event_types import Event, EventType
-from madsci.common.types.location_types import (
-    LocationArgument,
-)
 from madsci.common.types.node_types import (
     AdminCommands,
     NodeCapabilities,
@@ -56,12 +53,12 @@ from madsci.common.types.node_types import (
     NodeStatus,
 )
 from madsci.common.utils import (
-    is_optional,
     pretty_type_repr,
     repeat_on_interval,
     threaded_daemon,
     to_snake_case,
 )
+from madsci.node_module.type_analyzer import analyze_type
 from pydantic import BaseModel, ValidationError
 from semver import Version
 
@@ -457,59 +454,38 @@ class AbstractNode(MadsciClientMixin):
         self.node_info.actions[action_name] = action_def
 
     def _is_file_type(self, type_hint: Any) -> bool:
-        """Check if a type hint represents a file parameter (Path or list[Path]).
+        """Check if a type hint represents a file parameter.
+
+        Now uses TypeAnalyzer for robust type detection at any nesting level.
 
         Args:
-            type_hint: The type hint to check (after extracting from Annotated/Optional)
+            type_hint: The type hint to check
 
         Returns:
             True if the type represents a file parameter (Path, list[Path], etc.)
         """
-        # Direct Path type
-        if getattr(type_hint, "__name__", None) in [
-            "Path",
-            "PurePath",
-            "PosixPath",
-            "WindowsPath",
-        ]:
-            return True
-
-        # Check for list[Path] - get_origin returns list, get_args returns (Path,)
-        origin = get_origin(type_hint)
-        if origin is list:
-            args = get_args(type_hint)
-            if args and getattr(args[0], "__name__", None) in [
-                "Path",
-                "PurePath",
-                "PosixPath",
-                "WindowsPath",
-            ]:
-                return True
-
-        return False
+        try:
+            type_info = analyze_type(type_hint)
+            return type_info.special_type == "file"
+        except (ValueError, TypeError):
+            return False
 
     def _contains_location_argument(self, type_hint: Any) -> bool:
-        """Check if a type hint contains LocationArgument either directly or within a Union.
+        """Check if a type hint contains LocationArgument.
+
+        Now uses TypeAnalyzer for robust type detection at any nesting level.
 
         Args:
-            type_hint: The type hint to check. Caller must unwrap Optional and Annotated types
-                      before calling this method.
+            type_hint: The type hint to check
 
         Returns:
-            True if the type is LocationArgument or a Union containing LocationArgument
+            True if the type is or contains LocationArgument
         """
-        # Direct LocationArgument type
-        if type_hint is LocationArgument:
-            return True
-
-        # Check for Union types that contain LocationArgument
-        origin = get_origin(type_hint)
-        if origin is Union:
-            args = get_args(type_hint)
-            # Check if any of the Union arguments is LocationArgument
-            return any(arg is LocationArgument for arg in args)
-
-        return False
+        try:
+            type_info = analyze_type(type_hint)
+            return type_info.special_type == "location"
+        except (ValueError, TypeError):
+            return False
 
     def _parse_action_arg(
         self,
@@ -518,81 +494,108 @@ class AbstractNode(MadsciClientMixin):
         parameter_name: str,
         parameter_type: Any,
     ) -> None:
-        """Parses a function argument of an action handler into a MADSci ArgumentDefinition"""
-        type_hint = parameter_type
[…]

[…]
