### 2025-12-11 8eba380
basic functionality seems to be working
