### 2025-12-12 151f174
added 20 min disconnect reset attempt timer
