+++ Configuration.md
@@ -118,6 +118,7 @@ Settings for the MADSci Workcell Manager.
 | `WORKCELL_REDIS_PASSWORD`                                                      | `string` \| `NoneType`           | `null`                                                   | The password for the redis server.                                                                                                                                                       | `null`                                                   |
 | `WORKCELL_SCHEDULER_UPDATE_INTERVAL`                                           | `number`                         | `5.0`                                                    | The interval at which the scheduler runs, in seconds. Must be >= node_update_interval                                                                                                    | `5.0`                                                    |
 | `WORKCELL_NODE_UPDATE_INTERVAL`                                                | `number`                         | `2.0`                                                    | The interval at which the workcell queries its node's states and status, in seconds. Must be <= scheduler_update_interval                                                                | `2.0`                                                    |
+| `WORKCELL_RECONNECT_ATTEMPT_INTERVAL`                                          | `number`                         | `1200.0`                                                 | The interval at which the workcell resets disconnected nodes, in seconds. Must be <= scheduler_update_interval                                                                           | `1200.0`                                                 |
 | `WORKCELL_NODE_INFO_UPDATE_INTERVAL`                                           | `number`                         | `60.0`                                                   | The interval at which the workcell queries its node's info, in seconds. Node info changes infrequently, so this can be much larger than node_update_interval to reduce network overhead. | `60.0`                                                   |
 | `WORKCELL_COLD_START_DELAY`                                                    | `integer`                        | `0`                                                      | How long the Workcell engine should sleep on startup                                                                                                                                     | `0`                                                      |
 | `WORKCELL_SCHEDULER`                                                           | `string`                         | `"madsci.workcell_manager.schedulers.default_scheduler"` | Scheduler module that contains a Scheduler class that inherits from AbstractScheduler to use                                                                                             | `"madsci.workcell_manager.schedulers.default_scheduler"` |
