### 2025-12-12 6875802
made node be initializing when it's status is reset
