### 2025-12-12 a45aace
Final testing and validation
