+++ src/madsci_node_module/tests/test_action_parsing_integration.py
@@ -1,0 +1,1321 @@
+"""
+Integration tests for action parsing fixes.
+
+This module contains end-to-end tests for the action parsing system,
+verifying that complex type hints work correctly in realistic scenarios.
+
+These tests cover:
+- Complete node creation with complex action types
+- Action execution with optional parameters
+- Action execution with ActionResult returns
+- OpenAPI schema generation
+- Error handling and validation
+- Real-world scenarios
+
+Test numbers: 180-244 (65 tests total)
+"""
+
+import tempfile
+import time
+from pathlib import Path
+from typing import Annotated, Optional, Union
+
+import pytest
+from madsci.common.types.action_types import (
+    ActionFailed,
+    ActionFiles,
+    ActionJSON,
+    ActionRequest,
+    ActionResult,
+    ActionRunning,
+    ActionStatus,
+    ActionSucceeded,
+)
+from madsci.common.types.base_types import Error
+from madsci.common.types.location_types import LocationArgument
+from madsci.common.types.node_types import NodeDefinition, RestNodeConfig
+from madsci.node_module.helpers import action
+from madsci.node_module.rest_node_module import RestNode
+from pydantic import BaseModel
+
+# ============================================================================
+# Test Fixtures and Helper Classes
+# ============================================================================
+
+
+class TestIntegrationNodeConfig(RestNodeConfig):
+    """Configuration for integration test nodes."""
+
+    __test__ = False
+    test_param: int = 1
+    update_node_files: bool = False
+
+
+def create_test_node_instance(node_class):
+    """Helper to create a node instance with proper config and definition."""
+    node_definition = NodeDefinition(
+        node_name="test_node",
+        module_name="test_module",
+        description="Integration test node",
+    )
+    node_config = TestIntegrationNodeConfig(test_required_param=1)
+    return node_class(node_definition=node_definition, node_config=node_config)
+
+
+class TestIntegrationNode(RestNode):
+    """Base test node for integration tests."""
+
+    __test__ = False
+
+
+class CustomDataModel(BaseModel):
+    """Custom Pydantic model for testing."""
+
+    value: int
+    description: str
+
+
+@pytest.fixture
+def node_with_complex_actions():
+    """Create a node with complex action types for integration testing."""
+
+    class ComplexTestNode(TestIntegrationNode):
+        """Test node with various complex action signatures."""
+
+        @action
+        def action_optional_location(
+            self, loc: Optional[Annotated[LocationArgument, "Source location"]]
+        ) -> None:
+            """Action with optional annotated location."""
+
+        @action
+        def action_optional_file(
+            self, file: Optional[Annotated[Path, "Config file"]] = None
+        ) -> None:
+            """Action with optional annotated file."""
+
+        @action
+        def action_list_files(self, files: list[Path]) -> None:
+            """Action with list of files."""
+
+        @action
+        def action_optional_list_files(
+            self, files: Optional[list[Path]] = None
+        ) -> None:
+            """Action with optional list of files."""
+
+        @action
+        def action_mixed_params(
+            self,
+            value: int,
+            loc: Optional[LocationArgument] = None,
+            file: Optional[Path] = None,
+        ) -> None:
+            """Action with mixed parameter types."""
+
+        @action
+        def action_returns_action_failed(self) -> Optional[ActionFailed]:
+            """Action that can return ActionFailed."""
+            return ActionFailed(errors=[Error(message="Test failure")])
+
+        @action
+        def action_returns_action_succeeded(self) -> Optional[ActionSucceeded]:
+            """Action that can return ActionSucceeded."""
+            return ActionSucceeded()
+
+        @action
+        def action_complex_nested(
+            self,
[…]
