+++ ruff.toml
@@ -115,7 +115,7 @@ unfixable = []
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
 [lint.per-file-ignores]
-"**/tests/**/*.py" = ["S101", "ANN", "D"]
+"**/tests/**/*.py" = ["S101", "ANN", "D", "ARG002"]
 "**/*.ipynb" = ["T20"]
 
 [format]

[…]
