### 2025-12-13 5671b8d
Update src/madsci_workcell_manager/madsci/workcell_manager/workcell_server.py
Co-authored-by: Ryan D. Lewis <ryan.lewis@anl.gov>
