### 2025-12-13 674533b
Update src/madsci_common/madsci/common/types/workcell_types.py
Co-authored-by: Ryan D. Lewis <ryan.lewis@anl.gov>
