+++ pyproject.toml
@@ -13,9 +13,9 @@ dev = [
     "ipykernel>=6.29.5",
     "pre-commit>=4.3.0",
     "pydantic-settings-export>=1.0.2",
-    "ruff>=0.13.0",
     "nbconvert>=7.16.6",
     "pytest-timeout>=2.4.0",
+    "ruff>=0.14.9",
 ]
 
 [tool.pytest.ini_options]
