### 2025-12-15 f1f6b57
Fix: test_disconnect_node_on_connection_failure bypassing asserts due to lack of nodes
