+++ src/madsci_workcell_manager/tests/test_workcell_engine.py
@@ -131,12 +131,18 @@ def test_disconnect_node_on_connection_failure(engine: Engine) -> None:
         "madsci.client.node.rest_node_client.RestNodeClient.get_status",
         side_effect=Exception("Connection failed"),
     ):
+        engine.state_handler.set_node(
+            node_name="Test Node",
+            node=test_node,
+        )
+
         engine.update_active_nodes(engine.state_handler)
         for node in engine.state_handler.get_nodes().values():
             assert node.status.disconnected is True
         engine.reset_disconnects()
         for node in engine.state_handler.get_nodes().values():
-            assert node.status.disconnected is True
+            assert node.status.initializing is True
+            assert node.status.disconnected is False
 
 
 def test_run_next_step_with_ready_workflow(
