### 2025-12-15 c80987a
Update: engine test fixture now includes node by default
