+++ src/madsci_workcell_manager/tests/test_workcell_engine.py
@@ -110,7 +110,9 @@ def engine(state_handler: WorkcellStateHandler) -> Engine:
         mock_location_client.return_value = mock_location_client_instance
 
         warnings.simplefilter("ignore", UserWarning)
-        return Engine(state_handler=state_handler, data_client=DataClient())
+        engine = Engine(state_handler=state_handler, data_client=DataClient())
+        engine.state_handler.set_node(node_name="node1", node=test_node)
+        return engine
 
 
 def test_engine_initialization(engine: Engine) -> None:
@@ -131,11 +133,6 @@ def test_disconnect_node_on_connection_failure(engine: Engine) -> None:
         "madsci.client.node.rest_node_client.RestNodeClient.get_status",
         side_effect=Exception("Connection failed"),
     ):
-        engine.state_handler.set_node(
-            node_name="Test Node",
-            node=test_node,
-        )
-
         engine.update_active_nodes(engine.state_handler)
         for node in engine.state_handler.get_nodes().values():
             assert node.status.disconnected is True
@@ -175,10 +172,6 @@ def test_run_single_step(engine: Engine, state_handler: WorkcellStateHandler) ->
         status=WorkflowStatus(running=True),
     )
     state_handler.set_active_workflow(workflow)
-    state_handler.set_node(
-        node_name="node1",
-        node=test_node,
-    )
     with patch(
         "madsci.workcell_manager.workcell_engine.find_node_client"
     ) as mock_client:
@@ -229,10 +222,6 @@ def test_run_single_step_with_update_parameters(
         status=WorkflowStatus(running=True),
     )
     state_handler.set_active_workflow(workflow)
-    state_handler.set_node(
-        node_name="node1",
-        node=test_node,
-    )
     with patch(
         "madsci.workcell_manager.workcell_engine.find_node_client"
     ) as mock_client:
@@ -263,10 +252,6 @@ def test_run_single_step_of_workflow_with_multiple_steps(
         status=WorkflowStatus(running=True),
     )
     state_handler.set_active_workflow(workflow)
-    state_handler.set_node(
-        node_name="node1",
-        node=test_node,
-    )
     with patch(
         "madsci.workcell_manager.workcell_engine.find_node_client"
     ) as mock_client:
@@ -330,9 +315,7 @@ def test_finalize_step_failure(
     assert finalized_workflow.steps[0].status == ActionStatus.FAILED
 
 
-def test_handle_data_and_files_with_data(
-    engine: Engine, state_handler: WorkcellStateHandler
-) -> None:
+def test_handle_data_and_files_with_data(engine: Engine) -> None:
     """Test handle_data_and_files with data points."""
     step = Step(
         name="Test Step",
@@ -344,10 +327,6 @@ def test_handle_data_and_files_with_data(
         name="Test Workflow",
         steps=[step],
         status=WorkflowStatus(running=True),
-    )
-    state_handler.set_node(
-        node_name="node1",
-        node=test_node,
     )
     action_result = ActionSucceeded(json_result=42)
 
@@ -372,9 +351,7 @@ def test_handle_data_and_files_with_data(
         )
 
 
-def test_handle_data_and_files_with_files(
-    engine: Engine, state_handler: WorkcellStateHandler
-) -> None:
+def test_handle_data_and_files_with_files(engine: Engine) -> None:
     """Test handle_data_and_files with file points."""
     step = Step(
         name="Test Step",
@@ -387,10 +364,6 @@ def test_handle_data_and_files_with_files(
         name="Test Workflow",
         steps=[step],
         status=WorkflowStatus(running=True),
-    )
-    state_handler.set_node(
-        node_name="node1",
-        node=test_node,
     )
[…]
