+++ README.md
@@ -12,7 +12,7 @@
 | src/madsci\_client/madsci/client/lab\_client.py                                               |       35 |       21 |     40% |31-43, 52-58, 67-73, 82-88, 97-103 |
 | src/madsci\_client/madsci/client/location\_client.py                                          |       98 |       39 |     60% |45, 62, 76, 86, 104-112, 132-140, 189-198, 218-226, 254-263, 318-327, 374-382, 410-426 |
 | src/madsci\_client/madsci/client/node/\_\_init\_\_.py                                         |        4 |        0 |    100% |           |
-| src/madsci\_client/madsci/client/node/abstract\_node\_client.py                               |       30 |        4 |     87% |27, 30, 100-101 |
+| src/madsci\_client/madsci/client/node/abstract\_node\_client.py                               |       30 |        2 |     93% |    27, 30 |
 | src/madsci\_client/madsci/client/node/rest\_node\_client.py                                   |      208 |       44 |     79% |63, 146, 282-293, 324-334, 356-391, 418-421, 528, 558, 560, 578 |
 | src/madsci\_client/madsci/client/resource\_client.py                                          |      636 |      200 |     69% |102, 127, 149, 176-179, 187, 244-247, 264, 330-334, 350-365, 394, 428-431, 484-491, 497, 514, 526-527, 576, 607, 654-657, 694-696, 738-739, 778-779, 815, 851-852, 888-889, 925-926, 962-963, 994-995, 1026-1031, 1062-1064, 1095-1148, 1200-1210, 1240-1243, 1294-1304, 1325, 1328-1339, 1371-1378, 1402-1405, 1463-1481, 1502-1508, 1557-1583, 1627-1665, 1696-1713, 1797-1801, 1815, 1838-1839, 1847-1851, 1867-1868, 1879, 1916, 1929, 1942-1952 |
 | src/madsci\_client/madsci/client/workcell\_client.py                                          |      198 |       45 |     77% |94, 120, 146, 160, 209-213, 219, 245, 249, 273-278, 388-406, 433-461, 503, 511, 519, 524-525, 638, 671 |
@@ -51,14 +51,14 @@
 | src/madsci\_common/madsci/common/types/location\_types.py                                     |      102 |        9 |     91% |30, 32, 34, 36, 41, 46, 51, 98, 170 |
 | src/madsci\_common/madsci/common/types/manager\_types.py                                      |       46 |        5 |     89% |     27-31 |
 | src/madsci\_common/madsci/common/types/mongodb\_migration\_types.py                           |      214 |      116 |     46% |84, 105, 129-130, 150, 159, 165, 191, 197, 200-203, 206, 218-227, 248-262, 266-271, 293-297, 303-319, 323-333, 338-348, 352-359, 363, 367, 371, 385-441, 458-490 |
-| src/madsci\_common/madsci/common/types/node\_types.py                                         |      144 |        7 |     95% |398, 406, 410, 419, 425, 429, 459 |
+| src/madsci\_common/madsci/common/types/node\_types.py                                         |      149 |        5 |     97% |403, 417, 426, 438, 468 |
 | src/madsci\_common/madsci/common/types/parameter\_types.py                                    |       37 |        8 |     78% |14, 23, 47, 49, 51, 68, 70, 72 |
 | src/madsci\_common/madsci/common/types/resource\_types/\_\_init\_\_.py                        |      296 |       46 |     84% |119, 163, 240, 250, 271, 276, 280, 284, 325, 338, 344, 366, 370, 404, 426, 430, 478-482, 512, 555, 560-561, 572-588, 591-600, 607-609, 640, 655, 684-687, 724-727, 764, 769, 779, 783 |
[…]
