### 2025-12-15 dfdb0dd
Bump to v0.6.0-rc.6
