### 2025-12-15 f408733
Fix: actions incorrectly defaulting to non-blocking
