+++ README.md
@@ -79,7 +79,7 @@
 | src/madsci\_location\_manager/madsci/location\_manager/location\_state\_handler.py            |       76 |       15 |     80% |72, 84-90, 102-103, 114-115, 127-129, 145-146 |
 | src/madsci\_location\_manager/madsci/location\_manager/transfer\_planner.py                   |      175 |        2 |     99% |  282, 315 |
 | src/madsci\_node\_module/madsci/node\_module/\_\_init\_\_.py                                  |        4 |        0 |    100% |           |
-| src/madsci\_node\_module/madsci/node\_module/abstract\_node\_module.py                        |      521 |      125 |     76% |77, 81, 83, 85, 87, 89, 91, 95, 97, 99, 101, 112, 128, 139, 164-168, 197-199, 208, 239-247, 257-264, 292-304, 315-316, 327-337, 359-361, 375-377, 470-471, 484-488, 546, 637, 664, 689, 711, 742, 762, 765, 812-813, 842-843, 855-857, 863-864, 973, 985, 989, 999-1004, 1025, 1028-1040, 1057, 1070, 1076-1077, 1092, 1102-1103, 1109-1110, 1137-1145, 1161, 1177-1181, 1195-1199, 1213-1214, 1228-1229, 1250-1253 |
+| src/madsci\_node\_module/madsci/node\_module/abstract\_node\_module.py                        |      521 |      123 |     76% |77, 81, 83, 85, 87, 89, 91, 95, 97, 99, 101, 112, 128, 139, 164-168, 197-199, 208, 239-247, 257-264, 292-304, 315-316, 327-337, 359-361, 375-377, 470-471, 484-488, 546, 637, 664, 689, 711, 742, 762, 765, 812-813, 842-843, 855-857, 863-864, 973, 985, 989, 999-1004, 1025, 1028-1040, 1076-1077, 1092, 1102-1103, 1109-1110, 1137-1145, 1161, 1177-1181, 1195-1199, 1213-1214, 1228-1229, 1250-1253 |
 | src/madsci\_node\_module/madsci/node\_module/helpers.py                                       |      157 |       62 |     61% |43, 70-113, 118-135, 149, 168-173, 189-192, 242, 261 |
 | src/madsci\_node\_module/madsci/node\_module/rest\_node\_module.py                            |      390 |       88 |     77% |64, 68, 70, 72, 82-96, 107-130, 165, 170-181, 220-224, 250, 261-263, 274-288, 303, 311-325, 329, 347-349, 383, 386, 409, 412, 476, 502, 515-520, 532-542, 572, 576, 579-581, 631, 658, 672, 688, 699-700, 798, 886 |
 | src/madsci\_node\_module/madsci/node\_module/type\_analyzer.py                                |      156 |       15 |     90% |122, 217, 270-271, 288-289, 305-306, 358-360, 368-370, 377-380 |
@@ -102,7 +102,7 @@
 | src/madsci\_workcell\_manager/madsci/workcell\_manager/workcell\_server.py                    |      270 |       73 |     73% |86-87, 123-158, 183-184, 231, 243-247, 271-272, 284, 292, 302-305, 313-324, 355, 370, 384, 398, 404, 433-435, 439-444, 469-471, 508-510, 525, 536, 567-572, 579-580 |
 | src/madsci\_workcell\_manager/madsci/workcell\_manager/workcell\_utils.py                     |        9 |        0 |    100% |           |
 | src/madsci\_workcell\_manager/madsci/workcell\_manager/workflow\_utils.py                     |      199 |       67 |     66% |40-41, 48-58, 72-77, 95, 113, 131, 144, 160, 165-176, 244, 251-252, 268, 279, 300, 305, 313, 330-342, 358-380, 396, 419-426, 433, 438-441, 447-448 |
[…]
