### 2025-12-15 7b95efb
Bump to v0.6.0
