### 2025-12-24 #9416 @lucask07
@LuckierDodge I'm working through the paper review and had the following thoughts.

**Terminology bloat (?)**. The summary attempts to define specific terminology such as nodes and managers. I would advise omitting these definitions from the summary and moving it to the main text (or a glossary in an appendix). Other terms I found unclear later in the main body were: workcell and the difference between a workflow, experiment run, and an experiment campaign. My advice would be to investigate simplifying the terminology and after simplification make edits that more completely define specific terms.

A persistent challenge for lab automation is the **instrument specific communication**. It is not clear how I would implement a real new device. The madsci_node_module README page is very detailed yet I suspect would be a challenge for a new user to add a new robot, liquid handler, or other device. In the Quick Start at https://github.com/AD-SDL/MADSci/tree/main/src/madsci_node_module where does 'MyDeviceInterface' come from?
Perhaps I am misunderstanding but I think it would be beneficial to have an example with an **actual** instrument setup as a node; the example_modules such as https://github.com/AD-SDL/MADSci/blob/main/example_lab/example_modules/robotarm.py refer to a "fake robot arm". Can you add an example of a real device?

The text emphasizes a domain agnostic approach but then seems to make an **assumption of a wet lab (?)** in the resource management section. My EE lab uses software automation yet has no consumables or labware! This section also introduces the terminology containers, the necessity of this new term is not apparent to me.

This is a nicely done project and I will continue to work through my review.
