### 2025-12-27 e25b889
Fix experiment application automatic client init
