+++ src/madsci_experiment_application/tests/test_experiment_application.py
@@ -1,5 +1,6 @@
 """Unit tests for ExperimentApplication class."""
 
+from contextlib import contextmanager
 from typing import Any
 from unittest.mock import Mock, patch
 
@@ -25,6 +26,30 @@ from madsci.experiment_application import (
     ExperimentApplicationConfig,
 )
 from pydantic import AnyUrl
+
+
+@contextmanager
+def mock_all_clients():
+    """Context manager to mock all MADSci clients at the module level."""
+    with (
+        patch("madsci.client.event_client.EventClient") as mock_event_client,
+        patch(
+            "madsci.client.experiment_client.ExperimentClient"
+        ) as mock_experiment_client,
+        patch("madsci.client.workcell_client.WorkcellClient") as mock_workcell_client,
+        patch("madsci.client.location_client.LocationClient") as mock_location_client,
+        patch("madsci.client.resource_client.ResourceClient") as mock_resource_client,
+        patch("madsci.client.data_client.DataClient") as mock_data_client,
+    ):
+        # Set up mock instances
+        mock_event_client.return_value = Mock()
+        mock_experiment_client.return_value = Mock()
+        mock_workcell_client.return_value = Mock()
+        mock_location_client.return_value = Mock()
+        mock_resource_client.return_value = Mock()
+        mock_data_client.return_value = Mock()
+
+        yield
 
 
 class TestExperimentApplication(ExperimentApplication):
@@ -131,15 +156,7 @@ def experiment_app(
     experiment_design: ExperimentDesign,
 ) -> TestExperimentApplication:
     """Create a test ExperimentApplication instance."""
-    with patch.multiple(
-        "madsci.experiment_application.experiment_application",
-        EventClient=Mock,
-        ExperimentClient=Mock,
-        WorkcellClient=Mock,
-        LocationClient=Mock,
-        ResourceClient=Mock,
-        DataClient=Mock,
-    ):
+    with mock_all_clients():
         app = TestExperimentApplication(
             node_definition=node_definition,
             node_config=app_config,
@@ -182,17 +199,77 @@ def experiment_app_with_mocks(
 class TestExperimentApplicationInit:
     """Test ExperimentApplication initialization."""
 
+    def test_client_attributes_are_not_shadowed_by_type_annotations(
+        self,
+    ) -> None:
+        """Test that type annotations don't shadow MadsciClientMixin properties (issue #205)."""
+
+        # Check that ExperimentApplication class attributes don't shadow the mixin properties
+        # If they do, accessing these attributes will return the class itself, not trigger the property
+        assert (
+            not hasattr(ExperimentApplication, "__dict__")
+            or "experiment_client" not in ExperimentApplication.__dict__
+        ), (
+            "experiment_client should not be a class attribute - it should be inherited as a property from MadsciClientMixin"
+        )
+        assert (
+            not hasattr(ExperimentApplication, "__dict__")
+            or "workcell_client" not in ExperimentApplication.__dict__
+        ), (
+            "workcell_client should not be a class attribute - it should be inherited as a property from MadsciClientMixin"
+        )
+
+        # Verify that the properties are actually properties from the mixin
+        assert isinstance(
+            type(ExperimentApplication).__dict__.get("experiment_client"),
+            (property, type(None)),
+        ), "experiment_client should be a property or inherited"
+        assert isinstance(
+            type(ExperimentApplication).__dict__.get("workcell_client"),
+            (property, type(None)),
+        ), "workcell_client should be a property or inherited"
+
+    def test_all_client_properties_accessible_on_instances(
+        self, experiment_app: TestExperimentApplication
+    ) -> None:
+        """Test that all client properties are accessible on instances (issue #205).
+
+        This is an integration test that verifies the fix for issue #205 where
+        ExperimentApplication instances were getting AttributeError when accessing
[…]
