### 2025-12-31 a454516
Make service configuration a little easier
