+++ compose.yaml
@@ -25,8 +25,8 @@ x-madsci-service: &madsci-service
     - GROUP_ID=${GROUP_ID:-1000}
   network_mode: host
   env_file:
-    - .env # Optional: Load environment variables from a .env file
     - example_lab/.env # Load environment variables from the example_lab directory
+    - .env # Optional: Load environment variables from a .env file
   volumes:
     - ${REPO_PATH:-.}/example_lab:/home/madsci/example_lab
     - ${REPO_PATH:-.}/.madsci:/home/madsci/.madsci
@@ -35,25 +35,25 @@ x-madsci-service: &madsci-service
   working_dir: /home/madsci/example_lab
 
 services:
-  mongodb:
-    container_name: mongodb
+  madsci_mongodb:
+    container_name: madsci_mongodb
     image: mongo:8.0
     restart: unless-stopped
     ports:
-      - 27017:27017
+      - ${MONGODB_PORT:-27017}:27017
     volumes:
       - ${REPO_PATH:-.}/.madsci/mongodb:/data/db
-  redis:
-    container_name: redis
+  madsci_redis:
+    container_name: madsci_redis
     image: redis:7.4
     restart: unless-stopped
     ports:
-      - 6379:6379
+      - ${REDIS_PORT:-6379}:6379
     volumes:
       - ${REPO_PATH:-.}/.madsci/redis:/data
     command: redis-server --appendonly yes --save 900 1 --save 300 10 --save 60 10000
-  postgres:
-    container_name: postgres
+  madsci_postgres:
+    container_name: madsci_postgres
     image: postgres:17
     restart: unless-stopped
     environment:
@@ -61,19 +61,19 @@ services:
       - POSTGRES_PASSWORD=madsci
       - POSTGRES_DB=resources
     ports:
-      - 5432:5432
+      - ${POSTGRES_PORT:-5432}:5432
     volumes:
       - ${REPO_PATH:-.}/.madsci/postgresql/data:/var/lib/postgresql/data
-  minio:
-    container_name: minio
+  madsci_minio:
+    container_name: madsci_minio
     image: minio/minio
     restart: unless-stopped
     environment:
       - MINIO_ROOT_USER=minioadmin
       - MINIO_ROOT_PASSWORD=minioadmin
     ports:
-      - 9000:9000
-      - 9001:9001
+      - ${MINIO_PORT:-9000}:9000
+      - ${MINIO_CONSOLE_PORT:-9001}:9001
     volumes:
       - ${REPO_PATH:-.}/.madsci/minio:/home/madsci/.madsci/minio
     command: server /home/madsci/.madsci/minio --console-address ":9001"
@@ -92,7 +92,7 @@ services:
         - ghcr.io/ad-sdl/madsci_dashboard:dev
     command: python -m madsci.squid.lab_server
     healthcheck:
-      test: ["CMD", "curl", "-f", "http://localhost:8000/health", "||", "exit", "1"]
+      test: ["CMD", "curl", "-f", "${LAB_SERVER_URL:-http://localhost:8000}/health", "||", "exit", "1"]
       interval: 30s
       timeout: 3s
       retries: 3
@@ -105,14 +105,14 @@ services:
     container_name: event_manager
     command: python -m madsci.event_manager.event_server
     depends_on:
-        - mongodb
+        - madsci_mongodb
 
   experiment_manager:
     <<: *madsci-service
     container_name: experiment_manager
     command: python -m madsci.experiment_manager.experiment_server
     depends_on:
-      - mongodb
+      - madsci_mongodb
       - lab_manager
 
   resource_manager:
@@ -120,7 +120,7 @@ services:
     container_name: resource_manager
     command: python -m madsci.resource_manager.resource_server
     depends_on:
-      - postgres
+      - madsci_postgres
       - event_manager
 
[…]
