### 2026-01-01 a15d342
Add myst and pdoc for documentation site
