### 2026-01-01 9f7b8c4
Bump to v0.6.1; update lockfile
