### 2026-01-01 #9416 @LuckierDodge
Hey @lucask07, thank you for the valuable feedback!

- Agreed on the terminology, will open a PR with some rewrites to make that clearer and move definitions out of the summary section. In terms of term simplification overall...it's been an ongoing struggle for us, but we'll work on it 😅
- We actually have a number of examples of real robot integrations in separate repos, (see https://github.com/AD-SDL#madsci-node-modules for an incomplete list). Would adding links/references to those in appropriate places in the documentation/example lab suffice?
- Agreed that we could make clearer where the Device Interface comes from/it's purpose, will update to address
- Will update to make it clear that the resource management is optional, and while it's design and implementation was certainly motivated for bio and chem labs with a lot of labware/consumables, it is both flexible enough to be used for other purposes, and not required to gain the other benefits of the system
