### 2026-01-01 3be3919
Update text based on feedback for clarity and better definition of terms
