+++ paper.md
@@ -46,25 +46,22 @@ bibliography: paper.bib
 # Summary
 
 The Modular Autonomous Discovery for Science (MADSci) toolkit enables laboratory automation, high-throughput experimentation, and self-driving labs.
-Laboratory operators define workcells as collections of **Nodes** controlling individual instruments, robots, or devices.
-These nodes, combined with **Managers** providing lab functionality and coordination, form flexible laboratories.
-Users create and run experimental campaigns using Python applications and YAML workflow definitions.
-This design separates concerns between device integrators, lab operators, and domain scientists.
-MADSci enables science in biology, chemistry, materials science, and quantum science.
+MADSci provides a microservices architecture integrating laboratory instruments, robots, and devices with coordination services for workflows, data management, and resource tracking.
+Users define and execute experiments using Python applications and YAML-based workflow definitions.
+This modular design separates concerns between device integrators, lab operators, and domain scientists.
+MADSci supports research in biology, chemistry, materials science, and quantum science.
 
 # Statement of Need
 
 The lab automation ecosystem is fragmented, with many proprietary, costly, or narrowly domain-specific tools.
-MADSci provides an open-source, extensible, domain-agnostic toolkit for labs.
-It integrates equipment, sensors, and robots as "nodes," with managers for workflows, resources, experiments, logging, and data collection.
-Built on a microservices architecture with developer-friendly RESTful APIs and Python clients, MADSci leverages standard databases (PostgreSQL, MongoDB, Redis, MiniIO) and open-source libraries (FastAPI, Pydantic, SQLModel) to ensure adaptability and extensibility.
+MADSci provides an open-source, extensible, domain-agnostic toolkit integrating equipment, sensors, and robots as "nodes" with managers for workflows, resources, experiments, logging, and data collection.
+Built on a microservices architecture with RESTful APIs and Python clients, MADSci leverages standard databases (PostgreSQL, MongoDB, Redis, MiniIO) and open-source libraries (FastAPI, Pydantic, SQLModel).
 
-Laboratory automation software systems exhibit distinct architectural approaches and domain focus areas that influence their applicability across research environments.
-Commercial platforms such as Chemspeed's Autosuite [@seifrid2024chemspyd] and Retisoft Genera [@retisoft2024genera] provide comprehensive functionality but operate under proprietary licensing models that may limit accessibility in academic settings.
-Open-source alternatives including AlabOS [@fei2024alabos], ChemOS [@sim2024chemos; @roch2020chemos], and Bluesky [@allan2019bluesky] demonstrate strong capabilities within specific research domains but exhibit varying degrees of cross-disciplinary transferability.
-Specialized systems such as Polybot [@wang2025autonomous; @vriza2023selfdriving] showcase advanced automation features tailored to particular scientific applications.
-The Workcell Execution Interface [@vescovi2023modular], MADSci's predecessor, has a similar focus on instrument modularity but lacks some of the advanced features and microservices principles at the management layer.
-MADSci represents an open-source, modular framework designed to provide domain-agnostic laboratory automation capabilities while maintaining compatibility with diverse instrumentation ecosystems and institutional deployment requirements.
+Commercial platforms such as Chemspeed's Autosuite [@seifrid2024chemspyd] and Retisoft Genera [@retisoft2024genera] provide comprehensive functionality but operate under proprietary licensing that may limit academic accessibility.
+Open-source alternatives including AlabOS [@fei2024alabos], ChemOS [@sim2024chemos; @roch2020chemos], and Bluesky [@allan2019bluesky] demonstrate strong capabilities within specific domains but exhibit varying cross-disciplinary transferability.
[…]
