### 2026-01-01 80d4928
Add detail on location management
