### 2026-01-21 96378e7
Pause/resume functionality working so far.. Logging admin command notes to file for me, need to remove once confirmed working
