+++ src/madsci_common/madsci/common/types/admin_command_types.py
@@ -20,7 +20,6 @@ class AdminCommands(str, Enum):
     UNLOCK = "unlock"
     GET_LOCATION = "get_location"
 
-
 class AdminCommandResponse(MadsciBaseModel):
     """Response from an Admin Command"""
 
