+++ src/madsci_workcell_manager/madsci/workcell_manager/workcell_server.py
@@ -4,6 +4,7 @@ import json
 import traceback
 from contextlib import asynccontextmanager
 from typing import Annotated, Any, AsyncGenerator, Optional, Union
+from datetime import datetime
 
 from classy_fastapi import get, post
 from fastapi import FastAPI, Form, HTTPException, UploadFile
@@ -34,12 +35,12 @@ from madsci.workcell_manager.workflow_utils import (
     check_parameters,
     create_workflow,
     save_workflow_files,
-    cancel_workflow
 )
 from pymongo.synchronous.database import Database
 from ulid import ULID
 import threading
 import time
+
 
 # Module-level constants for Body() calls to avoid B008 linting errors
 LOOKUP_VAL_BODY = Body(...)
@@ -245,6 +246,8 @@ class WorkcellManager(
     ) -> AdminCommandResponse:
         """Send admin command to a node."""
         node = self.state_handler.get_node(node)
+        with open("notes.txt", "a", encoding="utf-8") as f:
+            f.write("\nworkcell_server.py: (send_admin_command_to_node) Posting admin command to node. " + datetime.now().strftime("%h:%M:%S"))
         if command in node.info.capabilities.admin_commands:
             client = find_node_client(node.node_url)
             return client.send_admin_command(command)
@@ -271,49 +274,72 @@ class WorkcellManager(
     def get_workflow(self, workflow_id: str) -> Workflow:
         """Get info on a specific workflow."""
         return self.state_handler.get_workflow(workflow_id)
-
+    
     @post("/workflow/{workflow_id}/pause")
     def pause_workflow(self, workflow_id: str) -> Workflow:
-        """Pause a specific workflow."""
-        with self.state_handler.wc_state_lock():
-            wf = self.state_handler.get_active_workflow(workflow_id)
-            if wf.status.active:
-                if wf.status.running:
-                    self.send_admin_command_to_node(
-                        "pause", wf.steps[wf.status.current_step_index].node
-                    )
-                wf.status.paused = True
-                self.state_handler.set_active_workflow(wf)
+        """Test background polling"""
+        pause_thread = threading.Thread(
+            target=self._pause_workflow_background,
+            args=(workflow_id,),
+            daemon=True
+        )
+        with open("notes.txt", "a", encoding="utf-8") as f:
+            f.write("\nworkcell_server.py: (pause_workflow) Pause workflow thread starting. " + datetime.now().strftime("%h:%M:%S"))
+        pause_thread.start()
+        return self.state_handler.get_workflow(workflow_id)
+    
+    def _pause_workflow_background(self, workflow_id: str) -> None:
+        max_retries = 100
+        retry_count = 0
+        poll_interval = 0.5
 
-        return self.state_handler.get_active_workflow(workflow_id)
+        while retry_count < max_retries:
+            try:
+                with self.state_handler.wc_state_lock():
+                    wf = self.state_handler.get_workflow(workflow_id)
+                    if not wf.status.active:
+                        self.logger.log(f"Workflow {workflow_id} already terminal or paused. Pause complete.")
+                        return
+                    if wf.status.active and 0 <= wf.status.current_step_index < len(wf.steps):
+                        current_node = wf.steps[wf.status.current_step_index].node
+                        try:
+                            response = self.send_admin_command_to_node("pause", current_node)
+                            if response.success:
+                                with open("notes.txt", "a", encoding="utf-8") as f:
+                                    f.write("\nworkcell_server.py: (_pause_workflow_background) Pause thread response successful. " + datetime.now().strftime("%h:%M:%S"))
+                                self.state_handler.set_active_workflow(wf)
+                                self.logger.log(f"Workflow {workflow_id} paused on node {current_node}.")
+                                return
+                            else:
[…]
