### 2026-01-27 cccbaa0
Update docs/github pages to include 'Enabled by MADSci' section
