+++ docs/madsci_powered/Modules.md
@@ -1,0 +1,33 @@
+# MADSci Node Modules
+
+An incomplete list of integrated devices and services available for use in your automated and autonomous laboratories:
+
+| Module Name              | Integrated Device(s)                          | Link |
+|--------------------------|-----------------------------------------------|------|
+| ot2_module               | Opentrons OT-2, Flex                          | [ot2_module](https://github.com/AD-SDL/ot2_module) |
+| pf400_module             | Precise Automation PF400                      | [pf400_module](https://github.com/AD-SDL/pf400_module) |
+| sciclops_module          | Hudson Robotics Sciclops                      | [sciclops_module](https://github.com/AD-SDL/sciclops_module) |
+| barty_module             | Barty the Bartending Robot                    | [barty_module](https://github.com/AD-SDL/barty_module) |
+| camera_module            | USB webcams                                   | [camera_module](https://github.com/AD-SDL/camera_module) |
+| ur_module                | Universal Robotics Arms, end-effectors        | [ur_module](https://github.com/AD-SDL/ur_module) |
+| hidex_module             | Hidex Sense microplate reader                 | [hidex_module](https://github.com/AD-SDL/hidex_module) |
+| big_kahuna_module        | Unchained Labs Big Kahuna                     | [big_kahuna_module](https://github.com/AD-SDL/big_kahuna_module) and [amews_big_kahuna_code](https://github.com/AD-SDL/amews_big_kahuna_code) |
+| hudson_solo_module       | Hudson SOLO                                   | [hudson_solo_module](https://github.com/AD-SDL/hudson_solo_module) |
+| hudson_platecrane_module | Hudson PlateCrane EX                          | [hudson_platecrane_module](https://github.com/AD-SDL/hudson_platecrane_module) |
+| a4s_sealer_module        | Azenta A4S Sealer                             | [a4s_sealer_module](https://github.com/AD-SDL/a4s_sealer_module) |
+| brooks_xpeel_module      | Brooks XPeel Peeler                           | [brooks_xpeel_module](https://github.com/AD-SDL/brooks_xpeel_module) |
+| pe_icp_module            | PerkinsElmer Syngistix ICP                    | [pe_icp_module](https://github.com/AD-SDL/pe_icp_module) |
+| sciclops_module          | Hudson Sciclops Platecrane                    | [sciclops_module](https://github.com/AD-SDL/sciclops_module) |
+| liconic_module           | Liconic STX Incubator                         | [liconic_module](https://github.com/AD-SDL/liconic_module) |
+| bmg_module               | BMG Vantastar Microplate Readerr              | [bmg_module](https://github.com/AD-SDL/bmg_module) |
+| vention_rail_module      | Vention Rack and Pinion Linear Rail           | [vention_rail_module](https://github.com/AD-SDL/vention_rail_module) |
+| labjack_module           | LabJack DAQ Devices                           | [labjack_module](https://github.com/AD-SDL/labjack_module) |
+| pal_module               | PAL System Automated Sample Preparation Robot | [pal_module](https://github.com/AD-SDL/pal_module) |
+| biometra_module          | Biometra TRobot II Thermocycler               | [biometra_module](https://github.com/AD-SDL/biometra_module) |
+| zigbee_module            | Low Cost Zigbee Sensors                       | [zigbee_module](https://github.com/AD-SDL/zigbee_module) |
+| inheco_incubator_module  | Inheco Single Plate Incubators                | [inheco_incubator_module](https://github.com/AD-SDL/inheco_incubator_module) |
+| metrohm_m101_module      | Metrohm M101 Multichannel Galvinometer        | [metrohm_m101_module](https://github.com/AD-SDL/metrohm_m101_module) |
+
+
[…]
