### 2026-01-27 7e64495
Add JOSS status badge to README
