### 2026-01-27 #9416 @RyanTheRobothead
- @lucask07 FYI here's the link to the new documentation for the (in)complete list of MADSci modules that we're aware of as of writing: https://ad-sdl.github.io/MADSci/modules/, and we updated the example lab to link to it at the top of the README.
- @arfon it seems that both @lucask07 and @ktahar have completed their review checklists (correct me if I'm wrong); may I enquire what the next step is?
