### 2026-01-29 de2ec4f
Initial plan document for event improvements
