+++ docs/dev/event_manager_improvement_plan.md
@@ -1,0 +1,1514 @@
+# EventClient/EventManager Improvements Development Plan
+
+> **Meta Issue:** [#218 - Meta: Event Client/Manager Improvements](https://github.com/AD-SDL/MADSci/issues/218)
+> **Created:** January 2026
+> **Status:** Planning
+
+## Overview
+
+This document outlines a phased, test-driven development plan for improving MADSci's EventClient and EventManager components. The work addresses four related GitHub issues:
+
+| Issue | Title | Type | Priority |
+|-------|-------|------|----------|
+| [#217](https://github.com/AD-SDL/MADSci/issues/217) | Feature: Automatic Logging Info | Enhancement | High |
+| [#216](https://github.com/AD-SDL/MADSci/issues/216) | Feature: StructLog Support | Enhancement | High |
+| [#215](https://github.com/AD-SDL/MADSci/issues/215) | Research: OpenTelemetry and the EventManager | Research | Medium |
+| [#80](https://github.com/AD-SDL/MADSci/issues/80) | Enhancement: Event/Log Rotation, Retention, and Backup | Enhancement | Medium |
+
+## Design Decisions
+
+The following decisions were made during planning:
+
+1. **OpenTelemetry Research Output**: Both a research document and proof-of-concept implementation
+2. **Event Retention Strategy**: Configurable soft-delete → hard-delete workflow using MongoDB TTL indexes
+3. **Structlog API**: Idiomatic structlog API as default (`logger.info()`, `logger.debug()`), with `log_*` aliases for backward compatibility
+4. **Structlog Configuration Scope**: Per-instance configuration using `structlog.wrap_logger()` for full isolation
+5. **Error Handling**: Configurable behavior via `fail_on_error` setting (default: silent logging)
+6. **Bulk Operations**: Batch limits with configurable size to prevent performance impact
+7. **Test Coverage Target**: Comprehensive (~80%+ coverage with unit, integration, and edge case tests)
+8. **Breaking Changes**: Acceptable where they improve functionality, usability, and feature set
+
+## Architecture Context
+
+### Key Files
+
+| File | Purpose |
+|------|---------|
+| `src/madsci_client/madsci/client/event_client.py` | EventClient implementation |
+| `src/madsci_event_manager/madsci/event_manager/event_server.py` | EventManager server |
+| `src/madsci_common/madsci/common/types/event_types.py` | Event, EventType, Settings, Config types |
+| `src/madsci_client/tests/test_event_client.py` | EventClient unit tests |
+| `src/madsci_event_manager/tests/test_event_server.py` | EventManager unit tests |
+
+### Current Architecture Notes
+
+- **EventClient is dual-purpose**: Both a Python logger AND an HTTP client to the EventManager
+- **Asynchronous sending**: Events are sent to server in background threads with retry logic
+- **Fallback behavior**: If no event server, events are still logged locally to files
+- **MongoDB storage**: Events stored in MongoDB with ULID as `_id`
+- **Bespoke context binding**: Current implementation has custom context handling that can be replaced by structlog's native features
+
+---
+
+## Phase 1: Foundation & Quick Wins
+
+**Target Issues:** [#217](https://github.com/AD-SDL/MADSci/issues/217) (Automatic Logging Info)
+**Estimated Effort:** Small (1-2 days)
+**Breaking Changes:** None
+
+### Goals
+
+- Log MADSci version, configuration, and environment info on EventClient initialization
+- Establish patterns for subsequent phases
+
+### Requirements
+
+The EventClient should automatically log useful details the first time it's initialized:
+
+- MADSci version (from package metadata)
+- Client name and configuration summary
+- Event server URL (if configured)
+- Log directory path
+- Python version and platform info
+
[…]
