### 2026-01-29 aecc137
Initial implementation of log rotation/compression functionality
