+++ src/madsci_client/madsci/client/otel_config.py
@@ -1,0 +1,339 @@
+"""OpenTelemetry configuration for MADSci EventClient.
+
+This module provides OpenTelemetry integration for the EventClient,
+including tracing, metrics, and log correlation support.
+
+Example usage:
+    from madsci.client.otel_config import OtelManager
+
+    # Initialize OTEL with console exporter (development)
+    otel = OtelManager(service_name="my-service", exporter_type="console")
+    otel.setup()
+
+    # Get a tracer for creating spans
+    tracer = otel.get_tracer("my-module")
+    with tracer.start_as_current_span("my-operation"):
+        # ... do work ...
+        pass
+
+    # Get a meter for recording metrics
+    meter = otel.get_meter("my-module")
+    counter = meter.create_counter("my.counter")
+    counter.add(1)
+"""
+
+import logging
+from typing import Literal, Optional
+
+from opentelemetry import metrics, trace
+from opentelemetry.sdk.metrics import MeterProvider
+from opentelemetry.sdk.metrics.export import (
+    ConsoleMetricExporter,
+    PeriodicExportingMetricReader,
+)
+from opentelemetry.sdk.resources import Resource
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
+
+# Type alias for exporter types
+ExporterType = Literal["console", "otlp", "none"]
+
+
+class OtelManager:
+    """Manages OpenTelemetry tracing and metrics configuration.
+
+    This class encapsulates OTEL setup to avoid global state and allow
+    for multiple configurations in testing scenarios.
+
+    Attributes:
+        service_name: The name of the service for resource identification.
+        service_version: The version of the service.
+        exporter_type: The type of exporter to use.
+        otlp_endpoint: The OTLP collector endpoint (for otlp exporter).
+
+    Example:
+        # Development setup with console output
+        otel = OtelManager(service_name="my-service", exporter_type="console")
+        otel.setup()
+
+        # Production setup with OTLP
+        otel = OtelManager(
+            service_name="my-service",
+            exporter_type="otlp",
+            otlp_endpoint="http://otel-collector:4317",
+        )
+        otel.setup()
+    """
+
+    def __init__(
+        self,
+        service_name: str = "madsci-client",
+        service_version: str = "unknown",
+        exporter_type: ExporterType = "console",
+        otlp_endpoint: Optional[str] = None,
+    ) -> None:
+        """Initialize the OtelManager.
+
+        Args:
+            service_name: The name of the service for resource identification.
+            service_version: The version of the service.
+            exporter_type: The type of exporter to use:
+                - "console": Print traces/metrics to stdout (development)
+                - "otlp": Send to OTLP collector (production)
+                - "none": Disable telemetry export
+            otlp_endpoint: The OTLP collector endpoint (required if exporter_type="otlp").
+        """
+        self.service_name = service_name
+        self.service_version = service_version
+        self.exporter_type = exporter_type
+        self.otlp_endpoint = otlp_endpoint
+        self._initialized = False
+        self._tracer_provider: Optional[TracerProvider] = None
+        self._meter_provider: Optional[MeterProvider] = None
+        self._logger = logging.getLogger(__name__)
+
+    def setup(self) -> None:
+        """Initialize OpenTelemetry tracing and metrics.
+
+        This method sets up the tracer and meter providers with the
+        specified configuration.
+
+        Raises:
+            ValueError: If exporter_type is "otlp" but otlp_endpoint is not provided.
+        """
+        if self._initialized:
+            self._logger.warning("OpenTelemetry already initialized, skipping setup")
+            return
+
[…]
