+++ src/madsci_client/pyproject.toml
@@ -16,7 +16,13 @@ dependencies = [
   "click>=8.1.7",
   "trogon>=0.6.0",
   "minio>=7.1.0",
+  "opentelemetry-api>=1.20.0",
+  "opentelemetry-sdk>=1.20.0",
+]
 
+[project.optional-dependencies]
+otel-exporters = [
+  "opentelemetry-exporter-otlp>=1.20.0",
 ]
 
 [project.urls]
