+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "madsci"
 version = "0.6.1"
-dependencies = ["pre-commit>=4.3.0"]
+dependencies = ["pre-commit>=4.3.0", "structlog>=25.5.0", "colorama>=0.4.6"]
 readme="README.md"
 requires-python = ">=3.9.1"
 

[…]
