### 2026-01-30 28dff19
Initial structlog impl
