+++ src/madsci_client/madsci/client/event_client.py
@@ -1,6 +1,7 @@
 """MADSci Event Handling."""
 
 import contextlib
+import copy
 import gzip
 import inspect
 import json
@@ -8,6 +9,7 @@ import logging
 import platform
 import queue
 import shutil
+import sys
 import time
 import traceback
 import warnings
@@ -19,10 +21,12 @@ from threading import Lock, Thread
 from typing import Any, Optional, Union
 
 import requests
+from madsci.client.structlog_config import create_instance_logger, get_log_level_value
 from madsci.common.context import get_current_madsci_context
 from madsci.common.types.event_types import (
     Event,
     EventClientConfig,
+    EventLogLevel,
     EventType,
 )
 from madsci.common.utils import create_http_session, threaded_task
@@ -43,14 +47,36 @@ def get_madsci_version() -> str:
 
 
 class EventClient:
-    """A logger and event handler for MADSci system components."""
+    """A logger and event handler for MADSci system components.
+
+    Uses structlog for structured logging with context binding support.
+    Each EventClient instance has its own isolated logger configuration.
+
+    Example:
+        # Basic usage
+        client = EventClient(name="my_module")
+        client.info("Starting process", step=1, total=10)
+
+        # Context binding
+        client = client.bind(workflow_id="wf-123")
+        client.info("Processing")  # Automatically includes workflow_id
+
+        # Nested binding
+        client = client.bind(node_id="node-456")
+        client.info("Action complete")  # Includes both workflow_id and node_id
+
+        # Multiple clients with different configs (fully isolated)
+        json_client = EventClient(name="json_logger", config=EventClientConfig(log_output_format="json"))
+        console_client = EventClient(name="console_logger", config=EventClientConfig(log_output_format="console"))
+    """
 
     config: Optional[EventClientConfig] = None
-    _event_buffer = queue.Queue()
-    _buffer_lock = Lock()
-    _retry_thread = None
-    _retrying = False
-    _shutdown = False
+    _event_buffer: queue.Queue = queue.Queue()
+    _buffer_lock: Lock = Lock()
+    _retry_thread: Optional[Thread] = None
+    _retrying: bool = False
+    _shutdown: bool = False
+    _bound_context: dict[str, Any]
 
     def __init__(
         self,
@@ -81,16 +107,32 @@ class EventClient:
                 # * No luck, name after EventClient
                 self.name = __name__
         self.name = str(self.name)
-        self.logger = logging.getLogger(self.name)
+
+        # Initialize bound context
+        self._bound_context = {}
+
+        # Set up log directory and file
         self.log_dir = Path(self.config.log_dir).expanduser()
         self.log_dir.mkdir(parents=True, exist_ok=True)
         self.logfile = self.log_dir / f"{self.name}.log"
-        self.logger.setLevel(self.config.log_level)
+
+        # Create the stdlib logger for file handling
+        self.logger = logging.getLogger(self.name)
+        self.logger.setLevel(get_log_level_value(self.config.log_level))
         for handler in self.logger.handlers:
             self.logger.removeHandler(handler)
         file_handler = self._create_file_handler()
         self.logger.addHandler(file_handler)
         self.logger.addHandler(RichHandler(rich_tracebacks=True, show_path=False))
+
+        # Create structlog logger for structured logging
+        self._structlog_logger = create_instance_logger(
+            name=self.name,
+            output_format=self.config.log_output_format,
+            log_level=get_log_level_value(self.config.log_level),
+            include_otel_context=self.config.otel_enabled,
+        )
+
         self.event_server = (
             self.config.event_server_url
             or get_current_madsci_context().event_server_url
@@ -197,6 +239,304 @@ class EventClient:
[…]
