+++ src/madsci_client/pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
   "minio>=7.1.0",
   "opentelemetry-api>=1.20.0",
   "opentelemetry-sdk>=1.20.0",
+  "structlog>=24.1.0",
 ]
 
 [project.optional-dependencies]

[…]
