### 2026-01-30 99cdf44
Initial implementation of event retention
