+++ src/madsci_event_manager/tests/test_event_server.py
@@ -5,7 +5,9 @@ Uses pytest-mock-resources to create a MongoDB fixture. Note that this _requires
 a working docker installation.
 """
 
-from unittest.mock import patch
+from datetime import datetime, timedelta
+from pathlib import Path
+from unittest.mock import MagicMock, patch
 
 import pytest
 from fastapi.testclient import TestClient
@@ -155,3 +157,385 @@ def test_health_endpoint(test_client: TestClient) -> None:
     assert health_data["db_connected"] is True
     assert isinstance(health_data["total_events"], int)
     assert health_data["total_events"] >= 0
+
+
+# =============================================================================
+# Event Retention Tests
+# =============================================================================
+
+
+class TestEventRetention:
+    """Test EventManager retention functionality."""
+
+    def test_archive_events_by_ids(self, test_client: TestClient) -> None:
+        """Test that specific events can be archived by their IDs."""
+        # Create test events
+        events = []
+        for i in range(3):
+            event = Event(
+                event_type=EventType.TEST,
+                event_data={"test": f"archive_test_{i}"},
+            )
+            test_client.post("/event", json=event.model_dump(mode="json"))
+            events.append(event)
+
+        # Archive specific events
+        event_ids = [events[0].event_id, events[1].event_id]
+        response = test_client.post("/events/archive", json={"event_ids": event_ids})
+        assert response.status_code == 200
+        result = response.json()
+        assert result["archived_count"] == 2
+
+        # Verify archived events
+        response = test_client.get(f"/event/{events[0].event_id}")
+        event_data = response.json()
+        assert event_data["archived"] is True
+        assert event_data["archived_at"] is not None
+
+        # Verify non-archived event
+        response = test_client.get(f"/event/{events[2].event_id}")
+        event_data = response.json()
+        assert event_data["archived"] is False
+
+    def test_archive_events_by_date(self, test_client: TestClient) -> None:
+        """Test that events before a certain date can be archived."""
+        # Create an old event (simulate by setting timestamp in past)
+        old_event = Event(
+            event_type=EventType.TEST,
+            event_data={"test": "old_event"},
+            event_timestamp=datetime.now() - timedelta(days=100),
+        )
+        test_client.post("/event", json=old_event.model_dump(mode="json"))
+
+        # Create a recent event
+        recent_event = Event(
+            event_type=EventType.TEST,
+            event_data={"test": "recent_event"},
+        )
+        test_client.post("/event", json=recent_event.model_dump(mode="json"))
+
+        # Archive events older than 50 days
+        before_date = (datetime.now() - timedelta(days=50)).isoformat()
+        response = test_client.post(
+            "/events/archive", json={"before_date": before_date}
+        )
+        assert response.status_code == 200
+        result = response.json()
+        assert result["archived_count"] >= 1
+
+    def test_archived_events_excluded_from_queries_by_default(
+        self, test_client: TestClient
+    ) -> None:
+        """Test that archived events are excluded from default queries."""
+        # Create and archive an event
+        event = Event(
+            event_type=EventType.TEST,
+            event_data={"test": "excluded_test"},
+        )
+        test_client.post("/event", json=event.model_dump(mode="json"))
+        test_client.post("/events/archive", json={"event_ids": [event.event_id]})
+
+        # Query events - archived should be excluded by default
+        response = test_client.get("/events", params={"number": 100})
+        result = response.json()
+
+        # The archived event should not appear in results
+        assert event.event_id not in result
+
+    def test_include_archived_events_option(self, test_client: TestClient) -> None:
+        """Test that archived events can be included via query option."""
+        # Create and archive an event
+        event = Event(
+            event_type=EventType.TEST,
[…]
