+++ src/madsci_event_manager/madsci/event_manager/event_server.py
@@ -1,18 +1,20 @@
 """Example Event Manager implementation using the new AbstractManagerBase class."""
 
-from datetime import datetime
+from datetime import datetime, timedelta
 from pathlib import Path
-from typing import Any, Dict, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 import pymongo
-from classy_fastapi import get, post
+from classy_fastapi import delete, get, post
 from fastapi import Query
 from fastapi.exceptions import HTTPException
 from fastapi.params import Body
 from fastapi.responses import Response
 from madsci.client.event_client import EventClient
+from madsci.common.backup_tools import MongoDBBackupTool
 from madsci.common.manager_base import AbstractManagerBase
 from madsci.common.mongodb_version_checker import MongoDBVersionChecker
+from madsci.common.types.backup_types import MongoDBBackupSettings
 from madsci.common.types.event_types import (
     Event,
     EventLogLevel,
@@ -25,8 +27,56 @@ from madsci.event_manager.events_csv_exporter import CSVExporter
 from madsci.event_manager.notifications import EmailAlerts
 from madsci.event_manager.time_series_analyzer import TimeSeriesAnalyzer
 from madsci.event_manager.utilization_analyzer import UtilizationAnalyzer
+from pydantic import BaseModel
 from pymongo import MongoClient, errors
 from pymongo.synchronous.database import Database
+
+# =============================================================================
+# Request/Response Models for new endpoints
+# =============================================================================
+
+
+class ArchiveEventsRequest(BaseModel):
+    """Request model for archiving events."""
+
+    event_ids: Optional[List[str]] = None
+    before_date: Optional[datetime] = None
+    batch_size: Optional[int] = None
+
+
+class ArchiveEventsResponse(BaseModel):
+    """Response model for archive operation."""
+
+    archived_count: int
+    message: str
+
+
+class PurgeEventsResponse(BaseModel):
+    """Response model for purge operation."""
+
+    deleted_count: int
+    message: str
+
+
+class BackupRequest(BaseModel):
+    """Request model for backup creation."""
+
+    description: Optional[str] = None
+
+
+class BackupResponse(BaseModel):
+    """Response model for backup operation."""
+
+    backup_path: str
+    status: str
+
+
+class BackupStatusResponse(BaseModel):
+    """Response model for backup status."""
+
+    backup_enabled: bool
+    backup_dir: str
+    available_backups: List[Dict[str, Any]]
 
 
 class EventManager(AbstractManagerBase[EventManagerSettings, EventManagerDefinition]):
@@ -100,6 +150,59 @@ class EventManager(AbstractManagerBase[EventManagerSettings, EventManagerDefinit
 
         self.events = self._db_connection[self.settings.collection_name]
 
+        # Setup indexes for retention and query performance
+        self._setup_indexes()
+
+    def _setup_indexes(self) -> None:
+        """Set up MongoDB indexes including TTL index for automatic hard-deletion."""
+        try:
+            # Standard indexes for query performance
+            self.events.create_index("event_timestamp")
+            self.events.create_index("log_level")
+            self.events.create_index("archived")
+            self.events.create_index([("archived", 1), ("archived_at", 1)])
+
+            # TTL index for automatic hard-deletion of archived events
+            # MongoDB will automatically delete documents where archived_at
+            # is older than hard_delete_after_days
+            if self.settings.retention_enabled:
+                ttl_seconds = self.settings.hard_delete_after_days * 24 * 60 * 60
+
+                # Note: TTL index only applies to documents where archived_at is set
+                # Non-archived events (archived_at=None) are not affected
+                # We need to check if the index already exists with a different TTL
[…]
