+++ Configuration.md
@@ -96,6 +96,17 @@ Handles settings and configuration for the Event Manager.
 | `EVENT_COLLECTION_NAME`                      | `string`                          | `"events"`                    | The name of the MongoDB collection where events are stored.                                                                | `"events"`                    |
 | `EVENT_ALERT_LEVEL`                          | `EventLogLevel`                   | `40`                          | The log level at which to send an alert.                                                                                   | `40`                          |
 | `EVENT_EMAIL_ALERTS`                         | `EmailAlertsConfig` \| `NoneType` | `null`                        | The configuration for sending email alerts.                                                                                | `null`                        |
+| `EVENT_RETENTION_ENABLED`                    | `boolean`                         | `false`                       | Whether automatic event retention is enabled.                                                                              | `false`                       |
+| `EVENT_SOFT_DELETE_AFTER_DAYS`               | `integer`                         | `90`                          | Days after which events are soft-deleted (archived).                                                                       | `90`                          |
+| `EVENT_HARD_DELETE_AFTER_DAYS`               | `integer`                         | `365`                         | Days after archive when events are permanently deleted via TTL index.                                                      | `365`                         |
+| `EVENT_RETENTION_CHECK_INTERVAL_HOURS`       | `integer`                         | `24`                          | How often to run soft-delete retention checks (in hours).                                                                  | `24`                          |
+| `EVENT_ARCHIVE_BATCH_SIZE`                   | `integer`                         | `1000`                        | Maximum number of events to archive in a single batch operation.                                                           | `1000`                        |
+| `EVENT_MAX_BATCHES_PER_RUN`                  | `integer`                         | `100`                         | Maximum number of batches to process per retention run (0 = unlimited).                                                    | `100`                         |
+| `EVENT_BACKUP_ENABLED`                       | `boolean`                         | `false`                       | Whether automatic event backups are enabled.                                                                               | `false`                       |
+| `EVENT_BACKUP_SCHEDULE`                      | `string` \| `NoneType`            | `null`                        | Cron expression for backup schedule (e.g., '0 2 * * *' for 2am daily).                                                     | `null`                        |
+| `EVENT_BACKUP_DIR`                           | `string` \| `Path`                | `"~/.madsci/backups/events"`  | Directory for event backups.                                                                                               | `"~/.madsci/backups/events"`  |
+| `EVENT_BACKUP_MAX_COUNT`                     | `integer`                         | `10`                          | Maximum number of backup files to keep.                                                                                    | `10`                          |
+| `EVENT_FAIL_ON_RETENTION_ERROR`              | `boolean`                         | `false`                       | If True, raise exceptions on retention failures. If False, log and continue.                                               | `false`                       |
 
 ## WorkcellManagerSettings
 

[…]
