### 2026-01-30 3859e53
Make root .env optional; add link to docker guide in example lab
