### 2026-01-30 c6d9161
Add rich traceback handler
