### 2026-02-02 aa63f58
Initial opentelemetry integration plan
