+++ docs/dev/opentelemetry_integration_plan/12_phase_2_managers.md
@@ -1,0 +1,404 @@
+# Phase 2: Manager-Level Tracing
+
+**Estimated Effort:** Large (5-7 days)
+**Breaking Changes:** None
+
+## Goals
+
+- Add OTEL tracing to all manager services
+- Extract trace context from incoming HTTP requests
+- Propagate trace context in outgoing HTTP requests
+- Create meaningful spans for manager operations
+
+Approach decision (this plan):
+
+- Use auto-instrumentation for the HTTP boundary (server + client).
+- Add manual, domain-level spans for lab logic (workflows, steps, node actions, resource operations).
+
+Non-goal: relying on `opentelemetry-api` alone for manager tracing. Managers should participate in the same OTEL SDK setup strategy as EventClient so exporting/sampling are consistent.
+
+## Definition: "OTEL Enabled" (Managers)
+
+For Phase 2, "OTEL enabled" means:
+
+- The manager process calls the shared OTEL bootstrap during startup (installing SDK providers/exporters).
+- The server request span is created automatically via ASGI/FastAPI auto-instrumentation (using inbound context).
+- Outbound HTTP calls create client spans automatically via HTTP client auto-instrumentation.
+- Manual domain spans are created as child spans under the active context.
+
+Obtaining a tracer via `trace.get_tracer(...)` without configuring SDK providers is treated as "not fully enabled"
+(it will generally behave as a no-op).
+
+## 2.0.0 Shared OTEL Bootstrap (Recommended)
+
+To avoid duplicating OTEL expertise across managers/nodes/clients, create a small common library in `madsci_common` responsible for:
+
+- Reading OTEL configuration (prefer upstream OTEL env vars)
+- Creating and installing SDK providers (tracing + metrics)
+- Creating a tracer/meter for the current service
+- Providing helpers for context injection/extraction and current trace correlation
+- Clean shutdown/flush
+
+Proposed location:
+
+- `src/madsci_common/madsci/common/otel/` (or similar), used by both `madsci_client` and manager services.
+
+The target end-state is a single canonical module in `madsci.common.otel`.
+
+All manager OTEL behavior must follow the configuration precedence rules defined in
+`docs/dev/opentelemetry_integration_plan/00_principles.md`.
+
+Recommendation: start this refactor in Phase 1 so both EventClient and managers rely on the same bootstrap and so SDK
+setup is not duplicated.
+
+Target end-state (required):
+
+- All OTEL SDK/provider/exporter configuration lives in `madsci.common.otel`.
+- All other packages (clients, managers, nodes) import OTEL functionality only from `madsci.common.otel`.
+- Any remaining `madsci_client` OTEL module should be removed after migration.
+
+### 2.0.0.1 Proposed API Surface
+
+The shared OTEL package should provide one small, stable interface that managers/clients can use without being OTEL experts.
+
+```python
+from dataclasses import dataclass
+from typing import Optional
+
+from opentelemetry import metrics, trace
+
+
+@dataclass
+class OtelRuntime:
+    """Handle to OTEL facilities for a running service."""
+
+    enabled: bool
+    service_name: str
+    tracer: Optional[trace.Tracer]
+    meter: Optional[metrics.Meter]
+
+    def shutdown(self) -> None:
+        """Flush and shutdown providers (TracerProvider and MeterProvider)."""
+
+
+def configure_otel(
+    *,
+    enabled: bool,
+    service_name: str,
+    service_version: str | None = None,
+) -> OtelRuntime:
+    """Configure OTEL SDK/providers using upstream env var conventions."""
+
+
+def current_trace_context() -> dict[str, str]:
+    """Return {"trace_id": "...", "span_id": "..."} if a span is active."""
+
+
+def inject_headers(headers: dict[str, str]) -> None:
+    """Inject W3C trace context into outbound headers."""
+
+
[…]
