+++ docs/dev/opentelemetry_integration_plan/11_phase_1_eventclient.md
@@ -1,0 +1,344 @@
+# Phase 1: EventClient OTEL Completion
+
+**Estimated Effort:** Medium (3-5 days)
+**Breaking Changes:** None (OTEL remains optional)
+
+## Goals
+
+- Complete OTEL integration in EventClient so that when `otel_enabled=True`, traces and metrics are properly captured
+- Automatically inject trace context into Event objects
+- Add HTTP context propagation for events sent to EventManager
+- Ensure graceful degradation when OTEL is misconfigured or unavailable
+
+## Definition: "OTEL Enabled" (EventClient)
+
+For Phase 1, "OTEL enabled" means:
+
+- The process configures OpenTelemetry SDK providers/exporters (not only `opentelemetry-api` no-ops).
+- EventClient injects trace correlation fields (`trace_id`, `span_id`) when a span is active.
+- EventClient propagates context on outbound HTTP requests (e.g., `traceparent`).
+
+It does *not* mean that each log call creates its own span.
+
+## 1.0 Implementation Principles (EventClient)
+
+## 1.0.0 Configuration Precedence
+
+EventClient must follow the shared precedence rules defined in
+`docs/dev/opentelemetry_integration_plan/00_principles.md`.
+
+In particular:
+
+- `otel_enabled` must be able to force-disable OTEL even if upstream `OTEL_*` env vars are set.
+- When enabled, MADSci settings may override env vars for local/dev/test, but behavior must be deterministic.
+
+### 1.0.1 OTLP Endpoint Format (gRPC vs HTTP)
+
+The OTLP exporter supports both gRPC and HTTP transports.
+
+- **gRPC** endpoints typically look like `http://otel-collector:4317` (no path).
+- **HTTP** endpoints typically look like `http://otel-collector:4318` (no path) and may use `/v1/traces`, `/v1/metrics`
+  internally depending on exporter configuration.
+
+**Decision: Use `str` for OTLP endpoints, not `AnyUrl`.**
+
+Pydantic's `AnyUrl` normalizes URLs and ensures a trailing `/`, which can break OTLP configurations that are sensitive
+to unexpected paths. MADSci will:
+
+1. Store `otel_endpoint` as `Optional[str]` in all settings classes
+2. Validate endpoint format at initialization time using a dedicated validator
+3. Support both gRPC and HTTP protocols via the `otel_protocol` field
+
+```python
+from pydantic import field_validator
+
+class OtelSettings:
+    otel_endpoint: Optional[str] = Field(
+        default=None,
+        description="OTLP endpoint (e.g., 'http://localhost:4317' for gRPC)",
+    )
+    otel_protocol: Literal["grpc", "http"] = Field(
+        default="grpc",
+        description="OTLP transport protocol",
+    )
+
+    @field_validator("otel_endpoint")
+    @classmethod
+    def validate_endpoint(cls, v: Optional[str]) -> Optional[str]:
+        if v is None:
+            return v
+        if not v.startswith(("http://", "https://")):
+            raise ValueError("OTLP endpoint must start with http:// or https://")
+        # Strip trailing slash to prevent path issues
+        return v.rstrip("/")
+```
+
+### 1.0.2 Canonical Trace Correlation Rule
+
+MADSci must have exactly one interpretation for `Event.trace_id` / `Event.span_id`:
+
+- They represent the *current active span context at the time the Event is created*.
+
+Implementation guidance:
+
+- Prefer using `madsci.common.otel.current_trace_context()` (and/or OTEL-aware structlog processors) rather than
+  formatting IDs by hand in multiple locations.
+- When no span is active (or OTEL is disabled), omit these fields or set them to `None`.
+
[…]

[…]
