### 2026-02-02 b4cc105
Some further refinements to the plan
