### 2026-02-02 90db2fd
Initial work on phase 0 and 1
