+++ src/madsci_common/madsci/common/otel/__init__.py
@@ -1,0 +1,19 @@
+"""OpenTelemetry bootstrap and helpers.
+
+This package provides a single canonical OpenTelemetry configuration surface
+for MADSci processes.
+"""
+
+from .bootstrap import (
+    OtelBootstrapConfig,
+    OtelRuntime,
+    configure_otel,
+    current_trace_context,
+)
+
+__all__ = [
+    "OtelBootstrapConfig",
+    "OtelRuntime",
+    "configure_otel",
+    "current_trace_context",
+]
