### 2026-02-04 a4c497c
Update OTEL plan progress
Mark Phase 1 complete and note Phase 2 progress/remaining work.
