### 2026-02-04 613b2ef
Add manager OTEL health fields
Expose OTEL status in /health output and add OTEL settings to ManagerSettings to drive manager-level tracing configuration.
