+++ Configuration.md
@@ -37,24 +37,29 @@ Settings for the MADSci Data Manager.
 
 **Environment Prefix**: `DATA_`
 
-| Name                                        | Type                    | Default                       | Description                                                                                                                | Example                       |
-|---------------------------------------------|-------------------------|-------------------------------|----------------------------------------------------------------------------------------------------------------------------|-------------------------------|
-| `DATA_SERVER_URL`                           | `AnyUrl`                | `"http://localhost:8004/"`    | The URL of the data manager server.                                                                                        | `"http://localhost:8004/"`    |
-| `DATA_MANAGER_DEFINITION`                   | `string` \| `Path`      | `"data.manager.yaml"`         | Path to the data manager definition file to use.                                                                           | `"data.manager.yaml"`         |
-| `DATA_RATE_LIMIT_ENABLED`                   | `boolean`               | `true`                        | Enable rate limiting for API endpoints.                                                                                    | `true`                        |
-| `DATA_RATE_LIMIT_REQUESTS`                  | `integer`               | `300`                         | Maximum number of requests allowed per long time window.                                                                   | `300`                         |
-| `DATA_RATE_LIMIT_WINDOW`                    | `integer`               | `60`                          | Long time window for rate limiting in seconds.                                                                             | `60`                          |
-| `DATA_RATE_LIMIT_SHORT_REQUESTS`            | `integer` \| `NoneType` | `50`                          | Maximum number of requests allowed per short time window for burst protection. If None, short window limiting is disabled. | `50`                          |
-| `DATA_RATE_LIMIT_SHORT_WINDOW`              | `integer` \| `NoneType` | `1`                           | Short time window for burst protection in seconds. If None, short window limiting is disabled.                             | `1`                           |
-| `DATA_RATE_LIMIT_CLEANUP_INTERVAL`          | `integer`               | `300`                         | Interval in seconds between cleanup operations to prevent memory leaks.                                                    | `300`                         |
-| `DATA_RATE_LIMIT_EXEMPT_IPS`                | `array` \| `NoneType`   | `null`                        | List of IP addresses exempt from rate limiting. Defaults to localhost IPs (127.0.0.1, ::1) if not specified.               | `null`                        |
-| `DATA_UVICORN_WORKERS`                      | `integer` \| `NoneType` | `null`                        | Number of uvicorn worker processes. If None, uses uvicorn default (1).                                                     | `null`                        |
-| `DATA_UVICORN_LIMIT_CONCURRENCY`            | `integer` \| `NoneType` | `null`                        | Maximum number of concurrent connections. If None, no limit is enforced.                                                   | `null`                        |
-| `DATA_UVICORN_LIMIT_MAX_REQUESTS`           | `integer` \| `NoneType` | `null`                        | Maximum number of requests a worker will process before restarting. Helps prevent memory leaks.                            | `null`                        |
-| `DATA_DATABASE_NAME`                        | `string`                | `"madsci_data"`               | The name of the MongoDB database where events are stored.                                                                  | `"madsci_data"`               |
-| `DATA_COLLECTION_NAME`                      | `string`                | `"datapoints"`                | The name of the MongoDB collection where data are stored.                                                                  | `"datapoints"`                |
-| `MONGO_DB_URL` \| `DATA_DB_URL` \| `DB_URL` | `AnyUrl`                | `"mongodb://localhost:27017"` | The URL of the MongoDB database used by the Data Manager.                                                                  | `"mongodb://localhost:27017"` |
-| `DATA_FILE_STORAGE_PATH`                    | `string` \| `Path`      | `"~/.madsci/datapoints"`      | The path where files are stored on the server.                                                                             | `"~/.madsci/datapoints"`      |
[…]
