### 2026-02-04 7de17cf
Instrument WorkcellManager workflow spans
Add domain spans for workflow execution, per-step execution, and node action dispatch to improve end-to-end trace visibility.
