### 2026-02-04 c7025f8
Update Phase 2 progress
Document WorkcellManager domain spans (workflow.execute/workflow.step/node.action) and update acceptance criteria notes.
