### 2026-02-04 c344298
Instrument DataManager datapoint spans
Wrap datapoint ingestion in a data.save span and update Phase 2 progress notes.
