### 2026-02-04 8d14fa4
Instrument Resource/Location manager spans
Add domain spans for resource lifecycle/query endpoints and for location creation, resource attachments, and transfer planning.
