### 2026-02-05 0e5b7e9
Instrument ExperimentManager spans
Add experiment.* domain spans around experiment lifecycle endpoints and update Phase 2 progress notes.
