+++ src/madsci_experiment_manager/madsci/experiment_manager/experiment_server.py
@@ -124,144 +124,175 @@ class ExperimentManager(
     @get("/experiment/{experiment_id}")
     async def get_experiment(self, experiment_id: str) -> Experiment:
         """Get an experiment by ID."""
-        experiment = self.experiments.find_one({"_id": experiment_id})
-        if not experiment:
-            raise HTTPException(status_code=404, detail="Experiment not found")
-        return Experiment.model_validate(experiment)
+        with self.span(
+            "experiment.get",
+            attributes={"experiment.id": experiment_id},
+        ):
+            experiment = self.experiments.find_one({"_id": experiment_id})
+            if not experiment:
+                raise HTTPException(status_code=404, detail="Experiment not found")
+            return Experiment.model_validate(experiment)
 
     @get("/experiments")
     async def get_experiments(self, number: int = 10) -> list[Experiment]:
         """Get the latest experiments."""
-        experiments_list = (
-            self.experiments.find().sort("started_at", -1).limit(number).to_list()
-        )
-        return [
-            Experiment.model_validate(experiment) for experiment in experiments_list
-        ]
+        with self.span("experiment.query", attributes={"experiment.limit": number}):
+            experiments_list = (
+                self.experiments.find().sort("started_at", -1).limit(number).to_list()
+            )
+            return [
+                Experiment.model_validate(experiment) for experiment in experiments_list
+            ]
 
     @post("/experiment")
     async def start_experiment(
         self, experiment_request: ExperimentRegistration
     ) -> Experiment:
         """Start a new experiment."""
-        experiment = Experiment.from_experiment_design(
-            run_name=experiment_request.run_name,
-            run_description=experiment_request.run_description,
-            experiment_design=experiment_request.experiment_design,
-        )
-        experiment.started_at = datetime.datetime.now()
-
-        self.experiments.insert_one(experiment.to_mongo())
-
-        # Log the experiment start event
-        self.logger.log(
-            event=Event(
-                event_type=EventType.EXPERIMENT_START,
-                event_data={"experiment": experiment},
+        with self.span(
+            "experiment.start",
+            attributes={
+                "experiment.run_name": experiment_request.run_name,
+            },
+        ):
+            experiment = Experiment.from_experiment_design(
+                run_name=experiment_request.run_name,
+                run_description=experiment_request.run_description,
+                experiment_design=experiment_request.experiment_design,
             )
-        )
-        return experiment
+            experiment.started_at = datetime.datetime.now()
+
+            self.experiments.insert_one(experiment.to_mongo())
+
+            # Log the experiment start event
+            self.logger.log(
+                event=Event(
+                    event_type=EventType.EXPERIMENT_START,
+                    event_data={"experiment": experiment},
+                )
+            )
+            return experiment
 
     @post("/experiment/{experiment_id}/end")
     async def end_experiment(self, experiment_id: str) -> Experiment:
         """End an experiment by ID."""
-        experiment = self.experiments.find_one({"_id": experiment_id})
-        if not experiment:
-            raise HTTPException(status_code=404, detail="Experiment not found")
-        experiment = Experiment.model_validate(experiment)
-        experiment.ended_at = datetime.datetime.now()
-        experiment.status = ExperimentStatus.COMPLETED
-        self.experiments.update_one(
-            {"_id": experiment_id},
-            {"$set": experiment.to_mongo()},
-        )
-        self.logger.log(
-            event=Event(
-                event_type=EventType.EXPERIMENT_COMPLETE,
-                event_data={"experiment": experiment},
+        with self.span(
+            "experiment.end",
+            attributes={"experiment.id": experiment_id},
+        ):
[…]
