### 2026-02-05 dcef10c
Instrument EventManager query/archive/backup spans
Add event.query, event.archive, and event.backup.* spans for key EventManager operations and update Phase 2 progress notes.
