+++ src/madsci_event_manager/madsci/event_manager/event_server.py
@@ -473,8 +473,9 @@ class EventManager(AbstractManagerBase[EventManagerSettings, EventManagerDefinit
     @post("/events/query")
     async def query_events(self, selector: Any = Body()) -> Dict[str, Event]:  # noqa: B008
         """Query events based on a selector. Note: this is a raw query, so be careful."""
-        event_list = self.events.find(selector).to_list()
-        return {event["_id"]: event for event in event_list}
+        with self.span("event.query"):
+            event_list = self.events.find(selector).to_list()
+            return {event["_id"]: event for event in event_list}
 
     # ==========================================================================
     # Event Retention Endpoints
@@ -498,53 +499,62 @@ class EventManager(AbstractManagerBase[EventManagerSettings, EventManagerDefinit
             Count of archived events
         """
         try:
-            query: Dict[str, Any] = {"archived": {"$ne": True}}
+            with self.span(
+                "event.archive",
+                attributes={
+                    "archive.by_ids": bool(request.event_ids),
+                    "archive.before_date": request.before_date.isoformat()
+                    if request.before_date
+                    else None,
+                },
+            ):
+                query: Dict[str, Any] = {"archived": {"$ne": True}}
 
-            # Build query based on request parameters
-            # Note: Pydantic model_validator ensures at least one of these is set
-            if request.event_ids:
-                query["_id"] = {"$in": request.event_ids}
-            elif request.before_date:
-                # Convert datetime to ISO string format to match MongoDB storage format
-                # (Events are stored with ISO string timestamps via to_mongo())
-                query["event_timestamp"] = {"$lt": request.before_date.isoformat()}
-
-            # Perform archive in batches to prevent performance impact
-            batch_size = request.batch_size or self.settings.archive_batch_size
-            total_archived = 0
-
-            while True:
-                # Find batch of events to archive
-                events_to_archive = list(
-                    self.events.find(query, {"_id": 1}).limit(batch_size)
-                )
-
-                if not events_to_archive:
-                    break
-
-                event_ids = [e["_id"] for e in events_to_archive]
-
-                # Archive this batch
-                result = self.events.update_many(
-                    {"_id": {"$in": event_ids}},
-                    {
-                        "$set": {
-                            "archived": True,
-                            "archived_at": datetime.now(timezone.utc),
-                        }
-                    },
-                )
-
-                total_archived += result.modified_count
-
-                # If archiving by IDs, we're done after one batch
+                # Build query based on request parameters
+                # Note: Pydantic model_validator ensures at least one of these is set
                 if request.event_ids:
-                    break
+                    query["_id"] = {"$in": request.event_ids}
+                elif request.before_date:
+                    # Convert datetime to ISO string format to match MongoDB storage format
+                    # (Events are stored with ISO string timestamps via to_mongo())
+                    query["event_timestamp"] = {"$lt": request.before_date.isoformat()}
 
-            return ArchiveEventsResponse(
-                archived_count=total_archived,
-                message=f"Successfully archived {total_archived} events",
-            )
+                # Perform archive in batches to prevent performance impact
+                batch_size = request.batch_size or self.settings.archive_batch_size
+                total_archived = 0
+
+                while True:
+                    # Find batch of events to archive
+                    events_to_archive = list(
+                        self.events.find(query, {"_id": 1}).limit(batch_size)
+                    )
+
+                    if not events_to_archive:
+                        break
+
+                    event_ids = [e["_id"] for e in events_to_archive]
+
+                    # Archive this batch
+                    result = self.events.update_many(
+                        {"_id": {"$in": event_ids}},
+                        {
+                            "$set": {
+                                "archived": True,
[…]
