### 2026-02-05 61eef74
Close out Phase 2 OTEL coverage
Add remaining DataManager read/query spans, verify process-global OTEL runtime reuse with tests, document docker example-lab OTLP exporter gap, and update EventManager health OTEL reporting.
