### 2026-02-05 b54d309
Bundle OTLP exporter into docker image
Install OpenTelemetry OTLP exporter and FastAPI/ASGI/requests instrumentation in the base docker image so example-lab services can export traces when otel_exporter=otlp.
