### 2026-02-05 d6c1b95
Add OTEL collector to example lab
Add an OpenTelemetry Collector service + config for OTLP traces/metrics, update Phase 2/4 plan docs with runtime validation, and keep example_lab env aligned with OTLP export.
