### 2026-02-05 649b8e0
phase3: logging ratchet
Add Phase 3A EventType expansions and code-governed descriptions; document logging guidelines; add a scoped pre-commit logging-pattern audit to iteratively improve structured logging across core managers. Disable Pyright type checking for reduced editor noise and relax Ruff rules for scripts.
