### 2026-02-05 d6476cb
phase3: audit squid lab_server logging
