### 2026-02-05 cfe3f1c
phase3: audit event_manager utilization analyzer logging
