+++ src/madsci_event_manager/madsci/event_manager/utilization_analyzer.py
@@ -25,9 +25,17 @@ class UtilizationAnalyzer:
         # Verify database connection
         try:
             count = events_collection.count_documents({})
-            logger.info(f"Connected to events collection with {count} total events")
-        except Exception as e:
-            logger.error(f"Error connecting to events collection: {e}")
+            logger.info(
+                "Connected to events collection",
+                event_type=EventType.LOG_INFO,
+                total_events=count,
+            )
+        except Exception:
+            logger.error(
+                "Error connecting to events collection",
+                event_type=EventType.LOG_ERROR,
+                exc_info=True,
+            )
             raise
 
         # Cache for name resolution to avoid repeated lookups
@@ -62,7 +70,8 @@ class UtilizationAnalyzer:
             # If no formal sessions found, create a default analysis session
             if not sessions:
                 logger.info(
-                    "No formal workcell sessions found, creating default analysis session"
+                    "No formal workcell sessions found, creating default analysis session",
+                    event_type=EventType.LOG_INFO,
                 )
                 sessions = [
                     {
@@ -105,7 +114,11 @@ class UtilizationAnalyzer:
             }
 
         except Exception as e:
-            logger.error(f"Error generating session-based report: {e}")
+            logger.error(
+                "Error generating session-based report",
+                event_type=EventType.LOG_ERROR,
+                exc_info=True,
+            )
             return {"error": f"Failed to generate report: {e!s}"}
 
     def generate_user_utilization_report(
@@ -122,7 +135,10 @@ class UtilizationAnalyzer:
             Dict containing user statistics, system summary, and metadata
         """
         try:
-            logger.info("Generating user utilization report")
+            logger.info(
+                "Generating user utilization report",
+                event_type=EventType.LOG_INFO,
+            )
 
             # Determine analysis timeframe
             analysis_start, analysis_end = self._determine_analysis_period(
@@ -178,7 +194,11 @@ class UtilizationAnalyzer:
             }
 
         except Exception as e:
-            logger.error(f"Error generating user utilization report: {e}")
+            logger.error(
+                "Error generating user utilization report",
+                event_type=EventType.LOG_ERROR,
+                exc_info=True,
+            )
             return {"error": f"Failed to generate user report: {e!s}"}
 
     def _determine_analysis_period(
@@ -191,13 +211,17 @@ class UtilizationAnalyzer:
 
         # If no timeframe provided, analyze ALL records in database
         if not start_time and not end_time:
-            logger.info("No timeframe provided - analyzing full database range")
+            logger.info(
+                "No timeframe provided - analyzing full database range",
+                event_type=EventType.LOG_INFO,
+            )
             try:
                 # Get total count first to check if we have any data
                 total_count = self.events_collection.count_documents({})
                 if total_count == 0:
                     logger.warning(
-                        "No events found in database, using default 24-hour period"
+                        "No events found in database, using default 24-hour period",
+                        event_type=EventType.LOG_WARNING,
                     )
                     now = datetime.now(timezone.utc).replace(tzinfo=None)
                     return now - timedelta(days=1), now
@@ -226,15 +250,28 @@ class UtilizationAnalyzer:
                         buffered_end = latest_time + timedelta(minutes=1)
 
                         logger.info(
-                            f"Found database range: {buffered_start} to {buffered_end}"
+                            "Found database range",
+                            event_type=EventType.LOG_INFO,
+                            buffered_start=buffered_start,
+                            buffered_end=buffered_end,
                         )
                         return buffered_start, buffered_end
[…]
