### 2026-02-05 b00505e
audit time_series_analyzer logging patterns
