+++ docs/dev/opentelemetry_integration_plan/13_phase_3_logging_and_event_types.md
@@ -34,6 +34,7 @@ Audited/cleaned (passes the logging-pattern hook):
 - `src/madsci_experiment_manager/madsci/experiment_manager/experiment_server.py`
 - `src/madsci_squid/madsci/squid/lab_server.py`
 - `src/madsci_event_manager/madsci/event_manager/utilization_analyzer.py`
+- `src/madsci_event_manager/madsci/event_manager/time_series_analyzer.py`
 
 Deferred/backlog (not yet audited; tracked for follow-up):
 
@@ -44,7 +45,7 @@ Deferred/backlog (not yet audited; tracked for follow-up):
 Update (February 2026):
 
 - `utilization_analyzer.py` is now audited and included in the hook scope.
-- `time_series_analyzer.py` remains deferred (large surface area; will likely need a follow-up PR).
+- `time_series_analyzer.py` is now audited and included in the hook scope.
 
 Operational notes:
 
