### 2026-02-19 #9416 @arfon
@ktahar, @lucask07 – I would appreciate an explicit "yes" that you are happy for this to proceed to publication.
