### 2026-03-05 #9416 @RyanTheRobothead
I'm pleased to report that we have published the latest release of MADSci, v0.7.0, and archived it to Zenodo.

- GitHub Release: https://github.com/AD-SDL/MADSci/releases/tag/v0.7.0
- Zenodo Archive: https://zenodo.org/records/18879108
- Zenodo DoI: https://doi.org/10.5281/zenodo.18879108

Let me know if there's anything else that is needed from our end!
